Coercion to Network-Valued Data Object

This function flags a list of igraph objects as an nvd object as defined in this package.

as_nvd(obj)

Arguments

obj

A list of igraph objects.

Value

An nvd object.

Examples

as_nvd(nvd("smallworld", 10))
#> [[1]]
#> IGRAPH 4e678b8 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 4e678b8:
#>  [1]  1-- 2  2-- 3  3--15  4-- 5  5-- 6  6-- 7  7-- 8  8-- 9  9--10  2--10
#> [11] 11--12 12--13 13--14 14--15 15--16  2-- 8 17--18 18--19 15--20 20--21
#> [21] 21--22 22--23 23--24 14--24  9--25  1-- 3  1--16  1-- 4  4--23  5--22
#> [31]  1--22  2--11  2-- 4 24--25  2-- 5  2--23  2-- 6  3-- 5  3--25  3-- 6
#> [41]  3--24  3-- 7  6--18  2-- 7  4--25  4-- 8  5-- 7  5-- 8  5-- 9  8--24
#> [51]  1--11  6--10  7-- 9  7--10  7--11  8--10  8--11  8--12  9--11 12--23
#> [61]  9--13 10--12 10--13 10--14 11--13 11--14  5--11 12--14 12--15 16--25
#> + ... omitted several edges
#> 
#> [[2]]
#> IGRAPH d5cacb6 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from d5cacb6:
#>  [1]  1-- 2  2-- 3  3-- 4  4-- 5  6--24  6--12  8--23  8-- 9  9--10 11--20
#> [11] 12--19 12--13 13--14 14--23 15--16  6--16 17--18  2--18 19--20 20--21
#> [21] 21--22 16--23 23--24 10--24  1--25  1-- 3  1--24  4--20  4--23  1-- 5
#> [31]  1--22  2--25  2-- 4  2--24  2-- 5  2--23  2-- 6  3-- 5  3--25  3-- 6
#> [41]  3--22  3-- 7  4-- 6  4-- 7  4--25  4--19  5-- 7  5-- 8 14--19  6-- 8
#> [51]  6--11  6--10  7-- 9  7--20  7--11 10--19  8--11  8--12  9--11  9--12
#> [61]  8--13 10--12 10--13 10--14 13--24 11--14 11--15  5--23 12--15 12--16
#> + ... omitted several edges
#> 
#> [[3]]
#> IGRAPH 1d87f9d U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 1d87f9d:
#>  [1]  1-- 2  2--19  3--19  4--11  5-- 6  7--22  2-- 8  8-- 9  9--10 10--11
#> [11] 11--12 12--13 13--14 15--23 15--16 16--17 17--18 18--19 19--20 20--21
#> [21] 21--22 22--23  5--23 24--25  1--25  1-- 4  1--24  1--21  1--23  1-- 5
#> [31]  1--22  2--25  2-- 4 23--24  2-- 5  2--23  2-- 6  3-- 5  3--25  3-- 6
#> [41]  3--22  3-- 7  4-- 6  4--18  4--25  4-- 8  5-- 7  5--22  5--17  6-- 8
#> [51]  6-- 9  6--10  7-- 9  7--10  7--12  8--16  8--11  8--12  9--11  9--12
#> [61]  9--13 10--12 11--13  3--14 13--19 11--14 11--15 12--14 12--15 14--16
#> + ... omitted several edges
#> 
#> [[4]]
#> IGRAPH 90fee1b U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 90fee1b:
#>  [1]  1-- 2  2-- 3  3-- 4  4-- 5  5-- 6  6-- 7  7--17  8--14  9--10 10--11
#> [11] 11--12 12--13 13--14 14--15 15--16 16--17 17--18 18--19  2--19 20--21
#> [21] 16--21 22--23  7--22 17--24  1--25  1--17  1--24  1--15  1--23  1-- 5
#> [31]  8--22  2--10  2--24  2--22  2-- 5  2--23  2-- 6  3-- 5  3--25  3-- 6
#> [41]  3--24  3--12  4-- 6  1-- 7  4--25  4-- 8  5-- 7  5-- 8  5--15  6-- 8
#> [51]  6--11  6--10  7--21  5--10  7--11  8--10  8--11  8--12  9--19  9--12
#> [61]  9--13 10--19 10--13 22--25 11--13 11--14  6--15 12--14 12--15 12--16
#> + ... omitted several edges
#> 
#> [[5]]
#> IGRAPH 461cc16 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 461cc16:
#>  [1]  1-- 2  2-- 3  4--10  4-- 5  5-- 6  4-- 7  8--19  8-- 9  9--10 11--17
#> [11] 11--12 12--13 13--14 14--15 15--16 16--17 17--18 18--19 19--20 20--21
#> [21] 21--22 22--23 23--24 24--25  1--14  1-- 3  1--24  1--19  1--23  5--25
#> [31]  1--22  2--25  2-- 4  2--15  2--11  2--23  2-- 6  3-- 5  3--25  3-- 6
#> [41]  3--24  3--14  4-- 6  4--22  4--25  4--17  5-- 7  5--13  5-- 9  6-- 8
#> [51]  1-- 9 10--23  7-- 9  7--10  7--11  8--10  8--18 12--20  4--23  9--12
#> [61]  9--18 10--12 10--13 10--14 13--25 11--14 11--15 12--14 12--15 16--23
#> + ... omitted several edges
#> 
#> [[6]]
#> IGRAPH 8b84cf6 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 8b84cf6:
#>  [1]  1-- 2  2-- 3  4--23  4-- 5  5-- 6  6-- 7  7-- 8  8-- 9  9--10 10--11
#> [11] 11--12  4--13 13--14 14--15 15--16 16--17 18--25 18--19 19--20 20--21
#> [21] 21--22 22--23 23--24 13--25  1--25  3--11  1--24  1-- 4  1--15  1-- 5
#> [31]  1--22 16--25  2-- 4  2--24  2-- 5  2-- 9  2-- 6  5--14 14--25  3--18
#> [41]  3--24  3-- 7  4-- 6  4-- 7  4--25  4-- 8  5-- 7  5-- 8  5-- 9  6-- 8
#> [51]  6-- 9 16--22  9--16  7--10  7--11  8--10  8--11  8--12 11--25  6--14
#> [61]  9--13 10--12 10--13 10--14 13--24 11--14 11--15  5--12 12--15 12--16
#> + ... omitted several edges
#> 
#> [[7]]
#> IGRAPH d5f0105 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from d5f0105:
#>  [1]  1-- 2  2-- 3  4--19  4--17  5-- 6  2-- 7  7-- 8  8-- 9  9--10 10--11
#> [11] 11--12  4--13  5--18 14--15 15--16  6--17  9--15 18--19 10--19 20--21
#> [21] 13--21 22--23 23--24 24--25 18--20  1-- 3  1--24  1-- 4 23--25  5--12
#> [31]  1--22 16--25  2-- 4  2--24  2--11  2--23  2-- 6  3-- 5  3--25  3--11
#> [41]  3--24  3-- 7  4-- 6  4--18  4--25  8--19  5-- 7  5-- 8  5-- 9  6-- 8
#> [51]  6-- 9  6--10  7-- 9 10--17  7--11  8--17 11--17  8--12  9--11  9--12
#> [61]  9--13  8--21 10--13 10--14 13--16 11--14 11--15 12--14 12--18 12--16
#> + ... omitted several edges
#> 
#> [[8]]
#> IGRAPH 57a303e U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 57a303e:
#>  [1]  1-- 2  2--12  3-- 4  4-- 5  5-- 6  6-- 7  7-- 8  8-- 9 10--23 10--11
#> [11] 11--12 13--23 13--14 14--15 15--16 16--17 17--18 18--19 19--20 20--21
#> [21] 21--22  8--23  3--23 13--24  1--25  1--17  1--24  1-- 4  1--23  1-- 5
#> [31]  1--22  2--25 14--22  2--24  2--16  2--23  2-- 6  3-- 5  3-- 8  3-- 6
#> [41]  3--11  3-- 7  4-- 6  4-- 7  4--25  4--17  5--21  5-- 8  5-- 9  6-- 8
#> [51]  6-- 9 10--15  7-- 9  7--10  7--25  8--10  8--11  8--12 11--23  9--12
#> [61] 13--20 10--24 10--13 10--14 11--13 11--14 15--20  7--24 12--15  6--12
#> + ... omitted several edges
#> 
#> [[9]]
#> IGRAPH b7c55c9 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from b7c55c9:
#>  [1]  1-- 2 14--19  3-- 4  4-- 5  5-- 6  6-- 7  7-- 8  8-- 9  9--10 10--11
#> [11] 11--12 12--13 13--20 14--15 15--16 16--17 17--18  6--19 10--19 20--21
#> [21] 12--21 10--16 23--24 24--25  1--18  1-- 3 17--24  1-- 4  1--23  1-- 5
#> [31]  1--22  2--25  2-- 4  2--24  2-- 5  2--23  2-- 6  3-- 5  3--22  6--14
#> [41]  3--24  3-- 7  1-- 6  4-- 7  4--25  4-- 8  5-- 7  5-- 8  5-- 9  8--22
#> [51]  6-- 9  6--10  9--24  7--10  1--11 10--23  8--11  8--12  9--11  9--12
#> [61]  9--13 10--12 10--15 10--14 11--13 11--14 11--15 12--14 12--15 12--16
#> + ... omitted several edges
#> 
#> [[10]]
#> IGRAPH 6946672 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 6946672:
#>  [1]  1-- 2  2-- 3  3-- 4  4-- 5  5-- 6  6-- 7  7-- 8  8-- 9  9--10 10--11
#> [11] 11--12 12--13 13--18 14--15 15--20 16--17  7--18 18--19 19--20 20--21
#> [21]  6--21 16--22 23--24 24--25  1--25  1-- 3  1--24  1-- 4  1--23  5-- 7
#> [31]  1--22  2--25  2-- 4  2--24  5--17  2--21  2-- 6  3-- 5  3--25  2--14
#> [41]  3--24  3-- 7  4-- 6  4-- 7  4--22  4-- 8  5--15  5-- 8  5-- 9  6--13
#> [51]  6-- 9  6--10  7-- 9  7--10 11--18  8--10  8--11  8--12  9--11  9--12
#> [61]  9--13 10--12 13--14 10--14 11--13 11--14 11--15 12--14 12--15  4--25
#> + ... omitted several edges
#> 
#> attr(,"class")
#> [1] "nvd"  "list"