Transform distance matrix in edge properties of minimal spanning tree

Transform distance matrix in edge properties of minimal spanning tree

edge_count_global_variables(d, n1, k = 1L)

Arguments

d
A matrix of dimension \((n1+n2)x(n1+n2)\) containing the distances between all the elements of the two samples put together.

n1
An integer giving the size of the first sample.

k
An integer specifying the density of the minimal spanning tree to generate.

Value

A list of edge properties of the minimal spanning tree.

Examples

n1 <- 30L
n2 <- 10L
x <- nvd("smallworld", n1)
y <- nvd("pa", n2)
d <- dist_nvd(x, y, representation = "laplacian", distance = "frobenius")
e <- edge_count_global_variables(d, n1, k = 5L)