Network-Valued Data Constructor

This is the constructor for objects of class nvd.

nvd(
  model = "smallworld",
  n = 0L,
  pref.matrix = NULL,
  lambda = NULL,
  size = NULL,
  prob = NULL
)

Arguments

model

A string specifying the model to be used for sampling networks (current choices are: "sbm", "k_regular", "gnp", "smallworld", "pa", "poisson" and "binomial"). Default is "smallworld".

n

An integer specifying the sample size (default: 0L).

pref.matrix

A matrix giving the Bernoulli rates for the SBM generator (see sample_sbm for details). Default is NULL. It is required for model == "sbm".

lambda

A numeric value specifying the mean value for the Poisson generator. Default is NULL. It is required for model == "poisson".

size

An integer value specifying the number of trials for the binomial distribution. Default is NULL. It is required for model == "binomial".

prob

A numeric value specifying the probability of success of each trial for the binomial distribution. Default is NULL. It is required for model == "binomial".

Value

A nvd object which is a list of igraphobjects.

Examples

nvd(n = 10L)
#> [[1]]
#> IGRAPH 330d279 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 330d279:
#>  [1]  1--16  2-- 3  3-- 8  4-- 5  5-- 6  6-- 7  7-- 8  8-- 9  9--22 10--11
#> [11] 12--20 12--13 14--20 14--15 15--16 16--17 18--20 18--19 19--20 11--21
#> [21] 21--22  1--15  3--23 24--25  1--25  1-- 8  1--24  4--15  1--23 10--18
#> [31]  1--22 13--25  2-- 4  2--24  2--12  2--23  2-- 6  3--18 19--25  3-- 6
#> [41]  3--24  3-- 7  1-- 6  2-- 7  4--25  4-- 8  3--14  5--11  3-- 9  6-- 8
#> [51]  6-- 9  6--10  7-- 9  7--10  7--11  8--10 10--17  8--12  9--21  9--12
#> [61]  9--13  6--12 10--19 10--14  1--13 11--14 15--24 12--14 12--15 12--16
#> + ... omitted several edges
#> 
#> [[2]]
#> IGRAPH 1778140 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 1778140:
#>  [1]  1-- 2  2--13  3-- 4  4-- 5  5-- 6  6-- 7  7-- 8  8-- 9  9--10 10--11
#> [11] 12--24 12--13 13--14 14--15 15--16 16--17 17--18 18--19 19--20 20--21
#> [21] 21--22 14--22 23--24 24--25  1--25  1-- 3  1--24  4--13  1--23 12--17
#> [31]  1--22  2--25  2-- 4  4--11  2-- 5  2--23  2-- 6  3-- 5  3--25  3-- 6
#> [41]  3--24  3--12  6--25  4-- 7  4--25  8--14  5-- 7  9--16  9--24  6-- 8
#> [51]  6-- 9  6--10  7-- 9  7--10  7--11  8--17 11--17  8--12  9--11  9--12
#> [61]  3-- 9 10--12 13--19 10--14 11--13 11--14 11--15 12--14 12--15 12--16
#> + ... omitted several edges
#> 
#> [[3]]
#> IGRAPH fc78a53 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from fc78a53:
#>  [1]  1-- 2  2-- 3  3-- 4  4-- 5  6--12  6-- 7  3-- 8  8-- 9  9--10 10--11
#> [11] 11--12  5--12 13--14 14--15 15--16 16--20 17--18 18--19  9--20 20--21
#> [21] 21--22 22--23 23--24 24--25  1--25  1-- 3  1--20  1-- 4  1--23  1-- 5
#> [31]  1--12  2--25  2-- 4  2--24  2-- 5 14--23  2-- 6  3--23  3--25  3-- 6
#> [41]  3--24  3-- 7  4-- 6  4-- 7  4--14  4-- 8  5-- 7  5-- 8  5-- 9  6-- 8
#> [51]  9--19  6--10  7-- 9  7--10  7--18 10--16  8--11  8--23  5--11  9--12
#> [61]  9--13 10--12 10--13  9--14 11--13 11--14  7--15  3--12  8--12 12--22
#> + ... omitted several edges
#> 
#> [[4]]
#> IGRAPH 57a132d U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 57a132d:
#>  [1]  1-- 2  2-- 3  3-- 4  4-- 5  5-- 6  5-- 7  7-- 8  8-- 9  9--10 11--25
#> [11] 11--12 12--13 14--21 14--15 15--16 16--17 17--18 18--19 19--20 20--21
#> [21] 21--22 22--23 23--24 24--25  1--25  1-- 3  1--24  1-- 4  1--23  5--20
#> [31]  1--22  2--25  2-- 4  2--24  2-- 5  2--23  2-- 6  5--12  3--12  3--13
#> [41]  5--24  3-- 7  4-- 6  4-- 7  4--25  4-- 8  6--25  5-- 8  5-- 9  6--24
#> [51]  6-- 9  6--10  7-- 9  7--14 12--25  8--10  8--11  8--12  9--11  9--12
#> [61]  9--13  4--10 10--13  2--14 11--13  1--14 11--22 12--24 12--15 16--23
#> + ... omitted several edges
#> 
#> [[5]]
#> IGRAPH d00a15e U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from d00a15e:
#>  [1]  2--21  2--22  3-- 4  4--22  5-- 6  6-- 7  4-- 7  8-- 9  9--10 10--11
#> [11] 11--25 12--13 13--14 14--15 15--16 16--17  5--18 18--19 19--20 14--21
#> [21] 21--22 22--23 23--24 24--25  1--25  3--10  1--15  1-- 4  1--23  1-- 5
#> [31] 16--22  2--25  2-- 4  2--24  5--20  2--23  2-- 6  3-- 5  3--25  8--17
#> [41]  3--24  3-- 7  4-- 6  1--20 15--25  4-- 8  5-- 7  8--19  5-- 9  6-- 8
#> [51]  6--21  5--10  7-- 9 10--24 11--12  8--10  8--11  8--12  9--11  9--12
#> [61]  9--13 10--12 10--13 10--14 11--13 11--14 11--15 12--14 12--15 12--16
#> + ... omitted several edges
#> 
#> [[6]]
#> IGRAPH f47bccc U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from f47bccc:
#>  [1]  1--17  2-- 3  3--13  4-- 5  6--15  6-- 7  7-- 8  8-- 9  9--19 10--11
#> [11] 11--12 12--13 13--14 15--22 15--20 16--17  7--17  1--19  4--20 20--21
#> [21] 21--22 22--23 23--24 24--25  1--25  1-- 3  1--24  4--18 11--23  1-- 5
#> [31]  1--22 18--25  2-- 4  2--24  2-- 5  2--23  2-- 6  5--20  3--25  3--10
#> [41]  3--24  3-- 7  4-- 6  4-- 7  4--25  4--23  5-- 7  5--19  5-- 9  6-- 8
#> [51]  6--21  6--10  4-- 9  7--10  7--11  8--10  8--11 12--19  9--11  9--12
#> [61] 13--21 12--20 10--13 10--14 11--13 11--16 11--15  6--12 12--15 12--16
#> + ... omitted several edges
#> 
#> [[7]]
#> IGRAPH 3615c9d U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 3615c9d:
#>  [1]  2-- 8  2-- 3  3-- 4  4-- 5  5-- 6  1-- 6  7-- 8  8-- 9  9--10 10--11
#> [11] 11--16 12--13  5--14 14--15 15--16  7--17 17--18 18--19 19--20 20--21
#> [21] 21--22  5--23 23--24 24--25  1--25  1-- 3  1--24  1--10  1--23  1-- 5
#> [31]  6--22 12--25  2-- 4  2--24  2-- 5  2--23  2-- 6  3-- 5  3--25  4--22
#> [41]  3--24  3--12  4--24  4-- 7  4--25  8--14  3-- 7  5-- 8  5-- 9  6-- 8
#> [51]  6-- 9  6--10  7-- 9 10--18  7--15  8--10  8--11  8--12  9--11  9--12
#> [61]  9--13 10--12  6--13 10--14 11--13 11--14 11--15 12--14 12--22 12--16
#> + ... omitted several edges
#> 
#> [[8]]
#> IGRAPH 7833fb8 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 7833fb8:
#>  [1]  1--14  2-- 3  3-- 4  4-- 5  5-- 6  6-- 7  7-- 8  8-- 9  9--10 10--11
#> [11] 11--12 12--13 13--14 14--15 15--16 16--17 17--18 18--23 14--19 20--21
#> [21] 21--22 22--23 23--24 24--25  1--25  1-- 3  1--24  1-- 4  1--23  1-- 5
#> [31]  1--22  2--18  2-- 4  2--24  2-- 5  2--23  2-- 6  3-- 5  3--25  3-- 6
#> [41]  3--24  3-- 7  4-- 6  4-- 7  4--25  4-- 8  5-- 7  5-- 8  9--20  6-- 8
#> [51]  6-- 9  6--10  7-- 9  7--10  7--11  8--10  8--11  8--12  9--11  9--12
#> [61]  9--13 10--12  4--13 10--14 11--13 11--14 11--15 12--14 12--15 16--25
#> + ... omitted several edges
#> 
#> [[9]]
#> IGRAPH 53b351b U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 53b351b:
#>  [1]  1-- 2  2-- 3  3-- 4  4-- 5  5-- 6  6--12  7-- 8  9--19  9--10 10--11
#> [11] 11--12 12--23 13--14 16--25 10--23 17--23 17--18 12--19 19--20 20--21
#> [21] 21--22 22--23 23--24 24--25  1--25  1-- 3  1--24  1-- 4  1--23  1-- 5
#> [31]  1--22 13--25  2--10  2-- 8  2-- 5  2--23  2--14  3-- 5  3--25  3-- 6
#> [41]  3--24  3-- 9  4-- 6  4-- 7  4--25  4-- 8  5-- 7  5-- 8  5-- 9  6-- 8
#> [51]  6-- 9  6--10  2-- 9  7--21  7--11  8--10  8--11  8--12  9--11  9--12
#> [61]  9--13 10--12 10--13 10--14 11--13 11--18 11--15  3-- 8 12--15 12--16
#> + ... omitted several edges
#> 
#> [[10]]
#> IGRAPH 7747883 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 7747883:
#>  [1]  1-- 2  2--21  3-- 4  4-- 5  5-- 6  6-- 7  7-- 8  8-- 9  9--10 10--11
#> [11] 12--18 12--13 13--14 14--15 15--16  6--17 17--18 17--19 19--20 20--21
#> [21]  8--24  4--22  5--23 24--25  1--23  1-- 3  1--24  1-- 4  1--21  1-- 5
#> [31]  1--22  2--25  2-- 4 11--24  2-- 5  2--23  2-- 6  3-- 5  3--25  3-- 6
#> [41]  3--24  3-- 7  6--15  4-- 7  4--25  4-- 8  5-- 7  5-- 8  5-- 9  6-- 8
#> [51]  6-- 9  6--19  7--25  7--10  7--23  8--10  8--11  8--12  9--11  9--12
#> [61] 13--23  3--10 10--13 10--14 11--13 11--14 11--15  1--12 15--22  9--21
#> + ... omitted several edges
#> 
#> attr(,"class")
#> [1] "nvd"  "list"