Network Representation Functions

This is a collection of functions that convert a graph stored as an igraph object into a desired matrix representation among adjacency matrix, graph laplacian, modularity matrix or graphon (edge probability matrix).

repr_adjacency(network, validate = TRUE)

repr_laplacian(network, validate = TRUE)

repr_modularity(network, validate = TRUE)

repr_graphon(network, validate = TRUE)

Arguments

network

An igraph object.

validate

A boolean specifying whether the function should check the class of its input (default: TRUE).

Value

A numeric square matrix giving the desired network representation recorded in the object's class.

Examples

g <- igraph::sample_smallworld(1, 25, 3, 0.05)
repr_adjacency(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0    1    1    1    0    0    0    0    0     0     0     0     0
#>  [2,]    1    0    1    1    1    0    0    0    0     0     0     0     0
#>  [3,]    1    1    0    1    1    1    0    0    0     0     0     0     0
#>  [4,]    1    1    1    0    1    1    1    0    0     0     0     0     0
#>  [5,]    0    1    1    1    0    1    1    1    0     0     0     0     0
#>  [6,]    0    0    1    1    1    0    1    1    1     0     0     0     0
#>  [7,]    0    0    0    1    1    1    0    1    1     1     0     0     0
#>  [8,]    0    0    0    0    1    1    1    0    1     1     1     0     0
#>  [9,]    0    0    0    0    0    1    1    1    0     1     1     0     0
#> [10,]    0    0    0    0    0    0    1    1    1     0     0     0     1
#> [11,]    0    0    0    0    0    0    0    1    1     0     0     1     1
#> [12,]    0    0    0    0    0    0    0    0    0     0     1     0     1
#> [13,]    0    0    0    0    0    0    0    0    0     1     1     1     0
#> [14,]    0    0    0    0    0    0    0    0    0     0     1     0     1
#> [15,]    0    0    1    0    0    0    0    0    1     0     0     1     1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0     1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     1     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    1    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     1     0
#> [22,]    0    0    0    0    0    0    0    0    1     0     0     0     0
#> [23,]    1    0    0    0    0    0    0    0    0     1     0     0     0
#> [24,]    1    1    0    0    0    0    0    0    0     0     0     0     0
#> [25,]    1    0    1    0    0    0    0    0    1     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0     1     1     1
#>  [2,]     0     0     0     0     0     0     0     0     0     0     1     0
#>  [3,]     0     1     0     0     0     0     0     0     0     0     0     1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     1     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     1     0     0     0     0     0     0     1     0     0     1
#> [10,]     0     0     0     0     1     0     0     0     0     1     0     0
#> [11,]     1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]     0     1     0     0     0     0     0     1     0     0     0     0
#> [13,]     1     1     1     0     0     0     0     0     0     0     0     0
#> [14,]     0     1     1     1     0     0     0     0     0     0     0     0
#> [15,]     1     0     0     1     1     0     0     0     0     0     0     0
#> [16,]     1     0     0     1     1     1     0     0     0     0     0     0
#> [17,]     1     1     1     0     1     1     1     0     0     0     0     0
#> [18,]     0     1     1     1     0     1     1     1     0     1     0     0
#> [19,]     0     0     1     1     1     0     0     1     0     0     0     0
#> [20,]     0     0     0     1     1     0     0     1     1     1     0     0
#> [21,]     0     0     0     0     1     1     1     0     1     1     1     0
#> [22,]     0     0     0     0     0     0     1     1     0     1     1     1
#> [23,]     0     0     0     0     1     0     1     1     1     0     0     1
#> [24,]     0     0     0     0     0     0     0     1     1     0     0     1
#> [25,]     0     0     0     0     0     0     0     0     1     1     1     0
#> attr(,"representation")
#> [1] "adjacency"
repr_laplacian(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    6   -1   -1   -1    0    0    0    0    0     0     0     0     0
#>  [2,]   -1    5   -1   -1   -1    0    0    0    0     0     0     0     0
#>  [3,]   -1   -1    7   -1   -1   -1    0    0    0     0     0     0     0
#>  [4,]   -1   -1   -1    6   -1   -1   -1    0    0     0     0     0     0
#>  [5,]    0   -1   -1   -1    7   -1   -1   -1    0     0     0     0     0
#>  [6,]    0    0   -1   -1   -1    6   -1   -1   -1     0     0     0     0
#>  [7,]    0    0    0   -1   -1   -1    6   -1   -1    -1     0     0     0
#>  [8,]    0    0    0    0   -1   -1   -1    6   -1    -1    -1     0     0
#>  [9,]    0    0    0    0    0   -1   -1   -1    8    -1    -1     0     0
#> [10,]    0    0    0    0    0    0   -1   -1   -1     6     0     0    -1
#> [11,]    0    0    0    0    0    0    0   -1   -1     0     5    -1    -1
#> [12,]    0    0    0    0    0    0    0    0    0     0    -1     4    -1
#> [13,]    0    0    0    0    0    0    0    0    0    -1    -1    -1     6
#> [14,]    0    0    0    0    0    0    0    0    0     0    -1     0    -1
#> [15,]    0    0   -1    0    0    0    0    0   -1     0     0    -1    -1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0    -1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0    -1     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0   -1    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0    -1     0
#> [22,]    0    0    0    0    0    0    0    0   -1     0     0     0     0
#> [23,]   -1    0    0    0    0    0    0    0    0    -1     0     0     0
#> [24,]   -1   -1    0    0    0    0    0    0    0     0     0     0     0
#> [25,]   -1    0   -1    0    0    0    0    0   -1     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0    -1    -1    -1
#>  [2,]     0     0     0     0     0     0     0     0     0     0    -1     0
#>  [3,]     0    -1     0     0     0     0     0     0     0     0     0    -1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0    -1     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0    -1     0     0     0     0     0     0    -1     0     0    -1
#> [10,]     0     0     0     0    -1     0     0     0     0    -1     0     0
#> [11,]    -1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]     0    -1     0     0     0     0     0    -1     0     0     0     0
#> [13,]    -1    -1    -1     0     0     0     0     0     0     0     0     0
#> [14,]     5    -1    -1    -1     0     0     0     0     0     0     0     0
#> [15,]    -1     7     0    -1    -1     0     0     0     0     0     0     0
#> [16,]    -1     0     5    -1    -1    -1     0     0     0     0     0     0
#> [17,]    -1    -1    -1     6    -1    -1    -1     0     0     0     0     0
#> [18,]     0    -1    -1    -1     8    -1    -1    -1     0    -1     0     0
#> [19,]     0     0    -1    -1    -1     4     0    -1     0     0     0     0
#> [20,]     0     0     0    -1    -1     0     6    -1    -1    -1     0     0
#> [21,]     0     0     0     0    -1    -1    -1     7    -1    -1    -1     0
#> [22,]     0     0     0     0     0     0    -1    -1     6    -1    -1    -1
#> [23,]     0     0     0     0    -1     0    -1    -1    -1     7     0    -1
#> [24,]     0     0     0     0     0     0     0    -1    -1     0     5    -1
#> [25,]     0     0     0     0     0     0     0     0    -1    -1    -1     6
#> attr(,"representation")
#> [1] "laplacian"
repr_modularity(g)
#>        [,1]       [,2]       [,3]  [,4]       [,5]  [,6]  [,7]  [,8]       [,9]
#>  [1,] -0.24  0.8000000  0.7200000  0.76 -0.2800000 -0.24 -0.24 -0.24 -0.3200000
#>  [2,]  0.80 -0.1666667  0.7666667  0.80  0.7666667 -0.20 -0.20 -0.20 -0.2666667
#>  [3,]  0.72  0.7666667 -0.3266667  0.72  0.6733333  0.72 -0.28 -0.28 -0.3733333
#>  [4,]  0.76  0.8000000  0.7200000 -0.24  0.7200000  0.76  0.76 -0.24 -0.3200000
#>  [5,] -0.28  0.7666667  0.6733333  0.72 -0.3266667  0.72  0.72  0.72 -0.3733333
#>  [6,] -0.24 -0.2000000  0.7200000  0.76  0.7200000 -0.24  0.76  0.76  0.6800000
#>  [7,] -0.24 -0.2000000 -0.2800000  0.76  0.7200000  0.76 -0.24  0.76  0.6800000
#>  [8,] -0.24 -0.2000000 -0.2800000 -0.24  0.7200000  0.76  0.76 -0.24  0.6800000
#>  [9,] -0.32 -0.2666667 -0.3733333 -0.32 -0.3733333  0.68  0.68  0.68 -0.4266667
#> [10,] -0.24 -0.2000000 -0.2800000 -0.24 -0.2800000 -0.24  0.76  0.76  0.6800000
#> [11,] -0.20 -0.1666667 -0.2333333 -0.20 -0.2333333 -0.20 -0.20  0.80  0.7333333
#> [12,] -0.16 -0.1333333 -0.1866667 -0.16 -0.1866667 -0.16 -0.16 -0.16 -0.2133333
#> [13,] -0.24 -0.2000000 -0.2800000 -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.3200000
#> [14,] -0.20 -0.1666667 -0.2333333 -0.20 -0.2333333 -0.20 -0.20 -0.20 -0.2666667
#> [15,] -0.28 -0.2333333  0.6733333 -0.28 -0.3266667 -0.28 -0.28 -0.28  0.6266667
#> [16,] -0.20 -0.1666667 -0.2333333 -0.20 -0.2333333 -0.20 -0.20 -0.20 -0.2666667
#> [17,] -0.24 -0.2000000 -0.2800000 -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.3200000
#> [18,] -0.32 -0.2666667 -0.3733333 -0.32 -0.3733333 -0.32 -0.32 -0.32 -0.4266667
#> [19,] -0.16 -0.1333333 -0.1866667 -0.16 -0.1866667 -0.16 -0.16 -0.16 -0.2133333
#> [20,] -0.24 -0.2000000 -0.2800000 -0.24  0.7200000 -0.24 -0.24 -0.24 -0.3200000
#> [21,] -0.28 -0.2333333 -0.3266667 -0.28 -0.3266667 -0.28 -0.28 -0.28 -0.3733333
#> [22,] -0.24 -0.2000000 -0.2800000 -0.24 -0.2800000 -0.24 -0.24 -0.24  0.6800000
#> [23,]  0.72 -0.2333333 -0.3266667 -0.28 -0.3266667 -0.28 -0.28 -0.28 -0.3733333
#> [24,]  0.80  0.8333333 -0.2333333 -0.20 -0.2333333 -0.20 -0.20 -0.20 -0.2666667
#> [25,]  0.76 -0.2000000  0.7200000 -0.24 -0.2800000 -0.24 -0.24 -0.24  0.6800000
#>       [,10]      [,11]      [,12] [,13]      [,14]      [,15]      [,16] [,17]
#>  [1,] -0.24 -0.2000000 -0.1600000 -0.24 -0.2000000 -0.2800000 -0.2000000 -0.24
#>  [2,] -0.20 -0.1666667 -0.1333333 -0.20 -0.1666667 -0.2333333 -0.1666667 -0.20
#>  [3,] -0.28 -0.2333333 -0.1866667 -0.28 -0.2333333  0.6733333 -0.2333333 -0.28
#>  [4,] -0.24 -0.2000000 -0.1600000 -0.24 -0.2000000 -0.2800000 -0.2000000 -0.24
#>  [5,] -0.28 -0.2333333 -0.1866667 -0.28 -0.2333333 -0.3266667 -0.2333333 -0.28
#>  [6,] -0.24 -0.2000000 -0.1600000 -0.24 -0.2000000 -0.2800000 -0.2000000 -0.24
#>  [7,]  0.76 -0.2000000 -0.1600000 -0.24 -0.2000000 -0.2800000 -0.2000000 -0.24
#>  [8,]  0.76  0.8000000 -0.1600000 -0.24 -0.2000000 -0.2800000 -0.2000000 -0.24
#>  [9,]  0.68  0.7333333 -0.2133333 -0.32 -0.2666667  0.6266667 -0.2666667 -0.32
#> [10,] -0.24 -0.2000000 -0.1600000  0.76 -0.2000000 -0.2800000 -0.2000000 -0.24
#> [11,] -0.20 -0.1666667  0.8666667  0.80  0.8333333 -0.2333333 -0.1666667 -0.20
#> [12,] -0.16  0.8666667 -0.1066667  0.84 -0.1333333  0.8133333 -0.1333333 -0.16
#> [13,]  0.76  0.8000000  0.8400000 -0.24  0.8000000  0.7200000  0.8000000 -0.24
#> [14,] -0.20  0.8333333 -0.1333333  0.80 -0.1666667  0.7666667  0.8333333  0.80
#> [15,] -0.28 -0.2333333  0.8133333  0.72  0.7666667 -0.3266667 -0.2333333  0.72
#> [16,] -0.20 -0.1666667 -0.1333333  0.80  0.8333333 -0.2333333 -0.1666667  0.80
#> [17,] -0.24 -0.2000000 -0.1600000 -0.24  0.8000000  0.7200000  0.8000000 -0.24
#> [18,]  0.68 -0.2666667 -0.2133333 -0.32 -0.2666667  0.6266667  0.7333333  0.68
#> [19,] -0.16 -0.1333333 -0.1066667 -0.16 -0.1333333 -0.1866667  0.8666667  0.84
#> [20,] -0.24 -0.2000000 -0.1600000 -0.24 -0.2000000 -0.2800000 -0.2000000  0.76
#> [21,] -0.28 -0.2333333  0.8133333 -0.28 -0.2333333 -0.3266667 -0.2333333 -0.28
#> [22,] -0.24 -0.2000000 -0.1600000 -0.24 -0.2000000 -0.2800000 -0.2000000 -0.24
#> [23,]  0.72 -0.2333333 -0.1866667 -0.28 -0.2333333 -0.3266667 -0.2333333 -0.28
#> [24,] -0.20 -0.1666667 -0.1333333 -0.20 -0.1666667 -0.2333333 -0.1666667 -0.20
#> [25,] -0.24 -0.2000000 -0.1600000 -0.24 -0.2000000 -0.2800000 -0.2000000 -0.24
#>            [,18]      [,19] [,20]      [,21] [,22]      [,23]      [,24] [,25]
#>  [1,] -0.3200000 -0.1600000 -0.24 -0.2800000 -0.24  0.7200000  0.8000000  0.76
#>  [2,] -0.2666667 -0.1333333 -0.20 -0.2333333 -0.20 -0.2333333  0.8333333 -0.20
#>  [3,] -0.3733333 -0.1866667 -0.28 -0.3266667 -0.28 -0.3266667 -0.2333333  0.72
#>  [4,] -0.3200000 -0.1600000 -0.24 -0.2800000 -0.24 -0.2800000 -0.2000000 -0.24
#>  [5,] -0.3733333 -0.1866667  0.72 -0.3266667 -0.28 -0.3266667 -0.2333333 -0.28
#>  [6,] -0.3200000 -0.1600000 -0.24 -0.2800000 -0.24 -0.2800000 -0.2000000 -0.24
#>  [7,] -0.3200000 -0.1600000 -0.24 -0.2800000 -0.24 -0.2800000 -0.2000000 -0.24
#>  [8,] -0.3200000 -0.1600000 -0.24 -0.2800000 -0.24 -0.2800000 -0.2000000 -0.24
#>  [9,] -0.4266667 -0.2133333 -0.32 -0.3733333  0.68 -0.3733333 -0.2666667  0.68
#> [10,]  0.6800000 -0.1600000 -0.24 -0.2800000 -0.24  0.7200000 -0.2000000 -0.24
#> [11,] -0.2666667 -0.1333333 -0.20 -0.2333333 -0.20 -0.2333333 -0.1666667 -0.20
#> [12,] -0.2133333 -0.1066667 -0.16  0.8133333 -0.16 -0.1866667 -0.1333333 -0.16
#> [13,] -0.3200000 -0.1600000 -0.24 -0.2800000 -0.24 -0.2800000 -0.2000000 -0.24
#> [14,] -0.2666667 -0.1333333 -0.20 -0.2333333 -0.20 -0.2333333 -0.1666667 -0.20
#> [15,]  0.6266667 -0.1866667 -0.28 -0.3266667 -0.28 -0.3266667 -0.2333333 -0.28
#> [16,]  0.7333333  0.8666667 -0.20 -0.2333333 -0.20 -0.2333333 -0.1666667 -0.20
#> [17,]  0.6800000  0.8400000  0.76 -0.2800000 -0.24 -0.2800000 -0.2000000 -0.24
#> [18,] -0.4266667  0.7866667  0.68  0.6266667 -0.32  0.6266667 -0.2666667 -0.32
#> [19,]  0.7866667 -0.1066667 -0.16  0.8133333 -0.16 -0.1866667 -0.1333333 -0.16
#> [20,]  0.6800000 -0.1600000 -0.24  0.7200000  0.76  0.7200000 -0.2000000 -0.24
#> [21,]  0.6266667  0.8133333  0.72 -0.3266667  0.72  0.6733333  0.7666667 -0.28
#> [22,] -0.3200000 -0.1600000  0.76  0.7200000 -0.24  0.7200000  0.8000000  0.76
#> [23,]  0.6266667 -0.1866667  0.72  0.6733333  0.72 -0.3266667 -0.2333333  0.72
#> [24,] -0.2666667 -0.1333333 -0.20  0.7666667  0.80 -0.2333333 -0.1666667  0.80
#> [25,] -0.3200000 -0.1600000 -0.24 -0.2800000  0.76  0.7200000  0.8000000 -0.24
#> attr(,"representation")
#> [1] "modularity"
repr_graphon(g)
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.55555556 0.59027778 0.61904762 0.52777778 0.38888889 0.30952381
#>  [2,] 0.59027778 0.50000000 0.66964286 0.56250000 0.64583333 0.33035714
#>  [3,] 0.61904762 0.66964286 0.57142857 0.69047619 0.61904762 0.57142857
#>  [4,] 0.52777778 0.56250000 0.69047619 0.66666667 0.83333333 0.61904762
#>  [5,] 0.38888889 0.64583333 0.61904762 0.83333333 0.66666667 0.69047619
#>  [6,] 0.30952381 0.33035714 0.57142857 0.61904762 0.69047619 0.57142857
#>  [7,] 0.33333333 0.29166667 0.36507937 0.61111111 0.58333333 0.63492063
#>  [8,] 0.18253968 0.19642857 0.28571429 0.30952381 0.53571429 0.57142857
#>  [9,] 0.30952381 0.33035714 0.14285714 0.38095238 0.38095238 0.64285714
#> [10,] 0.05555556 0.06250000 0.18253968 0.19444444 0.30555556 0.25396825
#> [11,] 0.00000000 0.00000000 0.11111111 0.19444444 0.11111111 0.12698413
#> [12,] 0.07142857 0.07142857 0.00000000 0.00000000 0.00000000 0.00000000
#> [13,] 0.00000000 0.00000000 0.00000000 0.05555556 0.11111111 0.18253968
#> [14,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [15,] 0.05555556 0.06250000 0.19642857 0.08333333 0.08333333 0.00000000
#> [16,] 0.00000000 0.00000000 0.05555556 0.00000000 0.05555556 0.00000000
#> [17,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [18,] 0.05555556 0.00000000 0.00000000 0.00000000 0.05555556 0.07142857
#> [19,] 0.07142857 0.07142857 0.00000000 0.00000000 0.07142857 0.00000000
#> [20,] 0.27777778 0.05555556 0.12698413 0.08333333 0.13888889 0.07142857
#> [21,] 0.27777778 0.06250000 0.15476190 0.00000000 0.08333333 0.00000000
#> [22,] 0.29861111 0.25000000 0.19642857 0.06250000 0.06250000 0.00000000
#> [23,] 0.36666667 0.22500000 0.07142857 0.00000000 0.18333333 0.07142857
#> [24,] 0.50793651 0.40178571 0.35714286 0.15476190 0.22619048 0.14285714
#> [25,] 0.44444444 0.35416667 0.49206349 0.25000000 0.38888889 0.05555556
#>             [,7]       [,8]      [,9]      [,10]      [,11]      [,12]
#>  [1,] 0.33333333 0.18253968 0.3095238 0.05555556 0.00000000 0.07142857
#>  [2,] 0.29166667 0.19642857 0.3303571 0.06250000 0.00000000 0.07142857
#>  [3,] 0.36507937 0.28571429 0.1428571 0.18253968 0.11111111 0.00000000
#>  [4,] 0.61111111 0.30952381 0.3809524 0.19444444 0.19444444 0.00000000
#>  [5,] 0.58333333 0.53571429 0.3809524 0.30555556 0.11111111 0.00000000
#>  [6,] 0.63492063 0.57142857 0.6428571 0.25396825 0.12698413 0.00000000
#>  [7,] 0.55555556 0.63492063 0.4920635 0.27777778 0.16666667 0.05555556
#>  [8,] 0.63492063 0.57142857 0.5000000 0.50793651 0.43650794 0.14285714
#>  [9,] 0.49206349 0.50000000 0.2857143 0.74603175 0.49206349 0.14285714
#> [10,] 0.27777778 0.50793651 0.7460317 0.22222222 0.16666667 0.12698413
#> [11,] 0.16666667 0.43650794 0.4920635 0.16666667 0.33333333 0.38095238
#> [12,] 0.05555556 0.14285714 0.1428571 0.12698413 0.38095238 0.28571429
#> [13,] 0.16666667 0.18253968 0.1666667 0.33333333 0.55555556 0.39682540
#> [14,] 0.05555556 0.14285714 0.1428571 0.18253968 0.43650794 0.35714286
#> [15,] 0.18055556 0.39285714 0.5892857 0.00000000 0.16666667 0.46428571
#> [16,] 0.05555556 0.18253968 0.2380952 0.00000000 0.22222222 0.32539683
#> [17,] 0.00000000 0.00000000 0.0000000 0.25396825 0.30952381 0.35714286
#> [18,] 0.05555556 0.07142857 0.1269841 0.33333333 0.27777778 0.25396825
#> [19,] 0.00000000 0.00000000 0.0000000 0.05555556 0.19841270 0.07142857
#> [20,] 0.11111111 0.05555556 0.3095238 0.11111111 0.00000000 0.05555556
#> [21,] 0.00000000 0.00000000 0.2380952 0.19444444 0.05555556 0.29761905
#> [22,] 0.06250000 0.13392857 0.3303571 0.17361111 0.00000000 0.13392857
#> [23,] 0.05555556 0.07142857 0.1000000 0.32222222 0.05555556 0.10000000
#> [24,] 0.12698413 0.00000000 0.1428571 0.18253968 0.00000000 0.00000000
#> [25,] 0.22222222 0.18253968 0.4365079 0.05555556 0.00000000 0.07142857
#>            [,13]      [,14]      [,15]      [,16]      [,17]      [,18]
#>  [1,] 0.00000000 0.00000000 0.05555556 0.00000000 0.00000000 0.05555556
#>  [2,] 0.00000000 0.00000000 0.06250000 0.00000000 0.00000000 0.00000000
#>  [3,] 0.00000000 0.00000000 0.19642857 0.05555556 0.00000000 0.00000000
#>  [4,] 0.05555556 0.00000000 0.08333333 0.00000000 0.00000000 0.00000000
#>  [5,] 0.11111111 0.00000000 0.08333333 0.05555556 0.00000000 0.05555556
#>  [6,] 0.18253968 0.00000000 0.00000000 0.00000000 0.00000000 0.07142857
#>  [7,] 0.16666667 0.05555556 0.18055556 0.05555556 0.00000000 0.05555556
#>  [8,] 0.18253968 0.14285714 0.39285714 0.18253968 0.00000000 0.07142857
#>  [9,] 0.16666667 0.14285714 0.58928571 0.23809524 0.00000000 0.12698413
#> [10,] 0.33333333 0.18253968 0.00000000 0.00000000 0.25396825 0.33333333
#> [11,] 0.55555556 0.43650794 0.16666667 0.22222222 0.30952381 0.27777778
#> [12,] 0.39682540 0.35714286 0.46428571 0.32539683 0.35714286 0.25396825
#> [13,] 0.44444444 0.50793651 0.72222222 0.55555556 0.30952381 0.22222222
#> [14,] 0.50793651 0.57142857 0.66071429 0.50793651 0.42857143 0.32539683
#> [15,] 0.72222222 0.66071429 0.00000000 0.16666667 0.53571429 0.59722222
#> [16,] 0.55555556 0.50793651 0.16666667 0.33333333 0.63492063 0.61111111
#> [17,] 0.30952381 0.42857143 0.53571429 0.63492063 0.57142857 0.50793651
#> [18,] 0.22222222 0.32539683 0.59722222 0.61111111 0.50793651 0.55555556
#> [19,] 0.32539683 0.28571429 0.33928571 0.32539683 0.57142857 0.45238095
#> [20,] 0.16666667 0.12698413 0.00000000 0.11111111 0.38095238 0.61111111
#> [21,] 0.11111111 0.22619048 0.08333333 0.25000000 0.22619048 0.61111111
#> [22,] 0.06250000 0.00000000 0.00000000 0.05555556 0.13392857 0.41666667
#> [23,] 0.00000000 0.00000000 0.12500000 0.11111111 0.24285714 0.52222222
#> [24,] 0.00000000 0.00000000 0.00000000 0.00000000 0.14285714 0.25396825
#> [25,] 0.05555556 0.00000000 0.00000000 0.00000000 0.05555556 0.22222222
#>            [,19]      [,20]      [,21]      [,22]      [,23]     [,24]
#>  [1,] 0.07142857 0.27777778 0.27777778 0.29861111 0.36666667 0.5079365
#>  [2,] 0.07142857 0.05555556 0.06250000 0.25000000 0.22500000 0.4017857
#>  [3,] 0.00000000 0.12698413 0.15476190 0.19642857 0.07142857 0.3571429
#>  [4,] 0.00000000 0.08333333 0.00000000 0.06250000 0.00000000 0.1547619
#>  [5,] 0.07142857 0.13888889 0.08333333 0.06250000 0.18333333 0.2261905
#>  [6,] 0.00000000 0.07142857 0.00000000 0.00000000 0.07142857 0.1428571
#>  [7,] 0.00000000 0.11111111 0.00000000 0.06250000 0.05555556 0.1269841
#>  [8,] 0.00000000 0.05555556 0.00000000 0.13392857 0.07142857 0.0000000
#>  [9,] 0.00000000 0.30952381 0.23809524 0.33035714 0.10000000 0.1428571
#> [10,] 0.05555556 0.11111111 0.19444444 0.17361111 0.32222222 0.1825397
#> [11,] 0.19841270 0.00000000 0.05555556 0.00000000 0.05555556 0.0000000
#> [12,] 0.07142857 0.05555556 0.29761905 0.13392857 0.10000000 0.0000000
#> [13,] 0.32539683 0.16666667 0.11111111 0.06250000 0.00000000 0.0000000
#> [14,] 0.28571429 0.12698413 0.22619048 0.00000000 0.00000000 0.0000000
#> [15,] 0.33928571 0.00000000 0.08333333 0.00000000 0.12500000 0.0000000
#> [16,] 0.32539683 0.11111111 0.25000000 0.05555556 0.11111111 0.0000000
#> [17,] 0.57142857 0.38095238 0.22619048 0.13392857 0.24285714 0.1428571
#> [18,] 0.45238095 0.61111111 0.61111111 0.41666667 0.52222222 0.2539683
#> [19,] 0.28571429 0.18253968 0.45238095 0.20535714 0.17142857 0.0000000
#> [20,] 0.18253968 0.33333333 0.61111111 0.46527778 0.57777778 0.3095238
#> [21,] 0.45238095 0.61111111 0.50000000 0.58333333 0.73333333 0.5357143
#> [22,] 0.20535714 0.46527778 0.58333333 0.62500000 0.77500000 0.5357143
#> [23,] 0.17142857 0.57777778 0.73333333 0.77500000 0.60000000 0.4857143
#> [24,] 0.00000000 0.30952381 0.53571429 0.53571429 0.48571429 0.4285714
#> [25,] 0.07142857 0.22222222 0.33333333 0.41666667 0.57777778 0.5079365
#>            [,25]
#>  [1,] 0.44444444
#>  [2,] 0.35416667
#>  [3,] 0.49206349
#>  [4,] 0.25000000
#>  [5,] 0.38888889
#>  [6,] 0.05555556
#>  [7,] 0.22222222
#>  [8,] 0.18253968
#>  [9,] 0.43650794
#> [10,] 0.05555556
#> [11,] 0.00000000
#> [12,] 0.07142857
#> [13,] 0.05555556
#> [14,] 0.00000000
#> [15,] 0.00000000
#> [16,] 0.00000000
#> [17,] 0.05555556
#> [18,] 0.22222222
#> [19,] 0.07142857
#> [20,] 0.22222222
#> [21,] 0.33333333
#> [22,] 0.41666667
#> [23,] 0.57777778
#> [24,] 0.50793651
#> [25,] 0.33333333
#> attr(,"representation")
#> [1] "graphon"