Coercion to Network-Valued Data Object

This function flags a list of igraph objects as an nvd object as defined in this package.

as_nvd(obj)

Arguments

obj

A list of igraph objects.

Value

An nvd object.

Examples

gnp_params <- list(p = 1/3)
as_nvd(nvd(model = "gnp", n = 10L, model_params = gnp_params))
#> [[1]]
#> IGRAPH c55c1a1 U--- 25 99 -- Erdos-Renyi (gnp) graph
#> + attr: name (g/c), type (g/c), loops (g/l), p (g/n)
#> + edges from c55c1a1:
#>  [1]  1-- 5  2-- 5  3-- 5  1-- 6  1-- 7  5-- 7  1-- 8  4-- 8  6-- 8  1-- 9
#> [11]  8-- 9  1--10  4--10  1--11  5--11  6--11  7--11  8--11 10--11  2--12
#> [21]  6--12 10--12  5--13  1--14  2--14  3--14  4--14  5--14  9--14 12--14
#> [31] 13--14  1--15  2--15  4--15  5--15  9--15 10--15 13--15  1--16  4--16
#> [41] 12--16  7--17 11--17  8--18  9--18 10--18 16--18 17--18  6--19 10--19
#> [51] 11--19 15--19 18--19  1--20  4--20  5--20  7--20 14--20 16--20 18--20
#> [61] 19--20  4--21  6--21  7--21  8--21 12--21 16--21  2--22  5--22  8--22
#> [71]  9--22 11--22 12--22 14--22 15--22 17--22 19--22 20--22  1--23  5--23
#> + ... omitted several edges
#> 
#> [[2]]
#> IGRAPH fc9e0c1 U--- 25 93 -- Erdos-Renyi (gnp) graph
#> + attr: name (g/c), type (g/c), loops (g/l), p (g/n)
#> + edges from fc9e0c1:
#>  [1]  1-- 3  1-- 4  1-- 5  2-- 5  2-- 6  3-- 6  5-- 6  1-- 7  2-- 7  3-- 8
#> [11]  4-- 8  7-- 8  1-- 9  2-- 9  5-- 9  1--10  6--10  4--11  6--11  8--11
#> [21]  3--12  6--12  7--12  8--12  9--12 11--12  2--13 10--13 11--13  4--14
#> [31]  6--14  6--15 10--15 12--15 13--15  4--16 10--16 11--16 12--16  3--17
#> [41] 13--17 15--17 16--17  1--18  5--18 10--18 12--18  1--19  8--19 13--19
#> [51] 18--19  4--20  5--20  9--20 11--20 13--20 16--20 18--20 19--20  3--21
#> [61]  7--21  9--21 11--21 16--21  2--22  3--22  4--22 11--22 12--22 17--22
#> [71] 21--22  1--23  2--23  3--23  5--23  6--23  8--23 15--23 17--23 22--23
#> + ... omitted several edges
#> 
#> [[3]]
#> IGRAPH f6d966b U--- 25 77 -- Erdos-Renyi (gnp) graph
#> + attr: name (g/c), type (g/c), loops (g/l), p (g/n)
#> + edges from f6d966b:
#>  [1]  3-- 4  3-- 5  4-- 5  3-- 6  3-- 7  4-- 8  6-- 8  3-- 9  4-- 9  8-- 9
#> [11]  2--10  8--10  9--10  9--11  2--12  7--12  9--12 11--12  6--13  9--13
#> [21]  6--14  7--14  8--14  1--15  6--15  9--15  1--16  3--16  4--16  5--16
#> [31] 10--16 12--16 13--16  1--17  2--17  4--17  6--17  7--17 12--17  1--18
#> [41]  2--18  4--18  6--18  8--18 11--18 16--18 17--18  1--19  4--19  8--19
#> [51] 13--19 16--19  4--20  5--20  7--20  8--20 10--20 12--20 13--20  2--21
#> [61]  7--21 19--21 20--21  3--22  7--22 18--22 20--22  3--23 16--23 17--23
#> [71]  4--24 12--24 23--24 11--25 17--25 18--25 20--25
#> 
#> [[4]]
#> IGRAPH 5330dca U--- 25 99 -- Erdos-Renyi (gnp) graph
#> + attr: name (g/c), type (g/c), loops (g/l), p (g/n)
#> + edges from 5330dca:
#>  [1]  1-- 2  1-- 4  2-- 4  2-- 5  4-- 5  1-- 6  2-- 6  3-- 6  4-- 6  5-- 6
#> [11]  3-- 7  5-- 7  3-- 8  4-- 8  5-- 8  5-- 9  2--10  5--10  6--11  1--12
#> [21]  2--12  4--12  5--12  6--12  8--12  9--12  1--13  9--13 10--13 12--13
#> [31]  2--14  4--14  6--14  7--14 10--14  1--15  3--15  8--15  9--15 10--15
#> [41] 14--15  2--16  4--16  5--16  8--16 12--16  2--17 10--17 14--17 16--17
#> [51]  1--18  4--18  5--18  6--18 11--18 13--18 15--18 16--18  1--19  3--19
#> [61]  6--19  8--19 11--19 13--19 14--19  1--20  2--20  6--20  7--20  9--20
#> [71] 13--20  3--21 13--21 19--21  4--22  8--22 10--22 11--22 12--22 14--22
#> + ... omitted several edges
#> 
#> [[5]]
#> IGRAPH 86cdf31 U--- 25 99 -- Erdos-Renyi (gnp) graph
#> + attr: name (g/c), type (g/c), loops (g/l), p (g/n)
#> + edges from 86cdf31:
#>  [1]  1-- 2  2-- 3  1-- 4  3-- 4  3-- 5  5-- 6  4-- 8  5-- 8  1-- 9  2-- 9
#> [11]  5-- 9  6-- 9  1--10  2--10  5--10  9--10  8--11 10--11  3--12  6--12
#> [21]  9--12  2--13  3--13 10--13 11--13  4--14 12--14 13--14  1--15  3--15
#> [31]  4--15  6--15  7--15  8--15 11--15 14--15  5--16  6--16 11--16  4--17
#> [41]  6--17  7--17  8--17  9--17 10--17  4--18  6--18  8--18 10--18 11--18
#> [51] 13--18 15--18  2--19  6--19  9--19 12--19 16--19 18--19  1--20  2--20
#> [61]  3--20  5--20  7--20  8--20 11--20 15--20  1--21  2--21  9--21 12--21
#> [71] 15--21 18--21 20--21  1--22 10--22 12--22 15--22 20--22 21--22  1--23
#> + ... omitted several edges
#> 
#> [[6]]
#> IGRAPH 07368bf U--- 25 100 -- Erdos-Renyi (gnp) graph
#> + attr: name (g/c), type (g/c), loops (g/l), p (g/n)
#> + edges from 07368bf:
#>  [1]  1-- 4  2-- 4  3-- 4  1-- 5  4-- 6  3-- 7  5-- 7  3-- 8  5-- 8  6-- 9
#> [11]  8-- 9  2--10  3--10  9--10  3--11  9--11  6--12  8--12  4--13  6--13
#> [21]  8--13 10--13  2--14  5--14  9--14  1--15  5--15  8--15  9--15 10--15
#> [31] 11--15  1--16 12--16  2--17  4--17  7--17  8--17 11--17 13--17 15--17
#> [41]  1--18  3--18  5--18 16--18 17--18  9--19 10--19 12--19 15--19 16--19
#> [51] 17--19  3--20  8--20  9--20 11--20 12--20 15--20 17--20 19--20  1--21
#> [61]  3--21  7--21 11--21 12--21 13--21 20--21  1--22  8--22 10--22 12--22
#> [71] 16--22 17--22 19--22 21--22  6--23  7--23 10--23 14--23 17--23 22--23
#> + ... omitted several edges
#> 
#> [[7]]
#> IGRAPH a005df5 U--- 25 100 -- Erdos-Renyi (gnp) graph
#> + attr: name (g/c), type (g/c), loops (g/l), p (g/n)
#> + edges from a005df5:
#>  [1]  2-- 3  1-- 5  3-- 6  4-- 6  5-- 6  1-- 7  2-- 7  4-- 8  7-- 9  3--10
#> [11]  1--11  2--11  4--11  7--11  9--11  1--12  4--12  7--12 10--12  1--13
#> [21]  5--13  8--13 11--13  6--14  8--14  9--14  2--15  6--15  7--15 10--15
#> [31]  3--16  6--16  9--16 14--16  5--17 13--17 16--17  4--18  6--18  7--18
#> [41] 10--18 12--18 13--18  2--19  3--19  5--19  6--19 15--19 16--19  1--20
#> [51]  2--20  5--20  6--20  8--20  9--20 12--20 15--20 16--20 19--20  1--21
#> [61]  2--21  3--21  4--21  7--21  9--21 11--21 15--21 16--21 18--21 20--21
#> [71]  3--22  6--22  9--22 14--22 17--22 19--22  2--23  4--23  9--23 11--23
#> + ... omitted several edges
#> 
#> [[8]]
#> IGRAPH 771b2ae U--- 25 106 -- Erdos-Renyi (gnp) graph
#> + attr: name (g/c), type (g/c), loops (g/l), p (g/n)
#> + edges from 771b2ae:
#>  [1]  1-- 2  1-- 4  4-- 5  3-- 6  6-- 7  4-- 8  6-- 8  7-- 8  4-- 9  7-- 9
#> [11]  3--10  6--10  9--10  1--11  8--11  9--11 10--11  1--12  3--12  7--12
#> [21]  2--13  3--13  4--13  5--13  6--13  7--13 10--13 11--13  2--14 10--14
#> [31] 11--14  4--15 10--15 11--15 13--15 14--15  1--16  8--16 15--16  3--17
#> [41]  4--17  6--17  9--17 10--17 11--17 13--17 15--17  2--18  3--18  8--18
#> [51] 10--18 13--18 14--18 15--18  9--19 11--19 12--19 17--19  8--20  9--20
#> [61] 12--20 14--20 15--20  1--21  2--21  4--21 15--21 16--21 17--21 19--21
#> [71]  1--22  5--22  6--22  7--22 11--22 12--22 14--22 16--22 17--22 18--22
#> + ... omitted several edges
#> 
#> [[9]]
#> IGRAPH add6f49 U--- 25 113 -- Erdos-Renyi (gnp) graph
#> + attr: name (g/c), type (g/c), loops (g/l), p (g/n)
#> + edges from add6f49:
#>  [1]  1-- 3  2-- 3  3-- 4  1-- 5  3-- 5  4-- 5  2-- 6  3-- 6  4-- 6  5-- 6
#> [11]  1-- 7  4-- 7  1-- 8  1-- 9  7-- 9  1--10  4--10  5--10  6--10  7--10
#> [21]  1--11  3--11  5--11  6--11  7--11  2--12  3--12  6--12  7--12  1--13
#> [31]  2--13  4--13  5--13  7--13  8--13  1--14  3--14  7--14  8--14  9--14
#> [41] 12--14  1--15  3--15  5--15  7--15  8--16 15--16  1--17  4--17  5--17
#> [51]  9--17 11--17 13--17 15--17 16--17  1--18 10--18 11--18 12--18  2--19
#> [61]  5--19  8--19  9--19 11--19 13--19 15--19  2--20  6--20 11--20 15--20
#> [71] 19--20  3--21  6--21  7--21 11--21 13--21  1--22  7--22  8--22  9--22
#> + ... omitted several edges
#> 
#> [[10]]
#> IGRAPH 57d2f03 U--- 25 105 -- Erdos-Renyi (gnp) graph
#> + attr: name (g/c), type (g/c), loops (g/l), p (g/n)
#> + edges from 57d2f03:
#>  [1]  1-- 2  2-- 3  2-- 5  3-- 5  1-- 6  4-- 6  2-- 7  3-- 7  2-- 9  4-- 9
#> [11]  6-- 9  1--10  2--10  3--10  4--10  6--10  2--11  3--11  4--11  5--11
#> [21]  7--11  8--11  9--11  2--12  3--12  5--12 10--12  1--13  2--13  8--13
#> [31] 12--13  1--14  4--14  5--14 11--14 13--14  6--15  8--15  9--15 10--15
#> [41] 11--15 13--15 14--15  1--16  2--16  4--16  8--16 10--16 13--16  3--17
#> [51]  4--17  9--17 10--17 13--17 14--17 15--17  6--18  8--18 11--18 14--18
#> [61] 16--18 17--18  3--19  5--19 11--19 12--19 15--19 16--19  5--20  7--20
#> [71]  9--20 10--20 12--20 15--20  2--21  3--21 14--21 16--21 18--21 19--21
#> + ... omitted several edges
#> 
#> attr(,"class")
#> [1] "nvd"  "list"