Coercion to Network-Valued Data Object

This function flags a list of igraph objects as an nvd object as defined in this package.

as_nvd(obj)

Arguments

obj
A list of igraph objects.

Value

An nvd object.

Examples

as_nvd(nvd("smallworld", 10))

#> [[1]]
#> IGRAPH 2a3d31a U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 2a3d31a:
#>  [1]  1-- 2  2-- 3  3-- 8  4-- 5  5-- 6  6-- 7  7--25  8--24 10--19 10--11
#> [11]  6--12 12--13 13--14 14--15 15--16 16--17 17--18 11--18 19--20 20--21
#> [21] 21--22 22--23 23--24 18--25  1--25  1-- 3  1--24  4--14  1--23  1-- 5
#> [31]  1--22  2--25  2-- 4  2--24  2-- 5  2--23  6--25  5--12  3--25  3-- 6
#> [41]  3--12  3-- 7  4-- 6  4-- 7  4--25  8--13  5-- 7  5-- 8  5--10  6-- 8
#> [51]  9--21  1--10  7-- 9  7--10 11--23  8--10 11--21 12--23  9--11  9--12
#> [61]  9--13 10--12 10--13  1--13 11--13 11--14 11--15 12--17 12--15 12--16
#> + ... omitted several edges
#> 
#> [[2]]
#> IGRAPH 93a219d U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 93a219d:
#>  [1]  1-- 2  2-- 3  3--18  4-- 5  5-- 6  6--25  7-- 8  8-- 9  9--10 10--11
#> [11] 11--12 12--13 13--14 14--15  6--16 16--17 17--18 18--19 19--20 20--21
#> [21] 21--22 12--22 23--24 16--25  1--25  1-- 3  1--24  1-- 4  1--23  1-- 5
#> [31]  7--22  2--25  2-- 4  2--14  2-- 5  4--23  2-- 6  3-- 5  3--25  6--20
#> [41]  3--24  3-- 7  4-- 6  4-- 7  4--25  4-- 8  5-- 7  5-- 8  5-- 9  6-- 8
#> [51]  6-- 9  6--10  7-- 9  7--10 11--24  8--10  8--11  8--12  9--11  9--12
#> [61]  9--13 10--12 10--13 10--14 11--13 11--14 11--15  6--14 12--15 12--16
#> + ... omitted several edges
#> 
#> [[3]]
#> IGRAPH 650034c U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 650034c:
#>  [1]  1-- 2  2--17  3-- 4  4--17  5--14  6-- 7  7-- 8  8--16  9--10 11--20
#> [11] 11--12 12--13 13--14 23--24 15--16 16--17 17--18 18--19 11--19 10--20
#> [21] 21--22 22--23 16--23 24--25  1--25  1-- 3  1--24  1-- 4  1--23  1-- 5
#> [31]  1--22  2--25  2--21  2--24  2--15  2--23  2-- 6  3-- 5  3--25  3-- 6
#> [41]  3--24  7--18  4-- 6  4-- 7  4--25  4-- 8  5-- 7  5-- 8  5-- 9  8-- 9
#> [51]  6-- 9  6--10  7-- 9  7--10  7--11  8--10  1--11  8--12  9--11 12--20
#> [61]  9--13 10--12 10--13 10--14 11--13 11--14 11--15 12--14  4--21  2--12
#> + ... omitted several edges
#> 
#> [[4]]
#> IGRAPH d33cb73 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from d33cb73:
#>  [1]  1-- 2  2--18  3-- 4  5--13  5-- 6  7--24  2-- 8  3-- 9  9--10 10--11
#> [11] 11--12 13--19 13--14 14--15 15--16 16--17 17--18 18--19 19--20 20--21
#> [21] 21--22  4--22 16--24 24--25  1--25  1-- 3  1--24  1-- 4  1--23  1-- 5
#> [31]  1--22  2--25  4-- 9  2--24  2-- 5  2--23  2-- 6  5--23  3--23  3--21
#> [41]  3--24  3--16  4-- 6  4-- 7  4--25  4-- 8  5-- 7  5-- 8  5--24  6-- 8
#> [51]  6-- 9  6--10  7-- 9  7--10  7--11  8--10  8--11  7--12  9--11  9--12
#> [61]  9--13 12--22 10--13 10--14 11--13 11--14 11--15 12--14 12--18 12--16
#> + ... omitted several edges
#> 
#> [[5]]
#> IGRAPH cc5d4af U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from cc5d4af:
#>  [1]  1-- 2  3--10  3-- 4  4-- 5  5-- 6  1-- 6  7-- 8  8-- 9  9--10 10--11
#> [11] 11--12 12--13 13--14 14--15 15--16 16--17 17--18 18--19 14--20 20--21
#> [21] 21--22 22--23 23--24 24--25 16--25  1-- 3  1--24  1-- 4  1--10  1-- 5
#> [31]  1--22  8--25  2-- 4  9--24  2-- 5  2--23  2-- 6  3-- 5  3--12  3--21
#> [41]  3--24  3-- 7  4--11  7--20  4--23  4-- 8  5-- 7  8--21  5-- 9  6-- 8
#> [51]  3-- 9 10--18  2-- 7  7--10  7--11 10--23  8--11  4--12  9--11  9--12
#> [61]  9--13 21--25 10--13 10--14 11--13 11--15 11--18 12--14 12--15 12--25
#> + ... omitted several edges
#> 
#> [[6]]
#> IGRAPH 1e62d46 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 1e62d46:
#>  [1]  1-- 2  2-- 3  3-- 4  5--23  5-- 6  6-- 7  7-- 8  9--17  9--10 10--11
#> [11]  5--12 13--24 13--14 14--19 15--16 16--17 17--18 18--19 19--20 20--21
#> [21] 21--22 22--23  7--13 23--24  9--25  1-- 3  8--24  1-- 4  1--23  1-- 5
#> [31]  1--22  2--25  2-- 4  7--24  2-- 5  2--23  6--13  3--22  3--11  3-- 6
#> [41]  3--24  3-- 7  4-- 6  4-- 7  4--11  4-- 8  5-- 7  8--22  5-- 9  6-- 8
#> [51]  6--14  6--10  7-- 9  7--10  7--11 10--24  8--11  8--12  9--11  9--12
#> [61]  9--13 12--25 10--13 10--16 11--13 11--14 11--15 12--14 12--15 12--16
#> + ... omitted several edges
#> 
#> [[7]]
#> IGRAPH 6084e33 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 6084e33:
#>  [1]  1-- 2  2-- 3  3--13  4-- 5  6--19  6-- 7  7-- 8  8--16  9--24 10--11
#> [11] 12--25 12--13 13--14 15--20 15--16 16--17 17--18 18--19 19--20 12--20
#> [21] 21--22 22--23  4--24 24--25 18--25  1-- 3 12--24  4--11  1--23  1-- 5
#> [31]  1--22  2--25  4--17 16--24 11--16  2--23  2-- 6  3-- 5  3--25  3-- 6
#> [41]  3--24  3-- 7  6--15  4-- 7  4--25  4-- 8  1-- 7  5-- 8  9--20  6-- 8
#> [51]  6-- 9  6--10  7-- 9  7--10  7--11  8--10  8--11  8--12  4--23  9--12
#> [61] 13--22  9--10 10--13  1--16 11--13  8--14 11--15 12--14  6--12 12--16
#> + ... omitted several edges
#> 
#> [[8]]
#> IGRAPH a9dff09 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from a9dff09:
#>  [1]  2--19  2-- 3  3-- 4  4-- 5  5-- 6  2-- 7  7-- 8  8-- 9  4--10 10--11
#> [11] 11--12 12--13 13--14  8--15 15--16 16--17 18--22 18--19 12--19 20--21
#> [21] 21--22 13--22 23--24 24--25  1--10  1-- 3  1--24  1-- 4  1--23  1-- 5
#> [31]  1--22  2--25  2-- 4  8--24  2-- 5  2--23  6--16  3-- 5  3--25  3-- 6
#> [41] 11--24  3-- 7  1-- 6  4-- 7  4--25  4-- 8  5-- 7  5-- 8  5-- 9  8--17
#> [51]  6-- 9  6--25  7--22  7--10  7--11  8--10  8--11  8--18  9--17  9--12
#> [61]  9--13 10--12 13--24 14--22 11--13 11--14 11--15 12--14 12--15 12--16
#> + ... omitted several edges
#> 
#> [[9]]
#> IGRAPH b0c7c65 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from b0c7c65:
#>  [1]  1--18  2-- 3  3-- 8  4-- 5  5-- 6  6-- 7  7-- 8  8-- 9  9--10 10--11
#> [11] 11--12 12--13  8--14 14--15 15--16 16--17 17--18 18--19 19--20 20--21
#> [21]  6--21 22--23 16--24 24--25  1--25  1-- 3  1--24  4-- 9  1--23  1-- 5
#> [31]  1--22  2--25  4--15  9--24  2-- 5  2--23  2--10  3--13  3--25  3-- 6
#> [41]  3--24  3-- 7  4-- 6  4-- 7  4--25  4-- 8  7--13  5-- 8  5-- 9  6-- 8
#> [51]  6-- 9  2-- 6  7--18  7--10  7--11 11--22  8--11  8--12  9--17  9--12
#> [61]  5--13 10--12  1--13 10--14 11--13 11--14 11--15 12--18 12--15 12--16
#> + ... omitted several edges
#> 
#> [[10]]
#> IGRAPH 212b7f6 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 212b7f6:
#>  [1]  1-- 2  2-- 3  3-- 4  5--20  5-- 6  6-- 7  7-- 8  8-- 9  9--10 10--11
#> [11] 11--12 12--13  3--13 14--15 15--16 12--17 17--18 18--19 19--20 12--20
#> [21] 21--22 11--23  7--24 16--24  1--25  1-- 3  1--24  1-- 4  1--23  1-- 5
#> [31]  1--22  2--25  4--21  2--24  2--22 16--23  2-- 6  3--23  3--25  3-- 6
#> [41]  3--24  3-- 7  4-- 6  4-- 7  4--25  4--14  7--11  5-- 8  5-- 9  6-- 8
#> [51]  6-- 9  6--10  7-- 9  2--20  7--19  5--10  8--11  8--12 11--22  9--12
#> [61]  7--13 10--12  4--13 14--20 11--13 11--14 11--15 12--14 12--15 12--16
#> + ... omitted several edges
#> 
#> attr(,"class")
#> [1] "nvd"  "list"