Pairwise Distance Matrix Between Two Samples of Networks

This function computes the matrix of pairwise distances between all the elements of the two samples put together. The cardinality of the fist sample is denoted by \(n1\) and that of the second one is denoted by \(n2\).

dist_nvd(x, y = NULL, representation = "adjacency", distance = "frobenius")

Arguments

x
A list of igraph objects or matrix representations of underlying networks from a given first population.

y
A list of igraph objects or matrix representations of underlying networks from a given second population.

representation
A string specifying the desired type of representation, among: "adjacency", "laplacian", "modularity" or "graphon". Default is "laplacian".

distance
A string specifying the chosen distance for calculating the test statistic, among: "hamming", "frobenius", "spectral" and "root-euclidean". Default is "frobenius".

Value

A matrix of dimension \((n1+n2) \times (n1+n2)\) containing the distances between all the elements of the two samples put together.

Examples

x <- nvd("smallworld", 10)
y <- nvd("pa", 10)
dist_nvd(x, y, "adjacency", "spectral")

#>            1         2         3         4         5         6         7
#> 2  1.0007768                                                            
#> 3  1.2547340 1.5466917                                                  
#> 4  1.1169646 0.9945358 1.4009844                                        
#> 5  1.0582401 0.9119501 1.1583463 1.0206689                              
#> 6  1.2925743 1.5747227 1.0627523 1.6896397 1.4061843                    
#> 7  0.8522725 0.8152810 1.4329624 0.9015339 1.0508128 1.4070863          
#> 8  1.0972702 1.2641164 1.0145594 1.3118260 1.0794563 1.0998601 1.1492968
#> 9  0.9154579 1.0332619 1.0348070 0.9621600 0.9226440 1.0446769 0.9111684
#> 10 1.2916594 1.3218818 0.8787413 1.2046472 1.1687503 1.1145917 1.2823113
#> 11 3.6523175 3.3542895 3.8331883 3.4831538 3.4236746 4.0059810 3.4321786
#> 12 3.3714076 3.2108553 3.6021123 3.3344429 3.3061208 3.8463913 3.2175668
#> 13 3.1050038 2.9921942 3.2736148 3.0262021 2.9202576 3.5529607 2.9923266
#> 14 3.1218742 2.7995780 3.2667746 2.9442897 2.8911313 3.4758313 2.8660398
#> 15 2.9482339 2.7287075 3.1867729 2.8356126 2.8386638 3.3808853 2.7159246
#> 16 3.1880146 3.0040759 3.3331349 3.0828212 3.0458908 3.5617881 3.0459748
#> 17 3.6825038 3.5274752 3.7419617 3.6684135 3.5222828 3.9453571 3.5721592
#> 18 3.6181664 3.5100010 3.8105941 3.5539275 3.5007036 4.0512362 3.4821672
#> 19 2.7952487 2.5021716 3.0959980 2.7275347 2.6273020 3.2779815 2.6891832
#> 20 3.3520366 3.1648039 3.5922332 3.3581166 3.2305809 3.7035344 3.2015625
#>            8         9        10        11        12        13        14
#> 2                                                                       
#> 3                                                                       
#> 4                                                                       
#> 5                                                                       
#> 6                                                                       
#> 7                                                                       
#> 8                                                                       
#> 9  0.8224552                                                            
#> 10 0.8304619 0.8303592                                                  
#> 11 3.8554517 3.8417040 4.0044336                                        
#> 12 3.7129192 3.6346627 3.8339888 1.1323717                              
#> 13 3.3979595 3.3872015 3.5169194 1.0685720 1.0566684                    
#> 14 3.3678246 3.2510326 3.4565086 1.0766940 1.0345037 1.2064370          
#> 15 3.1971033 3.1646878 3.3048455 1.2330693 0.9555771 1.0810493 1.0951012
#> 16 3.4382879 3.3859602 3.5187702 1.1763674 1.0618766 0.9889068 0.9933931
#> 17 3.8566874 3.9151933 4.0083176 1.0406491 1.2422766 1.1056048 1.4185265
#> 18 3.8886366 3.9017288 4.0800380 1.0798317 1.3499407 0.9874588 1.5540719
#> 19 3.1221091 3.0142184 3.1907978 1.3474596 1.2003604 1.2201466 1.0278564
#> 20 3.6082035 3.6284070 3.7872077 0.7804937 1.0489378 0.9591493 1.1400543
#>           15        16        17        18        19
#> 2                                                   
#> 3                                                   
#> 4                                                   
#> 5                                                   
#> 6                                                   
#> 7                                                   
#> 8                                                   
#> 9                                                   
#> 10                                                  
#> 11                                                  
#> 12                                                  
#> 13                                                  
#> 14                                                  
#> 15                                                  
#> 16 1.0185227                                        
#> 17 1.3543988 1.0698290                              
#> 18 1.5259018 1.4985811 1.1488816                    
#> 19 0.9590243 1.0567554 1.5113757 1.6927370          
#> 20 1.1397908 1.1429681 0.9274965 1.0456956 1.2300453