Fréchet Mean of Network-Valued Data

This function computes the sample Fréchet mean from an observed sample of network-valued random variables according to a specified matrix representation. It currently only supports the Euclidean geometry i.e. the sample Fréchet mean is obtained as the argmin of the sum of squared Frobenius distances.

# S3 method for nvd
mean(x, weights = rep(1, length(x)), representation = "adjacency", ...)

Arguments

x
An nvd object.

weights
A numeric vector specifying weights for each observation (default: equally weighted).

representation
A string specifying the graph representation to be used (choices: adjacency [default], laplacian, modularity, graphon).

...
Other argument to be parsed to the mean function.

Value

The mean network in the chosen matrix representation assuming Euclidean geometry for now.

Examples

d <- nvd(n = 10L)
mean(d)

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]  0.0  0.3  0.9  0.8  0.8  0.1  0.2  0.1  0.0   0.1   0.3   0.2   0.0
#>  [2,]  0.3  0.0  0.7  0.6  0.8  0.9  0.0  0.1  0.2   0.2   0.3   0.3   0.3
#>  [3,]  0.9  0.7  0.0  0.9  0.8  0.7  0.8  0.1  0.0   0.1   0.1   0.2   0.1
#>  [4,]  0.8  0.6  0.9  0.0  0.9  0.8  0.9  0.9  0.1   0.1   0.2   0.0   0.1
#>  [5,]  0.8  0.8  0.8  0.9  0.0  0.8  0.4  0.5  0.6   0.4   0.3   0.2   0.0
#>  [6,]  0.1  0.9  0.7  0.8  0.8  0.0  0.8  0.6  0.9   0.8   0.0   0.3   0.1
#>  [7,]  0.2  0.0  0.8  0.9  0.4  0.8  0.0  1.0  0.9   0.7   0.6   0.3   0.1
#>  [8,]  0.1  0.1  0.1  0.9  0.5  0.6  1.0  0.0  0.8   0.8   0.8   0.7   0.1
#>  [9,]  0.0  0.2  0.0  0.1  0.6  0.9  0.9  0.8  0.0   0.5   0.8   0.7   0.7
#> [10,]  0.1  0.2  0.1  0.1  0.4  0.8  0.7  0.8  0.5   0.0   0.6   0.8   1.0
#> [11,]  0.3  0.3  0.1  0.2  0.3  0.0  0.6  0.8  0.8   0.6   0.0   0.6   0.6
#> [12,]  0.2  0.3  0.2  0.0  0.2  0.3  0.3  0.7  0.7   0.8   0.6   0.0   0.8
#> [13,]  0.0  0.3  0.1  0.1  0.0  0.1  0.1  0.1  0.7   1.0   0.6   0.8   0.0
#> [14,]  0.0  0.2  0.2  0.0  0.0  0.0  0.0  0.3  0.1   0.7   0.7   0.8   0.9
#> [15,]  0.2  0.1  0.0  0.1  0.0  0.3  0.2  0.0  0.1   0.2   0.6   0.6   0.8
#> [16,]  0.0  0.3  0.0  0.0  0.0  0.0  0.0  0.2  0.0   0.1   0.1   1.0   0.6
#> [17,]  0.0  0.1  0.2  0.1  0.2  0.2  0.1  0.1  0.2   0.3   0.3   0.2   0.8
#> [18,]  0.1  0.1  0.1  0.2  0.0  0.0  0.0  0.4  0.0   0.1   0.1   0.3   0.0
#> [19,]  0.2  0.1  0.1  0.0  0.3  0.2  0.0  0.1  0.1   0.0   0.3   0.2   0.1
#> [20,]  0.2  0.2  0.1  0.2  0.2  0.0  0.1  0.2  0.1   0.0   0.2   0.1   0.2
#> [21,]  0.0  0.3  0.1  0.3  0.0  0.1  0.0  0.0  0.2   0.1   0.1   0.0   0.2
#> [22,]  1.0  0.1  0.2  0.3  0.1  0.2  0.1  0.2  0.1   0.4   0.2   0.1   0.0
#> [23,]  0.9  0.6  0.1  0.0  0.1  0.1  0.1  0.4  0.2   0.0   0.0   0.2   0.3
#> [24,]  0.7  0.9  0.9  0.1  0.2  0.1  0.1  0.0  0.2   0.2   0.0   0.0   0.1
#> [25,]  0.8  0.7  0.8  0.9  0.2  0.0  0.0  0.0  0.0   0.0   0.1   0.0   0.1
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]   0.0   0.2   0.0   0.0   0.1   0.2   0.2   0.0   1.0   0.9   0.7   0.8
#>  [2,]   0.2   0.1   0.3   0.1   0.1   0.1   0.2   0.3   0.1   0.6   0.9   0.7
#>  [3,]   0.2   0.0   0.0   0.2   0.1   0.1   0.1   0.1   0.2   0.1   0.9   0.8
#>  [4,]   0.0   0.1   0.0   0.1   0.2   0.0   0.2   0.3   0.3   0.0   0.1   0.9
#>  [5,]   0.0   0.0   0.0   0.2   0.0   0.3   0.2   0.0   0.1   0.1   0.2   0.2
#>  [6,]   0.0   0.3   0.0   0.2   0.0   0.2   0.0   0.1   0.2   0.1   0.1   0.0
#>  [7,]   0.0   0.2   0.0   0.1   0.0   0.0   0.1   0.0   0.1   0.1   0.1   0.0
#>  [8,]   0.3   0.0   0.2   0.1   0.4   0.1   0.2   0.0   0.2   0.4   0.0   0.0
#>  [9,]   0.1   0.1   0.0   0.2   0.0   0.1   0.1   0.2   0.1   0.2   0.2   0.0
#> [10,]   0.7   0.2   0.1   0.3   0.1   0.0   0.0   0.1   0.4   0.0   0.2   0.0
#> [11,]   0.7   0.6   0.1   0.3   0.1   0.3   0.2   0.1   0.2   0.0   0.0   0.1
#> [12,]   0.8   0.6   1.0   0.2   0.3   0.2   0.1   0.0   0.1   0.2   0.0   0.0
#> [13,]   0.9   0.8   0.6   0.8   0.0   0.1   0.2   0.2   0.0   0.3   0.1   0.1
#> [14,]   0.0   0.8   0.7   0.7   0.8   0.0   0.2   0.1   0.1   0.0   0.1   0.0
#> [15,]   0.8   0.0   0.7   0.8   0.7   0.8   0.0   0.1   0.1   0.2   0.0   0.0
#> [16,]   0.7   0.7   0.0   0.8   0.5   0.7   0.8   0.1   0.4   0.1   0.2   0.1
#> [17,]   0.7   0.8   0.8   0.0   0.6   0.9   0.8   0.7   0.0   0.0   0.2   0.2
#> [18,]   0.8   0.7   0.5   0.6   0.0   0.7   0.7   0.9   0.7   0.2   0.3   0.1
#> [19,]   0.0   0.8   0.7   0.9   0.7   0.0   0.7   0.7   0.9   0.8   0.1   0.0
#> [20,]   0.2   0.0   0.8   0.8   0.7   0.7   0.0   0.9   0.7   0.9   0.7   0.0
#> [21,]   0.1   0.1   0.1   0.7   0.9   0.7   0.9   0.0   0.9   0.6   0.8   0.9
#> [22,]   0.1   0.1   0.4   0.0   0.7   0.9   0.7   0.9   0.0   0.9   0.8   0.6
#> [23,]   0.0   0.2   0.1   0.0   0.2   0.8   0.9   0.6   0.9   0.0   0.9   0.6
#> [24,]   0.1   0.0   0.2   0.2   0.3   0.1   0.7   0.8   0.8   0.9   0.0   0.8
#> [25,]   0.0   0.0   0.1   0.2   0.1   0.0   0.0   0.9   0.6   0.6   0.8   0.0
#> attr(,"representation")
#> [1] "adjacency"