Network-Valued Data Constructor

This is the constructor for objects of class nvd.

nvd(
  model = "smallworld",
  n = 0L,
  pref.matrix = NULL,
  lambda = NULL,
  size = NULL,
  prob = NULL
)

Arguments

model
A string specifying the model to be used for sampling networks (current choices are: "sbm", "k_regular", "gnp", "smallworld", "pa", "poisson" and "binomial"). Default is "smallworld".

n
An integer specifying the sample size (default: 0L).

pref.matrix
A matrix giving the Bernoulli rates for the SBM generator (see sample_sbm for details). Default is NULL. It is required for model == "sbm".

lambda
A numeric value specifying the mean value for the Poisson generator. Default is NULL. It is required for model == "poisson".

size
An integer value specifying the number of trials for the binomial distribution. Default is NULL. It is required for model == "binomial".

prob
A numeric value specifying the probability of success of each trial for the binomial distribution. Default is NULL. It is required for model == "binomial".

Value

A nvd object which is a list of igraph objects.

Examples

nvd(n = 10L)

#> [[1]]
#> IGRAPH 094d996 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 094d996:
#>  [1]  1-- 2  2-- 3  3-- 4  4-- 5  5-- 6  6-- 7  7--19  8-- 9  9--10 10--11
#> [11] 11--12 12--13 11--13 14--15 15--16 16--17  7--17 18--19 19--20 20--21
#> [21] 15--21 22--23 23--24 12--24  1--25  1-- 3  1--24  1-- 4  1--23  1--13
#> [31]  1--22  2--25  2-- 4  6--18  2-- 5  2--23  2--21  5--21  3--25  3-- 6
#> [41]  3--24  3--23  4-- 6  4-- 7  4--25  4-- 8  7--21  5-- 8  5-- 9  6-- 8
#> [51]  6-- 9  6--10  2-- 9 10--23  7--11  5--10  8--11  8--12  9--11  2--12
#> [61]  9--13 10--12 10--13 14--19 11--25 11--14 11--15 12--14 12--15  1--16
#> + ... omitted several edges
#> 
#> [[2]]
#> IGRAPH dfc0ddb U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from dfc0ddb:
#>  [1]  1-- 2  2--14  3-- 4  4--13  5-- 6  6--13  7-- 8  9--14  9--10 17--25
#> [11] 12--18 12--13 13--14 14--15 15--16 16--18 17--18 18--19 19--20 20--21
#> [21] 21--22 22--23 23--24 11--24  1--25  1-- 3  1--24  1-- 4  1--23  1-- 5
#> [31]  1--13  2--25  4--24  2--24  2--12  4--23  2--16  3-- 5  3--25  3-- 6
#> [41]  3--24  3-- 7  4-- 6  4--12  6--25  4-- 8  5-- 7  5-- 8  5--25  6-- 8
#> [51]  6-- 9  6--10  9--23  7--10  7--11  8--24  8--11  8--12  9--11 12--25
#> [61]  7--13  3--12 10--13 10--14 11--13 11--14 11--15 12--14  5--12 12--16
#> + ... omitted several edges
#> 
#> [[3]]
#> IGRAPH 567b0e5 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 567b0e5:
#>  [1]  1-- 2  3-- 9 11--13  4--20  5-- 6  6-- 7  7-- 8  8-- 9  9--10 10--11
#> [11]  4--11 12--13 13--14 11--14 15--16 16--17 17--18 18--19  3--20 20--21
#> [21] 21--22 22--23 12--23 24--25  1--19  1-- 3  1--24  1--20  1--23  1-- 5
#> [31]  5--22  2--25  3-- 4  2--24  2-- 5  2-- 9  2-- 6  3-- 5  3--25  6--17
#> [41]  3--24  3-- 7  6--25  4-- 7  4--25  4-- 8  5-- 7  5-- 8  5-- 9  8--17
#> [51]  6--14  6--10  7--19  7--10  7--11 10--21  8--11  8--23  9--11  9--12
#> [61]  9--13 12--22 10--13 14--25 11--20  5--11  2--11 12--14 12--15 12--16
#> + ... omitted several edges
#> 
#> [[4]]
#> IGRAPH 23e63e2 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 23e63e2:
#>  [1]  1-- 2  2-- 3  3--17  4-- 5  5-- 6  6-- 7  7-- 8  8-- 9  9--10 10--11
#> [11] 11--23 12--13  1--13 14--15 15--16 16--17 17--18 18--19 19--20 20--21
#> [21] 21--22 22--23 20--24 24--25  1--25  1-- 3  1-- 7  1--21  1--23  1-- 5
#> [31]  1--22  2--25  2-- 4  2--24  5--15  2--23  2-- 6  5--23  3--25  3-- 6
#> [41]  3--24  3-- 7  4-- 6  7--14  4--25  4-- 8  5-- 7  8--20  5-- 9  6-- 8
#> [51]  6--13  6--10  7-- 9  3--10  7--11  8--10  8--11  8--12  9--11 12--17
#> [61]  9--13 10--12 10--13 10--14 11--21  5--14 11--15 12--14 15--20 12--16
#> + ... omitted several edges
#> 
#> [[5]]
#> IGRAPH 723260b U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 723260b:
#>  [1]  2--18  2-- 3  3-- 4  4-- 5  5-- 6  6-- 7  7-- 8  8-- 9 10--16  4--10
#> [11] 11--22 12--13 13--14 14--15 15--16 16--17  1--15 18--19 19--20 20--21
#> [21] 21--22  4--22 23--24 24--25  1--25  3--20  1--24  1--21  1--23  1-- 5
#> [31]  1--22  2--25  2-- 4  2--24  2-- 5  2--23  2-- 6  3-- 5  3--25  3-- 6
#> [41]  3--24  3-- 7 13--18  4-- 7  4--25  4-- 8  1-- 7  8--24  5-- 9  6-- 8
#> [51]  6--17  6--22  7-- 9  7--10  5--11  8--10  8--11  8--12  9--11 12--20
#> [61]  9--13 12--17 10--13  7--14 11--13 11--14 13--15 12--14 12--15 12--16
#> + ... omitted several edges
#> 
#> [[6]]
#> IGRAPH 18a6ef3 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 18a6ef3:
#>  [1]  2--10  2-- 3  3-- 4  4-- 5  5--13  6-- 7  7-- 8  8-- 9  9--10 10--11
#> [11] 11--12 12--20 13--14 14--15 15--16 11--16 17--18 18--19 13--20 14--20
#> [21] 21--22 22--23 23--24 24--25  1--10  1-- 3  1--24  1--15  1--23  1-- 5
#> [31]  1--22  2--25  2-- 4 10--19  2-- 5  2--23  6--12  3--10  3--25  3-- 6
#> [41]  3--24 16--23  4-- 6  4-- 7  4--25  4-- 8  5-- 7  5-- 8  5-- 9  6-- 8
#> [51]  6-- 9  6--10  7-- 9  7--10  7--11  8--10  8--11  8--12  9--11  9--12
#> [61] 11--13 10--12 10--13 10--14  3--13 11--14 11--15 12--14 12--15  6--16
#> + ... omitted several edges
#> 
#> [[7]]
#> IGRAPH 3d0d0c6 U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 3d0d0c6:
#>  [1]  1--13  2-- 3  3-- 4  4-- 5  5-- 6  1-- 7  7-- 8  7--25  4--10 10--11
#> [11] 11--12 12--13 13--14 14--15 15--16 16--17 17--22 18--19  6--20 10--21
#> [21] 15--21  4--22 23--24 24--25  1--25  1-- 3  1--24  1-- 8  1--23  1-- 5
#> [31] 22--24  2--25  2-- 4  2--10  2-- 5  2--23  2--12  2--18  3--25  3--19
#> [41]  3--24  3-- 7  4-- 6  4-- 7  4--25  4--14  5-- 7  5-- 8  5--20  6-- 8
#> [51]  6-- 9  6--10  7--23  7--10  7--11  8--10  8--11  8--17 11--23  9--12
#> [61]  9--13 10--12 10--13 10--14 11--13 11--14 11--15 12--14 12--15 12--16
#> + ... omitted several edges
#> 
#> [[8]]
#> IGRAPH 0aa457d U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 0aa457d:
#>  [1]  1-- 2  2-- 3  3-- 4  4-- 5  7--15  6-- 7  7-- 8  8-- 9  9--10 10--19
#> [11] 11--12 12--13 13--14 10--15  2--15 16--17 17--18 18--19 19--20  7--20
#> [21] 21--22 15--23 23--24 24--25  1--16  1-- 3 20--24  1-- 5  1--23  1-- 7
#> [31]  1--22  2--25  4-- 6  2--24  5--24  2--23  2-- 6  3-- 5  3--25  3-- 6
#> [41]  3--24  3-- 7 11--24  4-- 7  3--15  4-- 8  5-- 7  8--22  5-- 9  6-- 8
#> [51]  6-- 9 10--22  7--21  7--10  7--11  8--10  8--11  8--12  9--11  3--12
#> [61]  9--13 10--12 10--13  4--14 11--13 14--24 11--15 12--22 12--15 12--16
#> + ... omitted several edges
#> 
#> [[9]]
#> IGRAPH ec8e0bd U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from ec8e0bd:
#>  [1]  1-- 2  2-- 3  3--20  4-- 5  5-- 6  7--23  7-- 8  9--16  9--10 10--11
#> [11] 11--12 12--13 13--14 15--23 15--16 16--17 17--18 18--19 19--20 20--21
#> [21] 21--22 22--23 23--24 24--25  1--25  1-- 3  1--12  1--14  1--23  1-- 5
#> [31]  1--22  2--25  2-- 4  2--24  2-- 5  2-- 8  2-- 6  3-- 5  3--25  6--12
#> [41]  3--24  3-- 7  1-- 4  4-- 7  4--10  4-- 8  5-- 7  5-- 8  9--21  6-- 8
#> [51]  6--15  6--10 15--22  7--10  6--11 15--17  8--11  8--12  9--11  9--12
#> [61]  4-- 9 10--12 10--13 10--14 11--13 11--14 11--15 12--14  4--15 12--16
#> + ... omitted several edges
#> 
#> [[10]]
#> IGRAPH 3699a0c U--- 25 100 -- Watts-Strogatz random graph
#> + attr: name (g/c), dim (g/n), size (g/n), nei (g/n), p (g/n), loops
#> | (g/l), multiple (g/l)
#> + edges from 3699a0c:
#>  [1]  1-- 2 13--21  4--24  4-- 5  5-- 6  7--21  7-- 8  8-- 9  9--10  5--11
#> [11] 11--12 12--13 13--14 14--15 15--16 16--17 17--25 10--18 19--20 20--21
#> [21] 21--22 22--23 23--24 24--25  1--25  1-- 3  1--24  4--22 16--23  1-- 5
#> [31]  1--22  2--25  2-- 4  2--24  2--18  2--23  2-- 6  5--12  3--25  6--24
#> [41]  8--24  3-- 7  4-- 6  4--16  4--25  4-- 8  5-- 7  5--21  5-- 9  6-- 8
#> [51]  6-- 9  6--10  7-- 9  7--10  6--11  8--10  8--11  8--12  3-- 9  9--12
#> [61]  9--13 10--12 10--13  4--14 11--13 14--25 11--15 12--14 12--15 12--16
#> + ... omitted several edges
#> 
#> attr(,"class")
#> [1] "nvd"  "list"