Power Simulations for Permutation Tests

This function provides a Monte-Carlo estimate of the power of the permutation tests proposed in this package.

power2(
  model1 = "gnp",
  model2 = "k_regular",
  n1 = 20L,
  n2 = 20L,
  size1 = NULL,
  prob1 = NULL,
  size2 = NULL,
  prob2 = NULL,
  lambda1 = NULL,
  lambda2 = NULL,
  pref.matrix1 = NULL,
  pref.matrix2 = NULL,
  representation = "adjacency",
  distance = "frobenius",
  stats = c("flipr:t_ip", "flipr:f_ip"),
  B = 1000L,
  alpha = 0.05,
  test = "exact",
  k = 5L,
  R = 1000L,
  seed = NULL
)

Arguments

model1
A string specifying the model to be used for sample 1 among "sbm", "k_regular", "gnp", "smallworld", "pa", "poisson" and "binomial". Default is "gnp".

model2
A string specifying the model to be used for sample 2 among "sbm", "k_regular", "gnp", "smallworld", "pa", "poisson" and "binomial". Default is "k_regular".

n1
The size of sample 1 (default: 20L).

n2
The size of sample 2 (default: 20L).

size1
The number of trials for the binomial distribution of sample 1 (default: NULL). Must be specified if model == "binomial".

prob1
The probability of success of each trial for the binomial distribution of sample 1 (default: NULL). Must be specified if model == "binomial".

size2
The number of trials for the binomial distribution of sample 2 (default: NULL). Must be specified if model == "binomial".

prob2
The probability of success of each trial for the binomial distribution of sample 2 (default: NULL). Must be specified if model == "binomial".

lambda1
The mean of the Poisson distribution of sample 1 (default: NULL). Must be specified if model == "poisson".

lambda2
The mean of the Poisson distribution of sample 2 (default: NULL). Must be specified if model == "poisson".

pref.matrix1
A matrix giving the Bernoulli rates for the SBM generator of sample 1 (default: NULL). Must be specified if model == "sbm".

pref.matrix2
A matrix giving the Bernoulli rates for the SBM generator of sample 2 (default: NULL). Must be specified if model == "sbm".

representation
A string specifying the desired type of representation, among: "adjacency", "laplacian" and "modularity". Defaults to "adjacency".

distance
A string specifying the chosen distance for calculating the test statistic, among: "hamming", "frobenius", "spectral" and "root-euclidean". Defaults to "frobenius".

stats
A character vector specifying the chosen test statistic(s), among: "original_edge_count", "generalized_edge_count", "weighted_edge_count", "student_euclidean", "welch_euclidean" or any statistics based on inter-point distances available in the flipr package: "flipr:student_ip", "flipr:fisher_ip", "flipr:bg_ip", "flipr:energy_ip", "flipr:cq_ip". Defaults to c("flipr:student_ip", "flipr:fisher_ip").

B
The number of permutation or the tolerance. If this number is lower than 1, it is intended as a tolerance. Otherwise, it is intended as the number of required permutations. Defaults to 1000L.

alpha
Significance level for hypothesis testing (default: 0.05).

test
A character string specifying the formula to be used to compute the permutation p-value. Choices are "estimate", "upper_bound" and "exact". Defaults to "exact" which provides exact tests.

k
An integer specifying the density of the minimum spanning tree used for the edge count statistics. Defaults to 5L.

R
The number of Monte-Carlo runs used to estimate the power (default: 1000L).

seed
An integer specifying the seed to start randomness from (default: uses clock).

Value

A numeric value estimating the power of the test.

Details

Currently, six scenarios of pairs of populations are implemented. Scenario 0 allows to make sure that all our permutation tests are exact.

Examples

power2(R = 100, B = 100)

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> Warning: '.Random.seed' is not an integer vector but of type 'NULL', so ignored

#> ! Setting the seed for sampling permutations is mandatory for obtaining a continuous p-value function. Using `seed = 1234`.

#> [1] 0.01