Network Representation Functions

This is a collection of functions that convert a graph stored as an igraph object into a desired matrix representation among adjacency matrix, graph laplacian, modularity matrix or graphon (edge probability matrix).

Usage

repr_adjacency(network, validate = TRUE)

repr_laplacian(network, validate = TRUE)

repr_modularity(network, validate = TRUE)

repr_graphon(network, validate = TRUE)

Arguments

network

An igraph object.

validate

A boolean specifying whether the function should check the class of its input (default: TRUE).

Value

A numeric square matrix giving the desired network representation recorded in the object's class.

Examples

g <- igraph::sample_smallworld(1, 25, 3, 0.05)
repr_adjacency(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0    1    1    1    0    0    0    0    0     0     0     0     0
#>  [2,]    1    0    1    1    1    0    0    0    0     0     0     0     0
#>  [3,]    1    1    0    1    1    1    0    0    0     0     0     0     0
#>  [4,]    1    1    1    0    1    1    0    0    0     0     0     0     0
#>  [5,]    0    1    1    1    0    1    1    1    0     0     0     0     0
#>  [6,]    0    0    1    1    1    0    1    0    1     0     0     0     0
#>  [7,]    0    0    0    0    1    1    0    1    1     1     0     0     0
#>  [8,]    0    0    0    0    1    0    1    0    1     1     1     0     0
#>  [9,]    0    0    0    0    0    1    1    1    0     1     1     1     0
#> [10,]    0    0    0    0    0    0    1    1    1     0     1     1     1
#> [11,]    0    0    0    0    0    0    0    1    1     1     0     1     1
#> [12,]    0    0    0    0    0    0    0    0    1     1     1     0     1
#> [13,]    0    0    0    0    0    0    0    0    0     1     1     1     0
#> [14,]    0    0    0    0    0    0    0    0    0     0     1     1     1
#> [15,]    0    0    0    0    0    0    0    0    0     0     0     1     1
#> [16,]    0    0    1    0    0    0    0    0    0     0     0     0     1
#> [17,]    0    1    0    1    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    1    0    0    0     0     0     0     0
#> [23,]    1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]    1    1    0    0    0    0    0    0    0     0     0     0     0
#> [25,]    0    1    1    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0     1     1     0
#>  [2,]     0     0     0     1     0     0     0     0     0     0     1     1
#>  [3,]     0     0     1     0     0     0     0     0     0     0     0     1
#>  [4,]     0     0     0     1     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     1     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]     1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]     1     1     0     0     0     0     0     0     0     0     0     0
#> [13,]     1     1     1     0     0     0     0     0     0     0     0     0
#> [14,]     0     1     0     1     0     0     0     0     0     0     0     0
#> [15,]     1     0     1     1     1     0     0     0     0     0     0     0
#> [16,]     0     1     0     1     1     1     0     0     0     0     0     0
#> [17,]     1     1     1     0     1     0     1     0     0     0     0     0
#> [18,]     0     1     1     1     0     1     1     1     0     0     0     0
#> [19,]     0     0     1     0     1     0     1     1     1     0     0     0
#> [20,]     0     0     0     1     1     1     0     1     1     1     0     1
#> [21,]     0     0     0     0     1     1     1     0     1     1     1     1
#> [22,]     0     0     0     0     0     1     1     1     0     1     1     1
#> [23,]     0     0     0     0     0     0     1     1     1     0     1     0
#> [24,]     0     0     0     0     0     0     0     1     1     1     0     1
#> [25,]     0     0     0     0     0     0     1     1     1     0     1     0
#> attr(,"representation")
#> [1] "adjacency"
repr_laplacian(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    5   -1   -1   -1    0    0    0    0    0     0     0     0     0
#>  [2,]   -1    7   -1   -1   -1    0    0    0    0     0     0     0     0
#>  [3,]   -1   -1    7   -1   -1   -1    0    0    0     0     0     0     0
#>  [4,]   -1   -1   -1    6   -1   -1    0    0    0     0     0     0     0
#>  [5,]    0   -1   -1   -1    6   -1   -1   -1    0     0     0     0     0
#>  [6,]    0    0   -1   -1   -1    6   -1    0   -1     0     0     0     0
#>  [7,]    0    0    0    0   -1   -1    5   -1   -1    -1     0     0     0
#>  [8,]    0    0    0    0   -1    0   -1    5   -1    -1    -1     0     0
#>  [9,]    0    0    0    0    0   -1   -1   -1    6    -1    -1    -1     0
#> [10,]    0    0    0    0    0    0   -1   -1   -1     6    -1    -1    -1
#> [11,]    0    0    0    0    0    0    0   -1   -1    -1     6    -1    -1
#> [12,]    0    0    0    0    0    0    0    0   -1    -1    -1     6    -1
#> [13,]    0    0    0    0    0    0    0    0    0    -1    -1    -1     6
#> [14,]    0    0    0    0    0    0    0    0    0     0    -1    -1    -1
#> [15,]    0    0    0    0    0    0    0    0    0     0     0    -1    -1
#> [16,]    0    0   -1    0    0    0    0    0    0     0     0     0    -1
#> [17,]    0   -1    0   -1    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0   -1    0    0    0     0     0     0     0
#> [23,]   -1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]   -1   -1    0    0    0    0    0    0    0     0     0     0     0
#> [25,]    0   -1   -1    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0    -1    -1     0
#>  [2,]     0     0     0    -1     0     0     0     0     0     0    -1    -1
#>  [3,]     0     0    -1     0     0     0     0     0     0     0     0    -1
#>  [4,]     0     0     0    -1     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0    -1     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]    -1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]    -1    -1     0     0     0     0     0     0     0     0     0     0
#> [13,]    -1    -1    -1     0     0     0     0     0     0     0     0     0
#> [14,]     5    -1     0    -1     0     0     0     0     0     0     0     0
#> [15,]    -1     6    -1    -1    -1     0     0     0     0     0     0     0
#> [16,]     0    -1     6    -1    -1    -1     0     0     0     0     0     0
#> [17,]    -1    -1    -1     7    -1     0    -1     0     0     0     0     0
#> [18,]     0    -1    -1    -1     6    -1    -1    -1     0     0     0     0
#> [19,]     0     0    -1     0    -1     5    -1    -1    -1     0     0     0
#> [20,]     0     0     0    -1    -1    -1     7    -1    -1    -1     0    -1
#> [21,]     0     0     0     0    -1    -1    -1     7    -1    -1    -1    -1
#> [22,]     0     0     0     0     0    -1    -1    -1     7    -1    -1    -1
#> [23,]     0     0     0     0     0     0    -1    -1    -1     5    -1     0
#> [24,]     0     0     0     0     0     0     0    -1    -1    -1     6    -1
#> [25,]     0     0     0     0     0     0    -1    -1    -1     0    -1     6
#> attr(,"representation")
#> [1] "laplacian"
repr_modularity(g)
#>             [,1]       [,2]       [,3]  [,4]  [,5]  [,6]       [,7]       [,8]
#>  [1,] -0.1666667  0.7666667  0.7666667  0.80 -0.20 -0.20 -0.1666667 -0.1666667
#>  [2,]  0.7666667 -0.3266667  0.6733333  0.72  0.72 -0.28 -0.2333333 -0.2333333
#>  [3,]  0.7666667  0.6733333 -0.3266667  0.72  0.72  0.72 -0.2333333 -0.2333333
#>  [4,]  0.8000000  0.7200000  0.7200000 -0.24  0.76  0.76 -0.2000000 -0.2000000
#>  [5,] -0.2000000  0.7200000  0.7200000  0.76 -0.24  0.76  0.8000000  0.8000000
#>  [6,] -0.2000000 -0.2800000  0.7200000  0.76  0.76 -0.24  0.8000000 -0.2000000
#>  [7,] -0.1666667 -0.2333333 -0.2333333 -0.20  0.80  0.80 -0.1666667  0.8333333
#>  [8,] -0.1666667 -0.2333333 -0.2333333 -0.20  0.80 -0.20  0.8333333 -0.1666667
#>  [9,] -0.2000000 -0.2800000 -0.2800000 -0.24 -0.24  0.76  0.8000000  0.8000000
#> [10,] -0.2000000 -0.2800000 -0.2800000 -0.24 -0.24 -0.24  0.8000000  0.8000000
#> [11,] -0.2000000 -0.2800000 -0.2800000 -0.24 -0.24 -0.24 -0.2000000  0.8000000
#> [12,] -0.2000000 -0.2800000 -0.2800000 -0.24 -0.24 -0.24 -0.2000000 -0.2000000
#> [13,] -0.2000000 -0.2800000 -0.2800000 -0.24 -0.24 -0.24 -0.2000000 -0.2000000
#> [14,] -0.1666667 -0.2333333 -0.2333333 -0.20 -0.20 -0.20 -0.1666667 -0.1666667
#> [15,] -0.2000000 -0.2800000 -0.2800000 -0.24 -0.24 -0.24 -0.2000000 -0.2000000
#> [16,] -0.2000000 -0.2800000  0.7200000 -0.24 -0.24 -0.24 -0.2000000 -0.2000000
#> [17,] -0.2333333  0.6733333 -0.3266667  0.72 -0.28 -0.28 -0.2333333 -0.2333333
#> [18,] -0.2000000 -0.2800000 -0.2800000 -0.24 -0.24 -0.24 -0.2000000 -0.2000000
#> [19,] -0.1666667 -0.2333333 -0.2333333 -0.20 -0.20 -0.20 -0.1666667 -0.1666667
#> [20,] -0.2333333 -0.3266667 -0.3266667 -0.28 -0.28 -0.28 -0.2333333 -0.2333333
#> [21,] -0.2333333 -0.3266667 -0.3266667 -0.28 -0.28 -0.28 -0.2333333 -0.2333333
#> [22,] -0.2333333 -0.3266667 -0.3266667 -0.28 -0.28  0.72 -0.2333333 -0.2333333
#> [23,]  0.8333333 -0.2333333 -0.2333333 -0.20 -0.20 -0.20 -0.1666667 -0.1666667
#> [24,]  0.8000000  0.7200000 -0.2800000 -0.24 -0.24 -0.24 -0.2000000 -0.2000000
#> [25,] -0.2000000  0.7200000  0.7200000 -0.24 -0.24 -0.24 -0.2000000 -0.2000000
#>        [,9] [,10] [,11] [,12] [,13]      [,14] [,15] [,16]      [,17] [,18]
#>  [1,] -0.20 -0.20 -0.20 -0.20 -0.20 -0.1666667 -0.20 -0.20 -0.2333333 -0.20
#>  [2,] -0.28 -0.28 -0.28 -0.28 -0.28 -0.2333333 -0.28 -0.28  0.6733333 -0.28
#>  [3,] -0.28 -0.28 -0.28 -0.28 -0.28 -0.2333333 -0.28  0.72 -0.3266667 -0.28
#>  [4,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24  0.7200000 -0.24
#>  [5,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000 -0.24
#>  [6,]  0.76 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000 -0.24
#>  [7,]  0.80  0.80 -0.20 -0.20 -0.20 -0.1666667 -0.20 -0.20 -0.2333333 -0.20
#>  [8,]  0.80  0.80  0.80 -0.20 -0.20 -0.1666667 -0.20 -0.20 -0.2333333 -0.20
#>  [9,] -0.24  0.76  0.76  0.76 -0.24 -0.2000000 -0.24 -0.24 -0.2800000 -0.24
#> [10,]  0.76 -0.24  0.76  0.76  0.76 -0.2000000 -0.24 -0.24 -0.2800000 -0.24
#> [11,]  0.76  0.76 -0.24  0.76  0.76  0.8000000 -0.24 -0.24 -0.2800000 -0.24
#> [12,]  0.76  0.76  0.76 -0.24  0.76  0.8000000  0.76 -0.24 -0.2800000 -0.24
#> [13,] -0.24  0.76  0.76  0.76 -0.24  0.8000000  0.76  0.76 -0.2800000 -0.24
#> [14,] -0.20 -0.20  0.80  0.80  0.80 -0.1666667  0.80 -0.20  0.7666667 -0.20
#> [15,] -0.24 -0.24 -0.24  0.76  0.76  0.8000000 -0.24  0.76  0.7200000  0.76
#> [16,] -0.24 -0.24 -0.24 -0.24  0.76 -0.2000000  0.76 -0.24  0.7200000  0.76
#> [17,] -0.28 -0.28 -0.28 -0.28 -0.28  0.7666667  0.72  0.72 -0.3266667  0.72
#> [18,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000  0.76  0.76  0.7200000 -0.24
#> [19,] -0.20 -0.20 -0.20 -0.20 -0.20 -0.1666667 -0.20  0.80 -0.2333333  0.80
#> [20,] -0.28 -0.28 -0.28 -0.28 -0.28 -0.2333333 -0.28 -0.28  0.6733333  0.72
#> [21,] -0.28 -0.28 -0.28 -0.28 -0.28 -0.2333333 -0.28 -0.28 -0.3266667  0.72
#> [22,] -0.28 -0.28 -0.28 -0.28 -0.28 -0.2333333 -0.28 -0.28 -0.3266667 -0.28
#> [23,] -0.20 -0.20 -0.20 -0.20 -0.20 -0.1666667 -0.20 -0.20 -0.2333333 -0.20
#> [24,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000 -0.24
#> [25,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000 -0.24
#>            [,19]      [,20]      [,21]      [,22]      [,23] [,24] [,25]
#>  [1,] -0.1666667 -0.2333333 -0.2333333 -0.2333333  0.8333333  0.80 -0.20
#>  [2,] -0.2333333 -0.3266667 -0.3266667 -0.3266667 -0.2333333  0.72  0.72
#>  [3,] -0.2333333 -0.3266667 -0.3266667 -0.3266667 -0.2333333 -0.28  0.72
#>  [4,] -0.2000000 -0.2800000 -0.2800000 -0.2800000 -0.2000000 -0.24 -0.24
#>  [5,] -0.2000000 -0.2800000 -0.2800000 -0.2800000 -0.2000000 -0.24 -0.24
#>  [6,] -0.2000000 -0.2800000 -0.2800000  0.7200000 -0.2000000 -0.24 -0.24
#>  [7,] -0.1666667 -0.2333333 -0.2333333 -0.2333333 -0.1666667 -0.20 -0.20
#>  [8,] -0.1666667 -0.2333333 -0.2333333 -0.2333333 -0.1666667 -0.20 -0.20
#>  [9,] -0.2000000 -0.2800000 -0.2800000 -0.2800000 -0.2000000 -0.24 -0.24
#> [10,] -0.2000000 -0.2800000 -0.2800000 -0.2800000 -0.2000000 -0.24 -0.24
#> [11,] -0.2000000 -0.2800000 -0.2800000 -0.2800000 -0.2000000 -0.24 -0.24
#> [12,] -0.2000000 -0.2800000 -0.2800000 -0.2800000 -0.2000000 -0.24 -0.24
#> [13,] -0.2000000 -0.2800000 -0.2800000 -0.2800000 -0.2000000 -0.24 -0.24
#> [14,] -0.1666667 -0.2333333 -0.2333333 -0.2333333 -0.1666667 -0.20 -0.20
#> [15,] -0.2000000 -0.2800000 -0.2800000 -0.2800000 -0.2000000 -0.24 -0.24
#> [16,]  0.8000000 -0.2800000 -0.2800000 -0.2800000 -0.2000000 -0.24 -0.24
#> [17,] -0.2333333  0.6733333 -0.3266667 -0.3266667 -0.2333333 -0.28 -0.28
#> [18,]  0.8000000  0.7200000  0.7200000 -0.2800000 -0.2000000 -0.24 -0.24
#> [19,] -0.1666667  0.7666667  0.7666667  0.7666667 -0.1666667 -0.20 -0.20
#> [20,]  0.7666667 -0.3266667  0.6733333  0.6733333  0.7666667 -0.28  0.72
#> [21,]  0.7666667  0.6733333 -0.3266667  0.6733333  0.7666667  0.72  0.72
#> [22,]  0.7666667  0.6733333  0.6733333 -0.3266667  0.7666667  0.72  0.72
#> [23,] -0.1666667  0.7666667  0.7666667  0.7666667 -0.1666667  0.80 -0.20
#> [24,] -0.2000000 -0.2800000  0.7200000  0.7200000  0.8000000 -0.24  0.76
#> [25,] -0.2000000  0.7200000  0.7200000  0.7200000 -0.2000000  0.76 -0.24
#> attr(,"representation")
#> [1] "modularity"
repr_graphon(g)
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.44444444 0.57777778 0.72222222 0.44444444 0.27777778 0.22222222
#>  [2,] 0.57777778 0.60000000 0.81666667 0.73333333 0.67777778 0.42222222
#>  [3,] 0.72222222 0.81666667 0.50000000 0.66666667 0.52777778 0.58333333
#>  [4,] 0.44444444 0.73333333 0.66666667 0.55555556 0.50000000 0.44444444
#>  [5,] 0.27777778 0.67777778 0.52777778 0.50000000 0.44444444 0.55555556
#>  [6,] 0.22222222 0.42222222 0.58333333 0.44444444 0.55555556 0.33333333
#>  [7,] 0.16666667 0.21111111 0.25000000 0.22222222 0.44444444 0.44444444
#>  [8,] 0.05555556 0.00000000 0.13888889 0.16666667 0.38888889 0.27777778
#>  [9,] 0.05555556 0.15555556 0.05555556 0.16666667 0.27777778 0.38888889
#> [10,] 0.00000000 0.00000000 0.00000000 0.05555556 0.19444444 0.19444444
#> [11,] 0.00000000 0.00000000 0.00000000 0.00000000 0.22222222 0.22222222
#> [12,] 0.00000000 0.00000000 0.00000000 0.00000000 0.19444444 0.05555556
#> [13,] 0.05555556 0.00000000 0.00000000 0.00000000 0.11111111 0.11111111
#> [14,] 0.00000000 0.00000000 0.13888889 0.00000000 0.00000000 0.05555556
#> [15,] 0.05555556 0.00000000 0.13888889 0.00000000 0.00000000 0.00000000
#> [16,] 0.07142857 0.17142857 0.30952381 0.05555556 0.12698413 0.07142857
#> [17,] 0.25000000 0.36666667 0.16666667 0.27777778 0.13888889 0.19444444
#> [18,] 0.11805556 0.06250000 0.20833333 0.00000000 0.00000000 0.00000000
#> [19,] 0.16666667 0.16666667 0.05555556 0.05555556 0.00000000 0.00000000
#> [20,] 0.18253968 0.07142857 0.15476190 0.05555556 0.05555556 0.12698413
#> [21,] 0.30952381 0.14285714 0.15476190 0.11111111 0.05555556 0.12698413
#> [22,] 0.36507937 0.24285714 0.15476190 0.16666667 0.11111111 0.18253968
#> [23,] 0.23611111 0.28750000 0.20833333 0.17361111 0.05555556 0.11805556
#> [24,] 0.43650794 0.48571429 0.29761905 0.23809524 0.18253968 0.23809524
#> [25,] 0.27777778 0.47777778 0.52777778 0.33333333 0.16666667 0.16666667
#>             [,7]       [,8]       [,9]      [,10]     [,11]      [,12]
#>  [1,] 0.16666667 0.05555556 0.05555556 0.00000000 0.0000000 0.00000000
#>  [2,] 0.21111111 0.00000000 0.15555556 0.00000000 0.0000000 0.00000000
#>  [3,] 0.25000000 0.13888889 0.05555556 0.00000000 0.0000000 0.00000000
#>  [4,] 0.22222222 0.16666667 0.16666667 0.05555556 0.0000000 0.00000000
#>  [5,] 0.44444444 0.38888889 0.27777778 0.19444444 0.2222222 0.19444444
#>  [6,] 0.44444444 0.27777778 0.38888889 0.19444444 0.2222222 0.05555556
#>  [7,] 0.55555556 0.50000000 0.55555556 0.52777778 0.3888889 0.33333333
#>  [8,] 0.50000000 0.44444444 0.61111111 0.58333333 0.5555556 0.44444444
#>  [9,] 0.55555556 0.61111111 0.55555556 0.66666667 0.6111111 0.61111111
#> [10,] 0.52777778 0.58333333 0.66666667 0.66666667 0.7500000 0.66666667
#> [11,] 0.38888889 0.55555556 0.61111111 0.75000000 0.6666667 0.75000000
#> [12,] 0.33333333 0.44444444 0.61111111 0.66666667 0.7500000 0.66666667
#> [13,] 0.33333333 0.38888889 0.44444444 0.61111111 0.6111111 0.66666667
#> [14,] 0.22222222 0.33333333 0.33333333 0.38888889 0.5000000 0.58333333
#> [15,] 0.11111111 0.27777778 0.33333333 0.33333333 0.3888889 0.52777778
#> [16,] 0.00000000 0.05555556 0.12698413 0.07142857 0.2539683 0.22619048
#> [17,] 0.05555556 0.05555556 0.05555556 0.16666667 0.1944444 0.25000000
#> [18,] 0.00000000 0.00000000 0.00000000 0.06250000 0.1805556 0.18750000
#> [19,] 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000 0.05555556
#> [20,] 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000
#> [21,] 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000
#> [22,] 0.05555556 0.05555556 0.00000000 0.00000000 0.0000000 0.00000000
#> [23,] 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000
#> [24,] 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000
#> [25,] 0.11111111 0.05555556 0.05555556 0.00000000 0.0000000 0.00000000
#>            [,13]      [,14]      [,15]      [,16]      [,17]      [,18]
#>  [1,] 0.05555556 0.00000000 0.05555556 0.07142857 0.25000000 0.11805556
#>  [2,] 0.00000000 0.00000000 0.00000000 0.17142857 0.36666667 0.06250000
#>  [3,] 0.00000000 0.13888889 0.13888889 0.30952381 0.16666667 0.20833333
#>  [4,] 0.00000000 0.00000000 0.00000000 0.05555556 0.27777778 0.00000000
#>  [5,] 0.11111111 0.00000000 0.00000000 0.12698413 0.13888889 0.00000000
#>  [6,] 0.11111111 0.05555556 0.00000000 0.07142857 0.19444444 0.00000000
#>  [7,] 0.33333333 0.22222222 0.11111111 0.00000000 0.05555556 0.00000000
#>  [8,] 0.38888889 0.33333333 0.27777778 0.05555556 0.05555556 0.00000000
#>  [9,] 0.44444444 0.33333333 0.33333333 0.12698413 0.05555556 0.00000000
#> [10,] 0.61111111 0.38888889 0.33333333 0.07142857 0.16666667 0.06250000
#> [11,] 0.61111111 0.50000000 0.38888889 0.25396825 0.19444444 0.18055556
#> [12,] 0.66666667 0.58333333 0.52777778 0.22619048 0.25000000 0.18750000
#> [13,] 0.55555556 0.55555556 0.55555556 0.45238095 0.25000000 0.24305556
#> [14,] 0.55555556 0.44444444 0.50000000 0.30952381 0.47222222 0.23611111
#> [15,] 0.55555556 0.50000000 0.55555556 0.50793651 0.47222222 0.41666667
#> [16,] 0.45238095 0.30952381 0.50793651 0.28571429 0.92857143 0.46428571
#> [17,] 0.25000000 0.47222222 0.47222222 0.92857143 0.33333333 0.50000000
#> [18,] 0.24305556 0.23611111 0.41666667 0.46428571 0.50000000 0.37500000
#> [19,] 0.11111111 0.22222222 0.22222222 0.43650794 0.25000000 0.40277778
#> [20,] 0.12698413 0.05555556 0.25396825 0.21428571 0.46428571 0.53571429
#> [21,] 0.05555556 0.05555556 0.11111111 0.21428571 0.23809524 0.46428571
#> [22,] 0.00000000 0.00000000 0.12698413 0.14285714 0.22619048 0.33035714
#> [23,] 0.00000000 0.00000000 0.06250000 0.19642857 0.06250000 0.12500000
#> [24,] 0.00000000 0.00000000 0.00000000 0.07142857 0.14285714 0.26785714
#> [25,] 0.00000000 0.00000000 0.05555556 0.18253968 0.19444444 0.05555556
#>            [,19]      [,20]      [,21]      [,22]      [,23]      [,24]
#>  [1,] 0.16666667 0.18253968 0.30952381 0.36507937 0.23611111 0.43650794
#>  [2,] 0.16666667 0.07142857 0.14285714 0.24285714 0.28750000 0.48571429
#>  [3,] 0.05555556 0.15476190 0.15476190 0.15476190 0.20833333 0.29761905
#>  [4,] 0.05555556 0.05555556 0.11111111 0.16666667 0.17361111 0.23809524
#>  [5,] 0.00000000 0.05555556 0.05555556 0.11111111 0.05555556 0.18253968
#>  [6,] 0.00000000 0.12698413 0.12698413 0.18253968 0.11805556 0.23809524
#>  [7,] 0.00000000 0.00000000 0.00000000 0.05555556 0.00000000 0.00000000
#>  [8,] 0.00000000 0.00000000 0.00000000 0.05555556 0.00000000 0.00000000
#>  [9,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [10,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [11,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [12,] 0.05555556 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [13,] 0.11111111 0.12698413 0.05555556 0.00000000 0.00000000 0.00000000
#> [14,] 0.22222222 0.05555556 0.05555556 0.00000000 0.00000000 0.00000000
#> [15,] 0.22222222 0.25396825 0.11111111 0.12698413 0.06250000 0.00000000
#> [16,] 0.43650794 0.21428571 0.21428571 0.14285714 0.19642857 0.07142857
#> [17,] 0.25000000 0.46428571 0.23809524 0.22619048 0.06250000 0.14285714
#> [18,] 0.40277778 0.53571429 0.46428571 0.33035714 0.12500000 0.26785714
#> [19,] 0.33333333 0.69047619 0.54761905 0.61904762 0.35416667 0.23809524
#> [20,] 0.69047619 0.57142857 0.71428571 0.64285714 0.66071429 0.35714286
#> [21,] 0.54761905 0.71428571 0.85714286 0.85714286 0.66071429 0.64285714
#> [22,] 0.61904762 0.64285714 0.85714286 0.71428571 0.59821429 0.64285714
#> [23,] 0.35416667 0.66071429 0.66071429 0.59821429 0.50000000 0.52678571
#> [24,] 0.23809524 0.35714286 0.64285714 0.64285714 0.52678571 0.42857143
#> [25,] 0.22222222 0.50793651 0.56349206 0.56349206 0.29861111 0.30952381
#>            [,25]
#>  [1,] 0.27777778
#>  [2,] 0.47777778
#>  [3,] 0.52777778
#>  [4,] 0.33333333
#>  [5,] 0.16666667
#>  [6,] 0.16666667
#>  [7,] 0.11111111
#>  [8,] 0.05555556
#>  [9,] 0.05555556
#> [10,] 0.00000000
#> [11,] 0.00000000
#> [12,] 0.00000000
#> [13,] 0.00000000
#> [14,] 0.00000000
#> [15,] 0.05555556
#> [16,] 0.18253968
#> [17,] 0.19444444
#> [18,] 0.05555556
#> [19,] 0.22222222
#> [20,] 0.50793651
#> [21,] 0.56349206
#> [22,] 0.56349206
#> [23,] 0.29861111
#> [24,] 0.30952381
#> [25,] 0.11111111
#> attr(,"representation")
#> [1] "graphon"