Network Representation Functions

This is a collection of functions that convert a graph stored as an igraph object into a desired matrix representation among adjacency matrix, graph laplacian, modularity matrix or graphon (edge probability matrix).

repr_adjacency(network, validate = TRUE)

repr_laplacian(network, validate = TRUE)

repr_modularity(network, validate = TRUE)

repr_graphon(network, validate = TRUE)

Arguments

network
An igraph object.

validate
A boolean specifying whether the function should check the class of its input (default: TRUE).

Value

A numeric square matrix giving the desired network representation recorded in the object's class.

Examples

g <- igraph::sample_smallworld(1, 25, 3, 0.05)
repr_adjacency(g)

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0    1    1    1    0    0    0    0    0     0     0     0     0
#>  [2,]    1    0    0    1    1    0    0    0    0     0     0     0     0
#>  [3,]    1    0    0    1    1    1    0    0    0     0     0     1     0
#>  [4,]    1    1    1    0    1    1    1    0    0     0     0     0     0
#>  [5,]    0    1    1    1    0    1    1    1    0     0     0     0     0
#>  [6,]    0    0    1    1    1    0    1    1    1     0     0     0     1
#>  [7,]    0    0    0    1    1    1    0    1    1     1     0     0     0
#>  [8,]    0    0    0    0    1    1    1    0    1     1     0     0     0
#>  [9,]    0    0    0    0    0    1    1    1    0     1     1     0     0
#> [10,]    0    0    0    0    0    0    1    1    1     0     1     0     1
#> [11,]    0    0    0    0    0    0    0    0    1     1     0     1     1
#> [12,]    0    0    1    0    0    0    0    0    0     0     1     0     0
#> [13,]    0    0    0    0    0    1    0    0    0     1     1     0     0
#> [14,]    0    0    0    0    0    0    0    0    0     0     1     1     1
#> [15,]    0    0    0    0    0    0    0    0    0     0     0     1     1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0     1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     1     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     1     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]    1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]    1    1    0    0    0    0    0    0    1     0     1     0     0
#> [25,]    1    1    1    1    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0     1     1     1
#>  [2,]     0     0     0     0     0     0     0     0     0     0     1     1
#>  [3,]     0     0     0     0     0     0     0     0     0     0     0     1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     1
#>  [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     1     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]     1     0     0     0     0     1     0     0     0     0     1     0
#> [12,]     1     1     0     0     0     0     0     1     0     0     0     0
#> [13,]     1     1     1     0     0     0     0     0     0     0     0     0
#> [14,]     0     1     1     1     0     0     0     0     0     0     0     0
#> [15,]     1     0     1     1     1     0     0     0     0     0     0     0
#> [16,]     1     1     0     1     1     1     0     0     0     0     0     0
#> [17,]     1     1     1     0     1     1     1     0     0     0     0     0
#> [18,]     0     1     1     1     0     1     1     1     0     0     0     0
#> [19,]     0     0     1     1     1     0     1     0     1     0     0     0
#> [20,]     0     0     0     1     1     1     0     1     1     1     0     0
#> [21,]     0     0     0     0     1     0     1     0     1     1     1     0
#> [22,]     0     0     0     0     0     1     1     1     0     1     1     0
#> [23,]     0     0     0     0     0     0     1     1     1     0     1     1
#> [24,]     0     0     0     0     0     0     0     1     1     1     0     1
#> [25,]     0     0     0     0     0     0     0     0     0     1     1     0
#> attr(,"representation")
#> [1] "adjacency"

repr_laplacian(g)

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    6   -1   -1   -1    0    0    0    0    0     0     0     0     0
#>  [2,]   -1    5    0   -1   -1    0    0    0    0     0     0     0     0
#>  [3,]   -1    0    6   -1   -1   -1    0    0    0     0     0    -1     0
#>  [4,]   -1   -1   -1    7   -1   -1   -1    0    0     0     0     0     0
#>  [5,]    0   -1   -1   -1    6   -1   -1   -1    0     0     0     0     0
#>  [6,]    0    0   -1   -1   -1    7   -1   -1   -1     0     0     0    -1
#>  [7,]    0    0    0   -1   -1   -1    6   -1   -1    -1     0     0     0
#>  [8,]    0    0    0    0   -1   -1   -1    5   -1    -1     0     0     0
#>  [9,]    0    0    0    0    0   -1   -1   -1    6    -1    -1     0     0
#> [10,]    0    0    0    0    0    0   -1   -1   -1     5    -1     0    -1
#> [11,]    0    0    0    0    0    0    0    0   -1    -1     7    -1    -1
#> [12,]    0    0   -1    0    0    0    0    0    0     0    -1     5     0
#> [13,]    0    0    0    0    0   -1    0    0    0    -1    -1     0     6
#> [14,]    0    0    0    0    0    0    0    0    0     0    -1    -1    -1
#> [15,]    0    0    0    0    0    0    0    0    0     0     0    -1    -1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0    -1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0    -1     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0    -1     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]   -1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]   -1   -1    0    0    0    0    0    0   -1     0    -1     0     0
#> [25,]   -1   -1   -1   -1    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0    -1    -1    -1
#>  [2,]     0     0     0     0     0     0     0     0     0     0    -1    -1
#>  [3,]     0     0     0     0     0     0     0     0     0     0     0    -1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0    -1
#>  [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0    -1     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]    -1     0     0     0     0    -1     0     0     0     0    -1     0
#> [12,]    -1    -1     0     0     0     0     0    -1     0     0     0     0
#> [13,]    -1    -1    -1     0     0     0     0     0     0     0     0     0
#> [14,]     6    -1    -1    -1     0     0     0     0     0     0     0     0
#> [15,]    -1     6    -1    -1    -1     0     0     0     0     0     0     0
#> [16,]    -1    -1     6    -1    -1    -1     0     0     0     0     0     0
#> [17,]    -1    -1    -1     6    -1    -1    -1     0     0     0     0     0
#> [18,]     0    -1    -1    -1     6    -1    -1    -1     0     0     0     0
#> [19,]     0     0    -1    -1    -1     6    -1     0    -1     0     0     0
#> [20,]     0     0     0    -1    -1    -1     6    -1    -1    -1     0     0
#> [21,]     0     0     0     0    -1     0    -1     6    -1    -1    -1     0
#> [22,]     0     0     0     0     0    -1    -1    -1     5    -1    -1     0
#> [23,]     0     0     0     0     0     0    -1    -1    -1     6    -1    -1
#> [24,]     0     0     0     0     0     0     0    -1    -1    -1     8    -1
#> [25,]     0     0     0     0     0     0     0     0     0    -1    -1     6
#> attr(,"representation")
#> [1] "laplacian"

repr_modularity(g)

#>        [,1]       [,2]  [,3]       [,4]  [,5]       [,6]  [,7]       [,8]  [,9]
#>  [1,] -0.24  0.8000000  0.76  0.7200000 -0.24 -0.2800000 -0.24 -0.2000000 -0.24
#>  [2,]  0.80 -0.1666667 -0.20  0.7666667  0.80 -0.2333333 -0.20 -0.1666667 -0.20
#>  [3,]  0.76 -0.2000000 -0.24  0.7200000  0.76  0.7200000 -0.24 -0.2000000 -0.24
#>  [4,]  0.72  0.7666667  0.72 -0.3266667  0.72  0.6733333  0.72 -0.2333333 -0.28
#>  [5,] -0.24  0.8000000  0.76  0.7200000 -0.24  0.7200000  0.76  0.8000000 -0.24
#>  [6,] -0.28 -0.2333333  0.72  0.6733333  0.72 -0.3266667  0.72  0.7666667  0.72
#>  [7,] -0.24 -0.2000000 -0.24  0.7200000  0.76  0.7200000 -0.24  0.8000000  0.76
#>  [8,] -0.20 -0.1666667 -0.20 -0.2333333  0.80  0.7666667  0.80 -0.1666667  0.80
#>  [9,] -0.24 -0.2000000 -0.24 -0.2800000 -0.24  0.7200000  0.76  0.8000000 -0.24
#> [10,] -0.20 -0.1666667 -0.20 -0.2333333 -0.20 -0.2333333  0.80  0.8333333  0.80
#> [11,] -0.28 -0.2333333 -0.28 -0.3266667 -0.28 -0.3266667 -0.28 -0.2333333  0.72
#> [12,] -0.20 -0.1666667  0.80 -0.2333333 -0.20 -0.2333333 -0.20 -0.1666667 -0.20
#> [13,] -0.24 -0.2000000 -0.24 -0.2800000 -0.24  0.7200000 -0.24 -0.2000000 -0.24
#> [14,] -0.24 -0.2000000 -0.24 -0.2800000 -0.24 -0.2800000 -0.24 -0.2000000 -0.24
#> [15,] -0.24 -0.2000000 -0.24 -0.2800000 -0.24 -0.2800000 -0.24 -0.2000000 -0.24
#> [16,] -0.24 -0.2000000 -0.24 -0.2800000 -0.24 -0.2800000 -0.24 -0.2000000 -0.24
#> [17,] -0.24 -0.2000000 -0.24 -0.2800000 -0.24 -0.2800000 -0.24 -0.2000000 -0.24
#> [18,] -0.24 -0.2000000 -0.24 -0.2800000 -0.24 -0.2800000 -0.24 -0.2000000 -0.24
#> [19,] -0.24 -0.2000000 -0.24 -0.2800000 -0.24 -0.2800000 -0.24 -0.2000000 -0.24
#> [20,] -0.24 -0.2000000 -0.24 -0.2800000 -0.24 -0.2800000 -0.24 -0.2000000 -0.24
#> [21,] -0.24 -0.2000000 -0.24 -0.2800000 -0.24 -0.2800000 -0.24 -0.2000000 -0.24
#> [22,] -0.20 -0.1666667 -0.20 -0.2333333 -0.20 -0.2333333 -0.20 -0.1666667 -0.20
#> [23,]  0.76 -0.2000000 -0.24 -0.2800000 -0.24 -0.2800000 -0.24 -0.2000000 -0.24
#> [24,]  0.68  0.7333333 -0.32 -0.3733333 -0.32 -0.3733333 -0.32 -0.2666667  0.68
#> [25,]  0.76  0.8000000  0.76  0.7200000 -0.24 -0.2800000 -0.24 -0.2000000 -0.24
#>            [,10]      [,11]      [,12] [,13] [,14] [,15] [,16] [,17] [,18]
#>  [1,] -0.2000000 -0.2800000 -0.2000000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#>  [2,] -0.1666667 -0.2333333 -0.1666667 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20
#>  [3,] -0.2000000 -0.2800000  0.8000000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#>  [4,] -0.2333333 -0.3266667 -0.2333333 -0.28 -0.28 -0.28 -0.28 -0.28 -0.28
#>  [5,] -0.2000000 -0.2800000 -0.2000000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#>  [6,] -0.2333333 -0.3266667 -0.2333333  0.72 -0.28 -0.28 -0.28 -0.28 -0.28
#>  [7,]  0.8000000 -0.2800000 -0.2000000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#>  [8,]  0.8333333 -0.2333333 -0.1666667 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20
#>  [9,]  0.8000000  0.7200000 -0.2000000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#> [10,] -0.1666667  0.7666667 -0.1666667  0.80 -0.20 -0.20 -0.20 -0.20 -0.20
#> [11,]  0.7666667 -0.3266667  0.7666667  0.72  0.72 -0.28 -0.28 -0.28 -0.28
#> [12,] -0.1666667  0.7666667 -0.1666667 -0.20  0.80  0.80 -0.20 -0.20 -0.20
#> [13,]  0.8000000  0.7200000 -0.2000000 -0.24  0.76  0.76  0.76 -0.24 -0.24
#> [14,] -0.2000000  0.7200000  0.8000000  0.76 -0.24  0.76  0.76  0.76 -0.24
#> [15,] -0.2000000 -0.2800000  0.8000000  0.76  0.76 -0.24  0.76  0.76  0.76
#> [16,] -0.2000000 -0.2800000 -0.2000000  0.76  0.76  0.76 -0.24  0.76  0.76
#> [17,] -0.2000000 -0.2800000 -0.2000000 -0.24  0.76  0.76  0.76 -0.24  0.76
#> [18,] -0.2000000 -0.2800000 -0.2000000 -0.24 -0.24  0.76  0.76  0.76 -0.24
#> [19,] -0.2000000  0.7200000 -0.2000000 -0.24 -0.24 -0.24  0.76  0.76  0.76
#> [20,] -0.2000000 -0.2800000 -0.2000000 -0.24 -0.24 -0.24 -0.24  0.76  0.76
#> [21,] -0.2000000 -0.2800000  0.8000000 -0.24 -0.24 -0.24 -0.24 -0.24  0.76
#> [22,] -0.1666667 -0.2333333 -0.1666667 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20
#> [23,] -0.2000000 -0.2800000 -0.2000000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#> [24,] -0.2666667  0.6266667 -0.2666667 -0.32 -0.32 -0.32 -0.32 -0.32 -0.32
#> [25,] -0.2000000 -0.2800000 -0.2000000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#>       [,19] [,20] [,21]      [,22] [,23]      [,24] [,25]
#>  [1,] -0.24 -0.24 -0.24 -0.2000000  0.76  0.6800000  0.76
#>  [2,] -0.20 -0.20 -0.20 -0.1666667 -0.20  0.7333333  0.80
#>  [3,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.3200000  0.76
#>  [4,] -0.28 -0.28 -0.28 -0.2333333 -0.28 -0.3733333  0.72
#>  [5,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.3200000 -0.24
#>  [6,] -0.28 -0.28 -0.28 -0.2333333 -0.28 -0.3733333 -0.28
#>  [7,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.3200000 -0.24
#>  [8,] -0.20 -0.20 -0.20 -0.1666667 -0.20 -0.2666667 -0.20
#>  [9,] -0.24 -0.24 -0.24 -0.2000000 -0.24  0.6800000 -0.24
#> [10,] -0.20 -0.20 -0.20 -0.1666667 -0.20 -0.2666667 -0.20
#> [11,]  0.72 -0.28 -0.28 -0.2333333 -0.28  0.6266667 -0.28
#> [12,] -0.20 -0.20  0.80 -0.1666667 -0.20 -0.2666667 -0.20
#> [13,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.3200000 -0.24
#> [14,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.3200000 -0.24
#> [15,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.3200000 -0.24
#> [16,]  0.76 -0.24 -0.24 -0.2000000 -0.24 -0.3200000 -0.24
#> [17,]  0.76  0.76 -0.24 -0.2000000 -0.24 -0.3200000 -0.24
#> [18,]  0.76  0.76  0.76 -0.2000000 -0.24 -0.3200000 -0.24
#> [19,] -0.24  0.76 -0.24  0.8000000 -0.24 -0.3200000 -0.24
#> [20,]  0.76 -0.24  0.76  0.8000000  0.76 -0.3200000 -0.24
#> [21,] -0.24  0.76 -0.24  0.8000000  0.76  0.6800000 -0.24
#> [22,]  0.80  0.80  0.80 -0.1666667  0.80  0.7333333 -0.20
#> [23,] -0.24  0.76  0.76  0.8000000 -0.24  0.6800000  0.76
#> [24,] -0.32 -0.32  0.68  0.7333333  0.68 -0.4266667  0.68
#> [25,] -0.24 -0.24 -0.24 -0.2000000  0.76  0.6800000 -0.24
#> attr(,"representation")
#> [1] "modularity"

repr_graphon(g)

#>             [,1]       [,2]       [,3]      [,4]       [,5]       [,6]
#>  [1,] 0.55555556 0.50000000 0.50000000 0.5277778 0.30952381 0.25000000
#>  [2,] 0.50000000 0.33333333 0.38888889 0.5277778 0.56349206 0.38888889
#>  [3,] 0.50000000 0.38888889 0.44444444 0.6458333 0.63492063 0.52777778
#>  [4,] 0.52777778 0.52777778 0.64583333 0.7500000 0.66964286 0.50000000
#>  [5,] 0.30952381 0.56349206 0.63492063 0.6696429 0.42857143 0.61904762
#>  [6,] 0.25000000 0.38888889 0.52777778 0.5000000 0.61904762 0.66666667
#>  [7,] 0.22222222 0.16666667 0.27777778 0.5277778 0.56349206 0.75000000
#>  [8,] 0.18055556 0.11805556 0.17361111 0.3750000 0.53571429 0.70833333
#>  [9,] 0.00000000 0.19444444 0.19444444 0.2916667 0.30952381 0.75000000
#> [10,] 0.00000000 0.17361111 0.29861111 0.2500000 0.33035714 0.41666667
#> [11,] 0.00000000 0.00000000 0.00000000 0.0000000 0.07142857 0.08333333
#> [12,] 0.11111111 0.05555556 0.11805556 0.0625000 0.00000000 0.06250000
#> [13,] 0.00000000 0.00000000 0.10000000 0.0625000 0.07142857 0.36666667
#> [14,] 0.00000000 0.00000000 0.05555556 0.0000000 0.00000000 0.05555556
#> [15,] 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000 0.05555556
#> [16,] 0.00000000 0.00000000 0.05555556 0.0000000 0.00000000 0.05555556
#> [17,] 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000 0.08333333
#> [18,] 0.05555556 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000
#> [19,] 0.05555556 0.05555556 0.00000000 0.0000000 0.00000000 0.00000000
#> [20,] 0.22222222 0.11111111 0.05555556 0.0000000 0.00000000 0.00000000
#> [21,] 0.22222222 0.22222222 0.16666667 0.1111111 0.00000000 0.00000000
#> [22,] 0.23611111 0.18055556 0.18055556 0.1250000 0.00000000 0.00000000
#> [23,] 0.47222222 0.16666667 0.11111111 0.2500000 0.26785714 0.00000000
#> [24,] 0.66666667 0.61111111 0.47222222 0.3541667 0.29761905 0.08333333
#> [25,] 0.55555556 0.50000000 0.50000000 0.5277778 0.30952381 0.25000000
#>             [,7]      [,8]       [,9]     [,10]      [,11]      [,12]
#>  [1,] 0.22222222 0.1805556 0.00000000 0.0000000 0.00000000 0.11111111
#>  [2,] 0.16666667 0.1180556 0.19444444 0.1736111 0.00000000 0.05555556
#>  [3,] 0.27777778 0.1736111 0.19444444 0.2986111 0.00000000 0.11805556
#>  [4,] 0.52777778 0.3750000 0.29166667 0.2500000 0.00000000 0.06250000
#>  [5,] 0.56349206 0.5357143 0.30952381 0.3303571 0.07142857 0.00000000
#>  [6,] 0.75000000 0.7083333 0.75000000 0.4166667 0.08333333 0.06250000
#>  [7,] 0.55555556 0.5277778 0.55555556 0.5347222 0.16666667 0.11805556
#>  [8,] 0.52777778 0.5000000 0.52083333 0.5625000 0.18750000 0.12500000
#>  [9,] 0.55555556 0.5208333 0.50000000 0.5833333 0.41666667 0.06250000
#> [10,] 0.53472222 0.5625000 0.58333333 0.3750000 0.41666667 0.12500000
#> [11,] 0.16666667 0.1875000 0.41666667 0.4166667 0.00000000 0.58333333
#> [12,] 0.11805556 0.1250000 0.06250000 0.1250000 0.58333333 0.12500000
#> [13,] 0.21111111 0.1625000 0.08333333 0.3875000 0.90000000 0.28750000
#> [14,] 0.11111111 0.1180556 0.13888889 0.1180556 0.77777778 0.36111111
#> [15,] 0.05555556 0.0625000 0.08333333 0.1805556 0.33333333 0.54166667
#> [16,] 0.05555556 0.0625000 0.08333333 0.1180556 0.38888889 0.36111111
#> [17,] 0.00000000 0.0000000 0.00000000 0.1458333 0.50000000 0.27083333
#> [18,] 0.00000000 0.0000000 0.00000000 0.0000000 0.47619048 0.19642857
#> [19,] 0.00000000 0.0000000 0.00000000 0.0000000 0.47619048 0.46428571
#> [20,] 0.00000000 0.0000000 0.00000000 0.0000000 0.05555556 0.29861111
#> [21,] 0.00000000 0.0000000 0.00000000 0.0625000 0.05555556 0.29861111
#> [22,] 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000 0.06250000
#> [23,] 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000 0.18750000
#> [24,] 0.19444444 0.1250000 0.16666667 0.0625000 0.25000000 0.06250000
#> [25,] 0.22222222 0.1805556 0.00000000 0.0000000 0.00000000 0.11111111
#>            [,13]      [,14]      [,15]      [,16]      [,17]      [,18]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.05555556
#>  [2,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [3,] 0.10000000 0.05555556 0.00000000 0.05555556 0.00000000 0.00000000
#>  [4,] 0.06250000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [5,] 0.07142857 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [6,] 0.36666667 0.05555556 0.05555556 0.05555556 0.08333333 0.00000000
#>  [7,] 0.21111111 0.11111111 0.05555556 0.05555556 0.00000000 0.00000000
#>  [8,] 0.16250000 0.11805556 0.06250000 0.06250000 0.00000000 0.00000000
#>  [9,] 0.08333333 0.13888889 0.08333333 0.08333333 0.00000000 0.00000000
#> [10,] 0.38750000 0.11805556 0.18055556 0.11805556 0.14583333 0.00000000
#> [11,] 0.90000000 0.77777778 0.33333333 0.38888889 0.50000000 0.47619048
#> [12,] 0.28750000 0.36111111 0.54166667 0.36111111 0.27083333 0.19642857
#> [13,] 0.20000000 0.52222222 0.56666667 0.46666667 0.26666667 0.24285714
#> [14,] 0.52222222 0.55555556 0.55555556 0.61111111 0.61111111 0.36507937
#> [15,] 0.56666667 0.55555556 0.55555556 0.66666667 0.66666667 0.61904762
#> [16,] 0.46666667 0.61111111 0.66666667 0.66666667 0.75000000 0.56349206
#> [17,] 0.26666667 0.61111111 0.66666667 0.75000000 0.66666667 0.69047619
#> [18,] 0.24285714 0.36507937 0.61904762 0.56349206 0.69047619 0.57142857
#> [19,] 0.24285714 0.30952381 0.43650794 0.57936508 0.60714286 0.71428571
#> [20,] 0.21111111 0.33333333 0.38888889 0.38888889 0.52777778 0.61904762
#> [21,] 0.15555556 0.22222222 0.16666667 0.33333333 0.25000000 0.49206349
#> [22,] 0.00000000 0.11805556 0.29166667 0.23611111 0.20833333 0.33035714
#> [23,] 0.00000000 0.00000000 0.11111111 0.05555556 0.00000000 0.26785714
#> [24,] 0.10000000 0.00000000 0.05555556 0.00000000 0.08333333 0.15476190
#> [25,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.05555556
#>            [,19]      [,20]      [,21]     [,22]      [,23]      [,24]
#>  [1,] 0.05555556 0.22222222 0.22222222 0.2361111 0.47222222 0.66666667
#>  [2,] 0.05555556 0.11111111 0.22222222 0.1805556 0.16666667 0.61111111
#>  [3,] 0.00000000 0.05555556 0.16666667 0.1805556 0.11111111 0.47222222
#>  [4,] 0.00000000 0.00000000 0.11111111 0.1250000 0.25000000 0.35416667
#>  [5,] 0.00000000 0.00000000 0.00000000 0.0000000 0.26785714 0.29761905
#>  [6,] 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000 0.08333333
#>  [7,] 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000 0.19444444
#>  [8,] 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000 0.12500000
#>  [9,] 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000 0.16666667
#> [10,] 0.00000000 0.00000000 0.06250000 0.0000000 0.00000000 0.06250000
#> [11,] 0.47619048 0.05555556 0.05555556 0.0000000 0.00000000 0.25000000
#> [12,] 0.46428571 0.29861111 0.29861111 0.0625000 0.18750000 0.06250000
#> [13,] 0.24285714 0.21111111 0.15555556 0.0000000 0.00000000 0.10000000
#> [14,] 0.30952381 0.33333333 0.22222222 0.1180556 0.00000000 0.00000000
#> [15,] 0.43650794 0.38888889 0.16666667 0.2916667 0.11111111 0.05555556
#> [16,] 0.57936508 0.38888889 0.33333333 0.2361111 0.05555556 0.00000000
#> [17,] 0.60714286 0.52777778 0.25000000 0.2083333 0.00000000 0.08333333
#> [18,] 0.71428571 0.61904762 0.49206349 0.3303571 0.26785714 0.15476190
#> [19,] 0.42857143 0.56349206 0.30952381 0.4642857 0.26785714 0.15476190
#> [20,] 0.56349206 0.66666667 0.50000000 0.5347222 0.54166667 0.25000000
#> [21,] 0.30952381 0.50000000 0.33333333 0.4722222 0.52777778 0.52777778
#> [22,] 0.46428571 0.53472222 0.47222222 0.3750000 0.56250000 0.56250000
#> [23,] 0.26785714 0.54166667 0.52777778 0.5625000 0.50000000 0.83333333
#> [24,] 0.15476190 0.25000000 0.52777778 0.5625000 0.83333333 0.50000000
#> [25,] 0.05555556 0.22222222 0.22222222 0.2361111 0.47222222 0.66666667
#>            [,25]
#>  [1,] 0.55555556
#>  [2,] 0.50000000
#>  [3,] 0.50000000
#>  [4,] 0.52777778
#>  [5,] 0.30952381
#>  [6,] 0.25000000
#>  [7,] 0.22222222
#>  [8,] 0.18055556
#>  [9,] 0.00000000
#> [10,] 0.00000000
#> [11,] 0.00000000
#> [12,] 0.11111111
#> [13,] 0.00000000
#> [14,] 0.00000000
#> [15,] 0.00000000
#> [16,] 0.00000000
#> [17,] 0.00000000
#> [18,] 0.05555556
#> [19,] 0.05555556
#> [20,] 0.22222222
#> [21,] 0.22222222
#> [22,] 0.23611111
#> [23,] 0.47222222
#> [24,] 0.66666667
#> [25,] 0.55555556
#> attr(,"representation")
#> [1] "graphon"