Network Representation Functions

This is a collection of functions that convert a graph stored as an igraph object into a desired matrix representation among adjacency matrix, graph laplacian, modularity matrix or graphon (edge probability matrix).

repr_adjacency(network, validate = TRUE)

repr_laplacian(network, validate = TRUE)

repr_modularity(network, validate = TRUE)

repr_graphon(network, validate = TRUE)

Arguments

network

An igraph object.

validate

A boolean specifying whether the function should check the class of its input (default: TRUE).

Value

A numeric square matrix giving the desired network representation recorded in the object's class.

Examples

g <- igraph::sample_smallworld(1, 25, 3, 0.05)
repr_adjacency(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0    0    1    1    0    0    0    0    0     0     0     0     0
#>  [2,]    0    0    1    1    1    0    0    0    0     0     0     0     0
#>  [3,]    1    1    0    1    1    1    0    0    0     0     0     0     0
#>  [4,]    1    1    1    0    0    1    1    0    0     0     0     0     0
#>  [5,]    0    1    1    0    0    0    1    1    0     0     0     0     0
#>  [6,]    0    0    1    1    0    0    1    1    1     1     0     0     0
#>  [7,]    0    0    0    1    1    1    0    1    1     1     0     0     0
#>  [8,]    0    0    0    0    1    1    1    0    1     1     1     0     0
#>  [9,]    0    0    0    0    0    1    1    1    0     1     1     1     0
#> [10,]    0    0    0    0    0    1    1    1    1     0     1     1     1
#> [11,]    0    0    0    0    0    0    0    1    1     1     0     1     1
#> [12,]    0    0    0    0    0    0    0    0    1     1     1     0     1
#> [13,]    0    0    0    0    0    0    0    0    0     1     1     1     0
#> [14,]    0    0    0    0    0    0    0    0    0     0     0     1     1
#> [15,]    0    0    0    0    0    0    0    0    0     0     0     1     1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0     1
#> [17,]    0    0    0    0    0    0    0    0    0     0     1     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     1     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    1    0    0    0    0    0    0    0     0     0     0     0
#> [23,]    1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]    1    1    0    1    0    0    0    0    0     0     0     0     0
#> [25,]    1    1    1    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0     1     1     1
#>  [2,]     0     0     0     0     0     0     0     0     1     0     1     1
#>  [3,]     0     0     0     0     0     0     0     0     0     0     0     1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     1     0
#>  [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]     0     0     0     1     0     0     1     0     0     0     0     0
#> [12,]     1     1     0     0     0     0     0     0     0     0     0     0
#> [13,]     1     1     1     0     0     0     0     0     0     0     0     0
#> [14,]     0     1     1     1     0     0     0     0     1     0     0     0
#> [15,]     1     0     1     1     1     0     0     0     0     0     0     0
#> [16,]     1     1     0     1     1     1     0     0     0     0     0     0
#> [17,]     1     1     1     0     1     1     1     0     0     0     0     0
#> [18,]     0     1     1     1     0     1     1     1     0     0     0     0
#> [19,]     0     0     1     1     1     0     1     1     1     0     0     0
#> [20,]     0     0     0     1     1     1     0     0     1     1     0     0
#> [21,]     0     0     0     0     1     1     0     0     0     1     1     0
#> [22,]     1     0     0     0     0     1     1     0     0     1     1     1
#> [23,]     0     0     0     0     0     0     1     1     1     0     1     1
#> [24,]     0     0     0     0     0     0     0     1     1     1     0     1
#> [25,]     0     0     0     0     0     0     0     0     1     1     1     0
#> attr(,"representation")
#> [1] "adjacency"
repr_laplacian(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    5    0   -1   -1    0    0    0    0    0     0     0     0     0
#>  [2,]    0    6   -1   -1   -1    0    0    0    0     0     0     0     0
#>  [3,]   -1   -1    6   -1   -1   -1    0    0    0     0     0     0     0
#>  [4,]   -1   -1   -1    6    0   -1   -1    0    0     0     0     0     0
#>  [5,]    0   -1   -1    0    4    0   -1   -1    0     0     0     0     0
#>  [6,]    0    0   -1   -1    0    6   -1   -1   -1    -1     0     0     0
#>  [7,]    0    0    0   -1   -1   -1    6   -1   -1    -1     0     0     0
#>  [8,]    0    0    0    0   -1   -1   -1    6   -1    -1    -1     0     0
#>  [9,]    0    0    0    0    0   -1   -1   -1    6    -1    -1    -1     0
#> [10,]    0    0    0    0    0   -1   -1   -1   -1     7    -1    -1    -1
#> [11,]    0    0    0    0    0    0    0   -1   -1    -1     7    -1    -1
#> [12,]    0    0    0    0    0    0    0    0   -1    -1    -1     6    -1
#> [13,]    0    0    0    0    0    0    0    0    0    -1    -1    -1     6
#> [14,]    0    0    0    0    0    0    0    0    0     0     0    -1    -1
#> [15,]    0    0    0    0    0    0    0    0    0     0     0    -1    -1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0    -1
#> [17,]    0    0    0    0    0    0    0    0    0     0    -1     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0    -1     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [22,]    0   -1    0    0    0    0    0    0    0     0     0     0     0
#> [23,]   -1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]   -1   -1    0   -1    0    0    0    0    0     0     0     0     0
#> [25,]   -1   -1   -1    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0    -1    -1    -1
#>  [2,]     0     0     0     0     0     0     0     0    -1     0    -1    -1
#>  [3,]     0     0     0     0     0     0     0     0     0     0     0    -1
#>  [4,]     0     0     0     0     0     0     0     0     0     0    -1     0
#>  [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]     0     0     0    -1     0     0    -1     0     0     0     0     0
#> [12,]    -1    -1     0     0     0     0     0     0     0     0     0     0
#> [13,]    -1    -1    -1     0     0     0     0     0     0     0     0     0
#> [14,]     6    -1    -1    -1     0     0     0     0    -1     0     0     0
#> [15,]    -1     6    -1    -1    -1     0     0     0     0     0     0     0
#> [16,]    -1    -1     6    -1    -1    -1     0     0     0     0     0     0
#> [17,]    -1    -1    -1     7    -1    -1    -1     0     0     0     0     0
#> [18,]     0    -1    -1    -1     6    -1    -1    -1     0     0     0     0
#> [19,]     0     0    -1    -1    -1     6    -1    -1    -1     0     0     0
#> [20,]     0     0     0    -1    -1    -1     6     0    -1    -1     0     0
#> [21,]     0     0     0     0    -1    -1     0     4     0    -1    -1     0
#> [22,]    -1     0     0     0     0    -1    -1     0     7    -1    -1    -1
#> [23,]     0     0     0     0     0     0    -1    -1    -1     6    -1    -1
#> [24,]     0     0     0     0     0     0     0    -1    -1    -1     7    -1
#> [25,]     0     0     0     0     0     0     0     0    -1    -1    -1     6
#> attr(,"representation")
#> [1] "laplacian"
repr_modularity(g)
#>             [,1]  [,2]  [,3]  [,4]       [,5]  [,6]  [,7]  [,8]  [,9]
#>  [1,] -0.1666667 -0.20  0.80  0.80 -0.1333333 -0.20 -0.20 -0.20 -0.20
#>  [2,] -0.2000000 -0.24  0.76  0.76  0.8400000 -0.24 -0.24 -0.24 -0.24
#>  [3,]  0.8000000  0.76 -0.24  0.76  0.8400000  0.76 -0.24 -0.24 -0.24
#>  [4,]  0.8000000  0.76  0.76 -0.24 -0.1600000  0.76  0.76 -0.24 -0.24
#>  [5,] -0.1333333  0.84  0.84 -0.16 -0.1066667 -0.16  0.84  0.84 -0.16
#>  [6,] -0.2000000 -0.24  0.76  0.76 -0.1600000 -0.24  0.76  0.76  0.76
#>  [7,] -0.2000000 -0.24 -0.24  0.76  0.8400000  0.76 -0.24  0.76  0.76
#>  [8,] -0.2000000 -0.24 -0.24 -0.24  0.8400000  0.76  0.76 -0.24  0.76
#>  [9,] -0.2000000 -0.24 -0.24 -0.24 -0.1600000  0.76  0.76  0.76 -0.24
#> [10,] -0.2333333 -0.28 -0.28 -0.28 -0.1866667  0.72  0.72  0.72  0.72
#> [11,] -0.2333333 -0.28 -0.28 -0.28 -0.1866667 -0.28 -0.28  0.72  0.72
#> [12,] -0.2000000 -0.24 -0.24 -0.24 -0.1600000 -0.24 -0.24 -0.24  0.76
#> [13,] -0.2000000 -0.24 -0.24 -0.24 -0.1600000 -0.24 -0.24 -0.24 -0.24
#> [14,] -0.2000000 -0.24 -0.24 -0.24 -0.1600000 -0.24 -0.24 -0.24 -0.24
#> [15,] -0.2000000 -0.24 -0.24 -0.24 -0.1600000 -0.24 -0.24 -0.24 -0.24
#> [16,] -0.2000000 -0.24 -0.24 -0.24 -0.1600000 -0.24 -0.24 -0.24 -0.24
#> [17,] -0.2333333 -0.28 -0.28 -0.28 -0.1866667 -0.28 -0.28 -0.28 -0.28
#> [18,] -0.2000000 -0.24 -0.24 -0.24 -0.1600000 -0.24 -0.24 -0.24 -0.24
#> [19,] -0.2000000 -0.24 -0.24 -0.24 -0.1600000 -0.24 -0.24 -0.24 -0.24
#> [20,] -0.2000000 -0.24 -0.24 -0.24 -0.1600000 -0.24 -0.24 -0.24 -0.24
#> [21,] -0.1333333 -0.16 -0.16 -0.16 -0.1066667 -0.16 -0.16 -0.16 -0.16
#> [22,] -0.2333333  0.72 -0.28 -0.28 -0.1866667 -0.28 -0.28 -0.28 -0.28
#> [23,]  0.8000000 -0.24 -0.24 -0.24 -0.1600000 -0.24 -0.24 -0.24 -0.24
#> [24,]  0.7666667  0.72 -0.28  0.72 -0.1866667 -0.28 -0.28 -0.28 -0.28
#> [25,]  0.8000000  0.76  0.76 -0.24 -0.1600000 -0.24 -0.24 -0.24 -0.24
#>            [,10]      [,11] [,12] [,13] [,14] [,15] [,16]      [,17] [,18]
#>  [1,] -0.2333333 -0.2333333 -0.20 -0.20 -0.20 -0.20 -0.20 -0.2333333 -0.20
#>  [2,] -0.2800000 -0.2800000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#>  [3,] -0.2800000 -0.2800000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#>  [4,] -0.2800000 -0.2800000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#>  [5,] -0.1866667 -0.1866667 -0.16 -0.16 -0.16 -0.16 -0.16 -0.1866667 -0.16
#>  [6,]  0.7200000 -0.2800000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#>  [7,]  0.7200000 -0.2800000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#>  [8,]  0.7200000  0.7200000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#>  [9,]  0.7200000  0.7200000  0.76 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#> [10,] -0.3266667  0.6733333  0.72  0.72 -0.28 -0.28 -0.28 -0.3266667 -0.28
#> [11,]  0.6733333 -0.3266667  0.72  0.72 -0.28 -0.28 -0.28  0.6733333 -0.28
#> [12,]  0.7200000  0.7200000 -0.24  0.76  0.76  0.76 -0.24 -0.2800000 -0.24
#> [13,]  0.7200000  0.7200000  0.76 -0.24  0.76  0.76  0.76 -0.2800000 -0.24
#> [14,] -0.2800000 -0.2800000  0.76  0.76 -0.24  0.76  0.76  0.7200000 -0.24
#> [15,] -0.2800000 -0.2800000  0.76  0.76  0.76 -0.24  0.76  0.7200000  0.76
#> [16,] -0.2800000 -0.2800000 -0.24  0.76  0.76  0.76 -0.24  0.7200000  0.76
#> [17,] -0.3266667  0.6733333 -0.28 -0.28  0.72  0.72  0.72 -0.3266667  0.72
#> [18,] -0.2800000 -0.2800000 -0.24 -0.24 -0.24  0.76  0.76  0.7200000 -0.24
#> [19,] -0.2800000 -0.2800000 -0.24 -0.24 -0.24 -0.24  0.76  0.7200000  0.76
#> [20,] -0.2800000  0.7200000 -0.24 -0.24 -0.24 -0.24 -0.24  0.7200000  0.76
#> [21,] -0.1866667 -0.1866667 -0.16 -0.16 -0.16 -0.16 -0.16 -0.1866667  0.84
#> [22,] -0.3266667 -0.3266667 -0.28 -0.28  0.72 -0.28 -0.28 -0.3266667 -0.28
#> [23,] -0.2800000 -0.2800000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#> [24,] -0.3266667 -0.3266667 -0.28 -0.28 -0.28 -0.28 -0.28 -0.3266667 -0.28
#> [25,] -0.2800000 -0.2800000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#>       [,19] [,20]      [,21]      [,22] [,23]      [,24] [,25]
#>  [1,] -0.20 -0.20 -0.1333333 -0.2333333  0.80  0.7666667  0.80
#>  [2,] -0.24 -0.24 -0.1600000  0.7200000 -0.24  0.7200000  0.76
#>  [3,] -0.24 -0.24 -0.1600000 -0.2800000 -0.24 -0.2800000  0.76
#>  [4,] -0.24 -0.24 -0.1600000 -0.2800000 -0.24  0.7200000 -0.24
#>  [5,] -0.16 -0.16 -0.1066667 -0.1866667 -0.16 -0.1866667 -0.16
#>  [6,] -0.24 -0.24 -0.1600000 -0.2800000 -0.24 -0.2800000 -0.24
#>  [7,] -0.24 -0.24 -0.1600000 -0.2800000 -0.24 -0.2800000 -0.24
#>  [8,] -0.24 -0.24 -0.1600000 -0.2800000 -0.24 -0.2800000 -0.24
#>  [9,] -0.24 -0.24 -0.1600000 -0.2800000 -0.24 -0.2800000 -0.24
#> [10,] -0.28 -0.28 -0.1866667 -0.3266667 -0.28 -0.3266667 -0.28
#> [11,] -0.28  0.72 -0.1866667 -0.3266667 -0.28 -0.3266667 -0.28
#> [12,] -0.24 -0.24 -0.1600000 -0.2800000 -0.24 -0.2800000 -0.24
#> [13,] -0.24 -0.24 -0.1600000 -0.2800000 -0.24 -0.2800000 -0.24
#> [14,] -0.24 -0.24 -0.1600000  0.7200000 -0.24 -0.2800000 -0.24
#> [15,] -0.24 -0.24 -0.1600000 -0.2800000 -0.24 -0.2800000 -0.24
#> [16,]  0.76 -0.24 -0.1600000 -0.2800000 -0.24 -0.2800000 -0.24
#> [17,]  0.72  0.72 -0.1866667 -0.3266667 -0.28 -0.3266667 -0.28
#> [18,]  0.76  0.76  0.8400000 -0.2800000 -0.24 -0.2800000 -0.24
#> [19,] -0.24  0.76  0.8400000  0.7200000 -0.24 -0.2800000 -0.24
#> [20,]  0.76 -0.24 -0.1600000  0.7200000  0.76 -0.2800000 -0.24
#> [21,]  0.84 -0.16 -0.1066667 -0.1866667  0.84  0.8133333 -0.16
#> [22,]  0.72  0.72 -0.1866667 -0.3266667  0.72  0.6733333  0.72
#> [23,] -0.24  0.76  0.8400000  0.7200000 -0.24  0.7200000  0.76
#> [24,] -0.28 -0.28  0.8133333  0.6733333  0.72 -0.3266667  0.72
#> [25,] -0.24 -0.24 -0.1600000  0.7200000  0.76  0.7200000 -0.24
#> attr(,"representation")
#> [1] "modularity"
repr_graphon(g)
#>             [,1]       [,2]       [,3]      [,4]       [,5]       [,6]
#>  [1,] 0.44444444 0.50000000 0.61111111 0.4722222 0.25396825 0.16666667
#>  [2,] 0.50000000 0.44444444 0.50000000 0.5833333 0.38095238 0.27777778
#>  [3,] 0.61111111 0.50000000 0.44444444 0.6527778 0.57936508 0.50000000
#>  [4,] 0.47222222 0.58333333 0.65277778 0.5000000 0.26785714 0.34722222
#>  [5,] 0.25396825 0.38095238 0.57936508 0.2678571 0.00000000 0.30952381
#>  [6,] 0.16666667 0.27777778 0.50000000 0.3472222 0.30952381 0.44444444
#>  [7,] 0.22222222 0.16666667 0.22222222 0.4097222 0.50793651 0.55555556
#>  [8,] 0.22222222 0.22222222 0.16666667 0.2916667 0.38095238 0.55555556
#>  [9,] 0.11111111 0.22222222 0.27777778 0.1736111 0.18253968 0.50000000
#> [10,] 0.05555556 0.11111111 0.18253968 0.2053571 0.28571429 0.56349206
#> [11,] 0.00000000 0.00000000 0.11805556 0.1250000 0.13392857 0.35416667
#> [12,] 0.00000000 0.00000000 0.05555556 0.1111111 0.18253968 0.22222222
#> [13,] 0.00000000 0.00000000 0.00000000 0.0000000 0.05555556 0.16666667
#> [14,] 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000 0.05555556
#> [15,] 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000 0.05555556
#> [16,] 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000
#> [17,] 0.00000000 0.05555556 0.00000000 0.0000000 0.00000000 0.00000000
#> [18,] 0.11111111 0.00000000 0.00000000 0.0000000 0.07142857 0.00000000
#> [19,] 0.11805556 0.00000000 0.00000000 0.0000000 0.07142857 0.00000000
#> [20,] 0.11111111 0.11111111 0.11111111 0.0625000 0.00000000 0.00000000
#> [21,] 0.12698413 0.32539683 0.32539683 0.1339286 0.00000000 0.00000000
#> [22,] 0.38095238 0.43650794 0.36507937 0.3303571 0.07142857 0.05555556
#> [23,] 0.44444444 0.38888889 0.38888889 0.3472222 0.32539683 0.16666667
#> [24,] 0.61904762 0.56349206 0.49206349 0.5982143 0.42857143 0.23809524
#> [25,] 0.50000000 0.55555556 0.55555556 0.4722222 0.18253968 0.22222222
#>             [,7]       [,8]       [,9]      [,10]      [,11]      [,12]
#>  [1,] 0.22222222 0.22222222 0.11111111 0.05555556 0.00000000 0.00000000
#>  [2,] 0.16666667 0.22222222 0.22222222 0.11111111 0.00000000 0.00000000
#>  [3,] 0.22222222 0.16666667 0.27777778 0.18253968 0.11805556 0.05555556
#>  [4,] 0.40972222 0.29166667 0.17361111 0.20535714 0.12500000 0.11111111
#>  [5,] 0.50793651 0.38095238 0.18253968 0.28571429 0.13392857 0.18253968
#>  [6,] 0.55555556 0.55555556 0.50000000 0.56349206 0.35416667 0.22222222
#>  [7,] 0.44444444 0.55555556 0.66666667 0.61904762 0.40972222 0.33333333
#>  [8,] 0.55555556 0.44444444 0.55555556 0.69047619 0.59027778 0.33333333
#>  [9,] 0.66666667 0.55555556 0.55555556 0.76190476 0.59027778 0.38888889
#> [10,] 0.61904762 0.69047619 0.76190476 0.85714286 0.59821429 0.54761905
#> [11,] 0.40972222 0.59027778 0.59027778 0.59821429 0.37500000 0.53472222
#> [12,] 0.33333333 0.33333333 0.38888889 0.54761905 0.53472222 0.44444444
#> [13,] 0.27777778 0.16666667 0.22222222 0.43650794 0.59027778 0.44444444
#> [14,] 0.05555556 0.16666667 0.22222222 0.23809524 0.29166667 0.44444444
#> [15,] 0.05555556 0.11111111 0.17361111 0.19642857 0.31250000 0.40972222
#> [16,] 0.00000000 0.05555556 0.12698413 0.07142857 0.33928571 0.30952381
#> [17,] 0.05555556 0.00000000 0.05555556 0.18253968 0.40972222 0.27777778
#> [18,] 0.00000000 0.00000000 0.00000000 0.05555556 0.22916667 0.27777778
#> [19,] 0.00000000 0.00000000 0.00000000 0.00000000 0.12500000 0.18055556
#> [20,] 0.05555556 0.00000000 0.00000000 0.07142857 0.17361111 0.11111111
#> [21,] 0.07142857 0.07142857 0.00000000 0.00000000 0.07142857 0.00000000
#> [22,] 0.05555556 0.00000000 0.00000000 0.00000000 0.13392857 0.05555556
#> [23,] 0.05555556 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [24,] 0.05555556 0.05555556 0.05555556 0.00000000 0.00000000 0.00000000
#> [25,] 0.22222222 0.11111111 0.00000000 0.00000000 0.00000000 0.00000000
#>            [,13]      [,14]      [,15]      [,16]      [,17]      [,18]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.11111111
#>  [2,] 0.00000000 0.00000000 0.00000000 0.00000000 0.05555556 0.00000000
#>  [3,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [4,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [5,] 0.05555556 0.00000000 0.00000000 0.00000000 0.00000000 0.07142857
#>  [6,] 0.16666667 0.05555556 0.05555556 0.00000000 0.00000000 0.00000000
#>  [7,] 0.27777778 0.05555556 0.05555556 0.00000000 0.05555556 0.00000000
#>  [8,] 0.16666667 0.16666667 0.11111111 0.05555556 0.00000000 0.00000000
#>  [9,] 0.22222222 0.22222222 0.17361111 0.12698413 0.05555556 0.00000000
#> [10,] 0.43650794 0.23809524 0.19642857 0.07142857 0.18253968 0.05555556
#> [11,] 0.59027778 0.29166667 0.31250000 0.33928571 0.40972222 0.22916667
#> [12,] 0.44444444 0.44444444 0.40972222 0.30952381 0.27777778 0.27777778
#> [13,] 0.44444444 0.55555556 0.58333333 0.56349206 0.38888889 0.33333333
#> [14,] 0.55555556 0.44444444 0.52777778 0.56349206 0.72222222 0.44444444
#> [15,] 0.58333333 0.52777778 0.62500000 0.66964286 0.65277778 0.52777778
#> [16,] 0.56349206 0.56349206 0.66964286 0.57142857 0.63492063 0.61904762
#> [17,] 0.38888889 0.72222222 0.65277778 0.63492063 0.55555556 0.61111111
#> [18,] 0.33333333 0.44444444 0.52777778 0.61904762 0.61111111 0.55555556
#> [19,] 0.35416667 0.35416667 0.43750000 0.59821429 0.65277778 0.53472222
#> [20,] 0.27777778 0.33333333 0.34722222 0.42063492 0.55555556 0.55555556
#> [21,] 0.12698413 0.25396825 0.19642857 0.07142857 0.25396825 0.38095238
#> [22,] 0.05555556 0.30952381 0.25892857 0.21428571 0.11111111 0.36507937
#> [23,] 0.00000000 0.05555556 0.11805556 0.18253968 0.22222222 0.11111111
#> [24,] 0.00000000 0.07142857 0.00000000 0.00000000 0.05555556 0.05555556
#> [25,] 0.00000000 0.00000000 0.00000000 0.00000000 0.05555556 0.11111111
#>            [,19]      [,20]      [,21]      [,22]      [,23]      [,24]
#>  [1,] 0.11805556 0.11111111 0.12698413 0.38095238 0.44444444 0.61904762
#>  [2,] 0.00000000 0.11111111 0.32539683 0.43650794 0.38888889 0.56349206
#>  [3,] 0.00000000 0.11111111 0.32539683 0.36507937 0.38888889 0.49206349
#>  [4,] 0.00000000 0.06250000 0.13392857 0.33035714 0.34722222 0.59821429
#>  [5,] 0.07142857 0.00000000 0.00000000 0.07142857 0.32539683 0.42857143
#>  [6,] 0.00000000 0.00000000 0.00000000 0.05555556 0.16666667 0.23809524
#>  [7,] 0.00000000 0.05555556 0.07142857 0.05555556 0.05555556 0.05555556
#>  [8,] 0.00000000 0.00000000 0.07142857 0.00000000 0.00000000 0.05555556
#>  [9,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.05555556
#> [10,] 0.00000000 0.07142857 0.00000000 0.00000000 0.00000000 0.00000000
#> [11,] 0.12500000 0.17361111 0.07142857 0.13392857 0.00000000 0.00000000
#> [12,] 0.18055556 0.11111111 0.00000000 0.05555556 0.00000000 0.00000000
#> [13,] 0.35416667 0.27777778 0.12698413 0.05555556 0.00000000 0.00000000
#> [14,] 0.35416667 0.33333333 0.25396825 0.30952381 0.05555556 0.07142857
#> [15,] 0.43750000 0.34722222 0.19642857 0.25892857 0.11805556 0.00000000
#> [16,] 0.59821429 0.42063492 0.07142857 0.21428571 0.18253968 0.00000000
#> [17,] 0.65277778 0.55555556 0.25396825 0.11111111 0.22222222 0.05555556
#> [18,] 0.53472222 0.55555556 0.38095238 0.36507937 0.11111111 0.05555556
#> [19,] 0.50000000 0.52777778 0.47321429 0.46428571 0.11805556 0.13392857
#> [20,] 0.52777778 0.33333333 0.18253968 0.43650794 0.33333333 0.25396825
#> [21,] 0.47321429 0.18253968 0.00000000 0.28571429 0.57936508 0.42857143
#> [22,] 0.46428571 0.43650794 0.28571429 0.42857143 0.63492063 0.64285714
#> [23,] 0.11805556 0.33333333 0.57936508 0.63492063 0.33333333 0.56349206
#> [24,] 0.13392857 0.25396825 0.42857143 0.64285714 0.56349206 0.71428571
#> [25,] 0.11805556 0.05555556 0.25396825 0.50793651 0.50000000 0.76190476
#>            [,25]
#>  [1,] 0.50000000
#>  [2,] 0.55555556
#>  [3,] 0.55555556
#>  [4,] 0.47222222
#>  [5,] 0.18253968
#>  [6,] 0.22222222
#>  [7,] 0.22222222
#>  [8,] 0.11111111
#>  [9,] 0.00000000
#> [10,] 0.00000000
#> [11,] 0.00000000
#> [12,] 0.00000000
#> [13,] 0.00000000
#> [14,] 0.00000000
#> [15,] 0.00000000
#> [16,] 0.00000000
#> [17,] 0.05555556
#> [18,] 0.11111111
#> [19,] 0.11805556
#> [20,] 0.05555556
#> [21,] 0.25396825
#> [22,] 0.50793651
#> [23,] 0.50000000
#> [24,] 0.76190476
#> [25,] 0.55555556
#> attr(,"representation")
#> [1] "graphon"