Network Representation Functions

This is a collection of functions that convert a graph stored as an igraph object into a desired matrix representation among adjacency matrix, graph laplacian, modularity matrix or graphon (edge probability matrix).

repr_adjacency(network, validate = TRUE)

repr_laplacian(network, validate = TRUE)

repr_modularity(network, validate = TRUE)

repr_graphon(network, validate = TRUE)

Arguments

network

An igraph object.

validate

A boolean specifying whether the function should check the class of its input (default: TRUE).

Value

A numeric square matrix giving the desired network representation recorded in the object's class.

Examples

g <- igraph::sample_smallworld(1, 25, 3, 0.05)
repr_adjacency(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0    1    1    1    0    0    0    0    0     0     0     0     0
#>  [2,]    1    0    1    1    1    0    0    0    0     0     0     0     0
#>  [3,]    1    1    0    1    1    0    0    0    0     0     0     0     0
#>  [4,]    1    1    1    0    1    1    1    0    0     0     0     1     0
#>  [5,]    0    1    1    1    0    1    0    0    0     0     0     0     0
#>  [6,]    0    0    0    1    1    0    1    1    1     0     0     0     0
#>  [7,]    0    0    0    1    0    1    0    1    1     1     0     0     0
#>  [8,]    0    0    0    0    0    1    1    0    1     1     1     0     0
#>  [9,]    0    0    0    0    0    1    1    1    0     1     0     1     0
#> [10,]    0    0    0    0    0    0    1    1    1     0     1     1     1
#> [11,]    0    0    0    0    0    0    0    1    0     1     0     1     1
#> [12,]    0    0    0    1    0    0    0    0    1     1     1     0     1
#> [13,]    0    0    0    0    0    0    0    0    0     1     1     1     0
#> [14,]    0    0    0    0    0    0    0    0    0     0     1     0     1
#> [15,]    0    0    0    0    0    0    0    0    0     0     0     1     1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0     1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    1    0    0    1     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]    1    0    0    0    1    0    0    1    0     0     0     0     0
#> [24,]    1    1    0    0    0    0    0    0    0     0     0     0     0
#> [25,]    1    1    1    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0     1     1     1
#>  [2,]     0     0     0     0     0     0     0     0     0     0     1     1
#>  [3,]     0     0     0     0     0     0     0     0     0     0     0     1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     0     0     1     0     0
#>  [6,]     0     0     0     0     1     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     1     0     0
#>  [9,]     0     0     0     0     1     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]     1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]     0     1     0     0     0     0     0     0     0     0     0     0
#> [13,]     1     1     1     0     0     0     0     0     0     0     0     0
#> [14,]     0     1     1     1     0     0     0     0     0     0     0     0
#> [15,]     1     0     1     1     1     0     0     0     0     0     0     0
#> [16,]     1     1     0     1     1     1     0     0     0     0     0     0
#> [17,]     1     1     1     0     1     1     1     0     0     0     0     1
#> [18,]     0     1     1     1     0     1     1     1     0     0     0     0
#> [19,]     0     0     1     1     1     0     1     1     1     0     0     0
#> [20,]     0     0     0     1     1     1     0     1     1     1     0     0
#> [21,]     0     0     0     0     1     1     1     0     1     1     1     0
#> [22,]     0     0     0     0     0     1     1     1     0     1     1     0
#> [23,]     0     0     0     0     0     0     1     1     1     0     1     1
#> [24,]     0     0     0     0     0     0     0     1     1     1     0     1
#> [25,]     0     0     0     1     0     0     0     0     0     1     1     0
#> attr(,"representation")
#> [1] "adjacency"
repr_laplacian(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    6   -1   -1   -1    0    0    0    0    0     0     0     0     0
#>  [2,]   -1    6   -1   -1   -1    0    0    0    0     0     0     0     0
#>  [3,]   -1   -1    5   -1   -1    0    0    0    0     0     0     0     0
#>  [4,]   -1   -1   -1    7   -1   -1   -1    0    0     0     0    -1     0
#>  [5,]    0   -1   -1   -1    5   -1    0    0    0     0     0     0     0
#>  [6,]    0    0    0   -1   -1    6   -1   -1   -1     0     0     0     0
#>  [7,]    0    0    0   -1    0   -1    5   -1   -1    -1     0     0     0
#>  [8,]    0    0    0    0    0   -1   -1    6   -1    -1    -1     0     0
#>  [9,]    0    0    0    0    0   -1   -1   -1    6    -1     0    -1     0
#> [10,]    0    0    0    0    0    0   -1   -1   -1     6    -1    -1    -1
#> [11,]    0    0    0    0    0    0    0   -1    0    -1     5    -1    -1
#> [12,]    0    0    0   -1    0    0    0    0   -1    -1    -1     6    -1
#> [13,]    0    0    0    0    0    0    0    0    0    -1    -1    -1     6
#> [14,]    0    0    0    0    0    0    0    0    0     0    -1     0    -1
#> [15,]    0    0    0    0    0    0    0    0    0     0     0    -1    -1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0    -1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0   -1    0    0   -1     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]   -1    0    0    0   -1    0    0   -1    0     0     0     0     0
#> [24,]   -1   -1    0    0    0    0    0    0    0     0     0     0     0
#> [25,]   -1   -1   -1    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0    -1    -1    -1
#>  [2,]     0     0     0     0     0     0     0     0     0     0    -1    -1
#>  [3,]     0     0     0     0     0     0     0     0     0     0     0    -1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     0     0    -1     0     0
#>  [6,]     0     0     0     0    -1     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0    -1     0     0
#>  [9,]     0     0     0     0    -1     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]    -1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]     0    -1     0     0     0     0     0     0     0     0     0     0
#> [13,]    -1    -1    -1     0     0     0     0     0     0     0     0     0
#> [14,]     5    -1    -1    -1     0     0     0     0     0     0     0     0
#> [15,]    -1     6    -1    -1    -1     0     0     0     0     0     0     0
#> [16,]    -1    -1     6    -1    -1    -1     0     0     0     0     0     0
#> [17,]    -1    -1    -1     7    -1    -1    -1     0     0     0     0    -1
#> [18,]     0    -1    -1    -1     8    -1    -1    -1     0     0     0     0
#> [19,]     0     0    -1    -1    -1     6    -1    -1    -1     0     0     0
#> [20,]     0     0     0    -1    -1    -1     6    -1    -1    -1     0     0
#> [21,]     0     0     0     0    -1    -1    -1     6    -1    -1    -1     0
#> [22,]     0     0     0     0     0    -1    -1    -1     5    -1    -1     0
#> [23,]     0     0     0     0     0     0    -1    -1    -1     8    -1    -1
#> [24,]     0     0     0     0     0     0     0    -1    -1    -1     6    -1
#> [25,]     0     0     0    -1     0     0     0     0     0    -1    -1     6
#> attr(,"representation")
#> [1] "laplacian"
repr_modularity(g)
#>        [,1]  [,2]       [,3]       [,4]       [,5]  [,6]       [,7]  [,8]  [,9]
#>  [1,] -0.24  0.76  0.8000000  0.7200000 -0.2000000 -0.24 -0.2000000 -0.24 -0.24
#>  [2,]  0.76 -0.24  0.8000000  0.7200000  0.8000000 -0.24 -0.2000000 -0.24 -0.24
#>  [3,]  0.80  0.80 -0.1666667  0.7666667  0.8333333 -0.20 -0.1666667 -0.20 -0.20
#>  [4,]  0.72  0.72  0.7666667 -0.3266667  0.7666667  0.72  0.7666667 -0.28 -0.28
#>  [5,] -0.20  0.80  0.8333333  0.7666667 -0.1666667  0.80 -0.1666667 -0.20 -0.20
#>  [6,] -0.24 -0.24 -0.2000000  0.7200000  0.8000000 -0.24  0.8000000  0.76  0.76
#>  [7,] -0.20 -0.20 -0.1666667  0.7666667 -0.1666667  0.80 -0.1666667  0.80  0.80
#>  [8,] -0.24 -0.24 -0.2000000 -0.2800000 -0.2000000  0.76  0.8000000 -0.24  0.76
#>  [9,] -0.24 -0.24 -0.2000000 -0.2800000 -0.2000000  0.76  0.8000000  0.76 -0.24
#> [10,] -0.24 -0.24 -0.2000000 -0.2800000 -0.2000000 -0.24  0.8000000  0.76  0.76
#> [11,] -0.20 -0.20 -0.1666667 -0.2333333 -0.1666667 -0.20 -0.1666667  0.80 -0.20
#> [12,] -0.24 -0.24 -0.2000000  0.7200000 -0.2000000 -0.24 -0.2000000 -0.24  0.76
#> [13,] -0.24 -0.24 -0.2000000 -0.2800000 -0.2000000 -0.24 -0.2000000 -0.24 -0.24
#> [14,] -0.20 -0.20 -0.1666667 -0.2333333 -0.1666667 -0.20 -0.1666667 -0.20 -0.20
#> [15,] -0.24 -0.24 -0.2000000 -0.2800000 -0.2000000 -0.24 -0.2000000 -0.24 -0.24
#> [16,] -0.24 -0.24 -0.2000000 -0.2800000 -0.2000000 -0.24 -0.2000000 -0.24 -0.24
#> [17,] -0.28 -0.28 -0.2333333 -0.3266667 -0.2333333 -0.28 -0.2333333 -0.28 -0.28
#> [18,] -0.32 -0.32 -0.2666667 -0.3733333 -0.2666667  0.68 -0.2666667 -0.32  0.68
#> [19,] -0.24 -0.24 -0.2000000 -0.2800000 -0.2000000 -0.24 -0.2000000 -0.24 -0.24
#> [20,] -0.24 -0.24 -0.2000000 -0.2800000 -0.2000000 -0.24 -0.2000000 -0.24 -0.24
#> [21,] -0.24 -0.24 -0.2000000 -0.2800000 -0.2000000 -0.24 -0.2000000 -0.24 -0.24
#> [22,] -0.20 -0.20 -0.1666667 -0.2333333 -0.1666667 -0.20 -0.1666667 -0.20 -0.20
#> [23,]  0.68 -0.32 -0.2666667 -0.3733333  0.7333333 -0.32 -0.2666667  0.68 -0.32
#> [24,]  0.76  0.76 -0.2000000 -0.2800000 -0.2000000 -0.24 -0.2000000 -0.24 -0.24
#> [25,]  0.76  0.76  0.8000000 -0.2800000 -0.2000000 -0.24 -0.2000000 -0.24 -0.24
#>       [,10]      [,11] [,12] [,13]      [,14] [,15] [,16]      [,17]      [,18]
#>  [1,] -0.24 -0.2000000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000 -0.3200000
#>  [2,] -0.24 -0.2000000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000 -0.3200000
#>  [3,] -0.20 -0.1666667 -0.20 -0.20 -0.1666667 -0.20 -0.20 -0.2333333 -0.2666667
#>  [4,] -0.28 -0.2333333  0.72 -0.28 -0.2333333 -0.28 -0.28 -0.3266667 -0.3733333
#>  [5,] -0.20 -0.1666667 -0.20 -0.20 -0.1666667 -0.20 -0.20 -0.2333333 -0.2666667
#>  [6,] -0.24 -0.2000000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000  0.6800000
#>  [7,]  0.80 -0.1666667 -0.20 -0.20 -0.1666667 -0.20 -0.20 -0.2333333 -0.2666667
#>  [8,]  0.76  0.8000000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000 -0.3200000
#>  [9,]  0.76 -0.2000000  0.76 -0.24 -0.2000000 -0.24 -0.24 -0.2800000  0.6800000
#> [10,] -0.24  0.8000000  0.76  0.76 -0.2000000 -0.24 -0.24 -0.2800000 -0.3200000
#> [11,]  0.80 -0.1666667  0.80  0.80  0.8333333 -0.20 -0.20 -0.2333333 -0.2666667
#> [12,]  0.76  0.8000000 -0.24  0.76 -0.2000000  0.76 -0.24 -0.2800000 -0.3200000
#> [13,]  0.76  0.8000000  0.76 -0.24  0.8000000  0.76  0.76 -0.2800000 -0.3200000
#> [14,] -0.20  0.8333333 -0.20  0.80 -0.1666667  0.80  0.80  0.7666667 -0.2666667
#> [15,] -0.24 -0.2000000  0.76  0.76  0.8000000 -0.24  0.76  0.7200000  0.6800000
#> [16,] -0.24 -0.2000000 -0.24  0.76  0.8000000  0.76 -0.24  0.7200000  0.6800000
#> [17,] -0.28 -0.2333333 -0.28 -0.28  0.7666667  0.72  0.72 -0.3266667  0.6266667
#> [18,] -0.32 -0.2666667 -0.32 -0.32 -0.2666667  0.68  0.68  0.6266667 -0.4266667
#> [19,] -0.24 -0.2000000 -0.24 -0.24 -0.2000000 -0.24  0.76  0.7200000  0.6800000
#> [20,] -0.24 -0.2000000 -0.24 -0.24 -0.2000000 -0.24 -0.24  0.7200000  0.6800000
#> [21,] -0.24 -0.2000000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000  0.6800000
#> [22,] -0.20 -0.1666667 -0.20 -0.20 -0.1666667 -0.20 -0.20 -0.2333333 -0.2666667
#> [23,] -0.32 -0.2666667 -0.32 -0.32 -0.2666667 -0.32 -0.32 -0.3733333 -0.4266667
#> [24,] -0.24 -0.2000000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000 -0.3200000
#> [25,] -0.24 -0.2000000 -0.24 -0.24 -0.2000000 -0.24 -0.24  0.7200000 -0.3200000
#>       [,19] [,20] [,21]      [,22]      [,23] [,24] [,25]
#>  [1,] -0.24 -0.24 -0.24 -0.2000000  0.6800000  0.76  0.76
#>  [2,] -0.24 -0.24 -0.24 -0.2000000 -0.3200000  0.76  0.76
#>  [3,] -0.20 -0.20 -0.20 -0.1666667 -0.2666667 -0.20  0.80
#>  [4,] -0.28 -0.28 -0.28 -0.2333333 -0.3733333 -0.28 -0.28
#>  [5,] -0.20 -0.20 -0.20 -0.1666667  0.7333333 -0.20 -0.20
#>  [6,] -0.24 -0.24 -0.24 -0.2000000 -0.3200000 -0.24 -0.24
#>  [7,] -0.20 -0.20 -0.20 -0.1666667 -0.2666667 -0.20 -0.20
#>  [8,] -0.24 -0.24 -0.24 -0.2000000  0.6800000 -0.24 -0.24
#>  [9,] -0.24 -0.24 -0.24 -0.2000000 -0.3200000 -0.24 -0.24
#> [10,] -0.24 -0.24 -0.24 -0.2000000 -0.3200000 -0.24 -0.24
#> [11,] -0.20 -0.20 -0.20 -0.1666667 -0.2666667 -0.20 -0.20
#> [12,] -0.24 -0.24 -0.24 -0.2000000 -0.3200000 -0.24 -0.24
#> [13,] -0.24 -0.24 -0.24 -0.2000000 -0.3200000 -0.24 -0.24
#> [14,] -0.20 -0.20 -0.20 -0.1666667 -0.2666667 -0.20 -0.20
#> [15,] -0.24 -0.24 -0.24 -0.2000000 -0.3200000 -0.24 -0.24
#> [16,]  0.76 -0.24 -0.24 -0.2000000 -0.3200000 -0.24 -0.24
#> [17,]  0.72  0.72 -0.28 -0.2333333 -0.3733333 -0.28  0.72
#> [18,]  0.68  0.68  0.68 -0.2666667 -0.4266667 -0.32 -0.32
#> [19,] -0.24  0.76  0.76  0.8000000 -0.3200000 -0.24 -0.24
#> [20,]  0.76 -0.24  0.76  0.8000000  0.6800000 -0.24 -0.24
#> [21,]  0.76  0.76 -0.24  0.8000000  0.6800000  0.76 -0.24
#> [22,]  0.80  0.80  0.80 -0.1666667  0.7333333  0.80 -0.20
#> [23,] -0.32  0.68  0.68  0.7333333 -0.4266667  0.68  0.68
#> [24,] -0.24 -0.24  0.76  0.8000000  0.6800000 -0.24  0.76
#> [25,] -0.24 -0.24 -0.24 -0.2000000  0.6800000  0.76 -0.24
#> attr(,"representation")
#> [1] "modularity"
repr_graphon(g)
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.50000000 0.64583333 0.52777778 0.58333333 0.34722222 0.18750000
#>  [2,] 0.64583333 0.66666667 0.61111111 0.61111111 0.55555556 0.17361111
#>  [3,] 0.52777778 0.61111111 0.55555556 0.66666667 0.44444444 0.22916667
#>  [4,] 0.58333333 0.61111111 0.66666667 0.50000000 0.52777778 0.56250000
#>  [5,] 0.34722222 0.55555556 0.44444444 0.52777778 0.22222222 0.35416667
#>  [6,] 0.18750000 0.17361111 0.22916667 0.56250000 0.35416667 0.37500000
#>  [7,] 0.05555556 0.22222222 0.16666667 0.44444444 0.16666667 0.47222222
#>  [8,] 0.00000000 0.17361111 0.17361111 0.33333333 0.11805556 0.50000000
#>  [9,] 0.05555556 0.11111111 0.16666667 0.11111111 0.22222222 0.53472222
#> [10,] 0.00000000 0.00000000 0.05555556 0.25000000 0.19444444 0.41666667
#> [11,] 0.00000000 0.00000000 0.00000000 0.11111111 0.05555556 0.29861111
#> [12,] 0.00000000 0.05555556 0.05555556 0.33333333 0.05555556 0.23611111
#> [13,] 0.00000000 0.00000000 0.00000000 0.05555556 0.00000000 0.23611111
#> [14,] 0.00000000 0.00000000 0.00000000 0.06250000 0.00000000 0.00000000
#> [15,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.11805556
#> [16,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [17,] 0.06250000 0.05555556 0.05555556 0.08333333 0.05555556 0.07142857
#> [18,] 0.06250000 0.05555556 0.05555556 0.08333333 0.00000000 0.19642857
#> [19,] 0.18055556 0.05555556 0.00000000 0.00000000 0.05555556 0.05555556
#> [20,] 0.25000000 0.18055556 0.06250000 0.00000000 0.05555556 0.00000000
#> [21,] 0.23611111 0.27777778 0.16666667 0.05555556 0.11111111 0.00000000
#> [22,] 0.25000000 0.36111111 0.24305556 0.12500000 0.05555556 0.06250000
#> [23,] 0.66071429 0.36507937 0.29365079 0.30952381 0.54761905 0.06250000
#> [24,] 0.53472222 0.44444444 0.33333333 0.33333333 0.33333333 0.06250000
#> [25,] 0.47222222 0.55555556 0.44444444 0.38888889 0.33333333 0.18055556
#>             [,7]      [,8]       [,9]      [,10]      [,11]      [,12]
#>  [1,] 0.05555556 0.0000000 0.05555556 0.00000000 0.00000000 0.00000000
#>  [2,] 0.22222222 0.1736111 0.11111111 0.00000000 0.00000000 0.05555556
#>  [3,] 0.16666667 0.1736111 0.16666667 0.05555556 0.00000000 0.05555556
#>  [4,] 0.44444444 0.3333333 0.11111111 0.25000000 0.11111111 0.33333333
#>  [5,] 0.16666667 0.1180556 0.22222222 0.19444444 0.05555556 0.05555556
#>  [6,] 0.47222222 0.5000000 0.53472222 0.41666667 0.29861111 0.23611111
#>  [7,] 0.44444444 0.5277778 0.55555556 0.61111111 0.27777778 0.33333333
#>  [8,] 0.52777778 0.5000000 0.59027778 0.64583333 0.47222222 0.40972222
#>  [9,] 0.55555556 0.5902778 0.44444444 0.61111111 0.33333333 0.61111111
#> [10,] 0.61111111 0.6458333 0.61111111 0.66666667 0.52777778 0.52777778
#> [11,] 0.27777778 0.4722222 0.33333333 0.52777778 0.44444444 0.55555556
#> [12,] 0.33333333 0.4097222 0.61111111 0.52777778 0.55555556 0.33333333
#> [13,] 0.33333333 0.2916667 0.33333333 0.52777778 0.61111111 0.55555556
#> [14,] 0.11805556 0.1875000 0.29166667 0.27083333 0.47222222 0.41666667
#> [15,] 0.16666667 0.2916667 0.11111111 0.25000000 0.38888889 0.50000000
#> [16,] 0.00000000 0.0625000 0.11111111 0.00000000 0.26666667 0.21111111
#> [17,] 0.00000000 0.0000000 0.12698413 0.07142857 0.23809524 0.25396825
#> [18,] 0.11111111 0.1250000 0.23809524 0.16666667 0.23809524 0.11111111
#> [19,] 0.00000000 0.0000000 0.05555556 0.00000000 0.11111111 0.11111111
#> [20,] 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000 0.06250000
#> [21,] 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [22,] 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [23,] 0.11111111 0.1964286 0.05555556 0.08333333 0.00000000 0.05555556
#> [24,] 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [25,] 0.05555556 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000
#>            [,13]     [,14]      [,15]      [,16]      [,17]      [,18]
#>  [1,] 0.00000000 0.0000000 0.00000000 0.00000000 0.06250000 0.06250000
#>  [2,] 0.00000000 0.0000000 0.00000000 0.00000000 0.05555556 0.05555556
#>  [3,] 0.00000000 0.0000000 0.00000000 0.00000000 0.05555556 0.05555556
#>  [4,] 0.05555556 0.0625000 0.00000000 0.00000000 0.08333333 0.08333333
#>  [5,] 0.00000000 0.0000000 0.00000000 0.00000000 0.05555556 0.00000000
#>  [6,] 0.23611111 0.0000000 0.11805556 0.00000000 0.07142857 0.19642857
#>  [7,] 0.33333333 0.1180556 0.16666667 0.00000000 0.00000000 0.11111111
#>  [8,] 0.29166667 0.1875000 0.29166667 0.06250000 0.00000000 0.12500000
#>  [9,] 0.33333333 0.2916667 0.11111111 0.11111111 0.12698413 0.23809524
#> [10,] 0.52777778 0.2708333 0.25000000 0.00000000 0.07142857 0.16666667
#> [11,] 0.61111111 0.4722222 0.38888889 0.26666667 0.23809524 0.23809524
#> [12,] 0.55555556 0.4166667 0.50000000 0.21111111 0.25396825 0.11111111
#> [13,] 0.66666667 0.5972222 0.50000000 0.46666667 0.30952381 0.30952381
#> [14,] 0.59722222 0.5000000 0.47222222 0.55000000 0.53571429 0.33035714
#> [15,] 0.50000000 0.4722222 0.33333333 0.52222222 0.56349206 0.56349206
#> [16,] 0.46666667 0.5500000 0.52222222 0.60000000 0.75714286 0.68571429
#> [17,] 0.30952381 0.5357143 0.56349206 0.75714286 0.57142857 0.71428571
#> [18,] 0.30952381 0.3303571 0.56349206 0.68571429 0.71428571 0.57142857
#> [19,] 0.22222222 0.1736111 0.27777778 0.52222222 0.61904762 0.63492063
#> [20,] 0.12500000 0.1875000 0.11805556 0.22500000 0.59821429 0.59821429
#> [21,] 0.05555556 0.1736111 0.22222222 0.21111111 0.36507937 0.56349206
#> [22,] 0.00000000 0.0625000 0.11111111 0.16250000 0.33035714 0.33035714
#> [23,] 0.05555556 0.0000000 0.05555556 0.17142857 0.14285714 0.28571429
#> [24,] 0.00000000 0.0000000 0.00000000 0.05555556 0.23809524 0.23809524
#> [25,] 0.00000000 0.0000000 0.00000000 0.10000000 0.18253968 0.18253968
#>            [,19]      [,20]      [,21]      [,22]      [,23]      [,24]
#>  [1,] 0.18055556 0.25000000 0.23611111 0.25000000 0.66071429 0.53472222
#>  [2,] 0.05555556 0.18055556 0.27777778 0.36111111 0.36507937 0.44444444
#>  [3,] 0.00000000 0.06250000 0.16666667 0.24305556 0.29365079 0.33333333
#>  [4,] 0.00000000 0.00000000 0.05555556 0.12500000 0.30952381 0.33333333
#>  [5,] 0.05555556 0.05555556 0.11111111 0.05555556 0.54761905 0.33333333
#>  [6,] 0.05555556 0.00000000 0.00000000 0.06250000 0.06250000 0.06250000
#>  [7,] 0.00000000 0.00000000 0.00000000 0.00000000 0.11111111 0.00000000
#>  [8,] 0.00000000 0.00000000 0.00000000 0.00000000 0.19642857 0.00000000
#>  [9,] 0.05555556 0.00000000 0.00000000 0.00000000 0.05555556 0.00000000
#> [10,] 0.00000000 0.00000000 0.00000000 0.00000000 0.08333333 0.00000000
#> [11,] 0.11111111 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [12,] 0.11111111 0.06250000 0.00000000 0.00000000 0.05555556 0.00000000
#> [13,] 0.22222222 0.12500000 0.05555556 0.00000000 0.05555556 0.00000000
#> [14,] 0.17361111 0.18750000 0.17361111 0.06250000 0.00000000 0.00000000
#> [15,] 0.27777778 0.11805556 0.22222222 0.11111111 0.05555556 0.00000000
#> [16,] 0.52222222 0.22500000 0.21111111 0.16250000 0.17142857 0.05555556
#> [17,] 0.61904762 0.59821429 0.36507937 0.33035714 0.14285714 0.23809524
#> [18,] 0.63492063 0.59821429 0.56349206 0.33035714 0.28571429 0.23809524
#> [19,] 0.55555556 0.47222222 0.55555556 0.40972222 0.36507937 0.27777778
#> [20,] 0.47222222 0.37500000 0.47222222 0.43750000 0.59821429 0.35416667
#> [21,] 0.55555556 0.47222222 0.44444444 0.47222222 0.61904762 0.44444444
#> [22,] 0.40972222 0.43750000 0.47222222 0.50000000 0.72321429 0.47222222
#> [23,] 0.36507937 0.59821429 0.61904762 0.72321429 0.42857143 0.61904762
#> [24,] 0.27777778 0.35416667 0.44444444 0.47222222 0.61904762 0.55555556
#> [25,] 0.22222222 0.17361111 0.33333333 0.29166667 0.67460317 0.50000000
#>            [,25]
#>  [1,] 0.47222222
#>  [2,] 0.55555556
#>  [3,] 0.44444444
#>  [4,] 0.38888889
#>  [5,] 0.33333333
#>  [6,] 0.18055556
#>  [7,] 0.05555556
#>  [8,] 0.00000000
#>  [9,] 0.00000000
#> [10,] 0.00000000
#> [11,] 0.00000000
#> [12,] 0.00000000
#> [13,] 0.00000000
#> [14,] 0.00000000
#> [15,] 0.00000000
#> [16,] 0.10000000
#> [17,] 0.18253968
#> [18,] 0.18253968
#> [19,] 0.22222222
#> [20,] 0.17361111
#> [21,] 0.33333333
#> [22,] 0.29166667
#> [23,] 0.67460317
#> [24,] 0.50000000
#> [25,] 0.44444444
#> attr(,"representation")
#> [1] "graphon"