Network Representation Functions

This is a collection of functions that convert a graph stored as an igraph object into a desired matrix representation among adjacency matrix, graph laplacian, modularity matrix or graphon (edge probability matrix).

Usage

repr_adjacency(network, validate = TRUE)

repr_laplacian(network, validate = TRUE)

repr_modularity(network, validate = TRUE)

repr_graphon(network, validate = TRUE)

Arguments

network

An igraph object.

validate

A boolean specifying whether the function should check the class of its input (default: TRUE).

Value

A numeric square matrix giving the desired network representation recorded in the object's class.

Examples

g <- igraph::sample_smallworld(1, 25, 3, 0.05)
repr_adjacency(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0    1    1    0    0    0    0    0    0     0     0     1     0
#>  [2,]    1    0    1    1    1    0    0    0    0     0     0     0     0
#>  [3,]    1    1    0    1    1    1    0    0    0     0     0     0     0
#>  [4,]    0    1    1    0    1    1    0    0    0     0     0     1     0
#>  [5,]    0    1    1    1    0    1    1    1    0     0     0     0     0
#>  [6,]    0    0    1    1    1    0    1    1    0     0     0     0     0
#>  [7,]    0    0    0    0    1    1    0    1    1     1     0     0     0
#>  [8,]    0    0    0    0    1    1    1    0    1     1     1     0     0
#>  [9,]    0    0    0    0    0    0    1    1    0     1     1     1     0
#> [10,]    0    0    0    0    0    0    1    1    1     0     1     1     1
#> [11,]    0    0    0    0    0    0    0    1    1     1     0     1     1
#> [12,]    1    0    0    1    0    0    0    0    1     1     1     0     0
#> [13,]    0    0    0    0    0    0    0    0    0     1     1     0     0
#> [14,]    0    0    0    0    0    0    0    0    0     0     1     1     1
#> [15,]    0    0    0    0    0    0    1    0    0     0     0     1     0
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0     1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     1
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]    1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]    1    1    0    0    0    0    0    0    1     0     0     0     0
#> [25,]    1    1    1    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0     1     1     1
#>  [2,]     0     0     0     0     0     0     0     0     0     0     1     1
#>  [3,]     0     0     0     0     0     0     0     0     0     0     0     1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     1     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     1     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]     1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]     1     1     0     0     0     0     0     0     0     0     0     0
#> [13,]     1     0     1     0     1     0     0     0     0     0     0     0
#> [14,]     0     1     1     1     0     0     0     0     0     0     0     0
#> [15,]     1     0     1     1     1     0     0     0     0     0     0     0
#> [16,]     1     1     0     1     1     1     0     0     0     0     0     0
#> [17,]     1     1     1     0     1     1     1     0     0     0     0     0
#> [18,]     0     1     1     1     0     1     1     1     0     0     0     0
#> [19,]     0     0     1     1     1     0     1     1     1     0     0     0
#> [20,]     0     0     0     1     1     1     0     1     1     1     0     0
#> [21,]     0     0     0     0     1     1     1     0     1     1     1     0
#> [22,]     0     0     0     0     0     1     1     1     0     1     1     1
#> [23,]     0     0     0     0     0     0     1     1     1     0     1     1
#> [24,]     0     0     0     0     0     0     0     1     1     1     0     1
#> [25,]     0     0     0     0     0     0     0     0     1     1     1     0
#> attr(,"representation")
#> [1] "adjacency"
repr_laplacian(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    6   -1   -1    0    0    0    0    0    0     0     0    -1     0
#>  [2,]   -1    6   -1   -1   -1    0    0    0    0     0     0     0     0
#>  [3,]   -1   -1    6   -1   -1   -1    0    0    0     0     0     0     0
#>  [4,]    0   -1   -1    5   -1   -1    0    0    0     0     0    -1     0
#>  [5,]    0   -1   -1   -1    6   -1   -1   -1    0     0     0     0     0
#>  [6,]    0    0   -1   -1   -1    5   -1   -1    0     0     0     0     0
#>  [7,]    0    0    0    0   -1   -1    6   -1   -1    -1     0     0     0
#>  [8,]    0    0    0    0   -1   -1   -1    6   -1    -1    -1     0     0
#>  [9,]    0    0    0    0    0    0   -1   -1    6    -1    -1    -1     0
#> [10,]    0    0    0    0    0    0   -1   -1   -1     6    -1    -1    -1
#> [11,]    0    0    0    0    0    0    0   -1   -1    -1     6    -1    -1
#> [12,]   -1    0    0   -1    0    0    0    0   -1    -1    -1     7     0
#> [13,]    0    0    0    0    0    0    0    0    0    -1    -1     0     5
#> [14,]    0    0    0    0    0    0    0    0    0     0    -1    -1    -1
#> [15,]    0    0    0    0    0    0   -1    0    0     0     0    -1     0
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0    -1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0    -1
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]   -1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]   -1   -1    0    0    0    0    0    0   -1     0     0     0     0
#> [25,]   -1   -1   -1    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0    -1    -1    -1
#>  [2,]     0     0     0     0     0     0     0     0     0     0    -1    -1
#>  [3,]     0     0     0     0     0     0     0     0     0     0     0    -1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0    -1     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0    -1     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]    -1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]    -1    -1     0     0     0     0     0     0     0     0     0     0
#> [13,]    -1     0    -1     0    -1     0     0     0     0     0     0     0
#> [14,]     6    -1    -1    -1     0     0     0     0     0     0     0     0
#> [15,]    -1     6    -1    -1    -1     0     0     0     0     0     0     0
#> [16,]    -1    -1     6    -1    -1    -1     0     0     0     0     0     0
#> [17,]    -1    -1    -1     6    -1    -1    -1     0     0     0     0     0
#> [18,]     0    -1    -1    -1     7    -1    -1    -1     0     0     0     0
#> [19,]     0     0    -1    -1    -1     6    -1    -1    -1     0     0     0
#> [20,]     0     0     0    -1    -1    -1     6    -1    -1    -1     0     0
#> [21,]     0     0     0     0    -1    -1    -1     6    -1    -1    -1     0
#> [22,]     0     0     0     0     0    -1    -1    -1     6    -1    -1    -1
#> [23,]     0     0     0     0     0     0    -1    -1    -1     6    -1    -1
#> [24,]     0     0     0     0     0     0     0    -1    -1    -1     7    -1
#> [25,]     0     0     0     0     0     0     0     0    -1    -1    -1     6
#> attr(,"representation")
#> [1] "laplacian"
repr_modularity(g)
#>        [,1]  [,2]  [,3]       [,4]  [,5]       [,6]  [,7]  [,8]  [,9] [,10]
#>  [1,] -0.24  0.76  0.76 -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24
#>  [2,]  0.76 -0.24  0.76  0.8000000  0.76 -0.2000000 -0.24 -0.24 -0.24 -0.24
#>  [3,]  0.76  0.76 -0.24  0.8000000  0.76  0.8000000 -0.24 -0.24 -0.24 -0.24
#>  [4,] -0.20  0.80  0.80 -0.1666667  0.80  0.8333333 -0.20 -0.20 -0.20 -0.20
#>  [5,] -0.24  0.76  0.76  0.8000000 -0.24  0.8000000  0.76  0.76 -0.24 -0.24
#>  [6,] -0.20 -0.20  0.80  0.8333333  0.80 -0.1666667  0.80  0.80 -0.20 -0.20
#>  [7,] -0.24 -0.24 -0.24 -0.2000000  0.76  0.8000000 -0.24  0.76  0.76  0.76
#>  [8,] -0.24 -0.24 -0.24 -0.2000000  0.76  0.8000000  0.76 -0.24  0.76  0.76
#>  [9,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.2000000  0.76  0.76 -0.24  0.76
#> [10,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.2000000  0.76  0.76  0.76 -0.24
#> [11,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.2000000 -0.24  0.76  0.76  0.76
#> [12,]  0.72 -0.28 -0.28  0.7666667 -0.28 -0.2333333 -0.28 -0.28  0.72  0.72
#> [13,] -0.20 -0.20 -0.20 -0.1666667 -0.20 -0.1666667 -0.20 -0.20 -0.20  0.80
#> [14,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24
#> [15,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.2000000  0.76 -0.24 -0.24 -0.24
#> [16,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24
#> [17,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24
#> [18,] -0.28 -0.28 -0.28 -0.2333333 -0.28 -0.2333333 -0.28 -0.28 -0.28 -0.28
#> [19,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24
#> [20,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24
#> [21,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24
#> [22,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24
#> [23,]  0.76 -0.24 -0.24 -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24
#> [24,]  0.72  0.72 -0.28 -0.2333333 -0.28 -0.2333333 -0.28 -0.28  0.72 -0.28
#> [25,]  0.76  0.76  0.76 -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24
#>       [,11]      [,12]      [,13] [,14] [,15] [,16] [,17]      [,18] [,19]
#>  [1,] -0.24  0.7200000 -0.2000000 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#>  [2,] -0.24 -0.2800000 -0.2000000 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#>  [3,] -0.24 -0.2800000 -0.2000000 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#>  [4,] -0.20  0.7666667 -0.1666667 -0.20 -0.20 -0.20 -0.20 -0.2333333 -0.20
#>  [5,] -0.24 -0.2800000 -0.2000000 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#>  [6,] -0.20 -0.2333333 -0.1666667 -0.20 -0.20 -0.20 -0.20 -0.2333333 -0.20
#>  [7,] -0.24 -0.2800000 -0.2000000 -0.24  0.76 -0.24 -0.24 -0.2800000 -0.24
#>  [8,]  0.76 -0.2800000 -0.2000000 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#>  [9,]  0.76  0.7200000 -0.2000000 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#> [10,]  0.76  0.7200000  0.8000000 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#> [11,] -0.24  0.7200000  0.8000000  0.76 -0.24 -0.24 -0.24 -0.2800000 -0.24
#> [12,]  0.72 -0.3266667 -0.2333333  0.72  0.72 -0.28 -0.28 -0.3266667 -0.28
#> [13,]  0.80 -0.2333333 -0.1666667  0.80 -0.20  0.80 -0.20  0.7666667 -0.20
#> [14,]  0.76  0.7200000  0.8000000 -0.24  0.76  0.76  0.76 -0.2800000 -0.24
#> [15,] -0.24  0.7200000 -0.2000000  0.76 -0.24  0.76  0.76  0.7200000 -0.24
#> [16,] -0.24 -0.2800000  0.8000000  0.76  0.76 -0.24  0.76  0.7200000  0.76
#> [17,] -0.24 -0.2800000 -0.2000000  0.76  0.76  0.76 -0.24  0.7200000  0.76
#> [18,] -0.28 -0.3266667  0.7666667 -0.28  0.72  0.72  0.72 -0.3266667  0.72
#> [19,] -0.24 -0.2800000 -0.2000000 -0.24 -0.24  0.76  0.76  0.7200000 -0.24
#> [20,] -0.24 -0.2800000 -0.2000000 -0.24 -0.24 -0.24  0.76  0.7200000  0.76
#> [21,] -0.24 -0.2800000 -0.2000000 -0.24 -0.24 -0.24 -0.24  0.7200000  0.76
#> [22,] -0.24 -0.2800000 -0.2000000 -0.24 -0.24 -0.24 -0.24 -0.2800000  0.76
#> [23,] -0.24 -0.2800000 -0.2000000 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#> [24,] -0.28 -0.3266667 -0.2333333 -0.28 -0.28 -0.28 -0.28 -0.3266667 -0.28
#> [25,] -0.24 -0.2800000 -0.2000000 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#>       [,20] [,21] [,22] [,23]      [,24] [,25]
#>  [1,] -0.24 -0.24 -0.24  0.76  0.7200000  0.76
#>  [2,] -0.24 -0.24 -0.24 -0.24  0.7200000  0.76
#>  [3,] -0.24 -0.24 -0.24 -0.24 -0.2800000  0.76
#>  [4,] -0.20 -0.20 -0.20 -0.20 -0.2333333 -0.20
#>  [5,] -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#>  [6,] -0.20 -0.20 -0.20 -0.20 -0.2333333 -0.20
#>  [7,] -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#>  [8,] -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#>  [9,] -0.24 -0.24 -0.24 -0.24  0.7200000 -0.24
#> [10,] -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#> [11,] -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#> [12,] -0.28 -0.28 -0.28 -0.28 -0.3266667 -0.28
#> [13,] -0.20 -0.20 -0.20 -0.20 -0.2333333 -0.20
#> [14,] -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#> [15,] -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#> [16,] -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.24
#> [17,]  0.76 -0.24 -0.24 -0.24 -0.2800000 -0.24
#> [18,]  0.72  0.72 -0.28 -0.28 -0.3266667 -0.28
#> [19,]  0.76  0.76  0.76 -0.24 -0.2800000 -0.24
#> [20,] -0.24  0.76  0.76  0.76 -0.2800000 -0.24
#> [21,]  0.76 -0.24  0.76  0.76  0.7200000 -0.24
#> [22,]  0.76  0.76 -0.24  0.76  0.7200000  0.76
#> [23,]  0.76  0.76  0.76 -0.24  0.7200000  0.76
#> [24,] -0.28  0.72  0.72  0.72 -0.3266667  0.72
#> [25,] -0.24 -0.24  0.76  0.76  0.7200000 -0.24
#> attr(,"representation")
#> [1] "modularity"
repr_graphon(g)
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.40000000 0.52222222 0.65000000 0.26666667 0.32500000 0.21111111
#>  [2,] 0.52222222 0.55555556 0.70833333 0.50000000 0.59722222 0.33333333
#>  [3,] 0.65000000 0.70833333 0.75000000 0.70833333 0.75000000 0.59722222
#>  [4,] 0.26666667 0.50000000 0.70833333 0.44444444 0.70833333 0.50000000
#>  [5,] 0.32500000 0.59722222 0.75000000 0.70833333 0.75000000 0.70833333
#>  [6,] 0.21111111 0.33333333 0.59722222 0.50000000 0.70833333 0.55555556
#>  [7,] 0.05555556 0.27777778 0.29166667 0.33333333 0.52777778 0.55555556
#>  [8,] 0.00000000 0.11111111 0.12500000 0.16666667 0.50000000 0.52777778
#>  [9,] 0.00000000 0.00000000 0.00000000 0.11111111 0.12500000 0.29166667
#> [10,] 0.00000000 0.05555556 0.11111111 0.22222222 0.16666667 0.33333333
#> [11,] 0.00000000 0.00000000 0.00000000 0.05555556 0.11111111 0.27777778
#> [12,] 0.30000000 0.11111111 0.12500000 0.21111111 0.32500000 0.36666667
#> [13,] 0.00000000 0.00000000 0.00000000 0.00000000 0.05555556 0.11111111
#> [14,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [15,] 0.00000000 0.00000000 0.00000000 0.05555556 0.00000000 0.05555556
#> [16,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [17,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [18,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [19,] 0.17142857 0.05555556 0.00000000 0.00000000 0.00000000 0.00000000
#> [20,] 0.21111111 0.16666667 0.05555556 0.00000000 0.00000000 0.00000000
#> [21,] 0.21111111 0.22222222 0.11111111 0.00000000 0.00000000 0.00000000
#> [22,] 0.26666667 0.22222222 0.16666667 0.11111111 0.05555556 0.00000000
#> [23,] 0.58571429 0.30952381 0.14285714 0.18253968 0.07142857 0.00000000
#> [24,] 0.68571429 0.56349206 0.26785714 0.23809524 0.07142857 0.18253968
#> [25,] 0.63333333 0.52777778 0.50000000 0.25000000 0.20833333 0.11111111
#>             [,7]      [,8]       [,9]      [,10]      [,11]      [,12]
#>  [1,] 0.05555556 0.0000000 0.00000000 0.00000000 0.00000000 0.30000000
#>  [2,] 0.27777778 0.1111111 0.00000000 0.05555556 0.00000000 0.11111111
#>  [3,] 0.29166667 0.1250000 0.00000000 0.11111111 0.00000000 0.12500000
#>  [4,] 0.33333333 0.1666667 0.11111111 0.22222222 0.05555556 0.21111111
#>  [5,] 0.52777778 0.5000000 0.12500000 0.16666667 0.11111111 0.32500000
#>  [6,] 0.55555556 0.5277778 0.29166667 0.33333333 0.27777778 0.36666667
#>  [7,] 0.55555556 0.7083333 0.59722222 0.55555556 0.38888889 0.32222222
#>  [8,] 0.70833333 0.7500000 0.75000000 0.70833333 0.59722222 0.45000000
#>  [9,] 0.59722222 0.7500000 0.75000000 0.59722222 0.59722222 0.77500000
#> [10,] 0.55555556 0.7083333 0.59722222 0.44444444 0.50000000 0.77777778
#> [11,] 0.38888889 0.5972222 0.59722222 0.50000000 0.55555556 0.57777778
#> [12,] 0.32222222 0.4500000 0.77500000 0.77777778 0.57777778 0.20000000
#> [13,] 0.27777778 0.2361111 0.36111111 0.38888889 0.44444444 0.32222222
#> [14,] 0.16666667 0.1111111 0.23611111 0.22222222 0.38888889 0.52222222
#> [15,] 0.13888889 0.1250000 0.00000000 0.19444444 0.33333333 0.36666667
#> [16,] 0.08333333 0.0000000 0.00000000 0.11111111 0.25000000 0.26666667
#> [17,] 0.07142857 0.0000000 0.00000000 0.18253968 0.23809524 0.07142857
#> [18,] 0.00000000 0.0000000 0.00000000 0.05555556 0.23809524 0.17142857
#> [19,] 0.07142857 0.0000000 0.00000000 0.00000000 0.05555556 0.07142857
#> [20,] 0.05555556 0.0000000 0.00000000 0.00000000 0.00000000 0.05555556
#> [21,] 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000 0.05555556
#> [22,] 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000 0.05555556
#> [23,] 0.00000000 0.0000000 0.07142857 0.00000000 0.00000000 0.00000000
#> [24,] 0.05555556 0.1250000 0.19642857 0.05555556 0.05555556 0.07142857
#> [25,] 0.05555556 0.0000000 0.08333333 0.00000000 0.00000000 0.08333333
#>            [,13]     [,14]      [,15]      [,16]      [,17]      [,18]
#>  [1,] 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [3,] 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [4,] 0.00000000 0.0000000 0.05555556 0.00000000 0.00000000 0.00000000
#>  [5,] 0.05555556 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [6,] 0.11111111 0.0000000 0.05555556 0.00000000 0.00000000 0.00000000
#>  [7,] 0.27777778 0.1666667 0.13888889 0.08333333 0.07142857 0.00000000
#>  [8,] 0.23611111 0.1111111 0.12500000 0.00000000 0.00000000 0.00000000
#>  [9,] 0.36111111 0.2361111 0.00000000 0.00000000 0.00000000 0.00000000
#> [10,] 0.38888889 0.2222222 0.19444444 0.11111111 0.18253968 0.05555556
#> [11,] 0.44444444 0.3888889 0.33333333 0.25000000 0.23809524 0.23809524
#> [12,] 0.32222222 0.5222222 0.36666667 0.26666667 0.07142857 0.17142857
#> [13,] 0.33333333 0.5555556 0.33333333 0.52777778 0.36507937 0.49206349
#> [14,] 0.55555556 0.5555556 0.55555556 0.58333333 0.56349206 0.42063492
#> [15,] 0.33333333 0.5555556 0.50000000 0.75000000 0.54761905 0.70238095
#> [16,] 0.52777778 0.5833333 0.75000000 0.66666667 0.77380952 0.61904762
#> [17,] 0.36507937 0.5634921 0.54761905 0.77380952 0.57142857 0.78571429
#> [18,] 0.49206349 0.4206349 0.70238095 0.61904762 0.78571429 0.57142857
#> [19,] 0.18253968 0.3095238 0.30952381 0.61904762 0.57142857 0.71428571
#> [20,] 0.16666667 0.3333333 0.27777778 0.33333333 0.50793651 0.61904762
#> [21,] 0.11111111 0.2222222 0.19444444 0.27777778 0.30952381 0.56349206
#> [22,] 0.05555556 0.1111111 0.13888889 0.19444444 0.25396825 0.36507937
#> [23,] 0.00000000 0.0000000 0.00000000 0.15476190 0.14285714 0.28571429
#> [24,] 0.05555556 0.0000000 0.00000000 0.00000000 0.14285714 0.14285714
#> [25,] 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000 0.15476190
#>            [,19]      [,20]      [,21]      [,22]      [,23]      [,24]
#>  [1,] 0.17142857 0.21111111 0.21111111 0.26666667 0.58571429 0.68571429
#>  [2,] 0.05555556 0.16666667 0.22222222 0.22222222 0.30952381 0.56349206
#>  [3,] 0.00000000 0.05555556 0.11111111 0.16666667 0.14285714 0.26785714
#>  [4,] 0.00000000 0.00000000 0.00000000 0.11111111 0.18253968 0.23809524
#>  [5,] 0.00000000 0.00000000 0.00000000 0.05555556 0.07142857 0.07142857
#>  [6,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.18253968
#>  [7,] 0.07142857 0.05555556 0.00000000 0.00000000 0.00000000 0.05555556
#>  [8,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.12500000
#>  [9,] 0.00000000 0.00000000 0.00000000 0.00000000 0.07142857 0.19642857
#> [10,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.05555556
#> [11,] 0.05555556 0.00000000 0.00000000 0.00000000 0.00000000 0.05555556
#> [12,] 0.07142857 0.05555556 0.05555556 0.05555556 0.00000000 0.07142857
#> [13,] 0.18253968 0.16666667 0.11111111 0.05555556 0.00000000 0.05555556
#> [14,] 0.30952381 0.33333333 0.22222222 0.11111111 0.00000000 0.00000000
#> [15,] 0.30952381 0.27777778 0.19444444 0.13888889 0.00000000 0.00000000
#> [16,] 0.61904762 0.33333333 0.27777778 0.19444444 0.15476190 0.00000000
#> [17,] 0.57142857 0.50793651 0.30952381 0.25396825 0.14285714 0.14285714
#> [18,] 0.71428571 0.61904762 0.56349206 0.36507937 0.28571429 0.14285714
#> [19,] 0.57142857 0.63492063 0.56349206 0.50793651 0.28571429 0.28571429
#> [20,] 0.63492063 0.55555556 0.50000000 0.55555556 0.50793651 0.36507937
#> [21,] 0.56349206 0.50000000 0.44444444 0.50000000 0.56349206 0.56349206
#> [22,] 0.50793651 0.55555556 0.50000000 0.55555556 0.63492063 0.61904762
#> [23,] 0.28571429 0.50793651 0.56349206 0.63492063 0.57142857 0.71428571
#> [24,] 0.28571429 0.36507937 0.56349206 0.61904762 0.71428571 0.57142857
#> [25,] 0.15476190 0.27777778 0.33333333 0.55555556 0.61904762 0.77380952
#>            [,25]
#>  [1,] 0.63333333
#>  [2,] 0.52777778
#>  [3,] 0.50000000
#>  [4,] 0.25000000
#>  [5,] 0.20833333
#>  [6,] 0.11111111
#>  [7,] 0.05555556
#>  [8,] 0.00000000
#>  [9,] 0.08333333
#> [10,] 0.00000000
#> [11,] 0.00000000
#> [12,] 0.08333333
#> [13,] 0.00000000
#> [14,] 0.00000000
#> [15,] 0.00000000
#> [16,] 0.00000000
#> [17,] 0.00000000
#> [18,] 0.15476190
#> [19,] 0.15476190
#> [20,] 0.27777778
#> [21,] 0.33333333
#> [22,] 0.55555556
#> [23,] 0.61904762
#> [24,] 0.77380952
#> [25,] 0.66666667
#> attr(,"representation")
#> [1] "graphon"