Network Representation Functions

This is a collection of functions that convert a graph stored as an igraph object into a desired matrix representation among adjacency matrix, graph laplacian, modularity matrix or graphon (edge probability matrix).

repr_adjacency(network, validate = TRUE)

repr_laplacian(network, validate = TRUE)

repr_modularity(network, validate = TRUE)

repr_graphon(network, validate = TRUE)

Arguments

network
An igraph object.

validate
A boolean specifying whether the function should check the class of its input (default: TRUE).

Value

A numeric square matrix giving the desired network representation recorded in the object's class.

Examples

g <- igraph::sample_smallworld(1, 25, 3, 0.05)
repr_adjacency(g)

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0    1    1    1    0    0    0    0    0     0     0     0     0
#>  [2,]    1    0    1    1    1    0    0    0    0     0     0     0     0
#>  [3,]    1    1    0    0    1    1    0    0    0     0     0     0     0
#>  [4,]    1    1    0    0    1    1    1    0    0     0     0     0     0
#>  [5,]    0    1    1    1    0    0    1    1    0     0     1     0     0
#>  [6,]    0    0    1    1    0    0    0    1    1     0     0     0     0
#>  [7,]    0    0    0    1    1    0    0    1    1     1     0     0     0
#>  [8,]    0    0    0    0    1    1    1    0    1     1     1     0     0
#>  [9,]    0    0    0    0    0    1    1    1    0     1     1     1     0
#> [10,]    0    0    0    0    0    0    1    1    1     0     1     1     1
#> [11,]    0    0    0    0    1    0    0    1    1     1     0     1     1
#> [12,]    0    0    0    0    0    0    0    0    1     1     1     0     1
#> [13,]    0    0    0    0    0    0    0    0    0     1     1     1     0
#> [14,]    0    0    0    0    0    0    0    0    0     0     1     1     1
#> [15,]    0    0    0    0    0    0    0    0    0     0     0     1     1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0     1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]    1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]    1    1    1    0    0    0    0    0    0     0     0     0     0
#> [25,]    1    1    1    0    0    0    1    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0     1     1     1
#>  [2,]     0     0     0     0     0     0     0     0     0     0     1     1
#>  [3,]     0     0     0     0     0     0     0     0     0     0     1     1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     1
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]     1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]     1     1     0     0     0     0     0     0     0     0     0     0
#> [13,]     1     1     1     0     0     0     0     0     0     0     0     0
#> [14,]     0     1     1     1     0     0     0     0     0     0     0     0
#> [15,]     1     0     1     1     1     0     0     0     0     0     0     0
#> [16,]     1     1     0     1     1     1     0     0     0     0     0     0
#> [17,]     1     1     1     0     1     1     1     0     0     0     0     0
#> [18,]     0     1     1     1     0     1     1     1     0     0     0     0
#> [19,]     0     0     1     1     1     0     1     1     1     0     0     0
#> [20,]     0     0     0     1     1     1     0     1     1     1     0     0
#> [21,]     0     0     0     0     1     1     1     0     1     1     1     0
#> [22,]     0     0     0     0     0     1     1     1     0     1     1     1
#> [23,]     0     0     0     0     0     0     1     1     1     0     1     1
#> [24,]     0     0     0     0     0     0     0     1     1     1     0     1
#> [25,]     0     0     0     0     0     0     0     0     1     1     1     0
#> attr(,"representation")
#> [1] "adjacency"

repr_laplacian(g)

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    6   -1   -1   -1    0    0    0    0    0     0     0     0     0
#>  [2,]   -1    6   -1   -1   -1    0    0    0    0     0     0     0     0
#>  [3,]   -1   -1    6    0   -1   -1    0    0    0     0     0     0     0
#>  [4,]   -1   -1    0    5   -1   -1   -1    0    0     0     0     0     0
#>  [5,]    0   -1   -1   -1    6    0   -1   -1    0     0    -1     0     0
#>  [6,]    0    0   -1   -1    0    4    0   -1   -1     0     0     0     0
#>  [7,]    0    0    0   -1   -1    0    6   -1   -1    -1     0     0     0
#>  [8,]    0    0    0    0   -1   -1   -1    6   -1    -1    -1     0     0
#>  [9,]    0    0    0    0    0   -1   -1   -1    6    -1    -1    -1     0
#> [10,]    0    0    0    0    0    0   -1   -1   -1     6    -1    -1    -1
#> [11,]    0    0    0    0   -1    0    0   -1   -1    -1     7    -1    -1
#> [12,]    0    0    0    0    0    0    0    0   -1    -1    -1     6    -1
#> [13,]    0    0    0    0    0    0    0    0    0    -1    -1    -1     6
#> [14,]    0    0    0    0    0    0    0    0    0     0    -1    -1    -1
#> [15,]    0    0    0    0    0    0    0    0    0     0     0    -1    -1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0    -1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]   -1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]   -1   -1   -1    0    0    0    0    0    0     0     0     0     0
#> [25,]   -1   -1   -1    0    0    0   -1    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0    -1    -1    -1
#>  [2,]     0     0     0     0     0     0     0     0     0     0    -1    -1
#>  [3,]     0     0     0     0     0     0     0     0     0     0    -1    -1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0    -1
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]    -1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]    -1    -1     0     0     0     0     0     0     0     0     0     0
#> [13,]    -1    -1    -1     0     0     0     0     0     0     0     0     0
#> [14,]     6    -1    -1    -1     0     0     0     0     0     0     0     0
#> [15,]    -1     6    -1    -1    -1     0     0     0     0     0     0     0
#> [16,]    -1    -1     6    -1    -1    -1     0     0     0     0     0     0
#> [17,]    -1    -1    -1     6    -1    -1    -1     0     0     0     0     0
#> [18,]     0    -1    -1    -1     6    -1    -1    -1     0     0     0     0
#> [19,]     0     0    -1    -1    -1     6    -1    -1    -1     0     0     0
#> [20,]     0     0     0    -1    -1    -1     6    -1    -1    -1     0     0
#> [21,]     0     0     0     0    -1    -1    -1     6    -1    -1    -1     0
#> [22,]     0     0     0     0     0    -1    -1    -1     6    -1    -1    -1
#> [23,]     0     0     0     0     0     0    -1    -1    -1     6    -1    -1
#> [24,]     0     0     0     0     0     0     0    -1    -1    -1     7    -1
#> [25,]     0     0     0     0     0     0     0     0    -1    -1    -1     7
#> attr(,"representation")
#> [1] "laplacian"

repr_modularity(g)

#>        [,1]  [,2]  [,3]       [,4]  [,5]       [,6]  [,7]  [,8]  [,9] [,10]
#>  [1,] -0.24  0.76  0.76  0.8000000 -0.24 -0.1600000 -0.24 -0.24 -0.24 -0.24
#>  [2,]  0.76 -0.24  0.76  0.8000000  0.76 -0.1600000 -0.24 -0.24 -0.24 -0.24
#>  [3,]  0.76  0.76 -0.24 -0.2000000  0.76  0.8400000 -0.24 -0.24 -0.24 -0.24
#>  [4,]  0.80  0.80 -0.20 -0.1666667  0.80  0.8666667  0.80 -0.20 -0.20 -0.20
#>  [5,] -0.24  0.76  0.76  0.8000000 -0.24 -0.1600000  0.76  0.76 -0.24 -0.24
#>  [6,] -0.16 -0.16  0.84  0.8666667 -0.16 -0.1066667 -0.16  0.84  0.84 -0.16
#>  [7,] -0.24 -0.24 -0.24  0.8000000  0.76 -0.1600000 -0.24  0.76  0.76  0.76
#>  [8,] -0.24 -0.24 -0.24 -0.2000000  0.76  0.8400000  0.76 -0.24  0.76  0.76
#>  [9,] -0.24 -0.24 -0.24 -0.2000000 -0.24  0.8400000  0.76  0.76 -0.24  0.76
#> [10,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.1600000  0.76  0.76  0.76 -0.24
#> [11,] -0.28 -0.28 -0.28 -0.2333333  0.72 -0.1866667 -0.28  0.72  0.72  0.72
#> [12,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.1600000 -0.24 -0.24  0.76  0.76
#> [13,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.1600000 -0.24 -0.24 -0.24  0.76
#> [14,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.1600000 -0.24 -0.24 -0.24 -0.24
#> [15,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.1600000 -0.24 -0.24 -0.24 -0.24
#> [16,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.1600000 -0.24 -0.24 -0.24 -0.24
#> [17,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.1600000 -0.24 -0.24 -0.24 -0.24
#> [18,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.1600000 -0.24 -0.24 -0.24 -0.24
#> [19,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.1600000 -0.24 -0.24 -0.24 -0.24
#> [20,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.1600000 -0.24 -0.24 -0.24 -0.24
#> [21,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.1600000 -0.24 -0.24 -0.24 -0.24
#> [22,] -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.1600000 -0.24 -0.24 -0.24 -0.24
#> [23,]  0.76 -0.24 -0.24 -0.2000000 -0.24 -0.1600000 -0.24 -0.24 -0.24 -0.24
#> [24,]  0.72  0.72  0.72 -0.2333333 -0.28 -0.1866667 -0.28 -0.28 -0.28 -0.28
#> [25,]  0.72  0.72  0.72 -0.2333333 -0.28 -0.1866667  0.72 -0.28 -0.28 -0.28
#>            [,11] [,12] [,13] [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21]
#>  [1,] -0.2800000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#>  [2,] -0.2800000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#>  [3,] -0.2800000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#>  [4,] -0.2333333 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20
#>  [5,]  0.7200000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#>  [6,] -0.1866667 -0.16 -0.16 -0.16 -0.16 -0.16 -0.16 -0.16 -0.16 -0.16 -0.16
#>  [7,] -0.2800000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#>  [8,]  0.7200000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#>  [9,]  0.7200000  0.76 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#> [10,]  0.7200000  0.76  0.76 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#> [11,] -0.3266667  0.72  0.72  0.72 -0.28 -0.28 -0.28 -0.28 -0.28 -0.28 -0.28
#> [12,]  0.7200000 -0.24  0.76  0.76  0.76 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#> [13,]  0.7200000  0.76 -0.24  0.76  0.76  0.76 -0.24 -0.24 -0.24 -0.24 -0.24
#> [14,]  0.7200000  0.76  0.76 -0.24  0.76  0.76  0.76 -0.24 -0.24 -0.24 -0.24
#> [15,] -0.2800000  0.76  0.76  0.76 -0.24  0.76  0.76  0.76 -0.24 -0.24 -0.24
#> [16,] -0.2800000 -0.24  0.76  0.76  0.76 -0.24  0.76  0.76  0.76 -0.24 -0.24
#> [17,] -0.2800000 -0.24 -0.24  0.76  0.76  0.76 -0.24  0.76  0.76  0.76 -0.24
#> [18,] -0.2800000 -0.24 -0.24 -0.24  0.76  0.76  0.76 -0.24  0.76  0.76  0.76
#> [19,] -0.2800000 -0.24 -0.24 -0.24 -0.24  0.76  0.76  0.76 -0.24  0.76  0.76
#> [20,] -0.2800000 -0.24 -0.24 -0.24 -0.24 -0.24  0.76  0.76  0.76 -0.24  0.76
#> [21,] -0.2800000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24  0.76  0.76  0.76 -0.24
#> [22,] -0.2800000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24  0.76  0.76  0.76
#> [23,] -0.2800000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24  0.76  0.76
#> [24,] -0.3266667 -0.28 -0.28 -0.28 -0.28 -0.28 -0.28 -0.28 -0.28 -0.28  0.72
#> [25,] -0.3266667 -0.28 -0.28 -0.28 -0.28 -0.28 -0.28 -0.28 -0.28 -0.28 -0.28
#>       [,22] [,23]      [,24]      [,25]
#>  [1,] -0.24  0.76  0.7200000  0.7200000
#>  [2,] -0.24 -0.24  0.7200000  0.7200000
#>  [3,] -0.24 -0.24  0.7200000  0.7200000
#>  [4,] -0.20 -0.20 -0.2333333 -0.2333333
#>  [5,] -0.24 -0.24 -0.2800000 -0.2800000
#>  [6,] -0.16 -0.16 -0.1866667 -0.1866667
#>  [7,] -0.24 -0.24 -0.2800000  0.7200000
#>  [8,] -0.24 -0.24 -0.2800000 -0.2800000
#>  [9,] -0.24 -0.24 -0.2800000 -0.2800000
#> [10,] -0.24 -0.24 -0.2800000 -0.2800000
#> [11,] -0.28 -0.28 -0.3266667 -0.3266667
#> [12,] -0.24 -0.24 -0.2800000 -0.2800000
#> [13,] -0.24 -0.24 -0.2800000 -0.2800000
#> [14,] -0.24 -0.24 -0.2800000 -0.2800000
#> [15,] -0.24 -0.24 -0.2800000 -0.2800000
#> [16,] -0.24 -0.24 -0.2800000 -0.2800000
#> [17,] -0.24 -0.24 -0.2800000 -0.2800000
#> [18,] -0.24 -0.24 -0.2800000 -0.2800000
#> [19,]  0.76 -0.24 -0.2800000 -0.2800000
#> [20,]  0.76  0.76 -0.2800000 -0.2800000
#> [21,]  0.76  0.76  0.7200000 -0.2800000
#> [22,] -0.24  0.76  0.7200000  0.7200000
#> [23,]  0.76 -0.24  0.7200000  0.7200000
#> [24,]  0.72  0.72 -0.3266667  0.6733333
#> [25,]  0.72  0.72  0.6733333 -0.3266667
#> attr(,"representation")
#> [1] "modularity"

repr_graphon(g)

#>             [,1]       [,2]      [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.57142857 0.57936508 0.6071429 0.50000000 0.25396825 0.12698413
#>  [2,] 0.57936508 0.44444444 0.4097222 0.56349206 0.38888889 0.22222222
#>  [3,] 0.60714286 0.40972222 0.2500000 0.40178571 0.53472222 0.40972222
#>  [4,] 0.50000000 0.56349206 0.4017857 0.28571429 0.56349206 0.56349206
#>  [5,] 0.25396825 0.38888889 0.5347222 0.56349206 0.44444444 0.33333333
#>  [6,] 0.12698413 0.22222222 0.4097222 0.56349206 0.33333333 0.22222222
#>  [7,] 0.24285714 0.11111111 0.3250000 0.65714286 0.67777778 0.32222222
#>  [8,] 0.18253968 0.33333333 0.1180556 0.25396825 0.50000000 0.50000000
#>  [9,] 0.05555556 0.22222222 0.1805556 0.18253968 0.38888889 0.44444444
#> [10,] 0.00000000 0.12698413 0.1964286 0.21428571 0.30952381 0.36507937
#> [11,] 0.07142857 0.05555556 0.0625000 0.27678571 0.40972222 0.40277778
#> [12,] 0.00000000 0.00000000 0.0000000 0.07142857 0.25396825 0.18253968
#> [13,] 0.00000000 0.00000000 0.0000000 0.00000000 0.05555556 0.19444444
#> [14,] 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000 0.05555556
#> [15,] 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000 0.05555556
#> [16,] 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000
#> [17,] 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000
#> [18,] 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000
#> [19,] 0.14285714 0.05555556 0.0625000 0.00000000 0.00000000 0.00000000
#> [20,] 0.12698413 0.11111111 0.1250000 0.00000000 0.00000000 0.00000000
#> [21,] 0.19841270 0.16666667 0.1805556 0.05555556 0.00000000 0.00000000
#> [22,] 0.25396825 0.11111111 0.1736111 0.18253968 0.05555556 0.00000000
#> [23,] 0.57142857 0.25396825 0.2678571 0.14285714 0.19841270 0.12698413
#> [24,] 0.66964286 0.52777778 0.5625000 0.33928571 0.23611111 0.17361111
#> [25,] 0.69047619 0.66666667 0.6527778 0.30952381 0.33333333 0.27777778
#>            [,7]       [,8]       [,9]     [,10]      [,11]      [,12]
#>  [1,] 0.2428571 0.18253968 0.05555556 0.0000000 0.07142857 0.00000000
#>  [2,] 0.1111111 0.33333333 0.22222222 0.1269841 0.05555556 0.00000000
#>  [3,] 0.3250000 0.11805556 0.18055556 0.1964286 0.06250000 0.00000000
#>  [4,] 0.6571429 0.25396825 0.18253968 0.2142857 0.27678571 0.07142857
#>  [5,] 0.6777778 0.50000000 0.38888889 0.3095238 0.40972222 0.25396825
#>  [6,] 0.3222222 0.50000000 0.44444444 0.3650794 0.40277778 0.18253968
#>  [7,] 0.2000000 0.57777778 0.67777778 0.5857143 0.28750000 0.24285714
#>  [8,] 0.5777778 0.44444444 0.44444444 0.5634921 0.58333333 0.30952381
#>  [9,] 0.6777778 0.44444444 0.44444444 0.5634921 0.70138889 0.50793651
#> [10,] 0.5857143 0.56349206 0.56349206 0.5714286 0.80357143 0.57142857
#> [11,] 0.2875000 0.58333333 0.70138889 0.8035714 0.62500000 0.66964286
#> [12,] 0.2428571 0.30952381 0.50793651 0.5714286 0.66964286 0.57142857
#> [13,] 0.1833333 0.19444444 0.25000000 0.5357143 0.58333333 0.69047619
#> [14,] 0.1714286 0.18253968 0.25396825 0.2857143 0.53571429 0.57142857
#> [15,] 0.0000000 0.11111111 0.16666667 0.2261905 0.29166667 0.53571429
#> [16,] 0.0000000 0.05555556 0.18253968 0.1428571 0.26785714 0.28571429
#> [17,] 0.0000000 0.00000000 0.05555556 0.1428571 0.13392857 0.28571429
#> [18,] 0.0000000 0.00000000 0.00000000 0.0000000 0.13392857 0.14285714
#> [19,] 0.0000000 0.00000000 0.00000000 0.0000000 0.00000000 0.14285714
#> [20,] 0.0000000 0.00000000 0.00000000 0.0000000 0.00000000 0.05555556
#> [21,] 0.0000000 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000
#> [22,] 0.0000000 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000
#> [23,] 0.0000000 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000
#> [24,] 0.1625000 0.05555556 0.00000000 0.0000000 0.00000000 0.00000000
#> [25,] 0.3666667 0.16666667 0.05555556 0.0000000 0.11805556 0.00000000
#>            [,13]      [,14]      [,15]      [,16]      [,17]     [,18]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000
#>  [2,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000
#>  [3,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000
#>  [4,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000
#>  [5,] 0.05555556 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000
#>  [6,] 0.19444444 0.05555556 0.05555556 0.00000000 0.00000000 0.0000000
#>  [7,] 0.18333333 0.17142857 0.00000000 0.00000000 0.00000000 0.0000000
#>  [8,] 0.19444444 0.18253968 0.11111111 0.05555556 0.00000000 0.0000000
#>  [9,] 0.25000000 0.25396825 0.16666667 0.18253968 0.05555556 0.0000000
#> [10,] 0.53571429 0.28571429 0.22619048 0.14285714 0.14285714 0.0000000
#> [11,] 0.58333333 0.53571429 0.29166667 0.26785714 0.13392857 0.1339286
#> [12,] 0.69047619 0.57142857 0.53571429 0.28571429 0.28571429 0.1428571
#> [13,] 0.50000000 0.69047619 0.58333333 0.61904762 0.30952381 0.3095238
#> [14,] 0.69047619 0.57142857 0.69047619 0.57142857 0.57142857 0.2857143
#> [15,] 0.58333333 0.69047619 0.66666667 0.77380952 0.61904762 0.6190476
#> [16,] 0.61904762 0.57142857 0.77380952 0.57142857 0.71428571 0.5714286
#> [17,] 0.30952381 0.57142857 0.61904762 0.71428571 0.57142857 0.7142857
#> [18,] 0.30952381 0.28571429 0.61904762 0.57142857 0.71428571 0.5714286
#> [19,] 0.15476190 0.28571429 0.30952381 0.57142857 0.57142857 0.7142857
#> [20,] 0.19444444 0.23809524 0.33333333 0.36507937 0.56349206 0.6190476
#> [21,] 0.05555556 0.18253968 0.25000000 0.30952381 0.36507937 0.5634921
#> [22,] 0.00000000 0.05555556 0.19444444 0.23809524 0.30952381 0.3650794
#> [23,] 0.00000000 0.00000000 0.00000000 0.14285714 0.14285714 0.2857143
#> [24,] 0.00000000 0.00000000 0.00000000 0.00000000 0.13392857 0.1339286
#> [25,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.1269841
#>            [,19]      [,20]      [,21]      [,22]     [,23]      [,24]
#>  [1,] 0.14285714 0.12698413 0.19841270 0.25396825 0.5714286 0.66964286
#>  [2,] 0.05555556 0.11111111 0.16666667 0.11111111 0.2539683 0.52777778
#>  [3,] 0.06250000 0.12500000 0.18055556 0.17361111 0.2678571 0.56250000
#>  [4,] 0.00000000 0.00000000 0.05555556 0.18253968 0.1428571 0.33928571
#>  [5,] 0.00000000 0.00000000 0.00000000 0.05555556 0.1984127 0.23611111
#>  [6,] 0.00000000 0.00000000 0.00000000 0.00000000 0.1269841 0.17361111
#>  [7,] 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000 0.16250000
#>  [8,] 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000 0.05555556
#>  [9,] 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000
#> [10,] 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000
#> [11,] 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000
#> [12,] 0.14285714 0.05555556 0.00000000 0.00000000 0.0000000 0.00000000
#> [13,] 0.15476190 0.19444444 0.05555556 0.00000000 0.0000000 0.00000000
#> [14,] 0.28571429 0.23809524 0.18253968 0.05555556 0.0000000 0.00000000
#> [15,] 0.30952381 0.33333333 0.25000000 0.19444444 0.0000000 0.00000000
#> [16,] 0.57142857 0.36507937 0.30952381 0.23809524 0.1428571 0.00000000
#> [17,] 0.57142857 0.56349206 0.36507937 0.30952381 0.1428571 0.13392857
#> [18,] 0.71428571 0.61904762 0.56349206 0.36507937 0.2857143 0.13392857
#> [19,] 0.57142857 0.69047619 0.61904762 0.56349206 0.2857143 0.26785714
#> [20,] 0.69047619 0.66666667 0.61111111 0.55555556 0.4523810 0.29166667
#> [21,] 0.61904762 0.61111111 0.55555556 0.50000000 0.5079365 0.47222222
#> [22,] 0.56349206 0.55555556 0.50000000 0.44444444 0.5079365 0.52777778
#> [23,] 0.28571429 0.45238095 0.50793651 0.50793651 0.4285714 0.66964286
#> [24,] 0.26785714 0.29166667 0.47222222 0.52777778 0.6696429 0.75000000
#> [25,] 0.12698413 0.22222222 0.27777778 0.44444444 0.5793651 0.70833333
#>            [,25]
#>  [1,] 0.69047619
#>  [2,] 0.66666667
#>  [3,] 0.65277778
#>  [4,] 0.30952381
#>  [5,] 0.33333333
#>  [6,] 0.27777778
#>  [7,] 0.36666667
#>  [8,] 0.16666667
#>  [9,] 0.05555556
#> [10,] 0.00000000
#> [11,] 0.11805556
#> [12,] 0.00000000
#> [13,] 0.00000000
#> [14,] 0.00000000
#> [15,] 0.00000000
#> [16,] 0.00000000
#> [17,] 0.00000000
#> [18,] 0.12698413
#> [19,] 0.12698413
#> [20,] 0.22222222
#> [21,] 0.27777778
#> [22,] 0.44444444
#> [23,] 0.57936508
#> [24,] 0.70833333
#> [25,] 0.55555556
#> attr(,"representation")
#> [1] "graphon"