Network Representation Functions

This is a collection of functions that convert a graph stored as an igraph object into a desired matrix representation among adjacency matrix, graph laplacian, modularity matrix or graphon (edge probability matrix).

repr_adjacency(network, validate = TRUE)

repr_laplacian(network, validate = TRUE)

repr_modularity(network, validate = TRUE)

repr_graphon(network, validate = TRUE)

Arguments

network

An igraph object.

validate

A boolean specifying whether the function should check the class of its input (default: TRUE).

Value

A numeric square matrix giving the desired network representation recorded in the object's class.

Examples

g <- igraph::sample_smallworld(1, 25, 3, 0.05)
repr_adjacency(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0    1    0    0    0    0    0    0    0     0     0     0     0
#>  [2,]    1    0    1    1    1    0    0    0    0     0     0     0     0
#>  [3,]    0    1    0    0    1    1    0    0    0     0     0     0     0
#>  [4,]    0    1    0    0    1    1    1    0    0     1     0     0     0
#>  [5,]    0    1    1    1    0    1    1    1    0     0     0     0     0
#>  [6,]    0    0    1    1    1    0    1    1    1     0     0     0     0
#>  [7,]    0    0    0    1    1    1    0    1    1     0     0     0     0
#>  [8,]    0    0    0    0    1    1    1    0    1     0     1     0     0
#>  [9,]    0    0    0    0    0    1    1    1    0     1     0     1     0
#> [10,]    0    0    0    1    0    0    0    0    1     0     1     1     1
#> [11,]    0    0    0    0    0    0    0    1    0     1     0     1     1
#> [12,]    0    0    0    0    0    0    0    0    1     1     1     0     1
#> [13,]    0    0    0    0    0    0    0    0    0     1     1     1     0
#> [14,]    0    0    0    0    0    0    1    0    0     0     1     0     1
#> [15,]    0    0    1    0    0    0    0    0    1     0     0     0     1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0     1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     1     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    1    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     1     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]    1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]    1    1    1    0    0    0    0    0    0     0     0     0     0
#> [25,]    1    1    1    0    0    0    0    0    0     0     1     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0     1     1     1
#>  [2,]     0     0     0     0     0     0     0     0     0     0     1     1
#>  [3,]     0     1     0     0     0     0     0     0     0     0     1     1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     1     0     0     0     0     0     1     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     1     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]     1     0     0     0     1     0     0     1     0     0     0     1
#> [12,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [13,]     1     1     1     0     0     0     0     0     0     0     0     0
#> [14,]     0     1     1     1     0     0     0     0     1     0     0     0
#> [15,]     1     0     1     1     1     0     0     0     0     1     0     0
#> [16,]     1     1     0     1     1     1     0     0     0     0     0     0
#> [17,]     1     1     1     0     1     1     1     0     0     0     0     0
#> [18,]     0     1     1     1     0     1     0     0     0     0     0     0
#> [19,]     0     0     1     1     1     0     1     1     1     0     0     0
#> [20,]     0     0     0     1     0     1     0     1     1     1     0     0
#> [21,]     0     0     0     0     0     1     1     0     0     1     1     0
#> [22,]     1     0     0     0     0     1     1     0     0     1     1     1
#> [23,]     0     1     0     0     0     0     1     1     1     0     1     1
#> [24,]     0     0     0     0     0     0     0     1     1     1     0     1
#> [25,]     0     0     0     0     0     0     0     0     1     1     1     0
#> attr(,"representation")
#> [1] "adjacency"
repr_laplacian(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    4   -1    0    0    0    0    0    0    0     0     0     0     0
#>  [2,]   -1    6   -1   -1   -1    0    0    0    0     0     0     0     0
#>  [3,]    0   -1    6    0   -1   -1    0    0    0     0     0     0     0
#>  [4,]    0   -1    0    5   -1   -1   -1    0    0    -1     0     0     0
#>  [5,]    0   -1   -1   -1    6   -1   -1   -1    0     0     0     0     0
#>  [6,]    0    0   -1   -1   -1    6   -1   -1   -1     0     0     0     0
#>  [7,]    0    0    0   -1   -1   -1    7   -1   -1     0     0     0     0
#>  [8,]    0    0    0    0   -1   -1   -1    5   -1     0    -1     0     0
#>  [9,]    0    0    0    0    0   -1   -1   -1    6    -1     0    -1     0
#> [10,]    0    0    0   -1    0    0    0    0   -1     5    -1    -1    -1
#> [11,]    0    0    0    0    0    0    0   -1    0    -1     8    -1    -1
#> [12,]    0    0    0    0    0    0    0    0   -1    -1    -1     4    -1
#> [13,]    0    0    0    0    0    0    0    0    0    -1    -1    -1     6
#> [14,]    0    0    0    0    0    0   -1    0    0     0    -1     0    -1
#> [15,]    0    0   -1    0    0    0    0    0   -1     0     0     0    -1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0    -1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0    -1     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0   -1    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0    -1     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]   -1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]   -1   -1   -1    0    0    0    0    0    0     0     0     0     0
#> [25,]   -1   -1   -1    0    0    0    0    0    0     0    -1     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0    -1    -1    -1
#>  [2,]     0     0     0     0     0     0     0     0     0     0    -1    -1
#>  [3,]     0    -1     0     0     0     0     0     0     0     0    -1    -1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]    -1     0     0     0     0     0    -1     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0    -1     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]    -1     0     0     0    -1     0     0    -1     0     0     0    -1
#> [12,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [13,]    -1    -1    -1     0     0     0     0     0     0     0     0     0
#> [14,]     7    -1    -1    -1     0     0     0     0    -1     0     0     0
#> [15,]    -1     8    -1    -1    -1     0     0     0     0    -1     0     0
#> [16,]    -1    -1     6    -1    -1    -1     0     0     0     0     0     0
#> [17,]    -1    -1    -1     6    -1    -1    -1     0     0     0     0     0
#> [18,]     0    -1    -1    -1     5    -1     0     0     0     0     0     0
#> [19,]     0     0    -1    -1    -1     6    -1    -1    -1     0     0     0
#> [20,]     0     0     0    -1     0    -1     6    -1    -1    -1     0     0
#> [21,]     0     0     0     0     0    -1    -1     5     0    -1    -1     0
#> [22,]    -1     0     0     0     0    -1    -1     0     6    -1    -1    -1
#> [23,]     0    -1     0     0     0     0    -1    -1    -1     7    -1    -1
#> [24,]     0     0     0     0     0     0     0    -1    -1    -1     7    -1
#> [25,]     0     0     0     0     0     0     0     0    -1    -1    -1     7
#> attr(,"representation")
#> [1] "laplacian"
repr_modularity(g)
#>             [,1]  [,2]  [,3]       [,4]  [,5]  [,6]       [,7]       [,8]  [,9]
#>  [1,] -0.1066667  0.84 -0.16 -0.1333333 -0.16 -0.16 -0.1866667 -0.1333333 -0.16
#>  [2,]  0.8400000 -0.24  0.76  0.8000000  0.76 -0.24 -0.2800000 -0.2000000 -0.24
#>  [3,] -0.1600000  0.76 -0.24 -0.2000000  0.76  0.76 -0.2800000 -0.2000000 -0.24
#>  [4,] -0.1333333  0.80 -0.20 -0.1666667  0.80  0.80  0.7666667 -0.1666667 -0.20
#>  [5,] -0.1600000  0.76  0.76  0.8000000 -0.24  0.76  0.7200000  0.8000000 -0.24
#>  [6,] -0.1600000 -0.24  0.76  0.8000000  0.76 -0.24  0.7200000  0.8000000  0.76
#>  [7,] -0.1866667 -0.28 -0.28  0.7666667  0.72  0.72 -0.3266667  0.7666667  0.72
#>  [8,] -0.1333333 -0.20 -0.20 -0.1666667  0.80  0.80  0.7666667 -0.1666667  0.80
#>  [9,] -0.1600000 -0.24 -0.24 -0.2000000 -0.24  0.76  0.7200000  0.8000000 -0.24
#> [10,] -0.1333333 -0.20 -0.20  0.8333333 -0.20 -0.20 -0.2333333 -0.1666667  0.80
#> [11,] -0.2133333 -0.32 -0.32 -0.2666667 -0.32 -0.32 -0.3733333  0.7333333 -0.32
#> [12,] -0.1066667 -0.16 -0.16 -0.1333333 -0.16 -0.16 -0.1866667 -0.1333333  0.84
#> [13,] -0.1600000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000 -0.2000000 -0.24
#> [14,] -0.1866667 -0.28 -0.28 -0.2333333 -0.28 -0.28  0.6733333 -0.2333333 -0.28
#> [15,] -0.2133333 -0.32  0.68 -0.2666667 -0.32 -0.32 -0.3733333 -0.2666667  0.68
#> [16,] -0.1600000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000 -0.2000000 -0.24
#> [17,] -0.1600000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000 -0.2000000 -0.24
#> [18,] -0.1333333 -0.20 -0.20 -0.1666667 -0.20 -0.20 -0.2333333 -0.1666667 -0.20
#> [19,] -0.1600000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000 -0.2000000 -0.24
#> [20,] -0.1600000 -0.24 -0.24 -0.2000000 -0.24 -0.24  0.7200000 -0.2000000 -0.24
#> [21,] -0.1333333 -0.20 -0.20 -0.1666667 -0.20 -0.20 -0.2333333 -0.1666667 -0.20
#> [22,] -0.1600000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000 -0.2000000 -0.24
#> [23,]  0.8133333 -0.28 -0.28 -0.2333333 -0.28 -0.28 -0.3266667 -0.2333333 -0.28
#> [24,]  0.8133333  0.72  0.72 -0.2333333 -0.28 -0.28 -0.3266667 -0.2333333 -0.28
#> [25,]  0.8133333  0.72  0.72 -0.2333333 -0.28 -0.28 -0.3266667 -0.2333333 -0.28
#>            [,10]      [,11]      [,12] [,13]      [,14]      [,15] [,16] [,17]
#>  [1,] -0.1333333 -0.2133333 -0.1066667 -0.16 -0.1866667 -0.2133333 -0.16 -0.16
#>  [2,] -0.2000000 -0.3200000 -0.1600000 -0.24 -0.2800000 -0.3200000 -0.24 -0.24
#>  [3,] -0.2000000 -0.3200000 -0.1600000 -0.24 -0.2800000  0.6800000 -0.24 -0.24
#>  [4,]  0.8333333 -0.2666667 -0.1333333 -0.20 -0.2333333 -0.2666667 -0.20 -0.20
#>  [5,] -0.2000000 -0.3200000 -0.1600000 -0.24 -0.2800000 -0.3200000 -0.24 -0.24
#>  [6,] -0.2000000 -0.3200000 -0.1600000 -0.24 -0.2800000 -0.3200000 -0.24 -0.24
#>  [7,] -0.2333333 -0.3733333 -0.1866667 -0.28  0.6733333 -0.3733333 -0.28 -0.28
#>  [8,] -0.1666667  0.7333333 -0.1333333 -0.20 -0.2333333 -0.2666667 -0.20 -0.20
#>  [9,]  0.8000000 -0.3200000  0.8400000 -0.24 -0.2800000  0.6800000 -0.24 -0.24
#> [10,] -0.1666667  0.7333333  0.8666667  0.80 -0.2333333 -0.2666667 -0.20 -0.20
#> [11,]  0.7333333 -0.4266667  0.7866667  0.68  0.6266667 -0.4266667 -0.32 -0.32
#> [12,]  0.8666667  0.7866667 -0.1066667  0.84 -0.1866667 -0.2133333 -0.16 -0.16
#> [13,]  0.8000000  0.6800000  0.8400000 -0.24  0.7200000  0.6800000  0.76 -0.24
#> [14,] -0.2333333  0.6266667 -0.1866667  0.72 -0.3266667  0.6266667  0.72  0.72
#> [15,] -0.2666667 -0.4266667 -0.2133333  0.68  0.6266667 -0.4266667  0.68  0.68
#> [16,] -0.2000000 -0.3200000 -0.1600000  0.76  0.7200000  0.6800000 -0.24  0.76
#> [17,] -0.2000000 -0.3200000 -0.1600000 -0.24  0.7200000  0.6800000  0.76 -0.24
#> [18,] -0.1666667  0.7333333 -0.1333333 -0.20 -0.2333333  0.7333333  0.80  0.80
#> [19,] -0.2000000 -0.3200000 -0.1600000 -0.24 -0.2800000 -0.3200000  0.76  0.76
#> [20,] -0.2000000 -0.3200000 -0.1600000 -0.24 -0.2800000 -0.3200000 -0.24  0.76
#> [21,] -0.1666667  0.7333333 -0.1333333 -0.20 -0.2333333 -0.2666667 -0.20 -0.20
#> [22,] -0.2000000 -0.3200000 -0.1600000 -0.24  0.7200000 -0.3200000 -0.24 -0.24
#> [23,] -0.2333333 -0.3733333 -0.1866667 -0.28 -0.3266667  0.6266667 -0.28 -0.28
#> [24,] -0.2333333 -0.3733333 -0.1866667 -0.28 -0.3266667 -0.3733333 -0.28 -0.28
#> [25,] -0.2333333  0.6266667 -0.1866667 -0.28 -0.3266667 -0.3733333 -0.28 -0.28
#>            [,18] [,19] [,20]      [,21] [,22]      [,23]      [,24]      [,25]
#>  [1,] -0.1333333 -0.16 -0.16 -0.1333333 -0.16  0.8133333  0.8133333  0.8133333
#>  [2,] -0.2000000 -0.24 -0.24 -0.2000000 -0.24 -0.2800000  0.7200000  0.7200000
#>  [3,] -0.2000000 -0.24 -0.24 -0.2000000 -0.24 -0.2800000  0.7200000  0.7200000
#>  [4,] -0.1666667 -0.20 -0.20 -0.1666667 -0.20 -0.2333333 -0.2333333 -0.2333333
#>  [5,] -0.2000000 -0.24 -0.24 -0.2000000 -0.24 -0.2800000 -0.2800000 -0.2800000
#>  [6,] -0.2000000 -0.24 -0.24 -0.2000000 -0.24 -0.2800000 -0.2800000 -0.2800000
#>  [7,] -0.2333333 -0.28  0.72 -0.2333333 -0.28 -0.3266667 -0.3266667 -0.3266667
#>  [8,] -0.1666667 -0.20 -0.20 -0.1666667 -0.20 -0.2333333 -0.2333333 -0.2333333
#>  [9,] -0.2000000 -0.24 -0.24 -0.2000000 -0.24 -0.2800000 -0.2800000 -0.2800000
#> [10,] -0.1666667 -0.20 -0.20 -0.1666667 -0.20 -0.2333333 -0.2333333 -0.2333333
#> [11,]  0.7333333 -0.32 -0.32  0.7333333 -0.32 -0.3733333 -0.3733333  0.6266667
#> [12,] -0.1333333 -0.16 -0.16 -0.1333333 -0.16 -0.1866667 -0.1866667 -0.1866667
#> [13,] -0.2000000 -0.24 -0.24 -0.2000000 -0.24 -0.2800000 -0.2800000 -0.2800000
#> [14,] -0.2333333 -0.28 -0.28 -0.2333333  0.72 -0.3266667 -0.3266667 -0.3266667
#> [15,]  0.7333333 -0.32 -0.32 -0.2666667 -0.32  0.6266667 -0.3733333 -0.3733333
#> [16,]  0.8000000  0.76 -0.24 -0.2000000 -0.24 -0.2800000 -0.2800000 -0.2800000
#> [17,]  0.8000000  0.76  0.76 -0.2000000 -0.24 -0.2800000 -0.2800000 -0.2800000
#> [18,] -0.1666667  0.80 -0.20 -0.1666667 -0.20 -0.2333333 -0.2333333 -0.2333333
#> [19,]  0.8000000 -0.24  0.76  0.8000000  0.76 -0.2800000 -0.2800000 -0.2800000
#> [20,] -0.2000000  0.76 -0.24  0.8000000  0.76  0.7200000 -0.2800000 -0.2800000
#> [21,] -0.1666667  0.80  0.80 -0.1666667 -0.20  0.7666667  0.7666667 -0.2333333
#> [22,] -0.2000000  0.76  0.76 -0.2000000 -0.24  0.7200000  0.7200000  0.7200000
#> [23,] -0.2333333 -0.28  0.72  0.7666667  0.72 -0.3266667  0.6733333  0.6733333
#> [24,] -0.2333333 -0.28 -0.28  0.7666667  0.72  0.6733333 -0.3266667  0.6733333
#> [25,] -0.2333333 -0.28 -0.28 -0.2333333  0.72  0.6733333  0.6733333 -0.3266667
#> attr(,"representation")
#> [1] "modularity"
repr_graphon(g)
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.33333333 0.50000000 0.27777778 0.05555556 0.22916667 0.17361111
#>  [2,] 0.50000000 0.55555556 0.50000000 0.41666667 0.35416667 0.23611111
#>  [3,] 0.27777778 0.50000000 0.22222222 0.23611111 0.40972222 0.46527778
#>  [4,] 0.05555556 0.41666667 0.23611111 0.25000000 0.50000000 0.62500000
#>  [5,] 0.22916667 0.35416667 0.40972222 0.50000000 0.75000000 0.68750000
#>  [6,] 0.17361111 0.23611111 0.46527778 0.62500000 0.68750000 0.62500000
#>  [7,] 0.05555556 0.23611111 0.47222222 0.68750000 0.68750000 0.68750000
#>  [8,] 0.00000000 0.29861111 0.23611111 0.37500000 0.50000000 0.56250000
#>  [9,] 0.05555556 0.26984127 0.19841270 0.33035714 0.47321429 0.60714286
#> [10,] 0.11111111 0.00000000 0.18253968 0.33035714 0.26785714 0.19642857
#> [11,] 0.16666667 0.05555556 0.15555556 0.12500000 0.06250000 0.22500000
#> [12,] 0.00000000 0.12500000 0.05555556 0.12500000 0.18750000 0.25000000
#> [13,] 0.05555556 0.00000000 0.00000000 0.12500000 0.00000000 0.12500000
#> [14,] 0.05555556 0.05555556 0.05555556 0.00000000 0.06250000 0.06250000
#> [15,] 0.05555556 0.11111111 0.25000000 0.12500000 0.12500000 0.06250000
#> [16,] 0.00000000 0.00000000 0.05555556 0.00000000 0.00000000 0.00000000
#> [17,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [18,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [19,] 0.18253968 0.05555556 0.05555556 0.00000000 0.00000000 0.00000000
#> [20,] 0.17361111 0.17361111 0.17361111 0.00000000 0.06250000 0.06250000
#> [21,] 0.05555556 0.16666667 0.05555556 0.05555556 0.05555556 0.05555556
#> [22,] 0.22222222 0.27777778 0.11111111 0.05555556 0.11111111 0.05555556
#> [23,] 0.56349206 0.43650794 0.32539683 0.13392857 0.14285714 0.07142857
#> [24,] 0.66666667 0.58333333 0.52777778 0.20833333 0.20833333 0.06250000
#> [25,] 0.61111111 0.58333333 0.52777778 0.20833333 0.20833333 0.06250000
#>             [,7]      [,8]       [,9]      [,10]      [,11]      [,12]
#>  [1,] 0.05555556 0.0000000 0.05555556 0.11111111 0.16666667 0.00000000
#>  [2,] 0.23611111 0.2986111 0.26984127 0.00000000 0.05555556 0.12500000
#>  [3,] 0.47222222 0.2361111 0.19841270 0.18253968 0.15555556 0.05555556
#>  [4,] 0.68750000 0.3750000 0.33035714 0.33035714 0.12500000 0.12500000
#>  [5,] 0.68750000 0.5000000 0.47321429 0.26785714 0.06250000 0.18750000
#>  [6,] 0.68750000 0.5625000 0.60714286 0.19642857 0.22500000 0.25000000
#>  [7,] 0.75000000 0.6875000 0.73214286 0.14285714 0.22500000 0.12500000
#>  [8,] 0.68750000 0.3750000 0.46428571 0.25892857 0.38750000 0.12500000
#>  [9,] 0.73214286 0.4642857 0.28571429 0.50000000 0.24285714 0.33035714
#> [10,] 0.14285714 0.2589286 0.50000000 0.28571429 0.72857143 0.33035714
#> [11,] 0.22500000 0.3875000 0.24285714 0.72857143 0.20000000 0.67500000
#> [12,] 0.12500000 0.1250000 0.33035714 0.33035714 0.67500000 0.25000000
#> [13,] 0.12500000 0.1875000 0.12500000 0.33035714 0.61250000 0.31250000
#> [14,] 0.29166667 0.0000000 0.07142857 0.15476190 0.65000000 0.14583333
#> [15,] 0.08333333 0.2083333 0.29761905 0.22619048 0.38333333 0.20833333
#> [16,] 0.11111111 0.0000000 0.12698413 0.19841270 0.36666667 0.18055556
#> [17,] 0.00000000 0.0000000 0.00000000 0.14285714 0.41428571 0.13392857
#> [18,] 0.12500000 0.0000000 0.00000000 0.06250000 0.55000000 0.06250000
#> [19,] 0.14285714 0.0000000 0.00000000 0.14285714 0.24285714 0.12500000
#> [20,] 0.25000000 0.0000000 0.00000000 0.07142857 0.18750000 0.06250000
#> [21,] 0.11111111 0.0000000 0.16666667 0.05555556 0.37777778 0.05555556
#> [22,] 0.05555556 0.0000000 0.00000000 0.00000000 0.05555556 0.00000000
#> [23,] 0.00000000 0.0000000 0.00000000 0.07142857 0.17142857 0.12500000
#> [24,] 0.08333333 0.0625000 0.07142857 0.07142857 0.08333333 0.12500000
#> [25,] 0.00000000 0.0625000 0.07142857 0.00000000 0.26666667 0.06250000
#>            [,13]      [,14]      [,15]      [,16]      [,17]     [,18]
#>  [1,] 0.05555556 0.05555556 0.05555556 0.00000000 0.00000000 0.0000000
#>  [2,] 0.00000000 0.05555556 0.11111111 0.00000000 0.00000000 0.0000000
#>  [3,] 0.00000000 0.05555556 0.25000000 0.05555556 0.00000000 0.0000000
#>  [4,] 0.12500000 0.00000000 0.12500000 0.00000000 0.00000000 0.0000000
#>  [5,] 0.00000000 0.06250000 0.12500000 0.00000000 0.00000000 0.0000000
#>  [6,] 0.12500000 0.06250000 0.06250000 0.00000000 0.00000000 0.0000000
#>  [7,] 0.12500000 0.29166667 0.08333333 0.11111111 0.00000000 0.1250000
#>  [8,] 0.18750000 0.00000000 0.20833333 0.00000000 0.00000000 0.0000000
#>  [9,] 0.12500000 0.07142857 0.29761905 0.12698413 0.00000000 0.0000000
#> [10,] 0.33035714 0.15476190 0.22619048 0.19841270 0.14285714 0.0625000
#> [11,] 0.61250000 0.65000000 0.38333333 0.36666667 0.41428571 0.5500000
#> [12,] 0.31250000 0.14583333 0.20833333 0.18055556 0.13392857 0.0625000
#> [13,] 0.50000000 0.35416667 0.70833333 0.41666667 0.25892857 0.2500000
#> [14,] 0.35416667 0.33333333 0.66666667 0.61111111 0.70238095 0.3541667
#> [15,] 0.70833333 0.66666667 0.50000000 0.61111111 0.61904762 0.7291667
#> [16,] 0.41666667 0.61111111 0.61111111 0.66666667 0.61904762 0.5347222
#> [17,] 0.25892857 0.70238095 0.61904762 0.61904762 0.42857143 0.5267857
#> [18,] 0.25000000 0.35416667 0.72916667 0.53472222 0.52678571 0.3750000
#> [19,] 0.33035714 0.39285714 0.52380952 0.56349206 0.71428571 0.5267857
#> [20,] 0.18750000 0.20833333 0.35416667 0.35416667 0.59821429 0.3125000
#> [21,] 0.11111111 0.27777778 0.27777778 0.22222222 0.18253968 0.1805556
#> [22,] 0.11805556 0.41666667 0.38888889 0.22222222 0.18253968 0.2986111
#> [23,] 0.00000000 0.15476190 0.22619048 0.23809524 0.21428571 0.1964286
#> [24,] 0.00000000 0.00000000 0.08333333 0.05555556 0.22619048 0.0625000
#> [25,] 0.00000000 0.00000000 0.16666667 0.05555556 0.07142857 0.0000000
#>            [,19]      [,20]      [,21]      [,22]      [,23]      [,24]
#>  [1,] 0.18253968 0.17361111 0.05555556 0.22222222 0.56349206 0.66666667
#>  [2,] 0.05555556 0.17361111 0.16666667 0.27777778 0.43650794 0.58333333
#>  [3,] 0.05555556 0.17361111 0.05555556 0.11111111 0.32539683 0.52777778
#>  [4,] 0.00000000 0.00000000 0.05555556 0.05555556 0.13392857 0.20833333
#>  [5,] 0.00000000 0.06250000 0.05555556 0.11111111 0.14285714 0.20833333
#>  [6,] 0.00000000 0.06250000 0.05555556 0.05555556 0.07142857 0.06250000
#>  [7,] 0.14285714 0.25000000 0.11111111 0.05555556 0.00000000 0.08333333
#>  [8,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.06250000
#>  [9,] 0.00000000 0.00000000 0.16666667 0.00000000 0.00000000 0.07142857
#> [10,] 0.14285714 0.07142857 0.05555556 0.00000000 0.07142857 0.07142857
#> [11,] 0.24285714 0.18750000 0.37777778 0.05555556 0.17142857 0.08333333
#> [12,] 0.12500000 0.06250000 0.05555556 0.00000000 0.12500000 0.12500000
#> [13,] 0.33035714 0.18750000 0.11111111 0.11805556 0.00000000 0.00000000
#> [14,] 0.39285714 0.20833333 0.27777778 0.41666667 0.15476190 0.00000000
#> [15,] 0.52380952 0.35416667 0.27777778 0.38888889 0.22619048 0.08333333
#> [16,] 0.56349206 0.35416667 0.22222222 0.22222222 0.23809524 0.05555556
#> [17,] 0.71428571 0.59821429 0.18253968 0.18253968 0.21428571 0.22619048
#> [18,] 0.52678571 0.31250000 0.18055556 0.29861111 0.19642857 0.06250000
#> [19,] 0.57142857 0.52678571 0.49206349 0.56349206 0.14285714 0.15476190
#> [20,] 0.52678571 0.37500000 0.35416667 0.47222222 0.46428571 0.20833333
#> [21,] 0.49206349 0.35416667 0.11111111 0.16666667 0.43650794 0.41666667
#> [22,] 0.56349206 0.47222222 0.16666667 0.22222222 0.50793651 0.66666667
#> [23,] 0.14285714 0.46428571 0.43650794 0.50793651 0.42857143 0.69047619
#> [24,] 0.15476190 0.20833333 0.41666667 0.66666667 0.69047619 0.66666667
#> [25,] 0.15476190 0.29166667 0.27777778 0.52777778 0.69047619 0.75000000
#>            [,25]
#>  [1,] 0.61111111
#>  [2,] 0.58333333
#>  [3,] 0.52777778
#>  [4,] 0.20833333
#>  [5,] 0.20833333
#>  [6,] 0.06250000
#>  [7,] 0.00000000
#>  [8,] 0.06250000
#>  [9,] 0.07142857
#> [10,] 0.00000000
#> [11,] 0.26666667
#> [12,] 0.06250000
#> [13,] 0.00000000
#> [14,] 0.00000000
#> [15,] 0.16666667
#> [16,] 0.05555556
#> [17,] 0.07142857
#> [18,] 0.00000000
#> [19,] 0.15476190
#> [20,] 0.29166667
#> [21,] 0.27777778
#> [22,] 0.52777778
#> [23,] 0.69047619
#> [24,] 0.75000000
#> [25,] 0.66666667
#> attr(,"representation")
#> [1] "graphon"