Network Representation Functions

This is a collection of functions that convert a graph stored as an igraph object into a desired matrix representation among adjacency matrix, graph laplacian, modularity matrix or graphon (edge probability matrix).

repr_adjacency(network, validate = TRUE)

repr_laplacian(network, validate = TRUE)

repr_modularity(network, validate = TRUE)

repr_graphon(network, validate = TRUE)

Arguments

network

An igraph object.

validate

A boolean specifying whether the function should check the class of its input (default: TRUE).

Value

A numeric square matrix giving the desired network representation recorded in the object's class.

Examples

g <- igraph::sample_smallworld(1, 25, 3, 0.05)
repr_adjacency(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0    1    1    1    0    0    0    0    0     0     0     0     0
#>  [2,]    1    0    1    1    1    0    0    0    0     0     0     0     0
#>  [3,]    1    1    0    0    1    1    0    0    0     0     0     0     0
#>  [4,]    1    1    0    0    0    1    1    0    0     0     0     1     0
#>  [5,]    0    1    1    0    0    1    0    1    0     0     0     0     0
#>  [6,]    0    0    1    1    1    0    1    1    1     0     0     0     0
#>  [7,]    0    0    0    1    0    1    0    1    1     1     0     1     0
#>  [8,]    0    0    0    0    1    1    1    0    1     1     1     0     0
#>  [9,]    0    0    0    0    0    1    1    1    0     1     1     0     0
#> [10,]    0    0    0    0    0    0    1    1    1     0     1     0     0
#> [11,]    0    0    0    0    0    0    0    1    1     1     0     1     1
#> [12,]    0    0    0    1    0    0    1    0    0     0     1     0     1
#> [13,]    0    0    0    0    0    0    0    0    0     0     1     1     0
#> [14,]    0    0    0    0    0    0    0    0    0     0     1     1     1
#> [15,]    1    0    0    0    0    0    0    0    0     0     0     1     1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0     1
#> [17,]    0    0    1    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    1    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]    1    1    0    0    0    0    0    0    0     1     0     0     0
#> [25,]    1    1    1    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     1     0     0     0     0     0     0     0     0     1     1
#>  [2,]     0     0     0     0     0     0     0     0     0     0     1     1
#>  [3,]     0     0     0     1     0     0     0     0     0     0     0     1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     1     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     1     0
#> [11,]     1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]     1     1     0     0     0     0     0     0     0     0     0     0
#> [13,]     1     1     1     0     0     0     0     0     0     0     0     0
#> [14,]     0     0     1     1     0     0     0     0     0     0     0     0
#> [15,]     0     0     1     0     1     0     1     1     0     0     0     0
#> [16,]     1     1     0     1     1     1     0     0     0     0     0     0
#> [17,]     1     0     1     0     1     1     1     0     0     0     0     0
#> [18,]     0     1     1     1     0     1     1     1     0     0     0     0
#> [19,]     0     0     1     1     1     0     1     1     1     1     0     0
#> [20,]     0     1     0     1     1     1     0     1     1     1     0     0
#> [21,]     0     1     0     0     1     1     1     0     1     1     1     0
#> [22,]     0     0     0     0     0     1     1     1     0     1     1     1
#> [23,]     0     0     0     0     0     1     1     1     1     0     1     1
#> [24,]     0     0     0     0     0     0     0     1     1     1     0     1
#> [25,]     0     0     0     0     0     0     0     0     1     1     1     0
#> attr(,"representation")
#> [1] "adjacency"
repr_laplacian(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    6   -1   -1   -1    0    0    0    0    0     0     0     0     0
#>  [2,]   -1    6   -1   -1   -1    0    0    0    0     0     0     0     0
#>  [3,]   -1   -1    6    0   -1   -1    0    0    0     0     0     0     0
#>  [4,]   -1   -1    0    5    0   -1   -1    0    0     0     0    -1     0
#>  [5,]    0   -1   -1    0    5   -1    0   -1    0     0     0     0     0
#>  [6,]    0    0   -1   -1   -1    6   -1   -1   -1     0     0     0     0
#>  [7,]    0    0    0   -1    0   -1    6   -1   -1    -1     0    -1     0
#>  [8,]    0    0    0    0   -1   -1   -1    6   -1    -1    -1     0     0
#>  [9,]    0    0    0    0    0   -1   -1   -1    5    -1    -1     0     0
#> [10,]    0    0    0    0    0    0   -1   -1   -1     5    -1     0     0
#> [11,]    0    0    0    0    0    0    0   -1   -1    -1     6    -1    -1
#> [12,]    0    0    0   -1    0    0   -1    0    0     0    -1     6    -1
#> [13,]    0    0    0    0    0    0    0    0    0     0    -1    -1     5
#> [14,]    0    0    0    0    0    0    0    0    0     0    -1    -1    -1
#> [15,]   -1    0    0    0    0    0    0    0    0     0     0    -1    -1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0    -1
#> [17,]    0    0   -1    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0   -1    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]   -1   -1    0    0    0    0    0    0    0    -1     0     0     0
#> [25,]   -1   -1   -1    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0    -1     0     0     0     0     0     0     0     0    -1    -1
#>  [2,]     0     0     0     0     0     0     0     0     0     0    -1    -1
#>  [3,]     0     0     0    -1     0     0     0     0     0     0     0    -1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0    -1     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0    -1     0
#> [11,]    -1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]    -1    -1     0     0     0     0     0     0     0     0     0     0
#> [13,]    -1    -1    -1     0     0     0     0     0     0     0     0     0
#> [14,]     5     0    -1    -1     0     0     0     0     0     0     0     0
#> [15,]     0     7    -1     0    -1     0    -1    -1     0     0     0     0
#> [16,]    -1    -1     6    -1    -1    -1     0     0     0     0     0     0
#> [17,]    -1     0    -1     6    -1    -1    -1     0     0     0     0     0
#> [18,]     0    -1    -1    -1     6    -1    -1    -1     0     0     0     0
#> [19,]     0     0    -1    -1    -1     8    -1    -1    -1    -1     0     0
#> [20,]     0    -1     0    -1    -1    -1     7    -1    -1    -1     0     0
#> [21,]     0    -1     0     0    -1    -1    -1     7    -1    -1    -1     0
#> [22,]     0     0     0     0     0    -1    -1    -1     6    -1    -1    -1
#> [23,]     0     0     0     0     0    -1    -1    -1    -1     6    -1    -1
#> [24,]     0     0     0     0     0     0     0    -1    -1    -1     7    -1
#> [25,]     0     0     0     0     0     0     0     0    -1    -1    -1     6
#> attr(,"representation")
#> [1] "laplacian"
repr_modularity(g)
#>        [,1]  [,2]  [,3]       [,4]       [,5]  [,6]  [,7]  [,8]       [,9]
#>  [1,] -0.24  0.76  0.76  0.8000000 -0.2000000 -0.24 -0.24 -0.24 -0.2000000
#>  [2,]  0.76 -0.24  0.76  0.8000000  0.8000000 -0.24 -0.24 -0.24 -0.2000000
#>  [3,]  0.76  0.76 -0.24 -0.2000000  0.8000000  0.76 -0.24 -0.24 -0.2000000
#>  [4,]  0.80  0.80 -0.20 -0.1666667 -0.1666667  0.80  0.80 -0.20 -0.1666667
#>  [5,] -0.20  0.80  0.80 -0.1666667 -0.1666667  0.80 -0.20  0.80 -0.1666667
#>  [6,] -0.24 -0.24  0.76  0.8000000  0.8000000 -0.24  0.76  0.76  0.8000000
#>  [7,] -0.24 -0.24 -0.24  0.8000000 -0.2000000  0.76 -0.24  0.76  0.8000000
#>  [8,] -0.24 -0.24 -0.24 -0.2000000  0.8000000  0.76  0.76 -0.24  0.8000000
#>  [9,] -0.20 -0.20 -0.20 -0.1666667 -0.1666667  0.80  0.80  0.80 -0.1666667
#> [10,] -0.20 -0.20 -0.20 -0.1666667 -0.1666667 -0.20  0.80  0.80  0.8333333
#> [11,] -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.24  0.76  0.8000000
#> [12,] -0.24 -0.24 -0.24  0.8000000 -0.2000000 -0.24  0.76 -0.24 -0.2000000
#> [13,] -0.20 -0.20 -0.20 -0.1666667 -0.1666667 -0.20 -0.20 -0.20 -0.1666667
#> [14,] -0.20 -0.20 -0.20 -0.1666667 -0.1666667 -0.20 -0.20 -0.20 -0.1666667
#> [15,]  0.72 -0.28 -0.28 -0.2333333 -0.2333333 -0.28 -0.28 -0.28 -0.2333333
#> [16,] -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.24 -0.24 -0.2000000
#> [17,] -0.24 -0.24  0.76 -0.2000000 -0.2000000 -0.24 -0.24 -0.24 -0.2000000
#> [18,] -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.24 -0.24 -0.2000000
#> [19,] -0.32 -0.32 -0.32 -0.2666667  0.7333333 -0.32 -0.32 -0.32 -0.2666667
#> [20,] -0.28 -0.28 -0.28 -0.2333333 -0.2333333 -0.28 -0.28 -0.28 -0.2333333
#> [21,] -0.28 -0.28 -0.28 -0.2333333 -0.2333333 -0.28 -0.28 -0.28 -0.2333333
#> [22,] -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.24 -0.24 -0.2000000
#> [23,] -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.24 -0.24 -0.2000000
#> [24,]  0.72  0.72 -0.28 -0.2333333 -0.2333333 -0.28 -0.28 -0.28 -0.2333333
#> [25,]  0.76  0.76  0.76 -0.2000000 -0.2000000 -0.24 -0.24 -0.24 -0.2000000
#>            [,10] [,11] [,12]      [,13]      [,14]      [,15] [,16] [,17] [,18]
#>  [1,] -0.2000000 -0.24 -0.24 -0.2000000 -0.2000000  0.7200000 -0.24 -0.24 -0.24
#>  [2,] -0.2000000 -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.24 -0.24 -0.24
#>  [3,] -0.2000000 -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.24  0.76 -0.24
#>  [4,] -0.1666667 -0.20  0.80 -0.1666667 -0.1666667 -0.2333333 -0.20 -0.20 -0.20
#>  [5,] -0.1666667 -0.20 -0.20 -0.1666667 -0.1666667 -0.2333333 -0.20 -0.20 -0.20
#>  [6,] -0.2000000 -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.24 -0.24 -0.24
#>  [7,]  0.8000000 -0.24  0.76 -0.2000000 -0.2000000 -0.2800000 -0.24 -0.24 -0.24
#>  [8,]  0.8000000  0.76 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.24 -0.24 -0.24
#>  [9,]  0.8333333  0.80 -0.20 -0.1666667 -0.1666667 -0.2333333 -0.20 -0.20 -0.20
#> [10,] -0.1666667  0.80 -0.20 -0.1666667 -0.1666667 -0.2333333 -0.20 -0.20 -0.20
#> [11,]  0.8000000 -0.24  0.76  0.8000000  0.8000000 -0.2800000 -0.24 -0.24 -0.24
#> [12,] -0.2000000  0.76 -0.24  0.8000000  0.8000000  0.7200000 -0.24 -0.24 -0.24
#> [13,] -0.1666667  0.80  0.80 -0.1666667  0.8333333  0.7666667  0.80 -0.20 -0.20
#> [14,] -0.1666667  0.80  0.80  0.8333333 -0.1666667 -0.2333333  0.80  0.80 -0.20
#> [15,] -0.2333333 -0.28  0.72  0.7666667 -0.2333333 -0.3266667  0.72 -0.28  0.72
#> [16,] -0.2000000 -0.24 -0.24  0.8000000  0.8000000  0.7200000 -0.24  0.76  0.76
#> [17,] -0.2000000 -0.24 -0.24 -0.2000000  0.8000000 -0.2800000  0.76 -0.24  0.76
#> [18,] -0.2000000 -0.24 -0.24 -0.2000000 -0.2000000  0.7200000  0.76  0.76 -0.24
#> [19,] -0.2666667 -0.32 -0.32 -0.2666667 -0.2666667 -0.3733333  0.68  0.68  0.68
#> [20,] -0.2333333 -0.28 -0.28 -0.2333333 -0.2333333  0.6733333 -0.28  0.72  0.72
#> [21,] -0.2333333 -0.28 -0.28 -0.2333333 -0.2333333  0.6733333 -0.28 -0.28  0.72
#> [22,] -0.2000000 -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.24 -0.24 -0.24
#> [23,] -0.2000000 -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.24 -0.24 -0.24
#> [24,]  0.7666667 -0.28 -0.28 -0.2333333 -0.2333333 -0.3266667 -0.28 -0.28 -0.28
#> [25,] -0.2000000 -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.24 -0.24 -0.24
#>            [,19]      [,20]      [,21] [,22] [,23]      [,24] [,25]
#>  [1,] -0.3200000 -0.2800000 -0.2800000 -0.24 -0.24  0.7200000  0.76
#>  [2,] -0.3200000 -0.2800000 -0.2800000 -0.24 -0.24  0.7200000  0.76
#>  [3,] -0.3200000 -0.2800000 -0.2800000 -0.24 -0.24 -0.2800000  0.76
#>  [4,] -0.2666667 -0.2333333 -0.2333333 -0.20 -0.20 -0.2333333 -0.20
#>  [5,]  0.7333333 -0.2333333 -0.2333333 -0.20 -0.20 -0.2333333 -0.20
#>  [6,] -0.3200000 -0.2800000 -0.2800000 -0.24 -0.24 -0.2800000 -0.24
#>  [7,] -0.3200000 -0.2800000 -0.2800000 -0.24 -0.24 -0.2800000 -0.24
#>  [8,] -0.3200000 -0.2800000 -0.2800000 -0.24 -0.24 -0.2800000 -0.24
#>  [9,] -0.2666667 -0.2333333 -0.2333333 -0.20 -0.20 -0.2333333 -0.20
#> [10,] -0.2666667 -0.2333333 -0.2333333 -0.20 -0.20  0.7666667 -0.20
#> [11,] -0.3200000 -0.2800000 -0.2800000 -0.24 -0.24 -0.2800000 -0.24
#> [12,] -0.3200000 -0.2800000 -0.2800000 -0.24 -0.24 -0.2800000 -0.24
#> [13,] -0.2666667 -0.2333333 -0.2333333 -0.20 -0.20 -0.2333333 -0.20
#> [14,] -0.2666667 -0.2333333 -0.2333333 -0.20 -0.20 -0.2333333 -0.20
#> [15,] -0.3733333  0.6733333  0.6733333 -0.28 -0.28 -0.3266667 -0.28
#> [16,]  0.6800000 -0.2800000 -0.2800000 -0.24 -0.24 -0.2800000 -0.24
#> [17,]  0.6800000  0.7200000 -0.2800000 -0.24 -0.24 -0.2800000 -0.24
#> [18,]  0.6800000  0.7200000  0.7200000 -0.24 -0.24 -0.2800000 -0.24
#> [19,] -0.4266667  0.6266667  0.6266667  0.68  0.68 -0.3733333 -0.32
#> [20,]  0.6266667 -0.3266667  0.6733333  0.72  0.72 -0.3266667 -0.28
#> [21,]  0.6266667  0.6733333 -0.3266667  0.72  0.72  0.6733333 -0.28
#> [22,]  0.6800000  0.7200000  0.7200000 -0.24  0.76  0.7200000  0.76
#> [23,]  0.6800000  0.7200000  0.7200000  0.76 -0.24  0.7200000  0.76
#> [24,] -0.3733333 -0.3266667  0.6733333  0.72  0.72 -0.3266667  0.72
#> [25,] -0.3200000 -0.2800000 -0.2800000  0.76  0.76  0.7200000 -0.24
#> attr(,"representation")
#> [1] "modularity"
repr_graphon(g)
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.33333333 0.43650794 0.59027778 0.43650794 0.33333333 0.21111111
#>  [2,] 0.43650794 0.42857143 0.66071429 0.50000000 0.49206349 0.07142857
#>  [3,] 0.59027778 0.66071429 0.50000000 0.26785714 0.46527778 0.45000000
#>  [4,] 0.43650794 0.50000000 0.26785714 0.14285714 0.30952381 0.72857143
#>  [5,] 0.33333333 0.49206349 0.46527778 0.30952381 0.22222222 0.57777778
#>  [6,] 0.21111111 0.07142857 0.45000000 0.72857143 0.57777778 0.40000000
#>  [7,] 0.17361111 0.19642857 0.25000000 0.33928571 0.29166667 0.61250000
#>  [8,] 0.22222222 0.18253968 0.11805556 0.38095238 0.38888889 0.57777778
#>  [9,] 0.11111111 0.18253968 0.17361111 0.30952381 0.22222222 0.67777778
#> [10,] 0.00000000 0.00000000 0.19642857 0.42857143 0.25396825 0.41428571
#> [11,] 0.05555556 0.00000000 0.11805556 0.30952381 0.16666667 0.36666667
#> [12,] 0.00000000 0.00000000 0.06250000 0.30952381 0.11111111 0.10000000
#> [13,] 0.05555556 0.00000000 0.07142857 0.07142857 0.00000000 0.00000000
#> [14,] 0.11111111 0.05555556 0.00000000 0.05555556 0.05555556 0.11111111
#> [15,] 0.21111111 0.07142857 0.16250000 0.00000000 0.15555556 0.00000000
#> [16,] 0.00000000 0.00000000 0.05555556 0.05555556 0.05555556 0.00000000
#> [17,] 0.11111111 0.12698413 0.17361111 0.07142857 0.05555556 0.05555556
#> [18,] 0.00000000 0.00000000 0.05555556 0.00000000 0.11111111 0.00000000
#> [19,] 0.27777778 0.19841270 0.11805556 0.07142857 0.16666667 0.00000000
#> [20,] 0.17361111 0.20535714 0.00000000 0.00000000 0.11805556 0.00000000
#> [21,] 0.16666667 0.19841270 0.17361111 0.00000000 0.05555556 0.00000000
#> [22,] 0.23611111 0.26785714 0.37500000 0.19642857 0.11805556 0.00000000
#> [23,] 0.23611111 0.26785714 0.37500000 0.19642857 0.11805556 0.00000000
#> [24,] 0.52777778 0.60714286 0.27083333 0.07142857 0.19444444 0.28333333
#> [25,] 0.55555556 0.69047619 0.47222222 0.18253968 0.22222222 0.26666667
#>             [,7]       [,8]       [,9]      [,10]      [,11]      [,12]
#>  [1,] 0.17361111 0.22222222 0.11111111 0.00000000 0.05555556 0.00000000
#>  [2,] 0.19642857 0.18253968 0.18253968 0.00000000 0.00000000 0.00000000
#>  [3,] 0.25000000 0.11805556 0.17361111 0.19642857 0.11805556 0.06250000
#>  [4,] 0.33928571 0.38095238 0.30952381 0.42857143 0.30952381 0.30952381
#>  [5,] 0.29166667 0.38888889 0.22222222 0.25396825 0.16666667 0.11111111
#>  [6,] 0.61250000 0.57777778 0.67777778 0.41428571 0.36666667 0.10000000
#>  [7,] 0.62500000 0.70833333 0.64583333 0.60714286 0.46527778 0.35416667
#>  [8,] 0.70833333 0.55555556 0.61111111 0.57936508 0.50000000 0.16666667
#>  [9,] 0.64583333 0.61111111 0.55555556 0.57936508 0.50000000 0.16666667
#> [10,] 0.60714286 0.57936508 0.57936508 0.57142857 0.50793651 0.14285714
#> [11,] 0.46527778 0.50000000 0.50000000 0.50793651 0.66666667 0.55555556
#> [12,] 0.35416667 0.16666667 0.16666667 0.14285714 0.55555556 0.33333333
#> [13,] 0.13392857 0.18253968 0.18253968 0.21428571 0.38095238 0.54761905
#> [14,] 0.28472222 0.27777778 0.22222222 0.25396825 0.44444444 0.72222222
#> [15,] 0.00000000 0.05555556 0.05555556 0.07142857 0.21111111 0.70000000
#> [16,] 0.05555556 0.05555556 0.05555556 0.05555556 0.27777778 0.33333333
#> [17,] 0.00000000 0.11111111 0.00000000 0.00000000 0.11111111 0.27777778
#> [18,] 0.00000000 0.00000000 0.00000000 0.00000000 0.11111111 0.27777778
#> [19,] 0.06250000 0.00000000 0.05555556 0.05555556 0.00000000 0.22222222
#> [20,] 0.00000000 0.00000000 0.00000000 0.06250000 0.00000000 0.00000000
#> [21,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [22,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [23,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [24,] 0.06250000 0.05555556 0.05555556 0.15476190 0.05555556 0.00000000
#> [25,] 0.05555556 0.11111111 0.00000000 0.05555556 0.00000000 0.05555556
#>            [,13]      [,14]      [,15]      [,16]      [,17]      [,18]
#>  [1,] 0.05555556 0.11111111 0.21111111 0.00000000 0.11111111 0.00000000
#>  [2,] 0.00000000 0.05555556 0.07142857 0.00000000 0.12698413 0.00000000
#>  [3,] 0.07142857 0.00000000 0.16250000 0.05555556 0.17361111 0.05555556
#>  [4,] 0.07142857 0.05555556 0.00000000 0.05555556 0.07142857 0.00000000
#>  [5,] 0.00000000 0.05555556 0.15555556 0.05555556 0.05555556 0.11111111
#>  [6,] 0.00000000 0.11111111 0.00000000 0.00000000 0.05555556 0.00000000
#>  [7,] 0.13392857 0.28472222 0.00000000 0.05555556 0.00000000 0.00000000
#>  [8,] 0.18253968 0.27777778 0.05555556 0.05555556 0.11111111 0.00000000
#>  [9,] 0.18253968 0.22222222 0.05555556 0.05555556 0.00000000 0.00000000
#> [10,] 0.21428571 0.25396825 0.07142857 0.05555556 0.00000000 0.00000000
#> [11,] 0.38095238 0.44444444 0.21111111 0.27777778 0.11111111 0.11111111
#> [12,] 0.54761905 0.72222222 0.70000000 0.33333333 0.27777778 0.27777778
#> [13,] 0.57142857 0.57936508 0.48571429 0.50793651 0.32539683 0.25396825
#> [14,] 0.57936508 0.44444444 0.32222222 0.44444444 0.33333333 0.22222222
#> [15,] 0.48571429 0.32222222 0.20000000 0.83333333 0.52222222 0.52222222
#> [16,] 0.50793651 0.44444444 0.83333333 0.44444444 0.44444444 0.50000000
#> [17,] 0.32539683 0.33333333 0.52222222 0.44444444 0.22222222 0.50000000
#> [18,] 0.25396825 0.22222222 0.52222222 0.50000000 0.50000000 0.44444444
#> [19,] 0.26984127 0.16666667 0.36666667 0.50000000 0.55555556 0.61111111
#> [20,] 0.20535714 0.11805556 0.65000000 0.29166667 0.58333333 0.52777778
#> [21,] 0.12698413 0.16666667 0.52222222 0.27777778 0.44444444 0.55555556
#> [22,] 0.00000000 0.06250000 0.28750000 0.23611111 0.17361111 0.29166667
#> [23,] 0.00000000 0.06250000 0.28750000 0.23611111 0.17361111 0.29166667
#> [24,] 0.00000000 0.05555556 0.08333333 0.05555556 0.33333333 0.19444444
#> [25,] 0.00000000 0.00000000 0.05555556 0.00000000 0.16666667 0.11111111
#>            [,19]     [,20]      [,21]     [,22]     [,23]      [,24]      [,25]
#>  [1,] 0.27777778 0.1736111 0.16666667 0.2361111 0.2361111 0.52777778 0.55555556
#>  [2,] 0.19841270 0.2053571 0.19841270 0.2678571 0.2678571 0.60714286 0.69047619
#>  [3,] 0.11805556 0.0000000 0.17361111 0.3750000 0.3750000 0.27083333 0.47222222
#>  [4,] 0.07142857 0.0000000 0.00000000 0.1964286 0.1964286 0.07142857 0.18253968
#>  [5,] 0.16666667 0.1180556 0.05555556 0.1180556 0.1180556 0.19444444 0.22222222
#>  [6,] 0.00000000 0.0000000 0.00000000 0.0000000 0.0000000 0.28333333 0.26666667
#>  [7,] 0.06250000 0.0000000 0.00000000 0.0000000 0.0000000 0.06250000 0.05555556
#>  [8,] 0.00000000 0.0000000 0.00000000 0.0000000 0.0000000 0.05555556 0.11111111
#>  [9,] 0.05555556 0.0000000 0.00000000 0.0000000 0.0000000 0.05555556 0.00000000
#> [10,] 0.05555556 0.0625000 0.00000000 0.0000000 0.0000000 0.15476190 0.05555556
#> [11,] 0.00000000 0.0000000 0.00000000 0.0000000 0.0000000 0.05555556 0.00000000
#> [12,] 0.22222222 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.05555556
#> [13,] 0.26984127 0.2053571 0.12698413 0.0000000 0.0000000 0.00000000 0.00000000
#> [14,] 0.16666667 0.1180556 0.16666667 0.0625000 0.0625000 0.05555556 0.00000000
#> [15,] 0.36666667 0.6500000 0.52222222 0.2875000 0.2875000 0.08333333 0.05555556
#> [16,] 0.50000000 0.2916667 0.27777778 0.2361111 0.2361111 0.05555556 0.00000000
#> [17,] 0.55555556 0.5833333 0.44444444 0.1736111 0.1736111 0.33333333 0.16666667
#> [18,] 0.61111111 0.5277778 0.55555556 0.2916667 0.2916667 0.19444444 0.11111111
#> [19,] 0.66666667 0.7638889 0.77777778 0.5902778 0.5902778 0.41666667 0.33333333
#> [20,] 0.76388889 0.6250000 0.70833333 0.6250000 0.6250000 0.35416667 0.29861111
#> [21,] 0.77777778 0.7083333 0.55555556 0.4652778 0.4652778 0.55555556 0.27777778
#> [22,] 0.59027778 0.6250000 0.46527778 0.3750000 0.3750000 0.70833333 0.41666667
#> [23,] 0.59027778 0.6250000 0.46527778 0.3750000 0.3750000 0.70833333 0.41666667
#> [24,] 0.41666667 0.3541667 0.55555556 0.7083333 0.7083333 0.50000000 0.61111111
#> [25,] 0.33333333 0.2986111 0.27777778 0.4166667 0.4166667 0.61111111 0.66666667
#> attr(,"representation")
#> [1] "graphon"