Network Representation Functions

This is a collection of functions that convert a graph stored as an igraph object into a desired matrix representation among adjacency matrix, graph laplacian, modularity matrix or graphon (edge probability matrix).

repr_adjacency(network, validate = TRUE)

repr_laplacian(network, validate = TRUE)

repr_modularity(network, validate = TRUE)

repr_graphon(network, validate = TRUE)

Arguments

network
An igraph object.

validate
A boolean specifying whether the function should check the class of its input (default: TRUE).

Value

A numeric square matrix giving the desired network representation recorded in the object's class.

Examples

g <- igraph::sample_smallworld(1, 25, 3, 0.05)
repr_adjacency(g)

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0    1    1    1    0    0    0    0    0     0     0     1     0
#>  [2,]    1    0    1    1    1    0    0    0    0     0     0     0     0
#>  [3,]    1    1    0    1    0    1    1    0    0     0     0     0     0
#>  [4,]    1    1    1    0    1    1    0    0    0     0     0     0     0
#>  [5,]    0    1    0    1    0    1    1    1    0     0     0     0     0
#>  [6,]    0    0    1    1    1    0    1    1    1     0     0     0     0
#>  [7,]    0    0    1    0    1    1    0    1    1     1     0     1     0
#>  [8,]    0    0    0    0    1    1    1    0    1     1     1     0     0
#>  [9,]    0    0    0    0    0    1    1    1    0     1     1     1     0
#> [10,]    0    0    0    0    0    0    1    1    1     0     1     1     1
#> [11,]    0    0    0    0    0    0    0    1    1     1     0     0     1
#> [12,]    1    0    0    0    0    0    1    0    1     1     0     0     1
#> [13,]    0    0    0    0    0    0    0    0    0     1     1     1     0
#> [14,]    0    0    0    0    0    0    1    0    0     0     1     1     1
#> [15,]    0    0    0    0    0    0    0    0    0     0     0     1     1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0     1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    1    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]    1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]    0    1    0    0    0    0    0    0    0     0     0     0     0
#> [25,]    1    1    0    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0     1     0     1
#>  [2,]     0     0     0     0     0     0     0     0     0     0     1     1
#>  [3,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     1     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     1     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]     1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]     1     1     0     0     0     0     0     0     0     0     0     0
#> [13,]     1     1     1     0     0     0     0     0     0     0     0     0
#> [14,]     0     1     1     1     0     0     0     0     0     0     0     0
#> [15,]     1     0     1     1     1     0     0     0     0     0     0     0
#> [16,]     1     1     0     1     1     1     0     0     0     0     0     0
#> [17,]     1     1     1     0     1     1     1     0     0     0     0     0
#> [18,]     0     1     1     1     0     1     1     1     0     0     0     0
#> [19,]     0     0     1     1     1     0     1     1     1     0     0     0
#> [20,]     0     0     0     1     1     1     0     1     1     1     0     0
#> [21,]     0     0     0     0     1     1     1     0     1     1     1     0
#> [22,]     0     0     0     0     0     1     1     1     0     1     1     1
#> [23,]     0     0     0     0     0     0     1     1     1     0     1     1
#> [24,]     0     0     0     0     0     0     0     1     1     1     0     1
#> [25,]     0     0     0     0     0     0     0     0     1     1     1     0
#> attr(,"representation")
#> [1] "adjacency"

repr_laplacian(g)

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    6   -1   -1   -1    0    0    0    0    0     0     0    -1     0
#>  [2,]   -1    6   -1   -1   -1    0    0    0    0     0     0     0     0
#>  [3,]   -1   -1    5   -1    0   -1   -1    0    0     0     0     0     0
#>  [4,]   -1   -1   -1    5   -1   -1    0    0    0     0     0     0     0
#>  [5,]    0   -1    0   -1    6   -1   -1   -1    0     0     0     0     0
#>  [6,]    0    0   -1   -1   -1    6   -1   -1   -1     0     0     0     0
#>  [7,]    0    0   -1    0   -1   -1    8   -1   -1    -1     0    -1     0
#>  [8,]    0    0    0    0   -1   -1   -1    6   -1    -1    -1     0     0
#>  [9,]    0    0    0    0    0   -1   -1   -1    6    -1    -1    -1     0
#> [10,]    0    0    0    0    0    0   -1   -1   -1     6    -1    -1    -1
#> [11,]    0    0    0    0    0    0    0   -1   -1    -1     5     0    -1
#> [12,]   -1    0    0    0    0    0   -1    0   -1    -1     0     7    -1
#> [13,]    0    0    0    0    0    0    0    0    0    -1    -1    -1     6
#> [14,]    0    0    0    0    0    0   -1    0    0     0    -1    -1    -1
#> [15,]    0    0    0    0    0    0    0    0    0     0     0    -1    -1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0    -1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0   -1    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]   -1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]    0   -1    0    0    0    0    0    0    0     0     0     0     0
#> [25,]   -1   -1    0    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0    -1     0    -1
#>  [2,]     0     0     0     0     0     0     0     0     0     0    -1    -1
#>  [3,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0    -1     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]    -1     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]    -1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]    -1    -1     0     0     0     0     0     0     0     0     0     0
#> [13,]    -1    -1    -1     0     0     0     0     0     0     0     0     0
#> [14,]     7    -1    -1    -1     0     0     0     0     0     0     0     0
#> [15,]    -1     6    -1    -1    -1     0     0     0     0     0     0     0
#> [16,]    -1    -1     6    -1    -1    -1     0     0     0     0     0     0
#> [17,]    -1    -1    -1     6    -1    -1    -1     0     0     0     0     0
#> [18,]     0    -1    -1    -1     6    -1    -1    -1     0     0     0     0
#> [19,]     0     0    -1    -1    -1     6    -1    -1    -1     0     0     0
#> [20,]     0     0     0    -1    -1    -1     6    -1    -1    -1     0     0
#> [21,]     0     0     0     0    -1    -1    -1     7    -1    -1    -1     0
#> [22,]     0     0     0     0     0    -1    -1    -1     6    -1    -1    -1
#> [23,]     0     0     0     0     0     0    -1    -1    -1     6    -1    -1
#> [24,]     0     0     0     0     0     0     0    -1    -1    -1     5    -1
#> [25,]     0     0     0     0     0     0     0     0    -1    -1    -1     5
#> attr(,"representation")
#> [1] "laplacian"

repr_modularity(g)

#>        [,1]  [,2]       [,3]       [,4]  [,5]  [,6]       [,7]  [,8]  [,9]
#>  [1,] -0.24  0.76  0.8000000  0.8000000 -0.24 -0.24 -0.3200000 -0.24 -0.24
#>  [2,]  0.76 -0.24  0.8000000  0.8000000  0.76 -0.24 -0.3200000 -0.24 -0.24
#>  [3,]  0.80  0.80 -0.1666667  0.8333333 -0.20  0.80  0.7333333 -0.20 -0.20
#>  [4,]  0.80  0.80  0.8333333 -0.1666667  0.80  0.80 -0.2666667 -0.20 -0.20
#>  [5,] -0.24  0.76 -0.2000000  0.8000000 -0.24  0.76  0.6800000  0.76 -0.24
#>  [6,] -0.24 -0.24  0.8000000  0.8000000  0.76 -0.24  0.6800000  0.76  0.76
#>  [7,] -0.32 -0.32  0.7333333 -0.2666667  0.68  0.68 -0.4266667  0.68  0.68
#>  [8,] -0.24 -0.24 -0.2000000 -0.2000000  0.76  0.76  0.6800000 -0.24  0.76
#>  [9,] -0.24 -0.24 -0.2000000 -0.2000000 -0.24  0.76  0.6800000  0.76 -0.24
#> [10,] -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.24  0.6800000  0.76  0.76
#> [11,] -0.20 -0.20 -0.1666667 -0.1666667 -0.20 -0.20 -0.2666667  0.80  0.80
#> [12,]  0.72 -0.28 -0.2333333 -0.2333333 -0.28 -0.28  0.6266667 -0.28  0.72
#> [13,] -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.24 -0.3200000 -0.24 -0.24
#> [14,] -0.28 -0.28 -0.2333333 -0.2333333 -0.28 -0.28  0.6266667 -0.28 -0.28
#> [15,] -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.24 -0.3200000 -0.24 -0.24
#> [16,] -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.24 -0.3200000 -0.24 -0.24
#> [17,] -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.24 -0.3200000 -0.24 -0.24
#> [18,] -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.24 -0.3200000 -0.24 -0.24
#> [19,] -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.24 -0.3200000 -0.24 -0.24
#> [20,] -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.24 -0.3200000 -0.24 -0.24
#> [21,] -0.28 -0.28 -0.2333333 -0.2333333  0.72 -0.28 -0.3733333 -0.28 -0.28
#> [22,] -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.24 -0.3200000 -0.24 -0.24
#> [23,]  0.76 -0.24 -0.2000000 -0.2000000 -0.24 -0.24 -0.3200000 -0.24 -0.24
#> [24,] -0.20  0.80 -0.1666667 -0.1666667 -0.20 -0.20 -0.2666667 -0.20 -0.20
#> [25,]  0.80  0.80 -0.1666667 -0.1666667 -0.20 -0.20 -0.2666667 -0.20 -0.20
#>       [,10]      [,11]      [,12] [,13]      [,14] [,15] [,16] [,17] [,18]
#>  [1,] -0.24 -0.2000000  0.7200000 -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.24
#>  [2,] -0.24 -0.2000000 -0.2800000 -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.24
#>  [3,] -0.20 -0.1666667 -0.2333333 -0.20 -0.2333333 -0.20 -0.20 -0.20 -0.20
#>  [4,] -0.20 -0.1666667 -0.2333333 -0.20 -0.2333333 -0.20 -0.20 -0.20 -0.20
#>  [5,] -0.24 -0.2000000 -0.2800000 -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.24
#>  [6,] -0.24 -0.2000000 -0.2800000 -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.24
#>  [7,]  0.68 -0.2666667  0.6266667 -0.32  0.6266667 -0.32 -0.32 -0.32 -0.32
#>  [8,]  0.76  0.8000000 -0.2800000 -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.24
#>  [9,]  0.76  0.8000000  0.7200000 -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.24
#> [10,] -0.24  0.8000000  0.7200000  0.76 -0.2800000 -0.24 -0.24 -0.24 -0.24
#> [11,]  0.80 -0.1666667 -0.2333333  0.80  0.7666667 -0.20 -0.20 -0.20 -0.20
#> [12,]  0.72 -0.2333333 -0.3266667  0.72  0.6733333  0.72 -0.28 -0.28 -0.28
#> [13,]  0.76  0.8000000  0.7200000 -0.24  0.7200000  0.76  0.76 -0.24 -0.24
#> [14,] -0.28  0.7666667  0.6733333  0.72 -0.3266667  0.72  0.72  0.72 -0.28
#> [15,] -0.24 -0.2000000  0.7200000  0.76  0.7200000 -0.24  0.76  0.76  0.76
#> [16,] -0.24 -0.2000000 -0.2800000  0.76  0.7200000  0.76 -0.24  0.76  0.76
#> [17,] -0.24 -0.2000000 -0.2800000 -0.24  0.7200000  0.76  0.76 -0.24  0.76
#> [18,] -0.24 -0.2000000 -0.2800000 -0.24 -0.2800000  0.76  0.76  0.76 -0.24
#> [19,] -0.24 -0.2000000 -0.2800000 -0.24 -0.2800000 -0.24  0.76  0.76  0.76
#> [20,] -0.24 -0.2000000 -0.2800000 -0.24 -0.2800000 -0.24 -0.24  0.76  0.76
#> [21,] -0.28 -0.2333333 -0.3266667 -0.28 -0.3266667 -0.28 -0.28 -0.28  0.72
#> [22,] -0.24 -0.2000000 -0.2800000 -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.24
#> [23,] -0.24 -0.2000000 -0.2800000 -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.24
#> [24,] -0.20 -0.1666667 -0.2333333 -0.20 -0.2333333 -0.20 -0.20 -0.20 -0.20
#> [25,] -0.20 -0.1666667 -0.2333333 -0.20 -0.2333333 -0.20 -0.20 -0.20 -0.20
#>       [,19] [,20]      [,21] [,22] [,23]      [,24]      [,25]
#>  [1,] -0.24 -0.24 -0.2800000 -0.24  0.76 -0.2000000  0.8000000
#>  [2,] -0.24 -0.24 -0.2800000 -0.24 -0.24  0.8000000  0.8000000
#>  [3,] -0.20 -0.20 -0.2333333 -0.20 -0.20 -0.1666667 -0.1666667
#>  [4,] -0.20 -0.20 -0.2333333 -0.20 -0.20 -0.1666667 -0.1666667
#>  [5,] -0.24 -0.24  0.7200000 -0.24 -0.24 -0.2000000 -0.2000000
#>  [6,] -0.24 -0.24 -0.2800000 -0.24 -0.24 -0.2000000 -0.2000000
#>  [7,] -0.32 -0.32 -0.3733333 -0.32 -0.32 -0.2666667 -0.2666667
#>  [8,] -0.24 -0.24 -0.2800000 -0.24 -0.24 -0.2000000 -0.2000000
#>  [9,] -0.24 -0.24 -0.2800000 -0.24 -0.24 -0.2000000 -0.2000000
#> [10,] -0.24 -0.24 -0.2800000 -0.24 -0.24 -0.2000000 -0.2000000
#> [11,] -0.20 -0.20 -0.2333333 -0.20 -0.20 -0.1666667 -0.1666667
#> [12,] -0.28 -0.28 -0.3266667 -0.28 -0.28 -0.2333333 -0.2333333
#> [13,] -0.24 -0.24 -0.2800000 -0.24 -0.24 -0.2000000 -0.2000000
#> [14,] -0.28 -0.28 -0.3266667 -0.28 -0.28 -0.2333333 -0.2333333
#> [15,] -0.24 -0.24 -0.2800000 -0.24 -0.24 -0.2000000 -0.2000000
#> [16,]  0.76 -0.24 -0.2800000 -0.24 -0.24 -0.2000000 -0.2000000
#> [17,]  0.76  0.76 -0.2800000 -0.24 -0.24 -0.2000000 -0.2000000
#> [18,]  0.76  0.76  0.7200000 -0.24 -0.24 -0.2000000 -0.2000000
#> [19,] -0.24  0.76  0.7200000  0.76 -0.24 -0.2000000 -0.2000000
#> [20,]  0.76 -0.24  0.7200000  0.76  0.76 -0.2000000 -0.2000000
#> [21,]  0.72  0.72 -0.3266667  0.72  0.72  0.7666667 -0.2333333
#> [22,]  0.76  0.76  0.7200000 -0.24  0.76  0.8000000  0.8000000
#> [23,] -0.24  0.76  0.7200000  0.76 -0.24  0.8000000  0.8000000
#> [24,] -0.20 -0.20  0.7666667  0.80  0.80 -0.1666667  0.8333333
#> [25,] -0.20 -0.20 -0.2333333  0.80  0.80  0.8333333 -0.1666667
#> attr(,"representation")
#> [1] "modularity"

repr_graphon(g)

#>             [,1]       [,2]       [,3]       [,4]       [,5]      [,6]
#>  [1,] 0.44444444 0.64583333 0.44444444 0.50000000 0.27777778 0.2666667
#>  [2,] 0.64583333 0.50000000 0.52777778 0.58333333 0.46527778 0.2250000
#>  [3,] 0.44444444 0.52777778 0.44444444 0.50000000 0.38888889 0.4777778
#>  [4,] 0.50000000 0.58333333 0.50000000 0.55555556 0.44444444 0.5222222
#>  [5,] 0.27777778 0.46527778 0.38888889 0.44444444 0.33333333 0.5777778
#>  [6,] 0.26666667 0.22500000 0.47777778 0.52222222 0.57777778 0.4000000
#>  [7,] 0.27777778 0.18055556 0.44444444 0.27777778 0.55555556 0.6222222
#>  [8,] 0.22222222 0.22916667 0.27777778 0.27777778 0.38888889 0.5777778
#>  [9,] 0.11111111 0.17361111 0.22222222 0.27777778 0.33333333 0.5777778
#> [10,] 0.00000000 0.05555556 0.16666667 0.16666667 0.22222222 0.2666667
#> [11,] 0.05555556 0.00000000 0.16666667 0.11111111 0.27777778 0.3111111
#> [12,] 0.12698413 0.06250000 0.11111111 0.05555556 0.23809524 0.1714286
#> [13,] 0.00000000 0.00000000 0.00000000 0.00000000 0.05555556 0.2000000
#> [14,] 0.06250000 0.00000000 0.00000000 0.00000000 0.00000000 0.0625000
#> [15,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000
#> [16,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000
#> [17,] 0.00000000 0.00000000 0.00000000 0.00000000 0.05555556 0.0000000
#> [18,] 0.05555556 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000
#> [19,] 0.11111111 0.11805556 0.00000000 0.00000000 0.00000000 0.0000000
#> [20,] 0.05555556 0.19642857 0.00000000 0.00000000 0.07142857 0.0000000
#> [21,] 0.30952381 0.19642857 0.05555556 0.11111111 0.18253968 0.1000000
#> [22,] 0.19444444 0.20833333 0.05555556 0.11111111 0.13888889 0.0000000
#> [23,] 0.38888889 0.29861111 0.16666667 0.22222222 0.22222222 0.0000000
#> [24,] 0.29166667 0.50000000 0.17361111 0.23611111 0.00000000 0.1625000
#> [25,] 0.50000000 0.52777778 0.27777778 0.33333333 0.22222222 0.2111111
#>             [,7]       [,8]       [,9]      [,10]      [,11]      [,12]
#>  [1,] 0.27777778 0.22222222 0.11111111 0.00000000 0.05555556 0.12698413
#>  [2,] 0.18055556 0.22916667 0.17361111 0.05555556 0.00000000 0.06250000
#>  [3,] 0.44444444 0.27777778 0.22222222 0.16666667 0.16666667 0.11111111
#>  [4,] 0.27777778 0.27777778 0.27777778 0.16666667 0.11111111 0.05555556
#>  [5,] 0.55555556 0.38888889 0.33333333 0.22222222 0.27777778 0.23809524
#>  [6,] 0.62222222 0.57777778 0.57777778 0.26666667 0.31111111 0.17142857
#>  [7,] 0.55555556 0.61111111 0.66666667 0.66666667 0.50000000 0.50793651
#>  [8,] 0.61111111 0.55555556 0.55555556 0.50000000 0.44444444 0.30952381
#>  [9,] 0.66666667 0.55555556 0.55555556 0.50000000 0.50000000 0.50793651
#> [10,] 0.66666667 0.50000000 0.50000000 0.44444444 0.55555556 0.56349206
#> [11,] 0.50000000 0.44444444 0.50000000 0.55555556 0.44444444 0.50793651
#> [12,] 0.50793651 0.30952381 0.50793651 0.56349206 0.50793651 0.42857143
#> [13,] 0.36666667 0.21111111 0.11111111 0.42222222 0.57777778 0.75714286
#> [14,] 0.47222222 0.23611111 0.29166667 0.35416667 0.52777778 0.59821429
#> [15,] 0.19444444 0.05555556 0.05555556 0.19444444 0.33333333 0.53571429
#> [16,] 0.05555556 0.11805556 0.11805556 0.29166667 0.17361111 0.26785714
#> [17,] 0.00000000 0.00000000 0.00000000 0.16666667 0.16666667 0.25396825
#> [18,] 0.00000000 0.00000000 0.00000000 0.11111111 0.16666667 0.18253968
#> [19,] 0.00000000 0.00000000 0.00000000 0.00000000 0.05555556 0.12698413
#> [20,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [21,] 0.00000000 0.05555556 0.05555556 0.05555556 0.00000000 0.00000000
#> [22,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [23,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [24,] 0.06250000 0.06250000 0.00000000 0.00000000 0.00000000 0.06250000
#> [25,] 0.05555556 0.00000000 0.00000000 0.00000000 0.00000000 0.05555556
#>            [,13]     [,14]      [,15]      [,16]      [,17]      [,18]
#>  [1,] 0.00000000 0.0625000 0.00000000 0.00000000 0.00000000 0.05555556
#>  [2,] 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [3,] 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [4,] 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [5,] 0.05555556 0.0000000 0.00000000 0.00000000 0.05555556 0.00000000
#>  [6,] 0.20000000 0.0625000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [7,] 0.36666667 0.4722222 0.19444444 0.05555556 0.00000000 0.00000000
#>  [8,] 0.21111111 0.2361111 0.05555556 0.11805556 0.00000000 0.00000000
#>  [9,] 0.11111111 0.2916667 0.05555556 0.11805556 0.00000000 0.00000000
#> [10,] 0.42222222 0.3541667 0.19444444 0.29166667 0.16666667 0.11111111
#> [11,] 0.57777778 0.5277778 0.33333333 0.17361111 0.16666667 0.16666667
#> [12,] 0.75714286 0.5982143 0.53571429 0.26785714 0.25396825 0.18253968
#> [13,] 0.40000000 0.6125000 0.55000000 0.58750000 0.31111111 0.31111111
#> [14,] 0.61250000 0.5000000 0.64583333 0.56250000 0.47222222 0.34722222
#> [15,] 0.55000000 0.6458333 0.66666667 0.66666667 0.55555556 0.55555556
#> [16,] 0.58750000 0.5625000 0.66666667 0.62500000 0.59027778 0.59027778
#> [17,] 0.31111111 0.4722222 0.55555556 0.59027778 0.55555556 0.61111111
#> [18,] 0.31111111 0.3472222 0.55555556 0.59027778 0.61111111 0.55555556
#> [19,] 0.21111111 0.2916667 0.33333333 0.47222222 0.61111111 0.55555556
#> [20,] 0.17142857 0.1339286 0.30952381 0.33035714 0.56349206 0.56349206
#> [21,] 0.00000000 0.1339286 0.15476190 0.33035714 0.36507937 0.56349206
#> [22,] 0.00000000 0.0000000 0.16666667 0.20833333 0.33333333 0.33333333
#> [23,] 0.00000000 0.0000000 0.05555556 0.17361111 0.33333333 0.27777778
#> [24,] 0.00000000 0.0000000 0.00000000 0.06250000 0.23611111 0.23611111
#> [25,] 0.00000000 0.0000000 0.00000000 0.00000000 0.11111111 0.16666667
#>            [,19]      [,20]      [,21]      [,22]      [,23]     [,24]
#>  [1,] 0.11111111 0.05555556 0.30952381 0.19444444 0.38888889 0.2916667
#>  [2,] 0.11805556 0.19642857 0.19642857 0.20833333 0.29861111 0.5000000
#>  [3,] 0.00000000 0.00000000 0.05555556 0.05555556 0.16666667 0.1736111
#>  [4,] 0.00000000 0.00000000 0.11111111 0.11111111 0.22222222 0.2361111
#>  [5,] 0.00000000 0.07142857 0.18253968 0.13888889 0.22222222 0.0000000
#>  [6,] 0.00000000 0.00000000 0.10000000 0.00000000 0.00000000 0.1625000
#>  [7,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.0625000
#>  [8,] 0.00000000 0.00000000 0.05555556 0.00000000 0.00000000 0.0625000
#>  [9,] 0.00000000 0.00000000 0.05555556 0.00000000 0.00000000 0.0000000
#> [10,] 0.00000000 0.00000000 0.05555556 0.00000000 0.00000000 0.0000000
#> [11,] 0.05555556 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000
#> [12,] 0.12698413 0.00000000 0.00000000 0.00000000 0.00000000 0.0625000
#> [13,] 0.21111111 0.17142857 0.00000000 0.00000000 0.00000000 0.0000000
#> [14,] 0.29166667 0.13392857 0.13392857 0.00000000 0.00000000 0.0000000
#> [15,] 0.33333333 0.30952381 0.15476190 0.16666667 0.05555556 0.0000000
#> [16,] 0.47222222 0.33035714 0.33035714 0.20833333 0.17361111 0.0625000
#> [17,] 0.61111111 0.56349206 0.36507937 0.33333333 0.33333333 0.2361111
#> [18,] 0.55555556 0.56349206 0.56349206 0.33333333 0.27777778 0.2361111
#> [19,] 0.55555556 0.63492063 0.61904762 0.55555556 0.38888889 0.2361111
#> [20,] 0.63492063 0.57142857 0.71428571 0.61904762 0.56349206 0.3303571
#> [21,] 0.61904762 0.71428571 0.57142857 0.77380952 0.61904762 0.6607143
#> [22,] 0.55555556 0.61904762 0.77380952 0.66666667 0.66666667 0.5625000
#> [23,] 0.38888889 0.56349206 0.61904762 0.66666667 0.55555556 0.5902778
#> [24,] 0.23611111 0.33035714 0.66071429 0.56250000 0.59027778 0.3750000
#> [25,] 0.27777778 0.25396825 0.36507937 0.47222222 0.50000000 0.4722222
#>            [,25]
#>  [1,] 0.50000000
#>  [2,] 0.52777778
#>  [3,] 0.27777778
#>  [4,] 0.33333333
#>  [5,] 0.22222222
#>  [6,] 0.21111111
#>  [7,] 0.05555556
#>  [8,] 0.00000000
#>  [9,] 0.00000000
#> [10,] 0.00000000
#> [11,] 0.00000000
#> [12,] 0.05555556
#> [13,] 0.00000000
#> [14,] 0.00000000
#> [15,] 0.00000000
#> [16,] 0.00000000
#> [17,] 0.11111111
#> [18,] 0.16666667
#> [19,] 0.27777778
#> [20,] 0.25396825
#> [21,] 0.36507937
#> [22,] 0.47222222
#> [23,] 0.50000000
#> [24,] 0.47222222
#> [25,] 0.55555556
#> attr(,"representation")
#> [1] "graphon"