Network Representation Functions

This is a collection of functions that convert a graph stored as an igraph object into a desired matrix representation among adjacency matrix, graph laplacian, modularity matrix or graphon (edge probability matrix).

Usage

repr_adjacency(network, validate = TRUE)

repr_laplacian(network, validate = TRUE)

repr_modularity(network, validate = TRUE)

repr_graphon(network, validate = TRUE)

Arguments

network

An igraph object.

validate

A boolean specifying whether the function should check the class of its input (default: TRUE).

Value

A numeric square matrix giving the desired network representation recorded in the object's class.

Examples

g <- igraph::sample_smallworld(1, 25, 3, 0.05)
repr_adjacency(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0    1    1    1    1    0    0    0    0     0     0     0     0
#>  [2,]    1    0    0    0    1    0    0    0    0     0     0     1     0
#>  [3,]    1    0    0    1    1    1    0    0    0     0     0     0     0
#>  [4,]    1    0    1    0    1    1    1    0    0     0     0     0     0
#>  [5,]    1    1    1    1    0    1    0    1    0     0     0     0     0
#>  [6,]    0    0    1    1    1    0    1    1    1     0     0     0     0
#>  [7,]    0    0    0    1    0    1    0    1    1     1     0     0     0
#>  [8,]    0    0    0    0    1    1    1    0    1     1     1     0     0
#>  [9,]    0    0    0    0    0    1    1    1    0     1     1     0     0
#> [10,]    0    0    0    0    0    0    1    1    1     0     1     1     0
#> [11,]    0    0    0    0    0    0    0    1    1     1     0     1     1
#> [12,]    0    1    0    0    0    0    0    0    0     1     1     0     1
#> [13,]    0    0    0    0    0    0    0    0    0     0     1     1     0
#> [14,]    0    0    0    0    0    0    0    0    0     0     0     1     1
#> [15,]    0    0    0    0    0    0    0    0    0     0     0     1     0
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0     1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     1
#> [18,]    0    0    0    1    1    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     1     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    1    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     1     0     0
#> [23,]    1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [25,]    1    1    1    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0     1     0     1
#>  [2,]     0     0     0     0     0     0     0     0     0     0     0     1
#>  [3,]     0     0     0     0     0     0     0     1     0     0     0     1
#>  [4,]     0     0     0     0     1     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     1     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     1     0     0     0     0     0     0
#> [11,]     0     0     0     0     0     0     0     0     1     0     0     0
#> [12,]     1     1     0     0     0     0     0     0     0     0     0     0
#> [13,]     1     0     1     1     0     0     0     0     0     0     0     0
#> [14,]     0     1     1     0     0     0     0     0     0     0     1     0
#> [15,]     1     0     1     1     1     0     0     0     0     0     0     0
#> [16,]     1     1     0     1     1     1     0     0     0     0     0     0
#> [17,]     0     1     1     0     1     1     1     0     0     0     0     1
#> [18,]     0     1     1     1     0     1     1     1     0     0     0     0
#> [19,]     0     0     1     1     1     0     1     1     0     0     1     0
#> [20,]     0     0     0     1     1     1     0     1     1     1     0     0
#> [21,]     0     0     0     0     1     1     1     0     1     1     1     0
#> [22,]     0     0     0     0     0     0     1     1     0     1     1     1
#> [23,]     0     0     0     0     0     0     1     1     1     0     1     1
#> [24,]     1     0     0     0     0     1     0     1     1     1     0     1
#> [25,]     0     0     0     1     0     0     0     0     1     1     1     0
#> attr(,"representation")
#> [1] "adjacency"
repr_laplacian(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    6   -1   -1   -1   -1    0    0    0    0     0     0     0     0
#>  [2,]   -1    4    0    0   -1    0    0    0    0     0     0    -1     0
#>  [3,]   -1    0    6   -1   -1   -1    0    0    0     0     0     0     0
#>  [4,]   -1    0   -1    6   -1   -1   -1    0    0     0     0     0     0
#>  [5,]   -1   -1   -1   -1    7   -1    0   -1    0     0     0     0     0
#>  [6,]    0    0   -1   -1   -1    6   -1   -1   -1     0     0     0     0
#>  [7,]    0    0    0   -1    0   -1    5   -1   -1    -1     0     0     0
#>  [8,]    0    0    0    0   -1   -1   -1    6   -1    -1    -1     0     0
#>  [9,]    0    0    0    0    0   -1   -1   -1    5    -1    -1     0     0
#> [10,]    0    0    0    0    0    0   -1   -1   -1     6    -1    -1     0
#> [11,]    0    0    0    0    0    0    0   -1   -1    -1     6    -1    -1
#> [12,]    0   -1    0    0    0    0    0    0    0    -1    -1     6    -1
#> [13,]    0    0    0    0    0    0    0    0    0     0    -1    -1     5
#> [14,]    0    0    0    0    0    0    0    0    0     0     0    -1    -1
#> [15,]    0    0    0    0    0    0    0    0    0     0     0    -1     0
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0    -1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0    -1
#> [18,]    0    0    0   -1   -1    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0    -1     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0   -1    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0    -1     0     0
#> [23,]   -1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [25,]   -1   -1   -1    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0    -1     0    -1
#>  [2,]     0     0     0     0     0     0     0     0     0     0     0    -1
#>  [3,]     0     0     0     0     0     0     0    -1     0     0     0    -1
#>  [4,]     0     0     0     0    -1     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0    -1     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0    -1     0     0     0     0     0     0
#> [11,]     0     0     0     0     0     0     0     0    -1     0     0     0
#> [12,]    -1    -1     0     0     0     0     0     0     0     0     0     0
#> [13,]    -1     0    -1    -1     0     0     0     0     0     0     0     0
#> [14,]     5    -1    -1     0     0     0     0     0     0     0    -1     0
#> [15,]    -1     5    -1    -1    -1     0     0     0     0     0     0     0
#> [16,]    -1    -1     6    -1    -1    -1     0     0     0     0     0     0
#> [17,]     0    -1    -1     7    -1    -1    -1     0     0     0     0    -1
#> [18,]     0    -1    -1    -1     8    -1    -1    -1     0     0     0     0
#> [19,]     0     0    -1    -1    -1     7    -1    -1     0     0    -1     0
#> [20,]     0     0     0    -1    -1    -1     6    -1    -1    -1     0     0
#> [21,]     0     0     0     0    -1    -1    -1     7    -1    -1    -1     0
#> [22,]     0     0     0     0     0     0    -1    -1     6    -1    -1    -1
#> [23,]     0     0     0     0     0     0    -1    -1    -1     6    -1    -1
#> [24,]    -1     0     0     0     0    -1     0    -1    -1    -1     6    -1
#> [25,]     0     0     0    -1     0     0     0     0    -1    -1    -1     7
#> attr(,"representation")
#> [1] "laplacian"
repr_modularity(g)
#>        [,1]       [,2]  [,3]  [,4]       [,5]  [,6]       [,7]  [,8]       [,9]
#>  [1,] -0.24  0.8400000  0.76  0.76  0.7200000 -0.24 -0.2000000 -0.24 -0.2000000
#>  [2,]  0.84 -0.1066667 -0.16 -0.16  0.8133333 -0.16 -0.1333333 -0.16 -0.1333333
#>  [3,]  0.76 -0.1600000 -0.24  0.76  0.7200000  0.76 -0.2000000 -0.24 -0.2000000
#>  [4,]  0.76 -0.1600000  0.76 -0.24  0.7200000  0.76  0.8000000 -0.24 -0.2000000
#>  [5,]  0.72  0.8133333  0.72  0.72 -0.3266667  0.72 -0.2333333  0.72 -0.2333333
#>  [6,] -0.24 -0.1600000  0.76  0.76  0.7200000 -0.24  0.8000000  0.76  0.8000000
#>  [7,] -0.20 -0.1333333 -0.20  0.80 -0.2333333  0.80 -0.1666667  0.80  0.8333333
#>  [8,] -0.24 -0.1600000 -0.24 -0.24  0.7200000  0.76  0.8000000 -0.24  0.8000000
#>  [9,] -0.20 -0.1333333 -0.20 -0.20 -0.2333333  0.80  0.8333333  0.80 -0.1666667
#> [10,] -0.24 -0.1600000 -0.24 -0.24 -0.2800000 -0.24  0.8000000  0.76  0.8000000
#> [11,] -0.24 -0.1600000 -0.24 -0.24 -0.2800000 -0.24 -0.2000000  0.76  0.8000000
#> [12,] -0.24  0.8400000 -0.24 -0.24 -0.2800000 -0.24 -0.2000000 -0.24 -0.2000000
#> [13,] -0.20 -0.1333333 -0.20 -0.20 -0.2333333 -0.20 -0.1666667 -0.20 -0.1666667
#> [14,] -0.20 -0.1333333 -0.20 -0.20 -0.2333333 -0.20 -0.1666667 -0.20 -0.1666667
#> [15,] -0.20 -0.1333333 -0.20 -0.20 -0.2333333 -0.20 -0.1666667 -0.20 -0.1666667
#> [16,] -0.24 -0.1600000 -0.24 -0.24 -0.2800000 -0.24 -0.2000000 -0.24 -0.2000000
#> [17,] -0.28 -0.1866667 -0.28 -0.28 -0.3266667 -0.28 -0.2333333 -0.28 -0.2333333
#> [18,] -0.32 -0.2133333 -0.32  0.68  0.6266667 -0.32 -0.2666667 -0.32 -0.2666667
#> [19,] -0.28 -0.1866667 -0.28 -0.28 -0.3266667 -0.28 -0.2333333 -0.28 -0.2333333
#> [20,] -0.24 -0.1600000 -0.24 -0.24 -0.2800000 -0.24 -0.2000000 -0.24 -0.2000000
#> [21,] -0.28 -0.1866667  0.72 -0.28 -0.3266667 -0.28 -0.2333333 -0.28 -0.2333333
#> [22,] -0.24 -0.1600000 -0.24 -0.24 -0.2800000 -0.24 -0.2000000 -0.24 -0.2000000
#> [23,]  0.76 -0.1600000 -0.24 -0.24 -0.2800000 -0.24 -0.2000000 -0.24 -0.2000000
#> [24,] -0.24 -0.1600000 -0.24 -0.24 -0.2800000 -0.24 -0.2000000 -0.24 -0.2000000
#> [25,]  0.72  0.8133333  0.72 -0.28 -0.3266667 -0.28 -0.2333333 -0.28 -0.2333333
#>       [,10] [,11] [,12]      [,13]      [,14]      [,15] [,16]      [,17]
#>  [1,] -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.2000000 -0.24 -0.2800000
#>  [2,] -0.16 -0.16  0.84 -0.1333333 -0.1333333 -0.1333333 -0.16 -0.1866667
#>  [3,] -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.2000000 -0.24 -0.2800000
#>  [4,] -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.2000000 -0.24 -0.2800000
#>  [5,] -0.28 -0.28 -0.28 -0.2333333 -0.2333333 -0.2333333 -0.28 -0.3266667
#>  [6,] -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.2000000 -0.24 -0.2800000
#>  [7,]  0.80 -0.20 -0.20 -0.1666667 -0.1666667 -0.1666667 -0.20 -0.2333333
#>  [8,]  0.76  0.76 -0.24 -0.2000000 -0.2000000 -0.2000000 -0.24 -0.2800000
#>  [9,]  0.80  0.80 -0.20 -0.1666667 -0.1666667 -0.1666667 -0.20 -0.2333333
#> [10,] -0.24  0.76  0.76 -0.2000000 -0.2000000 -0.2000000 -0.24 -0.2800000
#> [11,]  0.76 -0.24  0.76  0.8000000 -0.2000000 -0.2000000 -0.24 -0.2800000
#> [12,]  0.76  0.76 -0.24  0.8000000  0.8000000  0.8000000 -0.24 -0.2800000
#> [13,] -0.20  0.80  0.80 -0.1666667  0.8333333 -0.1666667  0.80  0.7666667
#> [14,] -0.20 -0.20  0.80  0.8333333 -0.1666667  0.8333333  0.80 -0.2333333
#> [15,] -0.20 -0.20  0.80 -0.1666667  0.8333333 -0.1666667  0.80  0.7666667
#> [16,] -0.24 -0.24 -0.24  0.8000000  0.8000000  0.8000000 -0.24  0.7200000
#> [17,] -0.28 -0.28 -0.28  0.7666667 -0.2333333  0.7666667  0.72 -0.3266667
#> [18,] -0.32 -0.32 -0.32 -0.2666667 -0.2666667  0.7333333  0.68  0.6266667
#> [19,]  0.72 -0.28 -0.28 -0.2333333 -0.2333333 -0.2333333  0.72  0.6733333
#> [20,] -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.2000000 -0.24  0.7200000
#> [21,] -0.28 -0.28 -0.28 -0.2333333 -0.2333333 -0.2333333 -0.28 -0.3266667
#> [22,] -0.24  0.76 -0.24 -0.2000000 -0.2000000 -0.2000000 -0.24 -0.2800000
#> [23,] -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.2000000 -0.24 -0.2800000
#> [24,] -0.24 -0.24 -0.24 -0.2000000  0.8000000 -0.2000000 -0.24 -0.2800000
#> [25,] -0.28 -0.28 -0.28 -0.2333333 -0.2333333 -0.2333333 -0.28  0.6733333
#>            [,18]      [,19] [,20]      [,21] [,22] [,23] [,24]      [,25]
#>  [1,] -0.3200000 -0.2800000 -0.24 -0.2800000 -0.24  0.76 -0.24  0.7200000
#>  [2,] -0.2133333 -0.1866667 -0.16 -0.1866667 -0.16 -0.16 -0.16  0.8133333
#>  [3,] -0.3200000 -0.2800000 -0.24  0.7200000 -0.24 -0.24 -0.24  0.7200000
#>  [4,]  0.6800000 -0.2800000 -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.2800000
#>  [5,]  0.6266667 -0.3266667 -0.28 -0.3266667 -0.28 -0.28 -0.28 -0.3266667
#>  [6,] -0.3200000 -0.2800000 -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.2800000
#>  [7,] -0.2666667 -0.2333333 -0.20 -0.2333333 -0.20 -0.20 -0.20 -0.2333333
#>  [8,] -0.3200000 -0.2800000 -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.2800000
#>  [9,] -0.2666667 -0.2333333 -0.20 -0.2333333 -0.20 -0.20 -0.20 -0.2333333
#> [10,] -0.3200000  0.7200000 -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.2800000
#> [11,] -0.3200000 -0.2800000 -0.24 -0.2800000  0.76 -0.24 -0.24 -0.2800000
#> [12,] -0.3200000 -0.2800000 -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.2800000
#> [13,] -0.2666667 -0.2333333 -0.20 -0.2333333 -0.20 -0.20 -0.20 -0.2333333
#> [14,] -0.2666667 -0.2333333 -0.20 -0.2333333 -0.20 -0.20  0.80 -0.2333333
#> [15,]  0.7333333 -0.2333333 -0.20 -0.2333333 -0.20 -0.20 -0.20 -0.2333333
#> [16,]  0.6800000  0.7200000 -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.2800000
#> [17,]  0.6266667  0.6733333  0.72 -0.3266667 -0.28 -0.28 -0.28  0.6733333
#> [18,] -0.4266667  0.6266667  0.68  0.6266667 -0.32 -0.32 -0.32 -0.3733333
#> [19,]  0.6266667 -0.3266667  0.72  0.6733333 -0.28 -0.28  0.72 -0.3266667
#> [20,]  0.6800000  0.7200000 -0.24  0.7200000  0.76  0.76 -0.24 -0.2800000
#> [21,]  0.6266667  0.6733333  0.72 -0.3266667  0.72  0.72  0.72 -0.3266667
#> [22,] -0.3200000 -0.2800000  0.76  0.7200000 -0.24  0.76  0.76  0.7200000
#> [23,] -0.3200000 -0.2800000  0.76  0.7200000  0.76 -0.24  0.76  0.7200000
#> [24,] -0.3200000  0.7200000 -0.24  0.7200000  0.76  0.76 -0.24  0.7200000
#> [25,] -0.3733333 -0.3266667 -0.28 -0.3266667  0.72  0.72  0.72 -0.3266667
#> attr(,"representation")
#> [1] "modularity"
repr_graphon(g)
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.55555556 0.41666667 0.67777778 0.47222222 0.50000000 0.27777778
#>  [2,] 0.41666667 0.12500000 0.22500000 0.25000000 0.41666667 0.18055556
#>  [3,] 0.67777778 0.22500000 0.40000000 0.45000000 0.73333333 0.36666667
#>  [4,] 0.47222222 0.25000000 0.45000000 0.50000000 0.65277778 0.53472222
#>  [5,] 0.50000000 0.41666667 0.73333333 0.65277778 0.44444444 0.55555556
#>  [6,] 0.27777778 0.18055556 0.36666667 0.53472222 0.55555556 0.55555556
#>  [7,] 0.23611111 0.12500000 0.22500000 0.50000000 0.35416667 0.59027778
#>  [8,] 0.22222222 0.05555556 0.11111111 0.35416667 0.44444444 0.55555556
#>  [9,] 0.11111111 0.05555556 0.11111111 0.23611111 0.27777778 0.50000000
#> [10,] 0.00000000 0.00000000 0.07142857 0.26785714 0.25396825 0.43650794
#> [11,] 0.05555556 0.12500000 0.06250000 0.25000000 0.18055556 0.35416667
#> [12,] 0.05555556 0.32142857 0.10000000 0.06250000 0.07142857 0.25396825
#> [13,] 0.06250000 0.12500000 0.00000000 0.00000000 0.06250000 0.11805556
#> [14,] 0.11805556 0.18750000 0.00000000 0.00000000 0.06250000 0.00000000
#> [15,] 0.07142857 0.06250000 0.00000000 0.00000000 0.07142857 0.00000000
#> [16,] 0.00000000 0.25000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [17,] 0.05555556 0.12500000 0.08333333 0.08333333 0.13888889 0.00000000
#> [18,] 0.19444444 0.12500000 0.10000000 0.29166667 0.33333333 0.13888889
#> [19,] 0.11805556 0.00000000 0.06250000 0.06250000 0.11805556 0.05555556
#> [20,] 0.16666667 0.11805556 0.21111111 0.00000000 0.05555556 0.00000000
#> [21,] 0.29166667 0.18750000 0.32500000 0.06250000 0.05555556 0.05555556
#> [22,] 0.23611111 0.12500000 0.22500000 0.12500000 0.23611111 0.11805556
#> [23,] 0.44444444 0.11805556 0.21111111 0.11805556 0.27777778 0.11111111
#> [24,] 0.23611111 0.18750000 0.22500000 0.12500000 0.23611111 0.06250000
#> [25,] 0.52777778 0.50000000 0.73333333 0.35416667 0.19444444 0.19444444
#>            [,7]       [,8]       [,9]      [,10]      [,11]      [,12]
#>  [1,] 0.2361111 0.22222222 0.11111111 0.00000000 0.05555556 0.05555556
#>  [2,] 0.1250000 0.05555556 0.05555556 0.00000000 0.12500000 0.32142857
#>  [3,] 0.2250000 0.11111111 0.11111111 0.07142857 0.06250000 0.10000000
#>  [4,] 0.5000000 0.35416667 0.23611111 0.26785714 0.25000000 0.06250000
#>  [5,] 0.3541667 0.44444444 0.27777778 0.25396825 0.18055556 0.07142857
#>  [6,] 0.5902778 0.55555556 0.50000000 0.43650794 0.35416667 0.25396825
#>  [7,] 0.5000000 0.65277778 0.59027778 0.53571429 0.43750000 0.26785714
#>  [8,] 0.6527778 0.55555556 0.55555556 0.63492063 0.53472222 0.25396825
#>  [9,] 0.5902778 0.55555556 0.55555556 0.63492063 0.59027778 0.30952381
#> [10,] 0.5357143 0.63492063 0.63492063 0.57142857 0.59821429 0.50000000
#> [11,] 0.4375000 0.53472222 0.59027778 0.59821429 0.50000000 0.47321429
#> [12,] 0.2678571 0.25396825 0.30952381 0.50000000 0.47321429 0.42857143
#> [13,] 0.1875000 0.23611111 0.17361111 0.19642857 0.43750000 0.59821429
#> [14,] 0.0000000 0.11805556 0.17361111 0.19642857 0.18750000 0.52678571
#> [15,] 0.0000000 0.05555556 0.05555556 0.07142857 0.19642857 0.50000000
#> [16,] 0.0000000 0.00000000 0.05555556 0.13392857 0.18750000 0.33035714
#> [17,] 0.0000000 0.00000000 0.05555556 0.15476190 0.00000000 0.22619048
#> [18,] 0.0625000 0.05555556 0.05555556 0.08333333 0.00000000 0.07142857
#> [19,] 0.0625000 0.05555556 0.05555556 0.13392857 0.12500000 0.13392857
#> [20,] 0.0000000 0.00000000 0.00000000 0.05555556 0.05555556 0.05555556
#> [21,] 0.0625000 0.05555556 0.00000000 0.06250000 0.06250000 0.00000000
#> [22,] 0.0625000 0.05555556 0.05555556 0.13392857 0.12500000 0.07142857
#> [23,] 0.0000000 0.00000000 0.00000000 0.05555556 0.05555556 0.05555556
#> [24,] 0.0000000 0.00000000 0.00000000 0.00000000 0.12500000 0.07142857
#> [25,] 0.0625000 0.13888889 0.00000000 0.00000000 0.00000000 0.00000000
#>            [,13]      [,14]      [,15]      [,16]      [,17]      [,18]
#>  [1,] 0.06250000 0.11805556 0.07142857 0.00000000 0.05555556 0.19444444
#>  [2,] 0.12500000 0.18750000 0.06250000 0.25000000 0.12500000 0.12500000
#>  [3,] 0.00000000 0.00000000 0.00000000 0.00000000 0.08333333 0.10000000
#>  [4,] 0.00000000 0.00000000 0.00000000 0.00000000 0.08333333 0.29166667
#>  [5,] 0.06250000 0.06250000 0.07142857 0.00000000 0.13888889 0.33333333
#>  [6,] 0.11805556 0.00000000 0.00000000 0.00000000 0.00000000 0.13888889
#>  [7,] 0.18750000 0.00000000 0.00000000 0.00000000 0.00000000 0.06250000
#>  [8,] 0.23611111 0.11805556 0.05555556 0.00000000 0.00000000 0.05555556
#>  [9,] 0.17361111 0.17361111 0.05555556 0.05555556 0.05555556 0.05555556
#> [10,] 0.19642857 0.19642857 0.07142857 0.13392857 0.15476190 0.08333333
#> [11,] 0.43750000 0.18750000 0.19642857 0.18750000 0.00000000 0.00000000
#> [12,] 0.59821429 0.52678571 0.50000000 0.33035714 0.22619048 0.07142857
#> [13,] 0.37500000 0.56250000 0.33035714 0.43750000 0.43750000 0.29166667
#> [14,] 0.56250000 0.50000000 0.46428571 0.56250000 0.27083333 0.27083333
#> [15,] 0.33035714 0.46428571 0.28571429 0.53571429 0.69047619 0.53571429
#> [16,] 0.43750000 0.56250000 0.53571429 0.50000000 0.58333333 0.50000000
#> [17,] 0.43750000 0.27083333 0.69047619 0.58333333 0.50000000 0.66666667
#> [18,] 0.29166667 0.27083333 0.53571429 0.50000000 0.66666667 0.66666667
#> [19,] 0.18750000 0.18750000 0.26785714 0.50000000 0.58333333 0.64583333
#> [20,] 0.05555556 0.22916667 0.12698413 0.17361111 0.61111111 0.52777778
#> [21,] 0.06250000 0.06250000 0.20535714 0.25000000 0.35416667 0.58333333
#> [22,] 0.00000000 0.12500000 0.07142857 0.18750000 0.20833333 0.27083333
#> [23,] 0.00000000 0.05555556 0.07142857 0.18055556 0.25000000 0.25000000
#> [24,] 0.12500000 0.18750000 0.20535714 0.06250000 0.37500000 0.27083333
#> [25,] 0.06250000 0.12500000 0.07142857 0.12500000 0.25000000 0.41666667
#>            [,19]      [,20]      [,21]      [,22]      [,23]      [,24]
#>  [1,] 0.11805556 0.16666667 0.29166667 0.23611111 0.44444444 0.23611111
#>  [2,] 0.00000000 0.11805556 0.18750000 0.12500000 0.11805556 0.18750000
#>  [3,] 0.06250000 0.21111111 0.32500000 0.22500000 0.21111111 0.22500000
#>  [4,] 0.06250000 0.00000000 0.06250000 0.12500000 0.11805556 0.12500000
#>  [5,] 0.11805556 0.05555556 0.05555556 0.23611111 0.27777778 0.23611111
#>  [6,] 0.05555556 0.00000000 0.05555556 0.11805556 0.11111111 0.06250000
#>  [7,] 0.06250000 0.00000000 0.06250000 0.06250000 0.00000000 0.00000000
#>  [8,] 0.05555556 0.00000000 0.05555556 0.05555556 0.00000000 0.00000000
#>  [9,] 0.05555556 0.00000000 0.00000000 0.05555556 0.00000000 0.00000000
#> [10,] 0.13392857 0.05555556 0.06250000 0.13392857 0.05555556 0.00000000
#> [11,] 0.12500000 0.05555556 0.06250000 0.12500000 0.05555556 0.12500000
#> [12,] 0.13392857 0.05555556 0.00000000 0.07142857 0.05555556 0.07142857
#> [13,] 0.18750000 0.05555556 0.06250000 0.00000000 0.00000000 0.12500000
#> [14,] 0.18750000 0.22916667 0.06250000 0.12500000 0.05555556 0.18750000
#> [15,] 0.26785714 0.12698413 0.20535714 0.07142857 0.07142857 0.20535714
#> [16,] 0.50000000 0.17361111 0.25000000 0.18750000 0.18055556 0.06250000
#> [17,] 0.58333333 0.61111111 0.35416667 0.20833333 0.25000000 0.37500000
#> [18,] 0.64583333 0.52777778 0.58333333 0.27083333 0.25000000 0.27083333
#> [19,] 0.50000000 0.59722222 0.56250000 0.37500000 0.35416667 0.50000000
#> [20,] 0.59722222 0.44444444 0.59027778 0.52777778 0.50000000 0.46527778
#> [21,] 0.56250000 0.59027778 0.62500000 0.68750000 0.64583333 0.62500000
#> [22,] 0.37500000 0.52777778 0.68750000 0.50000000 0.59027778 0.50000000
#> [23,] 0.35416667 0.50000000 0.64583333 0.59027778 0.55555556 0.59027778
#> [24,] 0.50000000 0.46527778 0.62500000 0.50000000 0.59027778 0.37500000
#> [25,] 0.35416667 0.33333333 0.41666667 0.64583333 0.61111111 0.56250000
#>            [,25]
#>  [1,] 0.52777778
#>  [2,] 0.50000000
#>  [3,] 0.73333333
#>  [4,] 0.35416667
#>  [5,] 0.19444444
#>  [6,] 0.19444444
#>  [7,] 0.06250000
#>  [8,] 0.13888889
#>  [9,] 0.00000000
#> [10,] 0.00000000
#> [11,] 0.00000000
#> [12,] 0.00000000
#> [13,] 0.06250000
#> [14,] 0.12500000
#> [15,] 0.07142857
#> [16,] 0.12500000
#> [17,] 0.25000000
#> [18,] 0.41666667
#> [19,] 0.35416667
#> [20,] 0.33333333
#> [21,] 0.41666667
#> [22,] 0.64583333
#> [23,] 0.61111111
#> [24,] 0.56250000
#> [25,] 0.33333333
#> attr(,"representation")
#> [1] "graphon"