Network Representation Functions

This is a collection of functions that convert a graph stored as an igraph object into a desired matrix representation among adjacency matrix, graph laplacian, modularity matrix or graphon (edge probability matrix).

Usage

repr_adjacency(network, validate = TRUE)

repr_laplacian(network, validate = TRUE)

repr_modularity(network, validate = TRUE)

repr_graphon(network, validate = TRUE)

Arguments

network

An igraph object.

validate

A boolean specifying whether the function should check the class of its input (default: TRUE).

Value

A numeric square matrix giving the desired network representation recorded in the object's class.

Examples

g <- igraph::sample_smallworld(1, 25, 3, 0.05)
repr_adjacency(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0    1    0    1    0    0    0    0    0     0     0     1     0
#>  [2,]    1    0    1    1    1    0    0    0    0     0     0     0     0
#>  [3,]    0    1    0    1    1    0    0    1    1     0     0     0     0
#>  [4,]    1    1    1    0    0    1    1    0    0     0     0     0     0
#>  [5,]    0    1    1    0    0    1    1    1    0     0     0     0     0
#>  [6,]    0    0    0    1    1    0    1    1    1     0     0     0     0
#>  [7,]    0    0    0    1    1    1    0    1    1     0     0     0     1
#>  [8,]    0    0    1    0    1    1    1    0    1     1     1     0     0
#>  [9,]    0    0    1    0    0    1    1    1    0     1     1     1     0
#> [10,]    0    0    0    0    0    0    0    1    1     0     1     1     1
#> [11,]    0    0    0    0    0    0    0    1    1     1     0     1     1
#> [12,]    1    0    0    0    0    0    0    0    1     1     1     0     1
#> [13,]    0    0    0    0    0    0    1    0    0     1     1     1     0
#> [14,]    0    0    0    0    1    0    0    0    0     0     1     1     1
#> [15,]    0    0    0    0    0    0    0    0    0     0     0     1     1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0     1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    1    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]    1    0    0    1    0    0    0    0    0     1     0     0     0
#> [24,]    0    1    0    0    0    0    0    0    0     0     0     0     0
#> [25,]    1    1    1    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0     1     0     1
#>  [2,]     0     0     0     0     0     0     0     0     0     0     1     1
#>  [3,]     0     0     0     0     0     0     0     0     0     0     0     1
#>  [4,]     0     0     0     0     0     0     0     1     0     1     0     0
#>  [5,]     1     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     1     0     0
#> [11,]     1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]     1     1     0     0     0     0     0     0     0     0     0     0
#> [13,]     1     1     1     0     0     0     0     0     0     0     0     0
#> [14,]     0     1     1     0     0     0     0     0     0     0     0     1
#> [15,]     1     0     1     1     1     0     0     0     0     0     0     0
#> [16,]     1     1     0     1     1     1     0     0     0     0     0     0
#> [17,]     0     1     1     0     1     1     1     0     0     0     0     0
#> [18,]     0     1     1     1     0     1     1     1     0     0     0     0
#> [19,]     0     0     1     1     1     0     1     1     1     0     0     0
#> [20,]     0     0     0     1     1     1     0     1     1     0     0     0
#> [21,]     0     0     0     0     1     1     1     0     1     0     1     0
#> [22,]     0     0     0     0     0     1     1     1     0     1     1     1
#> [23,]     0     0     0     0     0     0     0     0     1     0     1     0
#> [24,]     0     0     0     0     0     0     0     1     1     1     0     1
#> [25,]     1     0     0     0     0     0     0     0     1     0     1     0
#> attr(,"representation")
#> [1] "adjacency"
repr_laplacian(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    5   -1    0   -1    0    0    0    0    0     0     0    -1     0
#>  [2,]   -1    6   -1   -1   -1    0    0    0    0     0     0     0     0
#>  [3,]    0   -1    6   -1   -1    0    0   -1   -1     0     0     0     0
#>  [4,]   -1   -1   -1    7    0   -1   -1    0    0     0     0     0     0
#>  [5,]    0   -1   -1    0    6   -1   -1   -1    0     0     0     0     0
#>  [6,]    0    0    0   -1   -1    5   -1   -1   -1     0     0     0     0
#>  [7,]    0    0    0   -1   -1   -1    6   -1   -1     0     0     0    -1
#>  [8,]    0    0   -1    0   -1   -1   -1    7   -1    -1    -1     0     0
#>  [9,]    0    0   -1    0    0   -1   -1   -1    7    -1    -1    -1     0
#> [10,]    0    0    0    0    0    0    0   -1   -1     6    -1    -1    -1
#> [11,]    0    0    0    0    0    0    0   -1   -1    -1     6    -1    -1
#> [12,]   -1    0    0    0    0    0    0    0   -1    -1    -1     7    -1
#> [13,]    0    0    0    0    0    0   -1    0    0    -1    -1    -1     7
#> [14,]    0    0    0    0   -1    0    0    0    0     0    -1    -1    -1
#> [15,]    0    0    0    0    0    0    0    0    0     0     0    -1    -1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0    -1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0   -1    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]   -1    0    0   -1    0    0    0    0    0    -1     0     0     0
#> [24,]    0   -1    0    0    0    0    0    0    0     0     0     0     0
#> [25,]   -1   -1   -1    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0    -1     0    -1
#>  [2,]     0     0     0     0     0     0     0     0     0     0    -1    -1
#>  [3,]     0     0     0     0     0     0     0     0     0     0     0    -1
#>  [4,]     0     0     0     0     0     0     0    -1     0    -1     0     0
#>  [5,]    -1     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0    -1     0     0
#> [11,]    -1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]    -1    -1     0     0     0     0     0     0     0     0     0     0
#> [13,]    -1    -1    -1     0     0     0     0     0     0     0     0     0
#> [14,]     7    -1    -1     0     0     0     0     0     0     0     0    -1
#> [15,]    -1     6    -1    -1    -1     0     0     0     0     0     0     0
#> [16,]    -1    -1     6    -1    -1    -1     0     0     0     0     0     0
#> [17,]     0    -1    -1     5    -1    -1    -1     0     0     0     0     0
#> [18,]     0    -1    -1    -1     6    -1    -1    -1     0     0     0     0
#> [19,]     0     0    -1    -1    -1     6    -1    -1    -1     0     0     0
#> [20,]     0     0     0    -1    -1    -1     5    -1    -1     0     0     0
#> [21,]     0     0     0     0    -1    -1    -1     6    -1     0    -1     0
#> [22,]     0     0     0     0     0    -1    -1    -1     6    -1    -1    -1
#> [23,]     0     0     0     0     0     0     0     0    -1     5    -1     0
#> [24,]     0     0     0     0     0     0     0    -1    -1    -1     5    -1
#> [25,]    -1     0     0     0     0     0     0     0    -1     0    -1     6
#> attr(,"representation")
#> [1] "laplacian"
repr_modularity(g)
#>             [,1]  [,2]  [,3]       [,4]  [,5]       [,6]  [,7]       [,8]
#>  [1,] -0.1666667  0.80 -0.20  0.7666667 -0.20 -0.1666667 -0.20 -0.2333333
#>  [2,]  0.8000000 -0.24  0.76  0.7200000  0.76 -0.2000000 -0.24 -0.2800000
#>  [3,] -0.2000000  0.76 -0.24  0.7200000  0.76 -0.2000000 -0.24  0.7200000
#>  [4,]  0.7666667  0.72  0.72 -0.3266667 -0.28  0.7666667  0.72 -0.3266667
#>  [5,] -0.2000000  0.76  0.76 -0.2800000 -0.24  0.8000000  0.76  0.7200000
#>  [6,] -0.1666667 -0.20 -0.20  0.7666667  0.80 -0.1666667  0.80  0.7666667
#>  [7,] -0.2000000 -0.24 -0.24  0.7200000  0.76  0.8000000 -0.24  0.7200000
#>  [8,] -0.2333333 -0.28  0.72 -0.3266667  0.72  0.7666667  0.72 -0.3266667
#>  [9,] -0.2333333 -0.28  0.72 -0.3266667 -0.28  0.7666667  0.72  0.6733333
#> [10,] -0.2000000 -0.24 -0.24 -0.2800000 -0.24 -0.2000000 -0.24  0.7200000
#> [11,] -0.2000000 -0.24 -0.24 -0.2800000 -0.24 -0.2000000 -0.24  0.7200000
#> [12,]  0.7666667 -0.28 -0.28 -0.3266667 -0.28 -0.2333333 -0.28 -0.3266667
#> [13,] -0.2333333 -0.28 -0.28 -0.3266667 -0.28 -0.2333333  0.72 -0.3266667
#> [14,] -0.2333333 -0.28 -0.28 -0.3266667  0.72 -0.2333333 -0.28 -0.3266667
#> [15,] -0.2000000 -0.24 -0.24 -0.2800000 -0.24 -0.2000000 -0.24 -0.2800000
#> [16,] -0.2000000 -0.24 -0.24 -0.2800000 -0.24 -0.2000000 -0.24 -0.2800000
#> [17,] -0.1666667 -0.20 -0.20 -0.2333333 -0.20 -0.1666667 -0.20 -0.2333333
#> [18,] -0.2000000 -0.24 -0.24 -0.2800000 -0.24 -0.2000000 -0.24 -0.2800000
#> [19,] -0.2000000 -0.24 -0.24 -0.2800000 -0.24 -0.2000000 -0.24 -0.2800000
#> [20,] -0.1666667 -0.20 -0.20 -0.2333333 -0.20 -0.1666667 -0.20 -0.2333333
#> [21,] -0.2000000 -0.24 -0.24  0.7200000 -0.24 -0.2000000 -0.24 -0.2800000
#> [22,] -0.2000000 -0.24 -0.24 -0.2800000 -0.24 -0.2000000 -0.24 -0.2800000
#> [23,]  0.8333333 -0.20 -0.20  0.7666667 -0.20 -0.1666667 -0.20 -0.2333333
#> [24,] -0.1666667  0.80 -0.20 -0.2333333 -0.20 -0.1666667 -0.20 -0.2333333
#> [25,]  0.8000000  0.76  0.76 -0.2800000 -0.24 -0.2000000 -0.24 -0.2800000
#>             [,9] [,10] [,11]      [,12]      [,13]      [,14] [,15] [,16]
#>  [1,] -0.2333333 -0.20 -0.20  0.7666667 -0.2333333 -0.2333333 -0.20 -0.20
#>  [2,] -0.2800000 -0.24 -0.24 -0.2800000 -0.2800000 -0.2800000 -0.24 -0.24
#>  [3,]  0.7200000 -0.24 -0.24 -0.2800000 -0.2800000 -0.2800000 -0.24 -0.24
#>  [4,] -0.3266667 -0.28 -0.28 -0.3266667 -0.3266667 -0.3266667 -0.28 -0.28
#>  [5,] -0.2800000 -0.24 -0.24 -0.2800000 -0.2800000  0.7200000 -0.24 -0.24
#>  [6,]  0.7666667 -0.20 -0.20 -0.2333333 -0.2333333 -0.2333333 -0.20 -0.20
#>  [7,]  0.7200000 -0.24 -0.24 -0.2800000  0.7200000 -0.2800000 -0.24 -0.24
#>  [8,]  0.6733333  0.72  0.72 -0.3266667 -0.3266667 -0.3266667 -0.28 -0.28
#>  [9,] -0.3266667  0.72  0.72  0.6733333 -0.3266667 -0.3266667 -0.28 -0.28
#> [10,]  0.7200000 -0.24  0.76  0.7200000  0.7200000 -0.2800000 -0.24 -0.24
#> [11,]  0.7200000  0.76 -0.24  0.7200000  0.7200000  0.7200000 -0.24 -0.24
#> [12,]  0.6733333  0.72  0.72 -0.3266667  0.6733333  0.6733333  0.72 -0.28
#> [13,] -0.3266667  0.72  0.72  0.6733333 -0.3266667  0.6733333  0.72  0.72
#> [14,] -0.3266667 -0.28  0.72  0.6733333  0.6733333 -0.3266667  0.72  0.72
#> [15,] -0.2800000 -0.24 -0.24  0.7200000  0.7200000  0.7200000 -0.24  0.76
#> [16,] -0.2800000 -0.24 -0.24 -0.2800000  0.7200000  0.7200000  0.76 -0.24
#> [17,] -0.2333333 -0.20 -0.20 -0.2333333 -0.2333333 -0.2333333  0.80  0.80
#> [18,] -0.2800000 -0.24 -0.24 -0.2800000 -0.2800000 -0.2800000  0.76  0.76
#> [19,] -0.2800000 -0.24 -0.24 -0.2800000 -0.2800000 -0.2800000 -0.24  0.76
#> [20,] -0.2333333 -0.20 -0.20 -0.2333333 -0.2333333 -0.2333333 -0.20 -0.20
#> [21,] -0.2800000 -0.24 -0.24 -0.2800000 -0.2800000 -0.2800000 -0.24 -0.24
#> [22,] -0.2800000 -0.24 -0.24 -0.2800000 -0.2800000 -0.2800000 -0.24 -0.24
#> [23,] -0.2333333  0.80 -0.20 -0.2333333 -0.2333333 -0.2333333 -0.20 -0.20
#> [24,] -0.2333333 -0.20 -0.20 -0.2333333 -0.2333333 -0.2333333 -0.20 -0.20
#> [25,] -0.2800000 -0.24 -0.24 -0.2800000 -0.2800000  0.7200000 -0.24 -0.24
#>            [,17] [,18] [,19]      [,20] [,21] [,22]      [,23]      [,24] [,25]
#>  [1,] -0.1666667 -0.20 -0.20 -0.1666667 -0.20 -0.20  0.8333333 -0.1666667  0.80
#>  [2,] -0.2000000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2000000  0.8000000  0.76
#>  [3,] -0.2000000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2000000 -0.2000000  0.76
#>  [4,] -0.2333333 -0.28 -0.28 -0.2333333  0.72 -0.28  0.7666667 -0.2333333 -0.28
#>  [5,] -0.2000000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2000000 -0.2000000 -0.24
#>  [6,] -0.1666667 -0.20 -0.20 -0.1666667 -0.20 -0.20 -0.1666667 -0.1666667 -0.20
#>  [7,] -0.2000000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2000000 -0.2000000 -0.24
#>  [8,] -0.2333333 -0.28 -0.28 -0.2333333 -0.28 -0.28 -0.2333333 -0.2333333 -0.28
#>  [9,] -0.2333333 -0.28 -0.28 -0.2333333 -0.28 -0.28 -0.2333333 -0.2333333 -0.28
#> [10,] -0.2000000 -0.24 -0.24 -0.2000000 -0.24 -0.24  0.8000000 -0.2000000 -0.24
#> [11,] -0.2000000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2000000 -0.2000000 -0.24
#> [12,] -0.2333333 -0.28 -0.28 -0.2333333 -0.28 -0.28 -0.2333333 -0.2333333 -0.28
#> [13,] -0.2333333 -0.28 -0.28 -0.2333333 -0.28 -0.28 -0.2333333 -0.2333333 -0.28
#> [14,] -0.2333333 -0.28 -0.28 -0.2333333 -0.28 -0.28 -0.2333333 -0.2333333  0.72
#> [15,]  0.8000000  0.76 -0.24 -0.2000000 -0.24 -0.24 -0.2000000 -0.2000000 -0.24
#> [16,]  0.8000000  0.76  0.76 -0.2000000 -0.24 -0.24 -0.2000000 -0.2000000 -0.24
#> [17,] -0.1666667  0.80  0.80  0.8333333 -0.20 -0.20 -0.1666667 -0.1666667 -0.20
#> [18,]  0.8000000 -0.24  0.76  0.8000000  0.76 -0.24 -0.2000000 -0.2000000 -0.24
#> [19,]  0.8000000  0.76 -0.24  0.8000000  0.76  0.76 -0.2000000 -0.2000000 -0.24
#> [20,]  0.8333333  0.80  0.80 -0.1666667  0.80  0.80 -0.1666667 -0.1666667 -0.20
#> [21,] -0.2000000  0.76  0.76  0.8000000 -0.24  0.76 -0.2000000  0.8000000 -0.24
#> [22,] -0.2000000 -0.24  0.76  0.8000000  0.76 -0.24  0.8000000  0.8000000  0.76
#> [23,] -0.1666667 -0.20 -0.20 -0.1666667 -0.20  0.80 -0.1666667  0.8333333 -0.20
#> [24,] -0.1666667 -0.20 -0.20 -0.1666667  0.80  0.80  0.8333333 -0.1666667  0.80
#> [25,] -0.2000000 -0.24 -0.24 -0.2000000 -0.24  0.76 -0.2000000  0.8000000 -0.24
#> attr(,"representation")
#> [1] "modularity"
repr_graphon(g)
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.25000000 0.35416667 0.11805556 0.47916667 0.29861111 0.11111111
#>  [2,] 0.35416667 0.33333333 0.33333333 0.66666667 0.44444444 0.22222222
#>  [3,] 0.11805556 0.33333333 0.22222222 0.61111111 0.50000000 0.22222222
#>  [4,] 0.47916667 0.66666667 0.61111111 0.22222222 0.27777778 0.55555556
#>  [5,] 0.29861111 0.44444444 0.50000000 0.27777778 0.44444444 0.50000000
#>  [6,] 0.11111111 0.22222222 0.22222222 0.55555556 0.50000000 0.22222222
#>  [7,] 0.12500000 0.18055556 0.29861111 0.47222222 0.59027778 0.29861111
#>  [8,] 0.31250000 0.29166667 0.58333333 0.23611111 0.52083333 0.64583333
#>  [9,] 0.25892857 0.30952381 0.56349206 0.05555556 0.34920635 0.56349206
#> [10,] 0.06250000 0.19841270 0.05555556 0.39682540 0.38095238 0.18253968
#> [11,] 0.11805556 0.11111111 0.11111111 0.27777778 0.27777778 0.16666667
#> [12,] 0.19642857 0.05555556 0.23809524 0.11111111 0.25396825 0.23809524
#> [13,] 0.13392857 0.05555556 0.30952381 0.00000000 0.25396825 0.30952381
#> [14,] 0.10000000 0.05555556 0.11111111 0.11111111 0.26666667 0.11111111
#> [15,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [16,] 0.05555556 0.00000000 0.00000000 0.05555556 0.00000000 0.00000000
#> [17,] 0.00000000 0.00000000 0.00000000 0.06250000 0.00000000 0.00000000
#> [18,] 0.00000000 0.11111111 0.00000000 0.05555556 0.00000000 0.00000000
#> [19,] 0.06250000 0.16666667 0.00000000 0.11111111 0.00000000 0.00000000
#> [20,] 0.06250000 0.16666667 0.00000000 0.11111111 0.00000000 0.00000000
#> [21,] 0.20833333 0.05555556 0.05555556 0.33333333 0.05555556 0.00000000
#> [22,] 0.18055556 0.27777778 0.11111111 0.27777778 0.11111111 0.05555556
#> [23,] 0.43750000 0.29861111 0.29166667 0.53472222 0.18055556 0.11111111
#> [24,] 0.35416667 0.55555556 0.16666667 0.33333333 0.11111111 0.16666667
#> [25,] 0.59821429 0.50793651 0.50793651 0.43650794 0.12698413 0.30952381
#>            [,7]       [,8]       [,9]      [,10]      [,11]      [,12]
#>  [1,] 0.1250000 0.31250000 0.25892857 0.06250000 0.11805556 0.19642857
#>  [2,] 0.1805556 0.29166667 0.30952381 0.19841270 0.11111111 0.05555556
#>  [3,] 0.2986111 0.58333333 0.56349206 0.05555556 0.11111111 0.23809524
#>  [4,] 0.4722222 0.23611111 0.05555556 0.39682540 0.27777778 0.11111111
#>  [5,] 0.5902778 0.52083333 0.34920635 0.38095238 0.27777778 0.25396825
#>  [6,] 0.2986111 0.64583333 0.56349206 0.18253968 0.16666667 0.23809524
#>  [7,] 0.3750000 0.75000000 0.73214286 0.19642857 0.23611111 0.26785714
#>  [8,] 0.7500000 0.50000000 0.53571429 0.60714286 0.58333333 0.26785714
#>  [9,] 0.7321429 0.53571429 0.42857143 0.64285714 0.61904762 0.57142857
#> [10,] 0.1964286 0.60714286 0.64285714 0.14285714 0.18253968 0.57142857
#> [11,] 0.2361111 0.58333333 0.61904762 0.18253968 0.22222222 0.52380952
#> [12,] 0.2678571 0.26785714 0.57142857 0.57142857 0.52380952 0.57142857
#> [13,] 0.4017857 0.26785714 0.35714286 0.64285714 0.69047619 0.71428571
#> [14,] 0.2250000 0.28750000 0.41428571 0.34285714 0.67777778 0.68571429
#> [15,] 0.0000000 0.25000000 0.19642857 0.06250000 0.17361111 0.47321429
#> [16,] 0.0000000 0.11805556 0.18253968 0.18253968 0.11111111 0.25396825
#> [17,] 0.0000000 0.00000000 0.00000000 0.07142857 0.11111111 0.13392857
#> [18,] 0.0000000 0.00000000 0.00000000 0.07142857 0.11111111 0.12698413
#> [19,] 0.0000000 0.00000000 0.00000000 0.00000000 0.05555556 0.12698413
#> [20,] 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000 0.05555556
#> [21,] 0.0000000 0.06250000 0.07142857 0.08333333 0.00000000 0.00000000
#> [22,] 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000 0.05555556
#> [23,] 0.1250000 0.12500000 0.20535714 0.20535714 0.05555556 0.13392857
#> [24,] 0.1180556 0.05555556 0.05555556 0.05555556 0.00000000 0.05555556
#> [25,] 0.2678571 0.07142857 0.00000000 0.21428571 0.05555556 0.07142857
#>            [,13]      [,14]     [,15]      [,16]      [,17]      [,18]
#>  [1,] 0.13392857 0.10000000 0.0000000 0.05555556 0.00000000 0.00000000
#>  [2,] 0.05555556 0.05555556 0.0000000 0.00000000 0.00000000 0.11111111
#>  [3,] 0.30952381 0.11111111 0.0000000 0.00000000 0.00000000 0.00000000
#>  [4,] 0.00000000 0.11111111 0.0000000 0.05555556 0.06250000 0.05555556
#>  [5,] 0.25396825 0.26666667 0.0000000 0.00000000 0.00000000 0.00000000
#>  [6,] 0.30952381 0.11111111 0.0000000 0.00000000 0.00000000 0.00000000
#>  [7,] 0.40178571 0.22500000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [8,] 0.26785714 0.28750000 0.2500000 0.11805556 0.00000000 0.00000000
#>  [9,] 0.35714286 0.41428571 0.1964286 0.18253968 0.00000000 0.00000000
#> [10,] 0.64285714 0.34285714 0.0625000 0.18253968 0.07142857 0.07142857
#> [11,] 0.69047619 0.67777778 0.1736111 0.11111111 0.11111111 0.11111111
#> [12,] 0.71428571 0.68571429 0.4732143 0.25396825 0.13392857 0.12698413
#> [13,] 0.57142857 0.68571429 0.4642857 0.43650794 0.19642857 0.18253968
#> [14,] 0.68571429 0.60000000 0.4875000 0.52222222 0.22500000 0.21111111
#> [15,] 0.46428571 0.48750000 0.3750000 0.47222222 0.50000000 0.47916667
#> [16,] 0.43650794 0.52222222 0.4722222 0.44444444 0.52777778 0.55555556
#> [17,] 0.19642857 0.22500000 0.5000000 0.52777778 0.62500000 0.65277778
#> [18,] 0.18253968 0.21111111 0.4791667 0.55555556 0.65277778 0.66666667
#> [19,] 0.11111111 0.11111111 0.4166667 0.50000000 0.65277778 0.66666667
#> [20,] 0.11111111 0.11111111 0.3541667 0.44444444 0.59027778 0.61111111
#> [21,] 0.00000000 0.00000000 0.2708333 0.33333333 0.41666667 0.61111111
#> [22,] 0.00000000 0.00000000 0.2361111 0.33333333 0.35416667 0.44444444
#> [23,] 0.00000000 0.16250000 0.0625000 0.00000000 0.12500000 0.23611111
#> [24,] 0.00000000 0.05555556 0.0000000 0.11111111 0.23611111 0.22222222
#> [25,] 0.07142857 0.24285714 0.0000000 0.00000000 0.06250000 0.18253968
#>            [,19]      [,20]      [,21]      [,22]      [,23]      [,24]
#>  [1,] 0.06250000 0.06250000 0.20833333 0.18055556 0.43750000 0.35416667
#>  [2,] 0.16666667 0.16666667 0.05555556 0.27777778 0.29861111 0.55555556
#>  [3,] 0.00000000 0.00000000 0.05555556 0.11111111 0.29166667 0.16666667
#>  [4,] 0.11111111 0.11111111 0.33333333 0.27777778 0.53472222 0.33333333
#>  [5,] 0.00000000 0.00000000 0.05555556 0.11111111 0.18055556 0.11111111
#>  [6,] 0.00000000 0.00000000 0.00000000 0.05555556 0.11111111 0.16666667
#>  [7,] 0.00000000 0.00000000 0.00000000 0.00000000 0.12500000 0.11805556
#>  [8,] 0.00000000 0.00000000 0.06250000 0.00000000 0.12500000 0.05555556
#>  [9,] 0.00000000 0.00000000 0.07142857 0.00000000 0.20535714 0.05555556
#> [10,] 0.00000000 0.00000000 0.08333333 0.00000000 0.20535714 0.05555556
#> [11,] 0.05555556 0.00000000 0.00000000 0.00000000 0.05555556 0.00000000
#> [12,] 0.12698413 0.05555556 0.00000000 0.05555556 0.13392857 0.05555556
#> [13,] 0.11111111 0.11111111 0.00000000 0.00000000 0.00000000 0.00000000
#> [14,] 0.11111111 0.11111111 0.00000000 0.00000000 0.16250000 0.05555556
#> [15,] 0.41666667 0.35416667 0.27083333 0.23611111 0.06250000 0.00000000
#> [16,] 0.50000000 0.44444444 0.33333333 0.33333333 0.00000000 0.11111111
#> [17,] 0.65277778 0.59027778 0.41666667 0.35416667 0.12500000 0.23611111
#> [18,] 0.66666667 0.61111111 0.61111111 0.44444444 0.23611111 0.22222222
#> [19,] 0.66666667 0.61111111 0.61111111 0.50000000 0.23611111 0.22222222
#> [20,] 0.61111111 0.55555556 0.61111111 0.50000000 0.17361111 0.22222222
#> [21,] 0.61111111 0.61111111 0.50000000 0.61111111 0.20833333 0.52777778
#> [22,] 0.50000000 0.50000000 0.61111111 0.44444444 0.47916667 0.44444444
#> [23,] 0.23611111 0.17361111 0.20833333 0.47916667 0.25000000 0.47222222
#> [24,] 0.22222222 0.22222222 0.52777778 0.44444444 0.47222222 0.33333333
#> [25,] 0.18253968 0.18253968 0.22619048 0.50793651 0.33035714 0.50793651
#>            [,25]
#>  [1,] 0.59821429
#>  [2,] 0.50793651
#>  [3,] 0.50793651
#>  [4,] 0.43650794
#>  [5,] 0.12698413
#>  [6,] 0.30952381
#>  [7,] 0.26785714
#>  [8,] 0.07142857
#>  [9,] 0.00000000
#> [10,] 0.21428571
#> [11,] 0.05555556
#> [12,] 0.07142857
#> [13,] 0.07142857
#> [14,] 0.24285714
#> [15,] 0.00000000
#> [16,] 0.00000000
#> [17,] 0.06250000
#> [18,] 0.18253968
#> [19,] 0.18253968
#> [20,] 0.18253968
#> [21,] 0.22619048
#> [22,] 0.50793651
#> [23,] 0.33035714
#> [24,] 0.50793651
#> [25,] 0.28571429
#> attr(,"representation")
#> [1] "graphon"