Network Representation Functions

This is a collection of functions that convert a graph stored as an igraph object into a desired matrix representation among adjacency matrix, graph laplacian, modularity matrix or graphon (edge probability matrix).

repr_adjacency(network, validate = TRUE)

repr_laplacian(network, validate = TRUE)

repr_modularity(network, validate = TRUE)

repr_graphon(network, validate = TRUE)

Arguments

network
An igraph object.

validate
A boolean specifying whether the function should check the class of its input (default: TRUE).

Value

A numeric square matrix giving the desired network representation recorded in the object's class.

Examples

g <- igraph::sample_smallworld(1, 25, 3, 0.05)
repr_adjacency(g)

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0    1    1    1    0    0    0    0    0     0     0     0     0
#>  [2,]    1    0    1    1    1    0    0    0    0     0     0     0     0
#>  [3,]    1    1    0    1    1    1    0    0    0     0     0     0     0
#>  [4,]    1    1    1    0    1    1    0    0    0     1     0     0     0
#>  [5,]    0    1    1    1    0    1    1    1    0     0     0     0     0
#>  [6,]    0    0    1    1    1    0    1    1    1     0     0     0     0
#>  [7,]    0    0    0    0    1    1    0    1    1     1     0     0     0
#>  [8,]    0    0    0    0    1    1    1    0    1     1     1     0     0
#>  [9,]    0    0    0    0    0    1    1    1    0     1     1     1     0
#> [10,]    0    0    0    1    0    0    1    1    1     0     1     1     1
#> [11,]    0    0    0    0    0    0    0    1    1     1     0     1     1
#> [12,]    0    0    0    0    0    0    0    0    1     1     1     0     1
#> [13,]    0    0    0    0    0    0    0    0    0     1     1     1     0
#> [14,]    0    0    0    0    0    0    0    0    0     0     1     1     1
#> [15,]    0    0    0    0    0    0    0    0    0     0     0     1     1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0     1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]    1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]    1    1    0    0    0    0    0    0    0     0     0     0     0
#> [25,]    1    1    1    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0     1     1     1
#>  [2,]     0     0     0     0     0     0     0     0     0     0     1     1
#>  [3,]     0     0     0     0     0     0     0     0     0     0     0     1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]     1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]     1     1     0     0     0     0     0     0     0     0     0     0
#> [13,]     1     1     1     0     0     0     0     0     0     0     0     0
#> [14,]     0     1     1     1     0     0     0     0     0     0     0     0
#> [15,]     1     0     1     1     1     0     0     0     0     0     0     0
#> [16,]     1     1     0     1     1     1     0     0     0     0     0     0
#> [17,]     1     1     1     0     1     1     1     0     0     0     0     0
#> [18,]     0     1     1     1     0     1     1     1     0     0     0     0
#> [19,]     0     0     1     1     1     0     1     1     1     0     0     0
#> [20,]     0     0     0     1     1     1     0     1     1     1     0     0
#> [21,]     0     0     0     0     1     1     1     0     1     1     1     0
#> [22,]     0     0     0     0     0     1     1     1     0     1     1     1
#> [23,]     0     0     0     0     0     0     1     1     1     0     1     1
#> [24,]     0     0     0     0     0     0     0     1     1     1     0     1
#> [25,]     0     0     0     0     0     0     0     0     1     1     1     0
#> attr(,"representation")
#> [1] "adjacency"

repr_laplacian(g)

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    6   -1   -1   -1    0    0    0    0    0     0     0     0     0
#>  [2,]   -1    6   -1   -1   -1    0    0    0    0     0     0     0     0
#>  [3,]   -1   -1    6   -1   -1   -1    0    0    0     0     0     0     0
#>  [4,]   -1   -1   -1    6   -1   -1    0    0    0    -1     0     0     0
#>  [5,]    0   -1   -1   -1    6   -1   -1   -1    0     0     0     0     0
#>  [6,]    0    0   -1   -1   -1    6   -1   -1   -1     0     0     0     0
#>  [7,]    0    0    0    0   -1   -1    5   -1   -1    -1     0     0     0
#>  [8,]    0    0    0    0   -1   -1   -1    6   -1    -1    -1     0     0
#>  [9,]    0    0    0    0    0   -1   -1   -1    6    -1    -1    -1     0
#> [10,]    0    0    0   -1    0    0   -1   -1   -1     7    -1    -1    -1
#> [11,]    0    0    0    0    0    0    0   -1   -1    -1     6    -1    -1
#> [12,]    0    0    0    0    0    0    0    0   -1    -1    -1     6    -1
#> [13,]    0    0    0    0    0    0    0    0    0    -1    -1    -1     6
#> [14,]    0    0    0    0    0    0    0    0    0     0    -1    -1    -1
#> [15,]    0    0    0    0    0    0    0    0    0     0     0    -1    -1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0    -1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]   -1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]   -1   -1    0    0    0    0    0    0    0     0     0     0     0
#> [25,]   -1   -1   -1    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0    -1    -1    -1
#>  [2,]     0     0     0     0     0     0     0     0     0     0    -1    -1
#>  [3,]     0     0     0     0     0     0     0     0     0     0     0    -1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]    -1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]    -1    -1     0     0     0     0     0     0     0     0     0     0
#> [13,]    -1    -1    -1     0     0     0     0     0     0     0     0     0
#> [14,]     6    -1    -1    -1     0     0     0     0     0     0     0     0
#> [15,]    -1     6    -1    -1    -1     0     0     0     0     0     0     0
#> [16,]    -1    -1     6    -1    -1    -1     0     0     0     0     0     0
#> [17,]    -1    -1    -1     6    -1    -1    -1     0     0     0     0     0
#> [18,]     0    -1    -1    -1     6    -1    -1    -1     0     0     0     0
#> [19,]     0     0    -1    -1    -1     6    -1    -1    -1     0     0     0
#> [20,]     0     0     0    -1    -1    -1     6    -1    -1    -1     0     0
#> [21,]     0     0     0     0    -1    -1    -1     6    -1    -1    -1     0
#> [22,]     0     0     0     0     0    -1    -1    -1     6    -1    -1    -1
#> [23,]     0     0     0     0     0     0    -1    -1    -1     6    -1    -1
#> [24,]     0     0     0     0     0     0     0    -1    -1    -1     6    -1
#> [25,]     0     0     0     0     0     0     0     0    -1    -1    -1     6
#> attr(,"representation")
#> [1] "laplacian"

repr_modularity(g)

#>        [,1]  [,2]  [,3]  [,4]  [,5]  [,6]       [,7]  [,8]  [,9]      [,10]
#>  [1,] -0.24  0.76  0.76  0.76 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000
#>  [2,]  0.76 -0.24  0.76  0.76  0.76 -0.24 -0.2000000 -0.24 -0.24 -0.2800000
#>  [3,]  0.76  0.76 -0.24  0.76  0.76  0.76 -0.2000000 -0.24 -0.24 -0.2800000
#>  [4,]  0.76  0.76  0.76 -0.24  0.76  0.76 -0.2000000 -0.24 -0.24  0.7200000
#>  [5,] -0.24  0.76  0.76  0.76 -0.24  0.76  0.8000000  0.76 -0.24 -0.2800000
#>  [6,] -0.24 -0.24  0.76  0.76  0.76 -0.24  0.8000000  0.76  0.76 -0.2800000
#>  [7,] -0.20 -0.20 -0.20 -0.20  0.80  0.80 -0.1666667  0.80  0.80  0.7666667
#>  [8,] -0.24 -0.24 -0.24 -0.24  0.76  0.76  0.8000000 -0.24  0.76  0.7200000
#>  [9,] -0.24 -0.24 -0.24 -0.24 -0.24  0.76  0.8000000  0.76 -0.24  0.7200000
#> [10,] -0.28 -0.28 -0.28  0.72 -0.28 -0.28  0.7666667  0.72  0.72 -0.3266667
#> [11,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000  0.76  0.76  0.7200000
#> [12,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24  0.76  0.7200000
#> [13,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24  0.7200000
#> [14,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000
#> [15,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000
#> [16,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000
#> [17,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000
#> [18,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000
#> [19,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000
#> [20,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000
#> [21,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000
#> [22,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000
#> [23,]  0.76 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000
#> [24,]  0.76  0.76 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000
#> [25,]  0.76  0.76  0.76 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.2800000
#>       [,11] [,12] [,13] [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22]
#>  [1,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#>  [2,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#>  [3,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#>  [4,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#>  [5,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#>  [6,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#>  [7,] -0.20 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20
#>  [8,]  0.76 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#>  [9,]  0.76  0.76 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#> [10,]  0.72  0.72  0.72 -0.28 -0.28 -0.28 -0.28 -0.28 -0.28 -0.28 -0.28 -0.28
#> [11,] -0.24  0.76  0.76  0.76 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#> [12,]  0.76 -0.24  0.76  0.76  0.76 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#> [13,]  0.76  0.76 -0.24  0.76  0.76  0.76 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#> [14,]  0.76  0.76  0.76 -0.24  0.76  0.76  0.76 -0.24 -0.24 -0.24 -0.24 -0.24
#> [15,] -0.24  0.76  0.76  0.76 -0.24  0.76  0.76  0.76 -0.24 -0.24 -0.24 -0.24
#> [16,] -0.24 -0.24  0.76  0.76  0.76 -0.24  0.76  0.76  0.76 -0.24 -0.24 -0.24
#> [17,] -0.24 -0.24 -0.24  0.76  0.76  0.76 -0.24  0.76  0.76  0.76 -0.24 -0.24
#> [18,] -0.24 -0.24 -0.24 -0.24  0.76  0.76  0.76 -0.24  0.76  0.76  0.76 -0.24
#> [19,] -0.24 -0.24 -0.24 -0.24 -0.24  0.76  0.76  0.76 -0.24  0.76  0.76  0.76
#> [20,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24  0.76  0.76  0.76 -0.24  0.76  0.76
#> [21,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24  0.76  0.76  0.76 -0.24  0.76
#> [22,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24  0.76  0.76  0.76 -0.24
#> [23,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24  0.76  0.76  0.76
#> [24,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24  0.76  0.76
#> [25,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24  0.76
#>       [,23] [,24] [,25]
#>  [1,]  0.76  0.76  0.76
#>  [2,] -0.24  0.76  0.76
#>  [3,] -0.24 -0.24  0.76
#>  [4,] -0.24 -0.24 -0.24
#>  [5,] -0.24 -0.24 -0.24
#>  [6,] -0.24 -0.24 -0.24
#>  [7,] -0.20 -0.20 -0.20
#>  [8,] -0.24 -0.24 -0.24
#>  [9,] -0.24 -0.24 -0.24
#> [10,] -0.28 -0.28 -0.28
#> [11,] -0.24 -0.24 -0.24
#> [12,] -0.24 -0.24 -0.24
#> [13,] -0.24 -0.24 -0.24
#> [14,] -0.24 -0.24 -0.24
#> [15,] -0.24 -0.24 -0.24
#> [16,] -0.24 -0.24 -0.24
#> [17,] -0.24 -0.24 -0.24
#> [18,] -0.24 -0.24 -0.24
#> [19,] -0.24 -0.24 -0.24
#> [20,]  0.76 -0.24 -0.24
#> [21,]  0.76  0.76 -0.24
#> [22,]  0.76  0.76  0.76
#> [23,] -0.24  0.76  0.76
#> [24,]  0.76 -0.24  0.76
#> [25,]  0.76  0.76 -0.24
#> attr(,"representation")
#> [1] "modularity"

repr_graphon(g)

#>            [,1]      [,2]       [,3]       [,4]      [,5]      [,6]       [,7]
#>  [1,] 0.5714286 0.7142857 0.61904762 0.56349206 0.3095238 0.2261905 0.18253968
#>  [2,] 0.7142857 0.5714286 0.77380952 0.61904762 0.6190476 0.2261905 0.23809524
#>  [3,] 0.6190476 0.7738095 0.66666667 0.61111111 0.5833333 0.5000000 0.25000000
#>  [4,] 0.5634921 0.6190476 0.61111111 0.44444444 0.6111111 0.6111111 0.27777778
#>  [5,] 0.3095238 0.6190476 0.58333333 0.61111111 0.5000000 0.6666667 0.52777778
#>  [6,] 0.2261905 0.2261905 0.50000000 0.61111111 0.6666667 0.5000000 0.66666667
#>  [7,] 0.1825397 0.2380952 0.25000000 0.27777778 0.5277778 0.6666667 0.44444444
#>  [8,] 0.1547619 0.1547619 0.25000000 0.19444444 0.5833333 0.6666667 0.52777778
#>  [9,] 0.0000000 0.1428571 0.07142857 0.25396825 0.2261905 0.6190476 0.56349206
#> [10,] 0.0000000 0.1339286 0.20833333 0.35416667 0.2916667 0.3541667 0.64583333
#> [11,] 0.0000000 0.0000000 0.00000000 0.12698413 0.1547619 0.3095238 0.30952381
#> [12,] 0.0000000 0.0000000 0.00000000 0.00000000 0.1666667 0.1666667 0.19444444
#> [13,] 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000 0.1547619 0.18253968
#> [14,] 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000 0.0000000 0.11111111
#> [15,] 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000 0.0000000 0.05555556
#> [16,] 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000 0.0000000 0.00000000
#> [17,] 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000 0.0000000 0.00000000
#> [18,] 0.0000000 0.0000000 0.00000000 0.00000000 0.0000000 0.0000000 0.00000000
#> [19,] 0.1428571 0.0000000 0.00000000 0.00000000 0.0000000 0.0000000 0.00000000
#> [20,] 0.1428571 0.1428571 0.00000000 0.00000000 0.0000000 0.0000000 0.00000000
#> [21,] 0.2857143 0.1428571 0.15476190 0.05555556 0.0000000 0.0000000 0.00000000
#> [22,] 0.2857143 0.2857143 0.15476190 0.18253968 0.0000000 0.0000000 0.00000000
#> [23,] 0.5714286 0.2857143 0.30952381 0.23809524 0.1547619 0.0000000 0.00000000
#> [24,] 0.5714286 0.5714286 0.30952381 0.30952381 0.1547619 0.1547619 0.00000000
#> [25,] 0.7142857 0.5714286 0.61904762 0.36507937 0.3095238 0.1547619 0.05555556
#>             [,8]       [,9]      [,10]     [,11]     [,12]     [,13]      [,14]
#>  [1,] 0.15476190 0.00000000 0.00000000 0.0000000 0.0000000 0.0000000 0.00000000
#>  [2,] 0.15476190 0.14285714 0.13392857 0.0000000 0.0000000 0.0000000 0.00000000
#>  [3,] 0.25000000 0.07142857 0.20833333 0.0000000 0.0000000 0.0000000 0.00000000
#>  [4,] 0.19444444 0.25396825 0.35416667 0.1269841 0.0000000 0.0000000 0.00000000
#>  [5,] 0.58333333 0.22619048 0.29166667 0.1547619 0.1666667 0.0000000 0.00000000
#>  [6,] 0.66666667 0.61904762 0.35416667 0.3095238 0.1666667 0.1547619 0.00000000
#>  [7,] 0.52777778 0.56349206 0.64583333 0.3095238 0.1944444 0.1825397 0.11111111
#>  [8,] 0.50000000 0.69047619 0.72916667 0.5357143 0.1666667 0.2261905 0.08333333
#>  [9,] 0.69047619 0.57142857 0.66964286 0.5714286 0.5357143 0.2857143 0.22619048
#> [10,] 0.72916667 0.66964286 0.50000000 0.6696429 0.5833333 0.5357143 0.29166667
#> [11,] 0.53571429 0.57142857 0.66964286 0.5714286 0.6904762 0.5714286 0.53571429
#> [12,] 0.16666667 0.53571429 0.58333333 0.6904762 0.5000000 0.6904762 0.58333333
#> [13,] 0.22619048 0.28571429 0.53571429 0.5714286 0.6904762 0.5714286 0.69047619
#> [14,] 0.08333333 0.22619048 0.29166667 0.5357143 0.5833333 0.6904762 0.66666667
#> [15,] 0.15476190 0.14285714 0.26785714 0.2857143 0.6190476 0.5714286 0.77380952
#> [16,] 0.00000000 0.14285714 0.13392857 0.2857143 0.3095238 0.5714286 0.61904762
#> [17,] 0.00000000 0.00000000 0.13392857 0.1428571 0.3095238 0.2857143 0.61904762
#> [18,] 0.00000000 0.00000000 0.00000000 0.1428571 0.1547619 0.2857143 0.30952381
#> [19,] 0.00000000 0.00000000 0.00000000 0.0000000 0.1547619 0.1428571 0.30952381
#> [20,] 0.00000000 0.00000000 0.00000000 0.0000000 0.0000000 0.1428571 0.15476190
#> [21,] 0.00000000 0.00000000 0.00000000 0.0000000 0.0000000 0.0000000 0.15476190
#> [22,] 0.00000000 0.00000000 0.00000000 0.0000000 0.0000000 0.0000000 0.00000000
#> [23,] 0.00000000 0.00000000 0.00000000 0.0000000 0.0000000 0.0000000 0.00000000
#> [24,] 0.00000000 0.00000000 0.00000000 0.0000000 0.0000000 0.0000000 0.00000000
#> [25,] 0.00000000 0.00000000 0.07142857 0.0000000 0.0000000 0.0000000 0.00000000
#>            [,15]     [,16]     [,17]     [,18]     [,19]     [,20]      [,21]
#>  [1,] 0.00000000 0.0000000 0.0000000 0.0000000 0.1428571 0.1428571 0.28571429
#>  [2,] 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.1428571 0.14285714
#>  [3,] 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.15476190
#>  [4,] 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.05555556
#>  [5,] 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#>  [6,] 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#>  [7,] 0.05555556 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#>  [8,] 0.15476190 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#>  [9,] 0.14285714 0.1428571 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> [10,] 0.26785714 0.1339286 0.1339286 0.0000000 0.0000000 0.0000000 0.00000000
#> [11,] 0.28571429 0.2857143 0.1428571 0.1428571 0.0000000 0.0000000 0.00000000
#> [12,] 0.61904762 0.3095238 0.3095238 0.1547619 0.1547619 0.0000000 0.00000000
#> [13,] 0.57142857 0.5714286 0.2857143 0.2857143 0.1428571 0.1428571 0.00000000
#> [14,] 0.77380952 0.6190476 0.6190476 0.3095238 0.3095238 0.1547619 0.15476190
#> [15,] 0.57142857 0.7142857 0.5714286 0.5714286 0.2857143 0.2857143 0.14285714
#> [16,] 0.71428571 0.5714286 0.7142857 0.5714286 0.5714286 0.2857143 0.28571429
#> [17,] 0.57142857 0.7142857 0.5714286 0.7142857 0.5714286 0.5714286 0.28571429
#> [18,] 0.57142857 0.5714286 0.7142857 0.5714286 0.7142857 0.5714286 0.57142857
#> [19,] 0.28571429 0.5714286 0.5714286 0.7142857 0.5714286 0.7142857 0.57142857
#> [20,] 0.28571429 0.2857143 0.5714286 0.5714286 0.7142857 0.5714286 0.71428571
#> [21,] 0.14285714 0.2857143 0.2857143 0.5714286 0.5714286 0.7142857 0.57142857
#> [22,] 0.14285714 0.1428571 0.2857143 0.2857143 0.5714286 0.5714286 0.71428571
#> [23,] 0.00000000 0.1428571 0.1428571 0.2857143 0.2857143 0.5714286 0.57142857
#> [24,] 0.00000000 0.0000000 0.1428571 0.1428571 0.2857143 0.2857143 0.57142857
#> [25,] 0.00000000 0.0000000 0.0000000 0.1428571 0.1428571 0.2857143 0.28571429
#>           [,22]     [,23]     [,24]      [,25]
#>  [1,] 0.2857143 0.5714286 0.5714286 0.71428571
#>  [2,] 0.2857143 0.2857143 0.5714286 0.57142857
#>  [3,] 0.1547619 0.3095238 0.3095238 0.61904762
#>  [4,] 0.1825397 0.2380952 0.3095238 0.36507937
#>  [5,] 0.0000000 0.1547619 0.1547619 0.30952381
#>  [6,] 0.0000000 0.0000000 0.1547619 0.15476190
#>  [7,] 0.0000000 0.0000000 0.0000000 0.05555556
#>  [8,] 0.0000000 0.0000000 0.0000000 0.00000000
#>  [9,] 0.0000000 0.0000000 0.0000000 0.00000000
#> [10,] 0.0000000 0.0000000 0.0000000 0.07142857
#> [11,] 0.0000000 0.0000000 0.0000000 0.00000000
#> [12,] 0.0000000 0.0000000 0.0000000 0.00000000
#> [13,] 0.0000000 0.0000000 0.0000000 0.00000000
#> [14,] 0.0000000 0.0000000 0.0000000 0.00000000
#> [15,] 0.1428571 0.0000000 0.0000000 0.00000000
#> [16,] 0.1428571 0.1428571 0.0000000 0.00000000
#> [17,] 0.2857143 0.1428571 0.1428571 0.00000000
#> [18,] 0.2857143 0.2857143 0.1428571 0.14285714
#> [19,] 0.5714286 0.2857143 0.2857143 0.14285714
#> [20,] 0.5714286 0.5714286 0.2857143 0.28571429
#> [21,] 0.7142857 0.5714286 0.5714286 0.28571429
#> [22,] 0.5714286 0.7142857 0.5714286 0.57142857
#> [23,] 0.7142857 0.5714286 0.7142857 0.57142857
#> [24,] 0.5714286 0.7142857 0.5714286 0.71428571
#> [25,] 0.5714286 0.5714286 0.7142857 0.57142857
#> attr(,"representation")
#> [1] "graphon"