Network Representation Functions

This is a collection of functions that convert a graph stored as an igraph object into a desired matrix representation among adjacency matrix, graph laplacian, modularity matrix or graphon (edge probability matrix).

repr_adjacency(network, validate = TRUE)

repr_laplacian(network, validate = TRUE)

repr_modularity(network, validate = TRUE)

repr_graphon(network, validate = TRUE)

Arguments

network

An igraph object.

validate

A boolean specifying whether the function should check the class of its input (default: TRUE).

Value

A numeric square matrix giving the desired network representation recorded in the object's class.

Examples

g <- igraph::sample_smallworld(1, 25, 3, 0.05)
repr_adjacency(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0    1    1    1    0    0    0    0    0     0     0     0     0
#>  [2,]    1    0    1    1    1    0    0    0    0     0     0     0     0
#>  [3,]    1    1    0    1    1    0    0    0    0     0     0     0     0
#>  [4,]    1    1    1    0    1    1    1    0    0     0     0     0     0
#>  [5,]    0    1    1    1    0    1    1    1    0     1     0     0     0
#>  [6,]    0    0    0    1    1    0    1    1    1     0     0     0     0
#>  [7,]    0    0    0    1    1    1    0    1    1     1     0     0     0
#>  [8,]    0    0    0    0    1    1    1    0    1     1     1     0     0
#>  [9,]    0    0    0    0    0    1    1    1    0     1     1     1     0
#> [10,]    0    0    0    0    1    0    1    1    1     0     1     1     0
#> [11,]    0    0    0    0    0    0    0    1    1     1     0     1     1
#> [12,]    0    0    0    0    0    0    0    0    1     1     1     0     1
#> [13,]    0    0    0    0    0    0    0    0    0     0     1     1     0
#> [14,]    0    0    1    0    0    0    0    0    0     0     1     1     1
#> [15,]    0    0    1    0    0    0    0    1    0     0     0     0     1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0     1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]    1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]    1    1    0    0    0    0    0    0    0     0     0     0     0
#> [25,]    1    1    1    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0     1     1     1
#>  [2,]     0     0     0     0     0     0     0     0     0     0     1     1
#>  [3,]     1     1     0     0     0     0     0     0     0     0     0     1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     1     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]     1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]     1     0     0     0     0     0     0     0     0     0     0     0
#> [13,]     1     1     1     0     0     0     0     0     0     0     0     0
#> [14,]     0     1     1     1     0     0     0     0     0     0     0     0
#> [15,]     1     0     0     1     1     0     0     0     0     0     0     0
#> [16,]     1     0     0     1     1     1     0     0     0     0     0     0
#> [17,]     1     1     1     0     1     1     1     0     0     0     0     0
#> [18,]     0     1     1     1     0     1     1     1     0     0     0     0
#> [19,]     0     0     1     1     1     0     1     1     1     0     0     0
#> [20,]     0     0     0     1     1     1     0     1     1     1     0     0
#> [21,]     0     0     0     0     1     1     1     0     1     1     1     0
#> [22,]     0     0     0     0     0     1     1     1     0     1     1     1
#> [23,]     0     0     0     0     0     0     1     1     1     0     1     1
#> [24,]     0     0     0     0     0     0     0     1     1     1     0     1
#> [25,]     0     0     0     0     0     0     0     0     1     1     1     0
#> attr(,"representation")
#> [1] "adjacency"
repr_laplacian(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    6   -1   -1   -1    0    0    0    0    0     0     0     0     0
#>  [2,]   -1    6   -1   -1   -1    0    0    0    0     0     0     0     0
#>  [3,]   -1   -1    7   -1   -1    0    0    0    0     0     0     0     0
#>  [4,]   -1   -1   -1    6   -1   -1   -1    0    0     0     0     0     0
#>  [5,]    0   -1   -1   -1    7   -1   -1   -1    0    -1     0     0     0
#>  [6,]    0    0    0   -1   -1    5   -1   -1   -1     0     0     0     0
#>  [7,]    0    0    0   -1   -1   -1    6   -1   -1    -1     0     0     0
#>  [8,]    0    0    0    0   -1   -1   -1    7   -1    -1    -1     0     0
#>  [9,]    0    0    0    0    0   -1   -1   -1    6    -1    -1    -1     0
#> [10,]    0    0    0    0   -1    0   -1   -1   -1     6    -1    -1     0
#> [11,]    0    0    0    0    0    0    0   -1   -1    -1     6    -1    -1
#> [12,]    0    0    0    0    0    0    0    0   -1    -1    -1     5    -1
#> [13,]    0    0    0    0    0    0    0    0    0     0    -1    -1     5
#> [14,]    0    0   -1    0    0    0    0    0    0     0    -1    -1    -1
#> [15,]    0    0   -1    0    0    0    0   -1    0     0     0     0    -1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0    -1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]   -1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]   -1   -1    0    0    0    0    0    0    0     0     0     0     0
#> [25,]   -1   -1   -1    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0    -1    -1    -1
#>  [2,]     0     0     0     0     0     0     0     0     0     0    -1    -1
#>  [3,]    -1    -1     0     0     0     0     0     0     0     0     0    -1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0    -1     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]    -1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]    -1     0     0     0     0     0     0     0     0     0     0     0
#> [13,]    -1    -1    -1     0     0     0     0     0     0     0     0     0
#> [14,]     7    -1    -1    -1     0     0     0     0     0     0     0     0
#> [15,]    -1     6     0    -1    -1     0     0     0     0     0     0     0
#> [16,]    -1     0     5    -1    -1    -1     0     0     0     0     0     0
#> [17,]    -1    -1    -1     6    -1    -1    -1     0     0     0     0     0
#> [18,]     0    -1    -1    -1     6    -1    -1    -1     0     0     0     0
#> [19,]     0     0    -1    -1    -1     6    -1    -1    -1     0     0     0
#> [20,]     0     0     0    -1    -1    -1     6    -1    -1    -1     0     0
#> [21,]     0     0     0     0    -1    -1    -1     6    -1    -1    -1     0
#> [22,]     0     0     0     0     0    -1    -1    -1     6    -1    -1    -1
#> [23,]     0     0     0     0     0     0    -1    -1    -1     6    -1    -1
#> [24,]     0     0     0     0     0     0     0    -1    -1    -1     6    -1
#> [25,]     0     0     0     0     0     0     0     0    -1    -1    -1     6
#> attr(,"representation")
#> [1] "laplacian"
repr_modularity(g)
#>        [,1]  [,2]       [,3]  [,4]       [,5]       [,6]  [,7]       [,8]  [,9]
#>  [1,] -0.24  0.76  0.7200000  0.76 -0.2800000 -0.2000000 -0.24 -0.2800000 -0.24
#>  [2,]  0.76 -0.24  0.7200000  0.76  0.7200000 -0.2000000 -0.24 -0.2800000 -0.24
#>  [3,]  0.72  0.72 -0.3266667  0.72  0.6733333 -0.2333333 -0.28 -0.3266667 -0.28
#>  [4,]  0.76  0.76  0.7200000 -0.24  0.7200000  0.8000000  0.76 -0.2800000 -0.24
#>  [5,] -0.28  0.72  0.6733333  0.72 -0.3266667  0.7666667  0.72  0.6733333 -0.28
#>  [6,] -0.20 -0.20 -0.2333333  0.80  0.7666667 -0.1666667  0.80  0.7666667  0.80
#>  [7,] -0.24 -0.24 -0.2800000  0.76  0.7200000  0.8000000 -0.24  0.7200000  0.76
#>  [8,] -0.28 -0.28 -0.3266667 -0.28  0.6733333  0.7666667  0.72 -0.3266667  0.72
#>  [9,] -0.24 -0.24 -0.2800000 -0.24 -0.2800000  0.8000000  0.76  0.7200000 -0.24
#> [10,] -0.24 -0.24 -0.2800000 -0.24  0.7200000 -0.2000000  0.76  0.7200000  0.76
#> [11,] -0.24 -0.24 -0.2800000 -0.24 -0.2800000 -0.2000000 -0.24  0.7200000  0.76
#> [12,] -0.20 -0.20 -0.2333333 -0.20 -0.2333333 -0.1666667 -0.20 -0.2333333  0.80
#> [13,] -0.20 -0.20 -0.2333333 -0.20 -0.2333333 -0.1666667 -0.20 -0.2333333 -0.20
#> [14,] -0.28 -0.28  0.6733333 -0.28 -0.3266667 -0.2333333 -0.28 -0.3266667 -0.28
#> [15,] -0.24 -0.24  0.7200000 -0.24 -0.2800000 -0.2000000 -0.24  0.7200000 -0.24
#> [16,] -0.20 -0.20 -0.2333333 -0.20 -0.2333333 -0.1666667 -0.20 -0.2333333 -0.20
#> [17,] -0.24 -0.24 -0.2800000 -0.24 -0.2800000 -0.2000000 -0.24 -0.2800000 -0.24
#> [18,] -0.24 -0.24 -0.2800000 -0.24 -0.2800000 -0.2000000 -0.24 -0.2800000 -0.24
#> [19,] -0.24 -0.24 -0.2800000 -0.24 -0.2800000 -0.2000000 -0.24 -0.2800000 -0.24
#> [20,] -0.24 -0.24 -0.2800000 -0.24 -0.2800000 -0.2000000 -0.24 -0.2800000 -0.24
#> [21,] -0.24 -0.24 -0.2800000 -0.24 -0.2800000 -0.2000000 -0.24 -0.2800000 -0.24
#> [22,] -0.24 -0.24 -0.2800000 -0.24 -0.2800000 -0.2000000 -0.24 -0.2800000 -0.24
#> [23,]  0.76 -0.24 -0.2800000 -0.24 -0.2800000 -0.2000000 -0.24 -0.2800000 -0.24
#> [24,]  0.76  0.76 -0.2800000 -0.24 -0.2800000 -0.2000000 -0.24 -0.2800000 -0.24
#> [25,]  0.76  0.76  0.7200000 -0.24 -0.2800000 -0.2000000 -0.24 -0.2800000 -0.24
#>       [,10] [,11]      [,12]      [,13]      [,14] [,15]      [,16] [,17] [,18]
#>  [1,] -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.24 -0.2000000 -0.24 -0.24
#>  [2,] -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.24 -0.2000000 -0.24 -0.24
#>  [3,] -0.28 -0.28 -0.2333333 -0.2333333  0.6733333  0.72 -0.2333333 -0.28 -0.28
#>  [4,] -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.24 -0.2000000 -0.24 -0.24
#>  [5,]  0.72 -0.28 -0.2333333 -0.2333333 -0.3266667 -0.28 -0.2333333 -0.28 -0.28
#>  [6,] -0.20 -0.20 -0.1666667 -0.1666667 -0.2333333 -0.20 -0.1666667 -0.20 -0.20
#>  [7,]  0.76 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.24 -0.2000000 -0.24 -0.24
#>  [8,]  0.72  0.72 -0.2333333 -0.2333333 -0.3266667  0.72 -0.2333333 -0.28 -0.28
#>  [9,]  0.76  0.76  0.8000000 -0.2000000 -0.2800000 -0.24 -0.2000000 -0.24 -0.24
#> [10,] -0.24  0.76  0.8000000 -0.2000000 -0.2800000 -0.24 -0.2000000 -0.24 -0.24
#> [11,]  0.76 -0.24  0.8000000  0.8000000  0.7200000 -0.24 -0.2000000 -0.24 -0.24
#> [12,]  0.80  0.80 -0.1666667  0.8333333  0.7666667 -0.20 -0.1666667 -0.20 -0.20
#> [13,] -0.20  0.80  0.8333333 -0.1666667  0.7666667  0.80  0.8333333 -0.20 -0.20
#> [14,] -0.28  0.72  0.7666667  0.7666667 -0.3266667  0.72  0.7666667  0.72 -0.28
#> [15,] -0.24 -0.24 -0.2000000  0.8000000  0.7200000 -0.24 -0.2000000  0.76  0.76
#> [16,] -0.20 -0.20 -0.1666667  0.8333333  0.7666667 -0.20 -0.1666667  0.80  0.80
#> [17,] -0.24 -0.24 -0.2000000 -0.2000000  0.7200000  0.76  0.8000000 -0.24  0.76
#> [18,] -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000  0.76  0.8000000  0.76 -0.24
#> [19,] -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.24  0.8000000  0.76  0.76
#> [20,] -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.24 -0.2000000  0.76  0.76
#> [21,] -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.24 -0.2000000 -0.24  0.76
#> [22,] -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.24 -0.2000000 -0.24 -0.24
#> [23,] -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.24 -0.2000000 -0.24 -0.24
#> [24,] -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.24 -0.2000000 -0.24 -0.24
#> [25,] -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.24 -0.2000000 -0.24 -0.24
#>       [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,] -0.24 -0.24 -0.24 -0.24  0.76  0.76  0.76
#>  [2,] -0.24 -0.24 -0.24 -0.24 -0.24  0.76  0.76
#>  [3,] -0.28 -0.28 -0.28 -0.28 -0.28 -0.28  0.72
#>  [4,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#>  [5,] -0.28 -0.28 -0.28 -0.28 -0.28 -0.28 -0.28
#>  [6,] -0.20 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20
#>  [7,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#>  [8,] -0.28 -0.28 -0.28 -0.28 -0.28 -0.28 -0.28
#>  [9,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#> [10,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#> [11,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#> [12,] -0.20 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20
#> [13,] -0.20 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20
#> [14,] -0.28 -0.28 -0.28 -0.28 -0.28 -0.28 -0.28
#> [15,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.24 -0.24
#> [16,]  0.80 -0.20 -0.20 -0.20 -0.20 -0.20 -0.20
#> [17,]  0.76  0.76 -0.24 -0.24 -0.24 -0.24 -0.24
#> [18,]  0.76  0.76  0.76 -0.24 -0.24 -0.24 -0.24
#> [19,] -0.24  0.76  0.76  0.76 -0.24 -0.24 -0.24
#> [20,]  0.76 -0.24  0.76  0.76  0.76 -0.24 -0.24
#> [21,]  0.76  0.76 -0.24  0.76  0.76  0.76 -0.24
#> [22,]  0.76  0.76  0.76 -0.24  0.76  0.76  0.76
#> [23,] -0.24  0.76  0.76  0.76 -0.24  0.76  0.76
#> [24,] -0.24 -0.24  0.76  0.76  0.76 -0.24  0.76
#> [25,] -0.24 -0.24 -0.24  0.76  0.76  0.76 -0.24
#> attr(,"representation")
#> [1] "modularity"
repr_graphon(g)
#>             [,1]       [,2]       [,3]      [,4]       [,5]      [,6]
#>  [1,] 0.62500000 0.66666667 0.65000000 0.5000000 0.19642857 0.1875000
#>  [2,] 0.66666667 0.66666667 0.73333333 0.6250000 0.46428571 0.1458333
#>  [3,] 0.65000000 0.73333333 0.60000000 0.8000000 0.51428571 0.2250000
#>  [4,] 0.50000000 0.62500000 0.80000000 0.5000000 0.53571429 0.4375000
#>  [5,] 0.19642857 0.46428571 0.51428571 0.5357143 0.57142857 0.6696429
#>  [6,] 0.18750000 0.14583333 0.22500000 0.4375000 0.66964286 0.3750000
#>  [7,] 0.06250000 0.15476190 0.17142857 0.4642857 0.64285714 0.5357143
#>  [8,] 0.06250000 0.12500000 0.12500000 0.3125000 0.59821429 0.6250000
#>  [9,] 0.05555556 0.11111111 0.11111111 0.1666667 0.49206349 0.5902778
#> [10,] 0.05555556 0.05555556 0.05555556 0.2361111 0.56349206 0.4722222
#> [11,] 0.00000000 0.00000000 0.05555556 0.1111111 0.30952381 0.2986111
#> [12,] 0.00000000 0.00000000 0.05555556 0.1111111 0.23809524 0.2986111
#> [13,] 0.00000000 0.00000000 0.05555556 0.0000000 0.05555556 0.1805556
#> [14,] 0.00000000 0.08333333 0.35000000 0.0000000 0.00000000 0.1250000
#> [15,] 0.00000000 0.08333333 0.22500000 0.0000000 0.07142857 0.0000000
#> [16,] 0.00000000 0.00000000 0.05555556 0.0000000 0.00000000 0.0000000
#> [17,] 0.00000000 0.00000000 0.05555556 0.0000000 0.00000000 0.0000000
#> [18,] 0.11805556 0.00000000 0.00000000 0.0000000 0.00000000 0.0000000
#> [19,] 0.20833333 0.00000000 0.00000000 0.0000000 0.00000000 0.0000000
#> [20,] 0.25892857 0.15476190 0.00000000 0.0000000 0.00000000 0.0000000
#> [21,] 0.33035714 0.15476190 0.17142857 0.0000000 0.00000000 0.0000000
#> [22,] 0.39285714 0.30952381 0.17142857 0.1964286 0.00000000 0.0000000
#> [23,] 0.59821429 0.30952381 0.34285714 0.1964286 0.14285714 0.0000000
#> [24,] 0.62500000 0.58333333 0.36666667 0.3333333 0.07142857 0.0625000
#> [25,] 0.59027778 0.55555556 0.62222222 0.2361111 0.18253968 0.1180556
#>             [,7]       [,8]       [,9]      [,10]      [,11]      [,12]
#>  [1,] 0.06250000 0.06250000 0.05555556 0.05555556 0.00000000 0.00000000
#>  [2,] 0.15476190 0.12500000 0.11111111 0.05555556 0.00000000 0.00000000
#>  [3,] 0.17142857 0.12500000 0.11111111 0.05555556 0.05555556 0.05555556
#>  [4,] 0.46428571 0.31250000 0.16666667 0.23611111 0.11111111 0.11111111
#>  [5,] 0.64285714 0.59821429 0.49206349 0.56349206 0.30952381 0.23809524
#>  [6,] 0.53571429 0.62500000 0.59027778 0.47222222 0.29861111 0.29861111
#>  [7,] 0.71428571 0.80357143 0.69047619 0.63492063 0.38095238 0.38095238
#>  [8,] 0.80357143 0.62500000 0.64583333 0.59027778 0.58333333 0.45833333
#>  [9,] 0.69047619 0.64583333 0.66666667 0.66666667 0.50000000 0.44444444
#> [10,] 0.63492063 0.59027778 0.66666667 0.55555556 0.50000000 0.44444444
#> [11,] 0.38095238 0.58333333 0.50000000 0.50000000 0.44444444 0.44444444
#> [12,] 0.38095238 0.45833333 0.44444444 0.44444444 0.44444444 0.44444444
#> [13,] 0.18253968 0.22916667 0.27777778 0.27777778 0.50000000 0.50000000
#> [14,] 0.07142857 0.18750000 0.11111111 0.16666667 0.52777778 0.52777778
#> [15,] 0.07142857 0.18750000 0.18055556 0.23611111 0.18055556 0.18055556
#> [16,] 0.00000000 0.11111111 0.11111111 0.16666667 0.11111111 0.11111111
#> [17,] 0.00000000 0.05555556 0.00000000 0.00000000 0.16666667 0.16666667
#> [18,] 0.00000000 0.00000000 0.00000000 0.00000000 0.16666667 0.16666667
#> [19,] 0.00000000 0.00000000 0.00000000 0.00000000 0.05555556 0.05555556
#> [20,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [21,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [22,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [23,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [24,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [25,] 0.05555556 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>            [,13]      [,14]      [,15]      [,16]      [,17]     [,18]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.1180556
#>  [2,] 0.00000000 0.08333333 0.08333333 0.00000000 0.00000000 0.0000000
#>  [3,] 0.05555556 0.35000000 0.22500000 0.05555556 0.05555556 0.0000000
#>  [4,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000
#>  [5,] 0.05555556 0.00000000 0.07142857 0.00000000 0.00000000 0.0000000
#>  [6,] 0.18055556 0.12500000 0.00000000 0.00000000 0.00000000 0.0000000
#>  [7,] 0.18253968 0.07142857 0.07142857 0.00000000 0.00000000 0.0000000
#>  [8,] 0.22916667 0.18750000 0.18750000 0.11111111 0.05555556 0.0000000
#>  [9,] 0.27777778 0.11111111 0.18055556 0.11111111 0.00000000 0.0000000
#> [10,] 0.27777778 0.16666667 0.23611111 0.16666667 0.00000000 0.0000000
#> [11,] 0.50000000 0.52777778 0.18055556 0.11111111 0.16666667 0.1666667
#> [12,] 0.50000000 0.52777778 0.18055556 0.11111111 0.16666667 0.1666667
#> [13,] 0.44444444 0.52777778 0.53472222 0.44444444 0.27777778 0.1666667
#> [14,] 0.52777778 0.50000000 0.68750000 0.65277778 0.59722222 0.2916667
#> [15,] 0.53472222 0.68750000 0.25000000 0.29861111 0.47916667 0.4791667
#> [16,] 0.44444444 0.65277778 0.29861111 0.33333333 0.50000000 0.5555556
#> [17,] 0.27777778 0.59722222 0.47916667 0.50000000 0.55555556 0.5555556
#> [18,] 0.16666667 0.29166667 0.47916667 0.55555556 0.55555556 0.5555556
#> [19,] 0.16666667 0.20833333 0.35416667 0.52777778 0.58333333 0.6666667
#> [20,] 0.18253968 0.19642857 0.19642857 0.36507937 0.56349206 0.6190476
#> [21,] 0.05555556 0.19642857 0.19642857 0.30952381 0.36507937 0.5634921
#> [22,] 0.00000000 0.00000000 0.19642857 0.23809524 0.30952381 0.3650794
#> [23,] 0.00000000 0.00000000 0.06250000 0.18253968 0.23809524 0.3095238
#> [24,] 0.00000000 0.00000000 0.00000000 0.05555556 0.19444444 0.2500000
#> [25,] 0.00000000 0.00000000 0.00000000 0.00000000 0.11111111 0.2222222
#>            [,19]     [,20]      [,21]     [,22]     [,23]      [,24]      [,25]
#>  [1,] 0.20833333 0.2589286 0.33035714 0.3928571 0.5982143 0.62500000 0.59027778
#>  [2,] 0.00000000 0.1547619 0.15476190 0.3095238 0.3095238 0.58333333 0.55555556
#>  [3,] 0.00000000 0.0000000 0.17142857 0.1714286 0.3428571 0.36666667 0.62222222
#>  [4,] 0.00000000 0.0000000 0.00000000 0.1964286 0.1964286 0.33333333 0.23611111
#>  [5,] 0.00000000 0.0000000 0.00000000 0.0000000 0.1428571 0.07142857 0.18253968
#>  [6,] 0.00000000 0.0000000 0.00000000 0.0000000 0.0000000 0.06250000 0.11805556
#>  [7,] 0.00000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.05555556
#>  [8,] 0.00000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.00000000
#>  [9,] 0.00000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.00000000
#> [10,] 0.00000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.00000000
#> [11,] 0.05555556 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.00000000
#> [12,] 0.05555556 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.00000000
#> [13,] 0.16666667 0.1825397 0.05555556 0.0000000 0.0000000 0.00000000 0.00000000
#> [14,] 0.20833333 0.1964286 0.19642857 0.0000000 0.0000000 0.00000000 0.00000000
#> [15,] 0.35416667 0.1964286 0.19642857 0.1964286 0.0625000 0.00000000 0.00000000
#> [16,] 0.52777778 0.3650794 0.30952381 0.2380952 0.1825397 0.05555556 0.00000000
#> [17,] 0.58333333 0.5634921 0.36507937 0.3095238 0.2380952 0.19444444 0.11111111
#> [18,] 0.66666667 0.6190476 0.56349206 0.3650794 0.3095238 0.25000000 0.22222222
#> [19,] 0.66666667 0.7738095 0.61904762 0.6190476 0.3095238 0.33333333 0.25000000
#> [20,] 0.77380952 0.5714286 0.71428571 0.5714286 0.5714286 0.30952381 0.30952381
#> [21,] 0.61904762 0.7142857 0.57142857 0.7142857 0.5714286 0.61904762 0.36507937
#> [22,] 0.61904762 0.5714286 0.71428571 0.5714286 0.7142857 0.61904762 0.56349206
#> [23,] 0.30952381 0.5714286 0.57142857 0.7142857 0.5714286 0.77380952 0.61904762
#> [24,] 0.33333333 0.3095238 0.61904762 0.6190476 0.7738095 0.66666667 0.66666667
#> [25,] 0.25000000 0.3095238 0.36507937 0.5634921 0.6190476 0.66666667 0.55555556
#> attr(,"representation")
#> [1] "graphon"