Network Representation Functions

This is a collection of functions that convert a graph stored as an igraph object into a desired matrix representation among adjacency matrix, graph laplacian, modularity matrix or graphon (edge probability matrix).

repr_adjacency(network, validate = TRUE)

repr_laplacian(network, validate = TRUE)

repr_modularity(network, validate = TRUE)

repr_graphon(network, validate = TRUE)

Arguments

network
An igraph object.

validate
A boolean specifying whether the function should check the class of its input (default: TRUE).

Value

A numeric square matrix giving the desired network representation recorded in the object's class.

Examples

g <- igraph::sample_smallworld(1, 25, 3, 0.05)
repr_adjacency(g)

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0    1    1    1    0    0    0    0    0     1     0     0     0
#>  [2,]    1    0    1    1    1    0    0    0    0     1     0     0     0
#>  [3,]    1    1    0    1    1    0    0    0    0     0     0     0     0
#>  [4,]    1    1    1    0    1    1    1    0    0     0     0     0     0
#>  [5,]    0    1    1    1    0    1    0    1    0     0     0     0     0
#>  [6,]    0    0    0    1    1    0    1    1    1     0     0     0     0
#>  [7,]    0    0    0    1    0    1    0    1    1     1     0     0     0
#>  [8,]    0    0    0    0    1    1    1    0    1     1     1     0     0
#>  [9,]    0    0    0    0    0    1    1    1    0     0     1     1     0
#> [10,]    1    1    0    0    0    0    1    1    0     0     0     1     1
#> [11,]    0    0    0    0    0    0    0    1    1     0     0     1     1
#> [12,]    0    0    0    0    0    0    0    0    1     1     1     0     1
#> [13,]    0    0    0    0    0    0    0    0    0     1     1     1     0
#> [14,]    0    0    0    0    0    0    0    0    0     0     1     1     1
#> [15,]    0    0    0    0    0    0    0    0    0     0     0     1     1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0     1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    1    0    0    0    0    0    0     0     0     0     0
#> [23,]    1    0    0    0    1    0    0    0    0     0     0     0     0
#> [24,]    1    1    0    0    0    0    0    0    0     0     0     0     0
#> [25,]    1    1    1    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0     1     1     1
#>  [2,]     0     0     0     0     0     0     0     0     0     0     1     1
#>  [3,]     0     0     0     0     0     0     0     0     1     0     0     1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     0     0     1     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]     1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]     1     1     0     0     0     0     0     0     0     0     0     0
#> [13,]     1     1     1     0     0     0     0     0     0     0     0     0
#> [14,]     0     1     1     1     0     0     0     0     0     0     0     0
#> [15,]     1     0     1     1     1     0     0     0     0     0     0     0
#> [16,]     1     1     0     1     1     1     0     0     0     0     0     0
#> [17,]     1     1     1     0     1     1     1     0     0     1     0     0
#> [18,]     0     1     1     1     0     1     1     0     0     0     0     1
#> [19,]     0     0     1     1     1     0     1     1     1     0     0     0
#> [20,]     0     0     0     1     1     1     0     1     1     0     0     0
#> [21,]     0     0     0     0     0     1     1     0     1     1     1     0
#> [22,]     0     0     0     0     0     1     1     1     0     1     1     1
#> [23,]     0     0     0     1     0     0     0     1     1     0     1     1
#> [24,]     0     0     0     0     0     0     0     1     1     1     0     1
#> [25,]     0     0     0     0     1     0     0     0     1     1     1     0
#> attr(,"representation")
#> [1] "adjacency"

repr_laplacian(g)

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    7   -1   -1   -1    0    0    0    0    0    -1     0     0     0
#>  [2,]   -1    7   -1   -1   -1    0    0    0    0    -1     0     0     0
#>  [3,]   -1   -1    6   -1   -1    0    0    0    0     0     0     0     0
#>  [4,]   -1   -1   -1    6   -1   -1   -1    0    0     0     0     0     0
#>  [5,]    0   -1   -1   -1    6   -1    0   -1    0     0     0     0     0
#>  [6,]    0    0    0   -1   -1    5   -1   -1   -1     0     0     0     0
#>  [7,]    0    0    0   -1    0   -1    5   -1   -1    -1     0     0     0
#>  [8,]    0    0    0    0   -1   -1   -1    6   -1    -1    -1     0     0
#>  [9,]    0    0    0    0    0   -1   -1   -1    5     0    -1    -1     0
#> [10,]   -1   -1    0    0    0    0   -1   -1    0     6     0    -1    -1
#> [11,]    0    0    0    0    0    0    0   -1   -1     0     5    -1    -1
#> [12,]    0    0    0    0    0    0    0    0   -1    -1    -1     6    -1
#> [13,]    0    0    0    0    0    0    0    0    0    -1    -1    -1     6
#> [14,]    0    0    0    0    0    0    0    0    0     0    -1    -1    -1
#> [15,]    0    0    0    0    0    0    0    0    0     0     0    -1    -1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0    -1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0   -1    0    0    0    0    0    0     0     0     0     0
#> [23,]   -1    0    0    0   -1    0    0    0    0     0     0     0     0
#> [24,]   -1   -1    0    0    0    0    0    0    0     0     0     0     0
#> [25,]   -1   -1   -1    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0    -1    -1    -1
#>  [2,]     0     0     0     0     0     0     0     0     0     0    -1    -1
#>  [3,]     0     0     0     0     0     0     0     0    -1     0     0    -1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     0     0    -1     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]    -1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]    -1    -1     0     0     0     0     0     0     0     0     0     0
#> [13,]    -1    -1    -1     0     0     0     0     0     0     0     0     0
#> [14,]     6    -1    -1    -1     0     0     0     0     0     0     0     0
#> [15,]    -1     6    -1    -1    -1     0     0     0     0     0     0     0
#> [16,]    -1    -1     6    -1    -1    -1     0     0     0     0     0     0
#> [17,]    -1    -1    -1     7    -1    -1    -1     0     0    -1     0     0
#> [18,]     0    -1    -1    -1     6    -1    -1     0     0     0     0    -1
#> [19,]     0     0    -1    -1    -1     6    -1    -1    -1     0     0     0
#> [20,]     0     0     0    -1    -1    -1     5    -1    -1     0     0     0
#> [21,]     0     0     0     0     0    -1    -1     5    -1    -1    -1     0
#> [22,]     0     0     0     0     0    -1    -1    -1     7    -1    -1    -1
#> [23,]     0     0     0    -1     0     0     0    -1    -1     7    -1    -1
#> [24,]     0     0     0     0     0     0     0    -1    -1    -1     6    -1
#> [25,]     0     0     0     0    -1     0     0     0    -1    -1    -1     7
#> attr(,"representation")
#> [1] "laplacian"

repr_modularity(g)

#>             [,1]       [,2]  [,3]  [,4]  [,5]       [,6]       [,7]  [,8]
#>  [1,] -0.3266667  0.6733333  0.72  0.72 -0.28 -0.2333333 -0.2333333 -0.28
#>  [2,]  0.6733333 -0.3266667  0.72  0.72  0.72 -0.2333333 -0.2333333 -0.28
#>  [3,]  0.7200000  0.7200000 -0.24  0.76  0.76 -0.2000000 -0.2000000 -0.24
#>  [4,]  0.7200000  0.7200000  0.76 -0.24  0.76  0.8000000  0.8000000 -0.24
#>  [5,] -0.2800000  0.7200000  0.76  0.76 -0.24  0.8000000 -0.2000000  0.76
#>  [6,] -0.2333333 -0.2333333 -0.20  0.80  0.80 -0.1666667  0.8333333  0.80
#>  [7,] -0.2333333 -0.2333333 -0.20  0.80 -0.20  0.8333333 -0.1666667  0.80
#>  [8,] -0.2800000 -0.2800000 -0.24 -0.24  0.76  0.8000000  0.8000000 -0.24
#>  [9,] -0.2333333 -0.2333333 -0.20 -0.20 -0.20  0.8333333  0.8333333  0.80
#> [10,]  0.7200000  0.7200000 -0.24 -0.24 -0.24 -0.2000000  0.8000000  0.76
#> [11,] -0.2333333 -0.2333333 -0.20 -0.20 -0.20 -0.1666667 -0.1666667  0.80
#> [12,] -0.2800000 -0.2800000 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24
#> [13,] -0.2800000 -0.2800000 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24
#> [14,] -0.2800000 -0.2800000 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24
#> [15,] -0.2800000 -0.2800000 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24
#> [16,] -0.2800000 -0.2800000 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24
#> [17,] -0.3266667 -0.3266667 -0.28 -0.28 -0.28 -0.2333333 -0.2333333 -0.28
#> [18,] -0.2800000 -0.2800000 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24
#> [19,] -0.2800000 -0.2800000 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24
#> [20,] -0.2333333 -0.2333333 -0.20 -0.20 -0.20 -0.1666667 -0.1666667 -0.20
#> [21,] -0.2333333 -0.2333333 -0.20 -0.20 -0.20 -0.1666667 -0.1666667 -0.20
#> [22,] -0.3266667 -0.3266667  0.72 -0.28 -0.28 -0.2333333 -0.2333333 -0.28
#> [23,]  0.6733333 -0.3266667 -0.28 -0.28  0.72 -0.2333333 -0.2333333 -0.28
#> [24,]  0.7200000  0.7200000 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24
#> [25,]  0.6733333  0.6733333  0.72 -0.28 -0.28 -0.2333333 -0.2333333 -0.28
#>             [,9] [,10]      [,11] [,12] [,13] [,14] [,15] [,16]      [,17]
#>  [1,] -0.2333333  0.72 -0.2333333 -0.28 -0.28 -0.28 -0.28 -0.28 -0.3266667
#>  [2,] -0.2333333  0.72 -0.2333333 -0.28 -0.28 -0.28 -0.28 -0.28 -0.3266667
#>  [3,] -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2800000
#>  [4,] -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2800000
#>  [5,] -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2800000
#>  [6,]  0.8333333 -0.20 -0.1666667 -0.20 -0.20 -0.20 -0.20 -0.20 -0.2333333
#>  [7,]  0.8333333  0.80 -0.1666667 -0.20 -0.20 -0.20 -0.20 -0.20 -0.2333333
#>  [8,]  0.8000000  0.76  0.8000000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2800000
#>  [9,] -0.1666667 -0.20  0.8333333  0.80 -0.20 -0.20 -0.20 -0.20 -0.2333333
#> [10,] -0.2000000 -0.24 -0.2000000  0.76  0.76 -0.24 -0.24 -0.24 -0.2800000
#> [11,]  0.8333333 -0.20 -0.1666667  0.80  0.80  0.80 -0.20 -0.20 -0.2333333
#> [12,]  0.8000000  0.76  0.8000000 -0.24  0.76  0.76  0.76 -0.24 -0.2800000
#> [13,] -0.2000000  0.76  0.8000000  0.76 -0.24  0.76  0.76  0.76 -0.2800000
#> [14,] -0.2000000 -0.24  0.8000000  0.76  0.76 -0.24  0.76  0.76  0.7200000
#> [15,] -0.2000000 -0.24 -0.2000000  0.76  0.76  0.76 -0.24  0.76  0.7200000
#> [16,] -0.2000000 -0.24 -0.2000000 -0.24  0.76  0.76  0.76 -0.24  0.7200000
#> [17,] -0.2333333 -0.28 -0.2333333 -0.28 -0.28  0.72  0.72  0.72 -0.3266667
#> [18,] -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.24  0.76  0.76  0.7200000
#> [19,] -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24  0.76  0.7200000
#> [20,] -0.1666667 -0.20 -0.1666667 -0.20 -0.20 -0.20 -0.20 -0.20  0.7666667
#> [21,] -0.1666667 -0.20 -0.1666667 -0.20 -0.20 -0.20 -0.20 -0.20 -0.2333333
#> [22,] -0.2333333 -0.28 -0.2333333 -0.28 -0.28 -0.28 -0.28 -0.28 -0.3266667
#> [23,] -0.2333333 -0.28 -0.2333333 -0.28 -0.28 -0.28 -0.28 -0.28  0.6733333
#> [24,] -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2800000
#> [25,] -0.2333333 -0.28 -0.2333333 -0.28 -0.28 -0.28 -0.28 -0.28 -0.3266667
#>       [,18] [,19]      [,20]      [,21]      [,22]      [,23] [,24]      [,25]
#>  [1,] -0.28 -0.28 -0.2333333 -0.2333333 -0.3266667  0.6733333  0.72  0.6733333
#>  [2,] -0.28 -0.28 -0.2333333 -0.2333333 -0.3266667 -0.3266667  0.72  0.6733333
#>  [3,] -0.24 -0.24 -0.2000000 -0.2000000  0.7200000 -0.2800000 -0.24  0.7200000
#>  [4,] -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.2800000 -0.24 -0.2800000
#>  [5,] -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000  0.7200000 -0.24 -0.2800000
#>  [6,] -0.20 -0.20 -0.1666667 -0.1666667 -0.2333333 -0.2333333 -0.20 -0.2333333
#>  [7,] -0.20 -0.20 -0.1666667 -0.1666667 -0.2333333 -0.2333333 -0.20 -0.2333333
#>  [8,] -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.2800000 -0.24 -0.2800000
#>  [9,] -0.20 -0.20 -0.1666667 -0.1666667 -0.2333333 -0.2333333 -0.20 -0.2333333
#> [10,] -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.2800000 -0.24 -0.2800000
#> [11,] -0.20 -0.20 -0.1666667 -0.1666667 -0.2333333 -0.2333333 -0.20 -0.2333333
#> [12,] -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.2800000 -0.24 -0.2800000
#> [13,] -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.2800000 -0.24 -0.2800000
#> [14,] -0.24 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.2800000 -0.24 -0.2800000
#> [15,]  0.76 -0.24 -0.2000000 -0.2000000 -0.2800000 -0.2800000 -0.24 -0.2800000
#> [16,]  0.76  0.76 -0.2000000 -0.2000000 -0.2800000 -0.2800000 -0.24 -0.2800000
#> [17,]  0.72  0.72  0.7666667 -0.2333333 -0.3266667  0.6733333 -0.28 -0.3266667
#> [18,] -0.24  0.76  0.8000000 -0.2000000 -0.2800000 -0.2800000 -0.24  0.7200000
#> [19,]  0.76 -0.24  0.8000000  0.8000000  0.7200000 -0.2800000 -0.24 -0.2800000
#> [20,]  0.80  0.80 -0.1666667  0.8333333  0.7666667 -0.2333333 -0.20 -0.2333333
#> [21,] -0.20  0.80  0.8333333 -0.1666667  0.7666667  0.7666667  0.80 -0.2333333
#> [22,] -0.28  0.72  0.7666667  0.7666667 -0.3266667  0.6733333  0.72  0.6733333
#> [23,] -0.28 -0.28 -0.2333333  0.7666667  0.6733333 -0.3266667  0.72  0.6733333
#> [24,] -0.24 -0.24 -0.2000000  0.8000000  0.7200000  0.7200000 -0.24  0.7200000
#> [25,]  0.72 -0.28 -0.2333333 -0.2333333  0.6733333  0.6733333  0.72 -0.3266667
#> attr(,"representation")
#> [1] "modularity"

repr_graphon(g)

#>             [,1]       [,2]      [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.57142857 0.71428571 0.7321429 0.53571429 0.36507937 0.23809524
#>  [2,] 0.71428571 0.57142857 0.6696429 0.52678571 0.61904762 0.29365079
#>  [3,] 0.73214286 0.66964286 0.3750000 0.37500000 0.47916667 0.23611111
#>  [4,] 0.53571429 0.52678571 0.3750000 0.37500000 0.52777778 0.47222222
#>  [5,] 0.36507937 0.61904762 0.4791667 0.52777778 0.33333333 0.44444444
#>  [6,] 0.23809524 0.29365079 0.2361111 0.47222222 0.44444444 0.55555556
#>  [7,] 0.00000000 0.14285714 0.2589286 0.52678571 0.36507937 0.56349206
#>  [8,] 0.14285714 0.07142857 0.2589286 0.39285714 0.49206349 0.61904762
#>  [9,] 0.11111111 0.11111111 0.1180556 0.23611111 0.27777778 0.44444444
#> [10,] 0.39285714 0.39285714 0.1875000 0.06250000 0.23611111 0.36111111
#> [11,] 0.07142857 0.07142857 0.0000000 0.26785714 0.12698413 0.25396825
#> [12,] 0.00000000 0.00000000 0.0625000 0.18750000 0.17361111 0.29166667
#> [13,] 0.00000000 0.00000000 0.0625000 0.06250000 0.11111111 0.22222222
#> [14,] 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000 0.05555556
#> [15,] 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000
#> [16,] 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000
#> [17,] 0.00000000 0.07142857 0.1250000 0.00000000 0.00000000 0.00000000
#> [18,] 0.07142857 0.07142857 0.0000000 0.00000000 0.00000000 0.00000000
#> [19,] 0.13392857 0.13392857 0.0000000 0.00000000 0.00000000 0.00000000
#> [20,] 0.19642857 0.13392857 0.0000000 0.00000000 0.06250000 0.00000000
#> [21,] 0.26785714 0.26785714 0.2500000 0.06250000 0.05555556 0.00000000
#> [22,] 0.43650794 0.38095238 0.4097222 0.23611111 0.22222222 0.05555556
#> [23,] 0.66964286 0.40178571 0.3125000 0.25000000 0.35416667 0.05555556
#> [24,] 0.61904762 0.63492063 0.2361111 0.05555556 0.22222222 0.05555556
#> [25,] 0.69047619 0.85714286 0.5833333 0.29166667 0.16666667 0.05555556
#>             [,7]       [,8]       [,9]      [,10]      [,11]      [,12]
#>  [1,] 0.00000000 0.14285714 0.11111111 0.39285714 0.07142857 0.00000000
#>  [2,] 0.14285714 0.07142857 0.11111111 0.39285714 0.07142857 0.00000000
#>  [3,] 0.25892857 0.25892857 0.11805556 0.18750000 0.00000000 0.06250000
#>  [4,] 0.52678571 0.39285714 0.23611111 0.06250000 0.26785714 0.18750000
#>  [5,] 0.36507937 0.49206349 0.27777778 0.23611111 0.12698413 0.17361111
#>  [6,] 0.56349206 0.61904762 0.44444444 0.36111111 0.25396825 0.29166667
#>  [7,] 0.42857143 0.64285714 0.63492063 0.58928571 0.42857143 0.20535714
#>  [8,] 0.64285714 0.42857143 0.56349206 0.66071429 0.64285714 0.26785714
#>  [9,] 0.63492063 0.56349206 0.44444444 0.29166667 0.50793651 0.52777778
#> [10,] 0.58928571 0.66071429 0.29166667 0.00000000 0.26785714 0.62500000
#> [11,] 0.42857143 0.64285714 0.50793651 0.26785714 0.28571429 0.53571429
#> [12,] 0.20535714 0.26785714 0.52777778 0.62500000 0.53571429 0.50000000
#> [13,] 0.25396825 0.18253968 0.33333333 0.41666667 0.56349206 0.59027778
#> [14,] 0.14285714 0.14285714 0.19444444 0.29166667 0.46428571 0.56250000
#> [15,] 0.07142857 0.22619048 0.19444444 0.08333333 0.23809524 0.58333333
#> [16,] 0.00000000 0.07142857 0.18253968 0.07142857 0.21428571 0.33035714
#> [17,] 0.00000000 0.00000000 0.05555556 0.06250000 0.13392857 0.31250000
#> [18,] 0.00000000 0.00000000 0.00000000 0.05555556 0.18253968 0.29166667
#> [19,] 0.00000000 0.00000000 0.00000000 0.00000000 0.06250000 0.18750000
#> [20,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.06250000
#> [21,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [22,] 0.00000000 0.00000000 0.00000000 0.11111111 0.00000000 0.00000000
#> [23,] 0.07142857 0.00000000 0.00000000 0.06250000 0.00000000 0.00000000
#> [24,] 0.11111111 0.05555556 0.00000000 0.00000000 0.00000000 0.05555556
#> [25,] 0.00000000 0.00000000 0.00000000 0.16666667 0.00000000 0.00000000
#>            [,13]      [,14]      [,15]      [,16]      [,17]      [,18]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.07142857
#>  [2,] 0.00000000 0.00000000 0.00000000 0.00000000 0.07142857 0.07142857
#>  [3,] 0.06250000 0.00000000 0.00000000 0.00000000 0.12500000 0.00000000
#>  [4,] 0.06250000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [5,] 0.11111111 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [6,] 0.22222222 0.05555556 0.00000000 0.00000000 0.00000000 0.00000000
#>  [7,] 0.25396825 0.14285714 0.07142857 0.00000000 0.00000000 0.00000000
#>  [8,] 0.18253968 0.14285714 0.22619048 0.07142857 0.00000000 0.00000000
#>  [9,] 0.33333333 0.19444444 0.19444444 0.18253968 0.05555556 0.00000000
#> [10,] 0.41666667 0.29166667 0.08333333 0.07142857 0.06250000 0.05555556
#> [11,] 0.56349206 0.46428571 0.23809524 0.21428571 0.13392857 0.18253968
#> [12,] 0.59027778 0.56250000 0.58333333 0.33035714 0.31250000 0.29166667
#> [13,] 0.55555556 0.61111111 0.61111111 0.50793651 0.34722222 0.27777778
#> [14,] 0.61111111 0.66666667 0.75000000 0.61904762 0.58333333 0.38888889
#> [15,] 0.61111111 0.75000000 0.50000000 0.69047619 0.58333333 0.52777778
#> [16,] 0.50793651 0.61904762 0.69047619 0.57142857 0.66964286 0.61904762
#> [17,] 0.34722222 0.58333333 0.58333333 0.66964286 0.50000000 0.58333333
#> [18,] 0.27777778 0.38888889 0.52777778 0.61904762 0.58333333 0.55555556
#> [19,] 0.29861111 0.29166667 0.27083333 0.53571429 0.68750000 0.52777778
#> [20,] 0.18055556 0.20833333 0.20833333 0.33035714 0.68750000 0.47222222
#> [21,] 0.00000000 0.06250000 0.20833333 0.25892857 0.31250000 0.36111111
#> [22,] 0.00000000 0.05555556 0.19444444 0.18253968 0.23611111 0.27777778
#> [23,] 0.00000000 0.00000000 0.00000000 0.13392857 0.37500000 0.29861111
#> [24,] 0.05555556 0.00000000 0.00000000 0.05555556 0.29166667 0.22222222
#> [25,] 0.05555556 0.08333333 0.00000000 0.07142857 0.12500000 0.22222222
#>           [,19]     [,20]      [,21]      [,22]      [,23]      [,24]
#>  [1,] 0.1339286 0.1964286 0.26785714 0.43650794 0.66964286 0.61904762
#>  [2,] 0.1339286 0.1339286 0.26785714 0.38095238 0.40178571 0.63492063
#>  [3,] 0.0000000 0.0000000 0.25000000 0.40972222 0.31250000 0.23611111
#>  [4,] 0.0000000 0.0000000 0.06250000 0.23611111 0.25000000 0.05555556
#>  [5,] 0.0000000 0.0625000 0.05555556 0.22222222 0.35416667 0.22222222
#>  [6,] 0.0000000 0.0000000 0.00000000 0.05555556 0.05555556 0.05555556
#>  [7,] 0.0000000 0.0000000 0.00000000 0.00000000 0.07142857 0.11111111
#>  [8,] 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000 0.05555556
#>  [9,] 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [10,] 0.0000000 0.0000000 0.00000000 0.11111111 0.06250000 0.00000000
#> [11,] 0.0625000 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [12,] 0.1875000 0.0625000 0.00000000 0.00000000 0.00000000 0.05555556
#> [13,] 0.2986111 0.1805556 0.00000000 0.00000000 0.00000000 0.05555556
#> [14,] 0.2916667 0.2083333 0.06250000 0.05555556 0.00000000 0.00000000
#> [15,] 0.2708333 0.2083333 0.20833333 0.19444444 0.00000000 0.00000000
#> [16,] 0.5357143 0.3303571 0.25892857 0.18253968 0.13392857 0.05555556
#> [17,] 0.6875000 0.6875000 0.31250000 0.23611111 0.37500000 0.29166667
#> [18,] 0.5277778 0.4722222 0.36111111 0.27777778 0.29861111 0.22222222
#> [19,] 0.5000000 0.4375000 0.50000000 0.47222222 0.25000000 0.17361111
#> [20,] 0.4375000 0.3750000 0.56250000 0.53472222 0.25000000 0.23611111
#> [21,] 0.5000000 0.5625000 0.50000000 0.47916667 0.62500000 0.40972222
#> [22,] 0.4722222 0.5347222 0.47916667 0.44444444 0.70833333 0.72222222
#> [23,] 0.2500000 0.2500000 0.62500000 0.70833333 0.37500000 0.41666667
#> [24,] 0.1736111 0.2361111 0.40972222 0.72222222 0.41666667 0.33333333
#> [25,] 0.1250000 0.1875000 0.35416667 0.66666667 0.81250000 0.50000000
#>            [,25]
#>  [1,] 0.69047619
#>  [2,] 0.85714286
#>  [3,] 0.58333333
#>  [4,] 0.29166667
#>  [5,] 0.16666667
#>  [6,] 0.05555556
#>  [7,] 0.00000000
#>  [8,] 0.00000000
#>  [9,] 0.00000000
#> [10,] 0.16666667
#> [11,] 0.00000000
#> [12,] 0.00000000
#> [13,] 0.05555556
#> [14,] 0.08333333
#> [15,] 0.00000000
#> [16,] 0.07142857
#> [17,] 0.12500000
#> [18,] 0.22222222
#> [19,] 0.12500000
#> [20,] 0.18750000
#> [21,] 0.35416667
#> [22,] 0.66666667
#> [23,] 0.81250000
#> [24,] 0.50000000
#> [25,] 0.66666667
#> attr(,"representation")
#> [1] "graphon"