Network Representation Functions

This is a collection of functions that convert a graph stored as an igraph object into a desired matrix representation among adjacency matrix, graph laplacian, modularity matrix or graphon (edge probability matrix).

repr_adjacency(network, validate = TRUE)

repr_laplacian(network, validate = TRUE)

repr_modularity(network, validate = TRUE)

repr_graphon(network, validate = TRUE)

Arguments

network
An igraph object.

validate
A boolean specifying whether the function should check the class of its input (default: TRUE).

Value

A numeric square matrix giving the desired network representation recorded in the object's class.

Examples

g <- igraph::sample_smallworld(1, 25, 3, 0.05)
repr_adjacency(g)

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0    1    1    1    0    0    0    0    0     0     0     0     0
#>  [2,]    1    0    1    1    1    0    0    0    0     0     0     0     0
#>  [3,]    1    1    0    0    1    1    0    0    0     0     0     0     0
#>  [4,]    1    1    0    0    1    1    1    0    0     0     0     0     0
#>  [5,]    0    1    1    1    0    1    1    1    0     0     0     0     0
#>  [6,]    0    0    1    1    1    0    1    1    1     0     0     0     0
#>  [7,]    0    0    0    1    1    1    0    1    1     1     0     0     0
#>  [8,]    0    0    0    0    1    1    1    0    1     1     0     0     0
#>  [9,]    0    0    0    0    0    1    1    1    0     1     1     1     0
#> [10,]    0    0    0    0    0    0    1    1    1     0     1     1     1
#> [11,]    0    0    0    0    0    0    0    0    1     1     0     1     1
#> [12,]    0    0    0    0    0    0    0    0    1     1     1     0     1
#> [13,]    0    0    0    0    0    0    0    0    0     1     1     1     0
#> [14,]    0    1    0    0    0    0    0    0    0     1     1     1     1
#> [15,]    0    0    1    0    0    0    0    0    0     0     0     1     1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0     1
#> [17,]    0    0    0    0    0    0    0    0    0     0     1     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]    1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]    1    1    0    0    0    0    0    0    0     0     0     0     0
#> [25,]    1    1    1    0    0    1    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0     1     1     1
#>  [2,]     1     0     0     0     0     0     0     0     0     0     1     1
#>  [3,]     0     1     0     0     0     0     0     0     0     0     0     1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     1
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     1     0     0     0     0     0     0     0     0     0     0     0
#> [11,]     1     0     0     1     0     0     0     0     0     0     0     0
#> [12,]     1     1     0     0     0     0     0     0     0     0     0     0
#> [13,]     1     1     1     0     0     0     0     0     0     0     0     0
#> [14,]     0     0     0     1     0     0     0     0     0     0     0     0
#> [15,]     0     0     1     1     1     0     0     0     0     0     0     0
#> [16,]     0     1     0     1     1     1     0     0     0     0     0     0
#> [17,]     1     1     1     0     1     0     1     0     0     0     0     0
#> [18,]     0     1     1     1     0     1     1     0     0     0     1     0
#> [19,]     0     0     1     0     1     0     1     1     1     0     0     0
#> [20,]     0     0     0     1     1     1     0     1     1     1     0     0
#> [21,]     0     0     0     0     0     1     1     0     1     1     1     0
#> [22,]     0     0     0     0     0     1     1     1     0     1     1     1
#> [23,]     0     0     0     0     0     0     1     1     1     0     1     1
#> [24,]     0     0     0     0     1     0     0     1     1     1     0     1
#> [25,]     0     0     0     0     0     0     0     0     1     1     1     0
#> attr(,"representation")
#> [1] "adjacency"

repr_laplacian(g)

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    6   -1   -1   -1    0    0    0    0    0     0     0     0     0
#>  [2,]   -1    7   -1   -1   -1    0    0    0    0     0     0     0     0
#>  [3,]   -1   -1    6    0   -1   -1    0    0    0     0     0     0     0
#>  [4,]   -1   -1    0    5   -1   -1   -1    0    0     0     0     0     0
#>  [5,]    0   -1   -1   -1    6   -1   -1   -1    0     0     0     0     0
#>  [6,]    0    0   -1   -1   -1    7   -1   -1   -1     0     0     0     0
#>  [7,]    0    0    0   -1   -1   -1    6   -1   -1    -1     0     0     0
#>  [8,]    0    0    0    0   -1   -1   -1    5   -1    -1     0     0     0
#>  [9,]    0    0    0    0    0   -1   -1   -1    6    -1    -1    -1     0
#> [10,]    0    0    0    0    0    0   -1   -1   -1     7    -1    -1    -1
#> [11,]    0    0    0    0    0    0    0    0   -1    -1     6    -1    -1
#> [12,]    0    0    0    0    0    0    0    0   -1    -1    -1     6    -1
#> [13,]    0    0    0    0    0    0    0    0    0    -1    -1    -1     6
#> [14,]    0   -1    0    0    0    0    0    0    0    -1    -1    -1    -1
#> [15,]    0    0   -1    0    0    0    0    0    0     0     0    -1    -1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0    -1
#> [17,]    0    0    0    0    0    0    0    0    0     0    -1     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]   -1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]   -1   -1    0    0    0    0    0    0    0     0     0     0     0
#> [25,]   -1   -1   -1    0    0   -1    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0    -1    -1    -1
#>  [2,]    -1     0     0     0     0     0     0     0     0     0    -1    -1
#>  [3,]     0    -1     0     0     0     0     0     0     0     0     0    -1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0    -1
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]    -1     0     0     0     0     0     0     0     0     0     0     0
#> [11,]    -1     0     0    -1     0     0     0     0     0     0     0     0
#> [12,]    -1    -1     0     0     0     0     0     0     0     0     0     0
#> [13,]    -1    -1    -1     0     0     0     0     0     0     0     0     0
#> [14,]     6     0     0    -1     0     0     0     0     0     0     0     0
#> [15,]     0     6    -1    -1    -1     0     0     0     0     0     0     0
#> [16,]     0    -1     5    -1    -1    -1     0     0     0     0     0     0
#> [17,]    -1    -1    -1     6    -1     0    -1     0     0     0     0     0
#> [18,]     0    -1    -1    -1     6    -1    -1     0     0     0    -1     0
#> [19,]     0     0    -1     0    -1     5    -1    -1    -1     0     0     0
#> [20,]     0     0     0    -1    -1    -1     6    -1    -1    -1     0     0
#> [21,]     0     0     0     0     0    -1    -1     5    -1    -1    -1     0
#> [22,]     0     0     0     0     0    -1    -1    -1     6    -1    -1    -1
#> [23,]     0     0     0     0     0     0    -1    -1    -1     6    -1    -1
#> [24,]     0     0     0     0    -1     0     0    -1    -1    -1     7    -1
#> [25,]     0     0     0     0     0     0     0     0    -1    -1    -1     7
#> attr(,"representation")
#> [1] "laplacian"

repr_modularity(g)

#>        [,1]       [,2]  [,3]       [,4]  [,5]       [,6]  [,7]       [,8]  [,9]
#>  [1,] -0.24  0.7200000  0.76  0.8000000 -0.24 -0.2800000 -0.24 -0.2000000 -0.24
#>  [2,]  0.72 -0.3266667  0.72  0.7666667  0.72 -0.3266667 -0.28 -0.2333333 -0.28
#>  [3,]  0.76  0.7200000 -0.24 -0.2000000  0.76  0.7200000 -0.24 -0.2000000 -0.24
#>  [4,]  0.80  0.7666667 -0.20 -0.1666667  0.80  0.7666667  0.80 -0.1666667 -0.20
#>  [5,] -0.24  0.7200000  0.76  0.8000000 -0.24  0.7200000  0.76  0.8000000 -0.24
#>  [6,] -0.28 -0.3266667  0.72  0.7666667  0.72 -0.3266667  0.72  0.7666667  0.72
#>  [7,] -0.24 -0.2800000 -0.24  0.8000000  0.76  0.7200000 -0.24  0.8000000  0.76
#>  [8,] -0.20 -0.2333333 -0.20 -0.1666667  0.80  0.7666667  0.80 -0.1666667  0.80
#>  [9,] -0.24 -0.2800000 -0.24 -0.2000000 -0.24  0.7200000  0.76  0.8000000 -0.24
#> [10,] -0.28 -0.3266667 -0.28 -0.2333333 -0.28 -0.3266667  0.72  0.7666667  0.72
#> [11,] -0.24 -0.2800000 -0.24 -0.2000000 -0.24 -0.2800000 -0.24 -0.2000000  0.76
#> [12,] -0.24 -0.2800000 -0.24 -0.2000000 -0.24 -0.2800000 -0.24 -0.2000000  0.76
#> [13,] -0.24 -0.2800000 -0.24 -0.2000000 -0.24 -0.2800000 -0.24 -0.2000000 -0.24
#> [14,] -0.24  0.7200000 -0.24 -0.2000000 -0.24 -0.2800000 -0.24 -0.2000000 -0.24
#> [15,] -0.24 -0.2800000  0.76 -0.2000000 -0.24 -0.2800000 -0.24 -0.2000000 -0.24
#> [16,] -0.20 -0.2333333 -0.20 -0.1666667 -0.20 -0.2333333 -0.20 -0.1666667 -0.20
#> [17,] -0.24 -0.2800000 -0.24 -0.2000000 -0.24 -0.2800000 -0.24 -0.2000000 -0.24
#> [18,] -0.24 -0.2800000 -0.24 -0.2000000 -0.24 -0.2800000 -0.24 -0.2000000 -0.24
#> [19,] -0.20 -0.2333333 -0.20 -0.1666667 -0.20 -0.2333333 -0.20 -0.1666667 -0.20
#> [20,] -0.24 -0.2800000 -0.24 -0.2000000 -0.24 -0.2800000 -0.24 -0.2000000 -0.24
#> [21,] -0.20 -0.2333333 -0.20 -0.1666667 -0.20 -0.2333333 -0.20 -0.1666667 -0.20
#> [22,] -0.24 -0.2800000 -0.24 -0.2000000 -0.24 -0.2800000 -0.24 -0.2000000 -0.24
#> [23,]  0.76 -0.2800000 -0.24 -0.2000000 -0.24 -0.2800000 -0.24 -0.2000000 -0.24
#> [24,]  0.72  0.6733333 -0.28 -0.2333333 -0.28 -0.3266667 -0.28 -0.2333333 -0.28
#> [25,]  0.72  0.6733333  0.72 -0.2333333 -0.28  0.6733333 -0.28 -0.2333333 -0.28
#>            [,10] [,11] [,12] [,13] [,14] [,15]      [,16] [,17] [,18]
#>  [1,] -0.2800000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24
#>  [2,] -0.3266667 -0.28 -0.28 -0.28  0.72 -0.28 -0.2333333 -0.28 -0.28
#>  [3,] -0.2800000 -0.24 -0.24 -0.24 -0.24  0.76 -0.2000000 -0.24 -0.24
#>  [4,] -0.2333333 -0.20 -0.20 -0.20 -0.20 -0.20 -0.1666667 -0.20 -0.20
#>  [5,] -0.2800000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24
#>  [6,] -0.3266667 -0.28 -0.28 -0.28 -0.28 -0.28 -0.2333333 -0.28 -0.28
#>  [7,]  0.7200000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24
#>  [8,]  0.7666667 -0.20 -0.20 -0.20 -0.20 -0.20 -0.1666667 -0.20 -0.20
#>  [9,]  0.7200000  0.76  0.76 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24
#> [10,] -0.3266667  0.72  0.72  0.72  0.72 -0.28 -0.2333333 -0.28 -0.28
#> [11,]  0.7200000 -0.24  0.76  0.76  0.76 -0.24 -0.2000000  0.76 -0.24
#> [12,]  0.7200000  0.76 -0.24  0.76  0.76  0.76 -0.2000000 -0.24 -0.24
#> [13,]  0.7200000  0.76  0.76 -0.24  0.76  0.76  0.8000000 -0.24 -0.24
#> [14,]  0.7200000  0.76  0.76  0.76 -0.24 -0.24 -0.2000000  0.76 -0.24
#> [15,] -0.2800000 -0.24  0.76  0.76 -0.24 -0.24  0.8000000  0.76  0.76
#> [16,] -0.2333333 -0.20 -0.20  0.80 -0.20  0.80 -0.1666667  0.80  0.80
#> [17,] -0.2800000  0.76 -0.24 -0.24  0.76  0.76  0.8000000 -0.24  0.76
#> [18,] -0.2800000 -0.24 -0.24 -0.24 -0.24  0.76  0.8000000  0.76 -0.24
#> [19,] -0.2333333 -0.20 -0.20 -0.20 -0.20 -0.20  0.8333333 -0.20  0.80
#> [20,] -0.2800000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000  0.76  0.76
#> [21,] -0.2333333 -0.20 -0.20 -0.20 -0.20 -0.20 -0.1666667 -0.20 -0.20
#> [22,] -0.2800000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24
#> [23,] -0.2800000 -0.24 -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.24 -0.24
#> [24,] -0.3266667 -0.28 -0.28 -0.28 -0.28 -0.28 -0.2333333 -0.28  0.72
#> [25,] -0.3266667 -0.28 -0.28 -0.28 -0.28 -0.28 -0.2333333 -0.28 -0.28
#>            [,19] [,20]      [,21] [,22] [,23]      [,24]      [,25]
#>  [1,] -0.2000000 -0.24 -0.2000000 -0.24  0.76  0.7200000  0.7200000
#>  [2,] -0.2333333 -0.28 -0.2333333 -0.28 -0.28  0.6733333  0.6733333
#>  [3,] -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.2800000  0.7200000
#>  [4,] -0.1666667 -0.20 -0.1666667 -0.20 -0.20 -0.2333333 -0.2333333
#>  [5,] -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.2800000 -0.2800000
#>  [6,] -0.2333333 -0.28 -0.2333333 -0.28 -0.28 -0.3266667  0.6733333
#>  [7,] -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.2800000 -0.2800000
#>  [8,] -0.1666667 -0.20 -0.1666667 -0.20 -0.20 -0.2333333 -0.2333333
#>  [9,] -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.2800000 -0.2800000
#> [10,] -0.2333333 -0.28 -0.2333333 -0.28 -0.28 -0.3266667 -0.3266667
#> [11,] -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.2800000 -0.2800000
#> [12,] -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.2800000 -0.2800000
#> [13,] -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.2800000 -0.2800000
#> [14,] -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.2800000 -0.2800000
#> [15,] -0.2000000 -0.24 -0.2000000 -0.24 -0.24 -0.2800000 -0.2800000
#> [16,]  0.8333333 -0.20 -0.1666667 -0.20 -0.20 -0.2333333 -0.2333333
#> [17,] -0.2000000  0.76 -0.2000000 -0.24 -0.24 -0.2800000 -0.2800000
#> [18,]  0.8000000  0.76 -0.2000000 -0.24 -0.24  0.7200000 -0.2800000
#> [19,] -0.1666667  0.80  0.8333333  0.80 -0.20 -0.2333333 -0.2333333
#> [20,]  0.8000000 -0.24  0.8000000  0.76  0.76 -0.2800000 -0.2800000
#> [21,]  0.8333333  0.80 -0.1666667  0.80  0.80  0.7666667 -0.2333333
#> [22,]  0.8000000  0.76  0.8000000 -0.24  0.76  0.7200000  0.7200000
#> [23,] -0.2000000  0.76  0.8000000  0.76 -0.24  0.7200000  0.7200000
#> [24,] -0.2333333 -0.28  0.7666667  0.72  0.72 -0.3266667  0.6733333
#> [25,] -0.2333333 -0.28 -0.2333333  0.72  0.72  0.6733333 -0.3266667
#> attr(,"representation")
#> [1] "modularity"

repr_graphon(g)

#>            [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.5000000 0.58333333 0.46428571 0.40178571 0.29166667 0.25892857
#>  [2,] 0.5833333 0.50000000 0.61904762 0.53571429 0.52777778 0.30952381
#>  [3,] 0.4642857 0.61904762 0.28571429 0.35714286 0.50793651 0.64285714
#>  [4,] 0.4017857 0.53571429 0.35714286 0.28571429 0.50793651 0.85714286
#>  [5,] 0.2916667 0.52777778 0.50793651 0.50793651 0.55555556 0.67460317
#>  [6,] 0.2589286 0.30952381 0.64285714 0.85714286 0.67460317 0.42857143
#>  [7,] 0.1736111 0.25000000 0.32539683 0.45238095 0.55555556 0.61904762
#>  [8,] 0.1736111 0.25000000 0.25396825 0.38095238 0.50000000 0.61904762
#>  [9,] 0.0000000 0.22619048 0.21428571 0.28571429 0.30952381 0.57142857
#> [10,] 0.0000000 0.07142857 0.21428571 0.28571429 0.23809524 0.21428571
#> [11,] 0.0000000 0.08333333 0.08333333 0.07142857 0.05555556 0.07142857
#> [12,] 0.0000000 0.08333333 0.00000000 0.07142857 0.05555556 0.07142857
#> [13,] 0.0000000 0.05555556 0.05555556 0.00000000 0.00000000 0.12698413
#> [14,] 0.0625000 0.14583333 0.06250000 0.00000000 0.00000000 0.20535714
#> [15,] 0.0625000 0.13888889 0.12698413 0.07142857 0.05555556 0.00000000
#> [16,] 0.0000000 0.06250000 0.06250000 0.00000000 0.00000000 0.00000000
#> [17,] 0.0000000 0.00000000 0.07142857 0.00000000 0.00000000 0.00000000
#> [18,] 0.0000000 0.00000000 0.06250000 0.00000000 0.00000000 0.00000000
#> [19,] 0.1805556 0.00000000 0.05555556 0.00000000 0.00000000 0.00000000
#> [20,] 0.2986111 0.13888889 0.12698413 0.00000000 0.00000000 0.00000000
#> [21,] 0.2361111 0.19444444 0.18253968 0.05555556 0.00000000 0.05555556
#> [22,] 0.3125000 0.29166667 0.13392857 0.13392857 0.05555556 0.00000000
#> [23,] 0.4166667 0.33333333 0.12698413 0.12698413 0.16666667 0.12698413
#> [24,] 0.6250000 0.58333333 0.26785714 0.13392857 0.17361111 0.26785714
#> [25,] 0.6458333 0.75000000 0.43650794 0.23809524 0.33333333 0.43650794
#>             [,7]       [,8]      [,9]      [,10]      [,11]      [,12]
#>  [1,] 0.17361111 0.17361111 0.0000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.25000000 0.25000000 0.2261905 0.07142857 0.08333333 0.08333333
#>  [3,] 0.32539683 0.25396825 0.2142857 0.21428571 0.08333333 0.00000000
#>  [4,] 0.45238095 0.38095238 0.2857143 0.28571429 0.07142857 0.07142857
#>  [5,] 0.55555556 0.50000000 0.3095238 0.23809524 0.05555556 0.05555556
#>  [6,] 0.61904762 0.61904762 0.5714286 0.21428571 0.07142857 0.07142857
#>  [7,] 0.55555556 0.50000000 0.5634921 0.49206349 0.19444444 0.19444444
#>  [8,] 0.50000000 0.44444444 0.5634921 0.49206349 0.19444444 0.19444444
#>  [9,] 0.56349206 0.56349206 0.4285714 0.64285714 0.53571429 0.53571429
#> [10,] 0.49206349 0.49206349 0.6428571 0.71428571 0.69047619 0.69047619
#> [11,] 0.19444444 0.19444444 0.5357143 0.69047619 0.66666667 0.83333333
#> [12,] 0.19444444 0.19444444 0.5357143 0.69047619 0.83333333 0.66666667
#> [13,] 0.22222222 0.22222222 0.3095238 0.56349206 0.75000000 0.66666667
#> [14,] 0.23611111 0.23611111 0.3303571 0.59821429 0.64583333 0.79166667
#> [15,] 0.11111111 0.11111111 0.1269841 0.30952381 0.33333333 0.47222222
#> [16,] 0.00000000 0.00000000 0.1964286 0.25892857 0.35416667 0.35416667
#> [17,] 0.05555556 0.05555556 0.1428571 0.28571429 0.46428571 0.30952381
#> [18,] 0.00000000 0.00000000 0.0000000 0.06250000 0.20833333 0.20833333
#> [19,] 0.00000000 0.00000000 0.0000000 0.00000000 0.05555556 0.05555556
#> [20,] 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000 0.13888889
#> [21,] 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000
#> [22,] 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000
#> [23,] 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000
#> [24,] 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000
#> [25,] 0.22222222 0.16666667 0.0000000 0.07142857 0.00000000 0.00000000
#>            [,13]      [,14]      [,15]     [,16]      [,17]     [,18]
#>  [1,] 0.00000000 0.06250000 0.06250000 0.0000000 0.00000000 0.0000000
#>  [2,] 0.05555556 0.14583333 0.13888889 0.0625000 0.00000000 0.0000000
#>  [3,] 0.05555556 0.06250000 0.12698413 0.0625000 0.07142857 0.0625000
#>  [4,] 0.00000000 0.00000000 0.07142857 0.0000000 0.00000000 0.0000000
#>  [5,] 0.00000000 0.00000000 0.05555556 0.0000000 0.00000000 0.0000000
#>  [6,] 0.12698413 0.20535714 0.00000000 0.0000000 0.00000000 0.0000000
#>  [7,] 0.22222222 0.23611111 0.11111111 0.0000000 0.05555556 0.0000000
#>  [8,] 0.22222222 0.23611111 0.11111111 0.0000000 0.05555556 0.0000000
#>  [9,] 0.30952381 0.33035714 0.12698413 0.1964286 0.14285714 0.0000000
#> [10,] 0.56349206 0.59821429 0.30952381 0.2589286 0.28571429 0.0625000
#> [11,] 0.75000000 0.64583333 0.33333333 0.3541667 0.46428571 0.2083333
#> [12,] 0.66666667 0.79166667 0.47222222 0.3541667 0.30952381 0.2083333
#> [13,] 0.55555556 0.65277778 0.44444444 0.5347222 0.43650794 0.2361111
#> [14,] 0.65277778 0.50000000 0.35416667 0.3125000 0.46428571 0.2500000
#> [15,] 0.44444444 0.35416667 0.44444444 0.5277778 0.69047619 0.5277778
#> [16,] 0.53472222 0.31250000 0.52777778 0.3750000 0.73214286 0.5625000
#> [17,] 0.43650794 0.46428571 0.69047619 0.7321429 0.42857143 0.4732143
#> [18,] 0.23611111 0.25000000 0.52777778 0.5625000 0.47321429 0.5000000
#> [19,] 0.16666667 0.11805556 0.33333333 0.4097222 0.36507937 0.5277778
#> [20,] 0.22222222 0.00000000 0.27777778 0.2916667 0.43650794 0.5902778
#> [21,] 0.05555556 0.00000000 0.22222222 0.1736111 0.23809524 0.2916667
#> [22,] 0.00000000 0.00000000 0.17361111 0.1875000 0.19642857 0.3125000
#> [23,] 0.00000000 0.00000000 0.16666667 0.1736111 0.11111111 0.2916667
#> [24,] 0.05555556 0.00000000 0.11805556 0.1250000 0.13392857 0.3750000
#> [25,] 0.00000000 0.05555556 0.05555556 0.0000000 0.00000000 0.1180556
#>            [,19]     [,20]      [,21]      [,22]     [,23]      [,24]
#>  [1,] 0.18055556 0.2986111 0.23611111 0.31250000 0.4166667 0.62500000
#>  [2,] 0.00000000 0.1388889 0.19444444 0.29166667 0.3333333 0.58333333
#>  [3,] 0.05555556 0.1269841 0.18253968 0.13392857 0.1269841 0.26785714
#>  [4,] 0.00000000 0.0000000 0.05555556 0.13392857 0.1269841 0.13392857
#>  [5,] 0.00000000 0.0000000 0.00000000 0.05555556 0.1666667 0.17361111
#>  [6,] 0.00000000 0.0000000 0.05555556 0.00000000 0.1269841 0.26785714
#>  [7,] 0.00000000 0.0000000 0.00000000 0.00000000 0.0000000 0.00000000
#>  [8,] 0.00000000 0.0000000 0.00000000 0.00000000 0.0000000 0.00000000
#>  [9,] 0.00000000 0.0000000 0.00000000 0.00000000 0.0000000 0.00000000
#> [10,] 0.00000000 0.0000000 0.00000000 0.00000000 0.0000000 0.00000000
#> [11,] 0.05555556 0.0000000 0.00000000 0.00000000 0.0000000 0.00000000
#> [12,] 0.05555556 0.1388889 0.00000000 0.00000000 0.0000000 0.00000000
#> [13,] 0.16666667 0.2222222 0.05555556 0.00000000 0.0000000 0.05555556
#> [14,] 0.11805556 0.0000000 0.00000000 0.00000000 0.0000000 0.00000000
#> [15,] 0.33333333 0.2777778 0.22222222 0.17361111 0.1666667 0.11805556
#> [16,] 0.40972222 0.2916667 0.17361111 0.18750000 0.1736111 0.12500000
#> [17,] 0.36507937 0.4365079 0.23809524 0.19642857 0.1111111 0.13392857
#> [18,] 0.52777778 0.5902778 0.29166667 0.31250000 0.2916667 0.37500000
#> [19,] 0.55555556 0.6111111 0.50000000 0.47222222 0.3888889 0.34722222
#> [20,] 0.61111111 0.5555556 0.55555556 0.59027778 0.5000000 0.40972222
#> [21,] 0.50000000 0.5555556 0.44444444 0.59027778 0.5555556 0.64583333
#> [22,] 0.47222222 0.5902778 0.59027778 0.62500000 0.5902778 0.62500000
#> [23,] 0.38888889 0.5000000 0.55555556 0.59027778 0.5555556 0.59027778
#> [24,] 0.34722222 0.4097222 0.64583333 0.62500000 0.5902778 0.50000000
#> [25,] 0.16666667 0.2777778 0.27777778 0.47222222 0.5000000 0.59027778
#>            [,25]
#>  [1,] 0.64583333
#>  [2,] 0.75000000
#>  [3,] 0.43650794
#>  [4,] 0.23809524
#>  [5,] 0.33333333
#>  [6,] 0.43650794
#>  [7,] 0.22222222
#>  [8,] 0.16666667
#>  [9,] 0.00000000
#> [10,] 0.07142857
#> [11,] 0.00000000
#> [12,] 0.00000000
#> [13,] 0.00000000
#> [14,] 0.05555556
#> [15,] 0.05555556
#> [16,] 0.00000000
#> [17,] 0.00000000
#> [18,] 0.11805556
#> [19,] 0.16666667
#> [20,] 0.27777778
#> [21,] 0.27777778
#> [22,] 0.47222222
#> [23,] 0.50000000
#> [24,] 0.59027778
#> [25,] 0.55555556
#> attr(,"representation")
#> [1] "graphon"