Network Representation Functions

This is a collection of functions that convert a graph stored as an igraph object into a desired matrix representation among adjacency matrix, graph laplacian, modularity matrix or graphon (edge probability matrix).

Usage

repr_adjacency(network, validate = TRUE)

repr_laplacian(network, validate = TRUE)

repr_modularity(network, validate = TRUE)

repr_graphon(network, validate = TRUE)

Arguments

network

An igraph object.

validate

A boolean specifying whether the function should check the class of its input (default: TRUE).

Value

A numeric square matrix giving the desired network representation recorded in the object's class.

Examples

g <- igraph::sample_smallworld(1, 25, 3, 0.05)
repr_adjacency(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0    1    1    1    0    0    0    0    0     0     0     0     0
#>  [2,]    1    0    1    1    1    0    0    0    0     0     0     0     0
#>  [3,]    1    1    0    1    1    0    0    0    0     0     0     0     0
#>  [4,]    1    1    1    0    1    1    1    0    0     0     0     0     0
#>  [5,]    0    1    1    1    0    1    0    1    0     1     0     0     0
#>  [6,]    0    0    0    1    1    0    1    1    1     0     0     0     0
#>  [7,]    0    0    0    1    0    1    0    1    1     1     0     0     0
#>  [8,]    0    0    0    0    1    1    1    0    0     1     1     0     0
#>  [9,]    0    0    0    0    0    1    1    0    0     1     1     1     0
#> [10,]    0    0    0    0    1    0    1    1    1     0     1     1     1
#> [11,]    0    0    0    0    0    0    0    1    1     1     0     1     1
#> [12,]    0    0    0    0    0    0    0    0    1     1     1     0     1
#> [13,]    0    0    0    0    0    0    0    0    0     1     1     1     0
#> [14,]    0    0    0    0    0    0    0    0    1     0     1     1     1
#> [15,]    0    0    0    0    0    1    0    0    0     0     0     1     1
#> [16,]    1    0    0    0    0    0    0    0    0     0     0     0     1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    1    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    1    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]    1    1    0    0    0    0    0    0    0     0     0     0     0
#> [25,]    1    1    0    0    0    1    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     1     0     0     0     0     0     0     0     1     1
#>  [2,]     0     0     0     0     0     0     0     0     0     0     1     1
#>  [3,]     0     0     0     0     0     1     0     1     0     0     0     0
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0     1     0     0     0     0     0     0     0     0     0     1
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     1     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]     1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]     1     1     0     0     0     0     0     0     0     0     0     0
#> [13,]     1     1     1     0     0     0     0     0     0     0     0     0
#> [14,]     0     1     1     1     0     0     0     0     0     0     0     0
#> [15,]     1     0     1     1     1     0     0     0     0     0     0     0
#> [16,]     1     1     0     1     1     0     0     0     0     0     0     0
#> [17,]     1     1     1     0     1     1     1     0     0     0     0     0
#> [18,]     0     1     1     1     0     1     1     1     0     0     0     0
#> [19,]     0     0     0     1     1     0     1     1     1     0     0     0
#> [20,]     0     0     0     1     1     1     0     1     1     1     0     0
#> [21,]     0     0     0     0     1     1     1     0     1     1     0     0
#> [22,]     0     0     0     0     0     1     1     1     0     1     1     1
#> [23,]     0     0     0     0     0     0     1     1     1     0     1     1
#> [24,]     0     0     0     0     0     0     0     0     1     1     0     1
#> [25,]     0     0     0     0     0     0     0     0     1     1     1     0
#> attr(,"representation")
#> [1] "adjacency"
repr_laplacian(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    6   -1   -1   -1    0    0    0    0    0     0     0     0     0
#>  [2,]   -1    6   -1   -1   -1    0    0    0    0     0     0     0     0
#>  [3,]   -1   -1    6   -1   -1    0    0    0    0     0     0     0     0
#>  [4,]   -1   -1   -1    6   -1   -1   -1    0    0     0     0     0     0
#>  [5,]    0   -1   -1   -1    6   -1    0   -1    0    -1     0     0     0
#>  [6,]    0    0    0   -1   -1    7   -1   -1   -1     0     0     0     0
#>  [7,]    0    0    0   -1    0   -1    5   -1   -1    -1     0     0     0
#>  [8,]    0    0    0    0   -1   -1   -1    5    0    -1    -1     0     0
#>  [9,]    0    0    0    0    0   -1   -1    0    6    -1    -1    -1     0
#> [10,]    0    0    0    0   -1    0   -1   -1   -1     7    -1    -1    -1
#> [11,]    0    0    0    0    0    0    0   -1   -1    -1     6    -1    -1
#> [12,]    0    0    0    0    0    0    0    0   -1    -1    -1     6    -1
#> [13,]    0    0    0    0    0    0    0    0    0    -1    -1    -1     6
#> [14,]    0    0    0    0    0    0    0    0   -1     0    -1    -1    -1
#> [15,]    0    0    0    0    0   -1    0    0    0     0     0    -1    -1
#> [16,]   -1    0    0    0    0    0    0    0    0     0     0     0    -1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0   -1    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0   -1    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]   -1   -1    0    0    0    0    0    0    0     0     0     0     0
#> [25,]   -1   -1    0    0    0   -1    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0    -1     0     0     0     0     0     0     0    -1    -1
#>  [2,]     0     0     0     0     0     0     0     0     0     0    -1    -1
#>  [3,]     0     0     0     0     0    -1     0    -1     0     0     0     0
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [6,]     0    -1     0     0     0     0     0     0     0     0     0    -1
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]    -1     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]    -1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]    -1    -1     0     0     0     0     0     0     0     0     0     0
#> [13,]    -1    -1    -1     0     0     0     0     0     0     0     0     0
#> [14,]     7    -1    -1    -1     0     0     0     0     0     0     0     0
#> [15,]    -1     7    -1    -1    -1     0     0     0     0     0     0     0
#> [16,]    -1    -1     6    -1    -1     0     0     0     0     0     0     0
#> [17,]    -1    -1    -1     6    -1    -1    -1     0     0     0     0     0
#> [18,]     0    -1    -1    -1     6    -1    -1    -1     0     0     0     0
#> [19,]     0     0     0    -1    -1     6    -1    -1    -1     0     0     0
#> [20,]     0     0     0    -1    -1    -1     6    -1    -1    -1     0     0
#> [21,]     0     0     0     0    -1    -1    -1     6    -1    -1     0     0
#> [22,]     0     0     0     0     0    -1    -1    -1     6    -1    -1    -1
#> [23,]     0     0     0     0     0     0    -1    -1    -1     5    -1    -1
#> [24,]     0     0     0     0     0     0     0     0    -1    -1     5    -1
#> [25,]     0     0     0     0     0     0     0     0    -1    -1    -1     6
#> attr(,"representation")
#> [1] "laplacian"
repr_modularity(g)
#>        [,1]  [,2]  [,3]  [,4]  [,5]       [,6]       [,7]       [,8]  [,9]
#>  [1,] -0.24  0.76  0.76  0.76 -0.24 -0.2800000 -0.2000000 -0.2000000 -0.24
#>  [2,]  0.76 -0.24  0.76  0.76  0.76 -0.2800000 -0.2000000 -0.2000000 -0.24
#>  [3,]  0.76  0.76 -0.24  0.76  0.76 -0.2800000 -0.2000000 -0.2000000 -0.24
#>  [4,]  0.76  0.76  0.76 -0.24  0.76  0.7200000  0.8000000 -0.2000000 -0.24
#>  [5,] -0.24  0.76  0.76  0.76 -0.24  0.7200000 -0.2000000  0.8000000 -0.24
#>  [6,] -0.28 -0.28 -0.28  0.72  0.72 -0.3266667  0.7666667  0.7666667  0.72
#>  [7,] -0.20 -0.20 -0.20  0.80 -0.20  0.7666667 -0.1666667  0.8333333  0.80
#>  [8,] -0.20 -0.20 -0.20 -0.20  0.80  0.7666667  0.8333333 -0.1666667 -0.20
#>  [9,] -0.24 -0.24 -0.24 -0.24 -0.24  0.7200000  0.8000000 -0.2000000 -0.24
#> [10,] -0.28 -0.28 -0.28 -0.28  0.72 -0.3266667  0.7666667  0.7666667  0.72
#> [11,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.2000000  0.8000000  0.76
#> [12,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.2000000 -0.2000000  0.76
#> [13,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.2000000 -0.2000000 -0.24
#> [14,] -0.28 -0.28 -0.28 -0.28 -0.28 -0.3266667 -0.2333333 -0.2333333  0.72
#> [15,] -0.28 -0.28 -0.28 -0.28 -0.28  0.6733333 -0.2333333 -0.2333333 -0.28
#> [16,]  0.76 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.2000000 -0.2000000 -0.24
#> [17,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.2000000 -0.2000000 -0.24
#> [18,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.2000000 -0.2000000 -0.24
#> [19,] -0.24 -0.24  0.76 -0.24 -0.24 -0.2800000 -0.2000000 -0.2000000 -0.24
#> [20,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.2000000 -0.2000000 -0.24
#> [21,] -0.24 -0.24  0.76 -0.24 -0.24 -0.2800000 -0.2000000 -0.2000000 -0.24
#> [22,] -0.24 -0.24 -0.24 -0.24 -0.24 -0.2800000 -0.2000000 -0.2000000 -0.24
#> [23,] -0.20 -0.20 -0.20 -0.20 -0.20 -0.2333333 -0.1666667 -0.1666667 -0.20
#> [24,]  0.80  0.80 -0.20 -0.20 -0.20 -0.2333333 -0.1666667 -0.1666667 -0.20
#> [25,]  0.76  0.76 -0.24 -0.24 -0.24  0.7200000 -0.2000000 -0.2000000 -0.24
#>            [,10] [,11] [,12] [,13]      [,14]      [,15] [,16] [,17] [,18]
#>  [1,] -0.2800000 -0.24 -0.24 -0.24 -0.2800000 -0.2800000  0.76 -0.24 -0.24
#>  [2,] -0.2800000 -0.24 -0.24 -0.24 -0.2800000 -0.2800000 -0.24 -0.24 -0.24
#>  [3,] -0.2800000 -0.24 -0.24 -0.24 -0.2800000 -0.2800000 -0.24 -0.24 -0.24
#>  [4,] -0.2800000 -0.24 -0.24 -0.24 -0.2800000 -0.2800000 -0.24 -0.24 -0.24
#>  [5,]  0.7200000 -0.24 -0.24 -0.24 -0.2800000 -0.2800000 -0.24 -0.24 -0.24
#>  [6,] -0.3266667 -0.28 -0.28 -0.28 -0.3266667  0.6733333 -0.28 -0.28 -0.28
#>  [7,]  0.7666667 -0.20 -0.20 -0.20 -0.2333333 -0.2333333 -0.20 -0.20 -0.20
#>  [8,]  0.7666667  0.80 -0.20 -0.20 -0.2333333 -0.2333333 -0.20 -0.20 -0.20
#>  [9,]  0.7200000  0.76  0.76 -0.24  0.7200000 -0.2800000 -0.24 -0.24 -0.24
#> [10,] -0.3266667  0.72  0.72  0.72 -0.3266667 -0.3266667 -0.28 -0.28 -0.28
#> [11,]  0.7200000 -0.24  0.76  0.76  0.7200000 -0.2800000 -0.24 -0.24 -0.24
#> [12,]  0.7200000  0.76 -0.24  0.76  0.7200000  0.7200000 -0.24 -0.24 -0.24
#> [13,]  0.7200000  0.76  0.76 -0.24  0.7200000  0.7200000  0.76 -0.24 -0.24
#> [14,] -0.3266667  0.72  0.72  0.72 -0.3266667  0.6733333  0.72  0.72 -0.28
#> [15,] -0.3266667 -0.28  0.72  0.72  0.6733333 -0.3266667  0.72  0.72  0.72
#> [16,] -0.2800000 -0.24 -0.24  0.76  0.7200000  0.7200000 -0.24  0.76  0.76
#> [17,] -0.2800000 -0.24 -0.24 -0.24  0.7200000  0.7200000  0.76 -0.24  0.76
#> [18,] -0.2800000 -0.24 -0.24 -0.24 -0.2800000  0.7200000  0.76  0.76 -0.24
#> [19,] -0.2800000 -0.24 -0.24 -0.24 -0.2800000 -0.2800000 -0.24  0.76  0.76
#> [20,] -0.2800000 -0.24 -0.24 -0.24 -0.2800000 -0.2800000 -0.24  0.76  0.76
#> [21,] -0.2800000 -0.24 -0.24 -0.24 -0.2800000 -0.2800000 -0.24 -0.24  0.76
#> [22,] -0.2800000 -0.24 -0.24 -0.24 -0.2800000 -0.2800000 -0.24 -0.24 -0.24
#> [23,] -0.2333333 -0.20 -0.20 -0.20 -0.2333333 -0.2333333 -0.20 -0.20 -0.20
#> [24,] -0.2333333 -0.20 -0.20 -0.20 -0.2333333 -0.2333333 -0.20 -0.20 -0.20
#> [25,] -0.2800000 -0.24 -0.24 -0.24 -0.2800000 -0.2800000 -0.24 -0.24 -0.24
#>       [,19] [,20] [,21] [,22]      [,23]      [,24] [,25]
#>  [1,] -0.24 -0.24 -0.24 -0.24 -0.2000000  0.8000000  0.76
#>  [2,] -0.24 -0.24 -0.24 -0.24 -0.2000000  0.8000000  0.76
#>  [3,]  0.76 -0.24  0.76 -0.24 -0.2000000 -0.2000000 -0.24
#>  [4,] -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24
#>  [5,] -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24
#>  [6,] -0.28 -0.28 -0.28 -0.28 -0.2333333 -0.2333333  0.72
#>  [7,] -0.20 -0.20 -0.20 -0.20 -0.1666667 -0.1666667 -0.20
#>  [8,] -0.20 -0.20 -0.20 -0.20 -0.1666667 -0.1666667 -0.20
#>  [9,] -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24
#> [10,] -0.28 -0.28 -0.28 -0.28 -0.2333333 -0.2333333 -0.28
#> [11,] -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24
#> [12,] -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24
#> [13,] -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24
#> [14,] -0.28 -0.28 -0.28 -0.28 -0.2333333 -0.2333333 -0.28
#> [15,] -0.28 -0.28 -0.28 -0.28 -0.2333333 -0.2333333 -0.28
#> [16,] -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24
#> [17,]  0.76  0.76 -0.24 -0.24 -0.2000000 -0.2000000 -0.24
#> [18,]  0.76  0.76  0.76 -0.24 -0.2000000 -0.2000000 -0.24
#> [19,] -0.24  0.76  0.76  0.76 -0.2000000 -0.2000000 -0.24
#> [20,]  0.76 -0.24  0.76  0.76  0.8000000 -0.2000000 -0.24
#> [21,]  0.76  0.76 -0.24  0.76  0.8000000 -0.2000000 -0.24
#> [22,]  0.76  0.76  0.76 -0.24  0.8000000  0.8000000  0.76
#> [23,] -0.20  0.80  0.80  0.80 -0.1666667  0.8333333  0.80
#> [24,] -0.20 -0.20 -0.20  0.80  0.8333333 -0.1666667  0.80
#> [25,] -0.24 -0.24 -0.24  0.76  0.8000000  0.8000000 -0.24
#> attr(,"representation")
#> [1] "modularity"
repr_graphon(g)
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.55555556 0.61111111 0.72222222 0.50000000 0.29166667 0.29166667
#>  [2,] 0.61111111 0.66666667 0.55555556 0.55555556 0.54166667 0.16666667
#>  [3,] 0.72222222 0.55555556 0.66666667 0.55555556 0.87500000 0.29166667
#>  [4,] 0.50000000 0.55555556 0.55555556 0.44444444 0.59722222 0.65277778
#>  [5,] 0.29166667 0.54166667 0.87500000 0.59722222 0.25000000 0.75000000
#>  [6,] 0.29166667 0.16666667 0.29166667 0.65277778 0.75000000 0.25000000
#>  [7,] 0.05555556 0.12698413 0.23809524 0.38095238 0.26785714 0.53571429
#>  [8,] 0.11805556 0.18055556 0.12500000 0.18055556 0.43750000 0.68750000
#>  [9,] 0.06250000 0.12500000 0.12500000 0.12500000 0.25000000 0.68750000
#> [10,] 0.05555556 0.05555556 0.00000000 0.23809524 0.46428571 0.26785714
#> [11,] 0.06250000 0.06250000 0.06250000 0.23611111 0.06250000 0.31250000
#> [12,] 0.05555556 0.00000000 0.00000000 0.11111111 0.05555556 0.29166667
#> [13,] 0.05555556 0.00000000 0.00000000 0.00000000 0.05555556 0.23611111
#> [14,] 0.06250000 0.00000000 0.00000000 0.05555556 0.06250000 0.06250000
#> [15,] 0.07142857 0.00000000 0.00000000 0.00000000 0.00000000 0.26785714
#> [16,] 0.11111111 0.05555556 0.05555556 0.05555556 0.12500000 0.05555556
#> [17,] 0.05555556 0.00000000 0.11111111 0.00000000 0.00000000 0.00000000
#> [18,] 0.11111111 0.05555556 0.11111111 0.05555556 0.05555556 0.00000000
#> [19,] 0.17361111 0.11111111 0.29166667 0.05555556 0.00000000 0.00000000
#> [20,] 0.11111111 0.11111111 0.20000000 0.00000000 0.00000000 0.00000000
#> [21,] 0.27777778 0.27777778 0.27777778 0.05555556 0.00000000 0.05555556
#> [22,] 0.16666667 0.16666667 0.10000000 0.11111111 0.00000000 0.00000000
#> [23,] 0.29166667 0.29166667 0.12500000 0.11111111 0.00000000 0.00000000
#> [24,] 0.61111111 0.61111111 0.16666667 0.16666667 0.12500000 0.29166667
#> [25,] 0.77777778 0.77777778 0.33333333 0.16666667 0.29166667 0.58333333
#>             [,7]      [,8]       [,9]      [,10]     [,11]      [,12]
#>  [1,] 0.05555556 0.1180556 0.06250000 0.05555556 0.0625000 0.05555556
#>  [2,] 0.12698413 0.1805556 0.12500000 0.05555556 0.0625000 0.00000000
#>  [3,] 0.23809524 0.1250000 0.12500000 0.00000000 0.0625000 0.00000000
#>  [4,] 0.38095238 0.1805556 0.12500000 0.23809524 0.2361111 0.11111111
#>  [5,] 0.26785714 0.4375000 0.25000000 0.46428571 0.0625000 0.05555556
#>  [6,] 0.53571429 0.6875000 0.68750000 0.26785714 0.3125000 0.29166667
#>  [7,] 0.42857143 0.5357143 0.47321429 0.71428571 0.3392857 0.25396825
#>  [8,] 0.53571429 0.3750000 0.37500000 0.58928571 0.5625000 0.29861111
#>  [9,] 0.47321429 0.3750000 0.37500000 0.79464286 0.5000000 0.40972222
#> [10,] 0.71428571 0.5892857 0.79464286 0.57142857 0.7321429 0.61904762
#> [11,] 0.33928571 0.5625000 0.50000000 0.73214286 0.5000000 0.52777778
#> [12,] 0.25396825 0.2986111 0.40972222 0.61904762 0.5277778 0.44444444
#> [13,] 0.25396825 0.1805556 0.29166667 0.56349206 0.5277778 0.44444444
#> [14,] 0.27678571 0.3125000 0.50000000 0.40178571 0.5625000 0.64583333
#> [15,] 0.21428571 0.1964286 0.26785714 0.42857143 0.4017857 0.63492063
#> [16,] 0.00000000 0.1180556 0.17361111 0.23809524 0.1736111 0.33333333
#> [17,] 0.00000000 0.0000000 0.05555556 0.18253968 0.1736111 0.16666667
#> [18,] 0.00000000 0.0000000 0.00000000 0.00000000 0.0625000 0.11111111
#> [19,] 0.00000000 0.0000000 0.00000000 0.00000000 0.0000000 0.05555556
#> [20,] 0.00000000 0.0000000 0.00000000 0.00000000 0.0000000 0.05555556
#> [21,] 0.00000000 0.0000000 0.00000000 0.00000000 0.0000000 0.00000000
#> [22,] 0.00000000 0.0000000 0.00000000 0.00000000 0.0000000 0.00000000
#> [23,] 0.00000000 0.0000000 0.00000000 0.00000000 0.0000000 0.00000000
#> [24,] 0.00000000 0.0000000 0.00000000 0.00000000 0.0000000 0.00000000
#> [25,] 0.23809524 0.0000000 0.00000000 0.07142857 0.0000000 0.00000000
#>            [,13]      [,14]      [,15]      [,16]      [,17]      [,18]
#>  [1,] 0.05555556 0.06250000 0.07142857 0.11111111 0.05555556 0.11111111
#>  [2,] 0.00000000 0.00000000 0.00000000 0.05555556 0.00000000 0.05555556
#>  [3,] 0.00000000 0.00000000 0.00000000 0.05555556 0.11111111 0.11111111
#>  [4,] 0.00000000 0.05555556 0.00000000 0.05555556 0.00000000 0.05555556
#>  [5,] 0.05555556 0.06250000 0.00000000 0.12500000 0.00000000 0.05555556
#>  [6,] 0.23611111 0.06250000 0.26785714 0.05555556 0.00000000 0.00000000
#>  [7,] 0.25396825 0.27678571 0.21428571 0.00000000 0.00000000 0.00000000
#>  [8,] 0.18055556 0.31250000 0.19642857 0.11805556 0.00000000 0.00000000
#>  [9,] 0.29166667 0.50000000 0.26785714 0.17361111 0.05555556 0.00000000
#> [10,] 0.56349206 0.40178571 0.42857143 0.23809524 0.18253968 0.00000000
#> [11,] 0.52777778 0.56250000 0.40178571 0.17361111 0.17361111 0.06250000
#> [12,] 0.44444444 0.64583333 0.63492063 0.33333333 0.16666667 0.11111111
#> [13,] 0.44444444 0.70833333 0.61904762 0.44444444 0.27777778 0.16666667
#> [14,] 0.70833333 0.62500000 0.74107143 0.58333333 0.47222222 0.23611111
#> [15,] 0.61904762 0.74107143 0.57142857 0.56349206 0.49206349 0.38095238
#> [16,] 0.44444444 0.58333333 0.56349206 0.44444444 0.44444444 0.38888889
#> [17,] 0.27777778 0.47222222 0.49206349 0.44444444 0.44444444 0.50000000
#> [18,] 0.16666667 0.23611111 0.38095238 0.38888889 0.50000000 0.55555556
#> [19,] 0.17361111 0.18750000 0.25892857 0.29166667 0.52777778 0.64583333
#> [20,] 0.11111111 0.06250000 0.17142857 0.26666667 0.57777778 0.63333333
#> [21,] 0.05555556 0.11805556 0.11111111 0.27777778 0.38888889 0.55555556
#> [22,] 0.00000000 0.00000000 0.10000000 0.21111111 0.36666667 0.42222222
#> [23,] 0.00000000 0.00000000 0.00000000 0.05555556 0.29166667 0.29166667
#> [24,] 0.00000000 0.00000000 0.00000000 0.00000000 0.05555556 0.11111111
#> [25,] 0.00000000 0.00000000 0.00000000 0.00000000 0.05555556 0.11111111
#>            [,19]      [,20]      [,21]     [,22]      [,23]      [,24]
#>  [1,] 0.17361111 0.11111111 0.27777778 0.1666667 0.29166667 0.61111111
#>  [2,] 0.11111111 0.11111111 0.27777778 0.1666667 0.29166667 0.61111111
#>  [3,] 0.29166667 0.20000000 0.27777778 0.1000000 0.12500000 0.16666667
#>  [4,] 0.05555556 0.00000000 0.05555556 0.1111111 0.11111111 0.16666667
#>  [5,] 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000 0.12500000
#>  [6,] 0.00000000 0.00000000 0.05555556 0.0000000 0.00000000 0.29166667
#>  [7,] 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000
#>  [8,] 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000
#>  [9,] 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000
#> [10,] 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000
#> [11,] 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000
#> [12,] 0.05555556 0.05555556 0.00000000 0.0000000 0.00000000 0.00000000
#> [13,] 0.17361111 0.11111111 0.05555556 0.0000000 0.00000000 0.00000000
#> [14,] 0.18750000 0.06250000 0.11805556 0.0000000 0.00000000 0.00000000
#> [15,] 0.25892857 0.17142857 0.11111111 0.1000000 0.00000000 0.00000000
#> [16,] 0.29166667 0.26666667 0.27777778 0.2111111 0.05555556 0.00000000
#> [17,] 0.52777778 0.57777778 0.38888889 0.3666667 0.29166667 0.05555556
#> [18,] 0.64583333 0.63333333 0.55555556 0.4222222 0.29166667 0.11111111
#> [19,] 0.62500000 0.77500000 0.59027778 0.6500000 0.31250000 0.12500000
#> [20,] 0.77500000 0.80000000 0.73333333 0.7000000 0.67500000 0.26666667
#> [21,] 0.59027778 0.73333333 0.55555556 0.7333333 0.65277778 0.33333333
#> [22,] 0.65000000 0.70000000 0.73333333 0.6000000 0.67500000 0.70000000
#> [23,] 0.31250000 0.67500000 0.65277778 0.6750000 0.50000000 0.58333333
#> [24,] 0.12500000 0.26666667 0.33333333 0.7000000 0.58333333 0.66666667
#> [25,] 0.12500000 0.10000000 0.16666667 0.5333333 0.70833333 0.50000000
#>            [,25]
#>  [1,] 0.77777778
#>  [2,] 0.77777778
#>  [3,] 0.33333333
#>  [4,] 0.16666667
#>  [5,] 0.29166667
#>  [6,] 0.58333333
#>  [7,] 0.23809524
#>  [8,] 0.00000000
#>  [9,] 0.00000000
#> [10,] 0.07142857
#> [11,] 0.00000000
#> [12,] 0.00000000
#> [13,] 0.00000000
#> [14,] 0.00000000
#> [15,] 0.00000000
#> [16,] 0.00000000
#> [17,] 0.05555556
#> [18,] 0.11111111
#> [19,] 0.12500000
#> [20,] 0.10000000
#> [21,] 0.16666667
#> [22,] 0.53333333
#> [23,] 0.70833333
#> [24,] 0.50000000
#> [25,] 0.33333333
#> attr(,"representation")
#> [1] "graphon"