Network Representation Functions

This is a collection of functions that convert a graph stored as an igraph object into a desired matrix representation among adjacency matrix, graph laplacian, modularity matrix or graphon (edge probability matrix).

Usage

repr_adjacency(network, validate = TRUE)

repr_laplacian(network, validate = TRUE)

repr_modularity(network, validate = TRUE)

repr_graphon(network, validate = TRUE)

Arguments

network

An igraph object.

validate

A boolean specifying whether the function should check the class of its input (default: TRUE).

Value

A numeric square matrix giving the desired network representation recorded in the object's class.

Examples

g <- igraph::sample_smallworld(1, 25, 3, 0.05)
repr_adjacency(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0    1    1    1    0    0    0    0    0     0     0     0     0
#>  [2,]    1    0    1    1    1    0    0    0    0     0     0     0     0
#>  [3,]    1    1    0    1    1    1    0    0    0     0     0     0     0
#>  [4,]    1    1    1    0    1    1    1    0    0     0     0     0     0
#>  [5,]    0    1    1    1    0    0    1    1    0     0     0     0     0
#>  [6,]    0    0    1    1    0    0    1    1    1     0     0     0     0
#>  [7,]    0    0    0    1    1    1    0    1    1     1     0     0     0
#>  [8,]    0    0    0    0    1    1    1    0    1     1     1     0     0
#>  [9,]    0    0    0    0    0    1    1    1    0     1     1     1     0
#> [10,]    0    0    0    0    0    0    1    1    1     0     1     1     1
#> [11,]    0    0    0    0    0    0    0    1    1     1     0     1     1
#> [12,]    0    0    0    0    0    0    0    0    1     1     1     0     1
#> [13,]    0    0    0    0    0    0    0    0    0     1     1     1     0
#> [14,]    0    0    0    0    0    0    0    0    0     0     1     1     0
#> [15,]    0    0    0    0    0    0    0    0    0     0     0     1     1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0     1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    1    0    0    0    0    0    0     0     0     0     1
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    1    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]    1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]    1    1    0    0    0    0    0    0    0     0     0     0     0
#> [25,]    1    1    1    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0     1     1     1
#>  [2,]     0     0     0     0     0     0     0     0     0     0     1     1
#>  [3,]     0     0     0     0     1     0     0     0     0     0     0     1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0     1     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]     1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]     1     1     0     0     0     0     0     0     0     0     0     0
#> [13,]     0     1     1     0     1     0     0     0     0     0     0     0
#> [14,]     0     1     1     1     0     1     0     0     0     0     0     0
#> [15,]     1     0     1     1     0     0     0     0     0     0     0     0
#> [16,]     1     1     0     1     1     0     0     0     0     0     0     0
#> [17,]     1     1     1     0     1     1     1     0     0     0     0     0
#> [18,]     0     0     1     1     0     1     1     1     0     0     0     0
#> [19,]     1     0     0     1     1     0     1     1     1     0     0     0
#> [20,]     0     0     0     1     1     1     0     1     1     1     0     0
#> [21,]     0     0     0     0     1     1     1     0     1     1     1     0
#> [22,]     0     0     0     0     0     1     1     1     0     1     1     1
#> [23,]     0     0     0     0     0     0     1     1     1     0     1     1
#> [24,]     0     0     0     0     0     0     0     1     1     1     0     1
#> [25,]     0     0     0     0     0     0     0     0     1     1     1     0
#> attr(,"representation")
#> [1] "adjacency"
repr_laplacian(g)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    6   -1   -1   -1    0    0    0    0    0     0     0     0     0
#>  [2,]   -1    6   -1   -1   -1    0    0    0    0     0     0     0     0
#>  [3,]   -1   -1    7   -1   -1   -1    0    0    0     0     0     0     0
#>  [4,]   -1   -1   -1    6   -1   -1   -1    0    0     0     0     0     0
#>  [5,]    0   -1   -1   -1    6    0   -1   -1    0     0     0     0     0
#>  [6,]    0    0   -1   -1    0    5   -1   -1   -1     0     0     0     0
#>  [7,]    0    0    0   -1   -1   -1    6   -1   -1    -1     0     0     0
#>  [8,]    0    0    0    0   -1   -1   -1    6   -1    -1    -1     0     0
#>  [9,]    0    0    0    0    0   -1   -1   -1    6    -1    -1    -1     0
#> [10,]    0    0    0    0    0    0   -1   -1   -1     6    -1    -1    -1
#> [11,]    0    0    0    0    0    0    0   -1   -1    -1     6    -1    -1
#> [12,]    0    0    0    0    0    0    0    0   -1    -1    -1     6    -1
#> [13,]    0    0    0    0    0    0    0    0    0    -1    -1    -1     6
#> [14,]    0    0    0    0    0    0    0    0    0     0    -1    -1     0
#> [15,]    0    0    0    0    0    0    0    0    0     0     0    -1    -1
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0    -1
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0   -1    0    0    0    0    0    0     0     0     0    -1
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0   -1    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]   -1    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]   -1   -1    0    0    0    0    0    0    0     0     0     0     0
#> [25,]   -1   -1   -1    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     0     0     0     0     0     0     0     0     0    -1    -1    -1
#>  [2,]     0     0     0     0     0     0     0     0     0     0    -1    -1
#>  [3,]     0     0     0     0    -1     0     0     0     0     0     0    -1
#>  [4,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0    -1     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [11,]    -1     0     0     0     0     0     0     0     0     0     0     0
#> [12,]    -1    -1     0     0     0     0     0     0     0     0     0     0
#> [13,]     0    -1    -1     0    -1     0     0     0     0     0     0     0
#> [14,]     6    -1    -1    -1     0    -1     0     0     0     0     0     0
#> [15,]    -1     5    -1    -1     0     0     0     0     0     0     0     0
#> [16,]    -1    -1     5    -1    -1     0     0     0     0     0     0     0
#> [17,]    -1    -1    -1     6    -1    -1    -1     0     0     0     0     0
#> [18,]     0     0    -1    -1     7    -1    -1    -1     0     0     0     0
#> [19,]    -1     0     0    -1    -1     6    -1    -1    -1     0     0     0
#> [20,]     0     0     0    -1    -1    -1     6    -1    -1    -1     0     0
#> [21,]     0     0     0     0    -1    -1    -1     7    -1    -1    -1     0
#> [22,]     0     0     0     0     0    -1    -1    -1     6    -1    -1    -1
#> [23,]     0     0     0     0     0     0    -1    -1    -1     6    -1    -1
#> [24,]     0     0     0     0     0     0     0    -1    -1    -1     6    -1
#> [25,]     0     0     0     0     0     0     0     0    -1    -1    -1     6
#> attr(,"representation")
#> [1] "laplacian"
repr_modularity(g)
#>        [,1]  [,2]       [,3]  [,4]  [,5]       [,6]  [,7]  [,8]  [,9] [,10]
#>  [1,] -0.24  0.76  0.7200000  0.76 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24
#>  [2,]  0.76 -0.24  0.7200000  0.76  0.76 -0.2000000 -0.24 -0.24 -0.24 -0.24
#>  [3,]  0.72  0.72 -0.3266667  0.72  0.72  0.7666667 -0.28 -0.28 -0.28 -0.28
#>  [4,]  0.76  0.76  0.7200000 -0.24  0.76  0.8000000  0.76 -0.24 -0.24 -0.24
#>  [5,] -0.24  0.76  0.7200000  0.76 -0.24 -0.2000000  0.76  0.76 -0.24 -0.24
#>  [6,] -0.20 -0.20  0.7666667  0.80 -0.20 -0.1666667  0.80  0.80  0.80 -0.20
#>  [7,] -0.24 -0.24 -0.2800000  0.76  0.76  0.8000000 -0.24  0.76  0.76  0.76
#>  [8,] -0.24 -0.24 -0.2800000 -0.24  0.76  0.8000000  0.76 -0.24  0.76  0.76
#>  [9,] -0.24 -0.24 -0.2800000 -0.24 -0.24  0.8000000  0.76  0.76 -0.24  0.76
#> [10,] -0.24 -0.24 -0.2800000 -0.24 -0.24 -0.2000000  0.76  0.76  0.76 -0.24
#> [11,] -0.24 -0.24 -0.2800000 -0.24 -0.24 -0.2000000 -0.24  0.76  0.76  0.76
#> [12,] -0.24 -0.24 -0.2800000 -0.24 -0.24 -0.2000000 -0.24 -0.24  0.76  0.76
#> [13,] -0.24 -0.24 -0.2800000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.24  0.76
#> [14,] -0.24 -0.24 -0.2800000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24
#> [15,] -0.20 -0.20 -0.2333333 -0.20 -0.20 -0.1666667 -0.20 -0.20 -0.20 -0.20
#> [16,] -0.20 -0.20 -0.2333333 -0.20 -0.20 -0.1666667 -0.20 -0.20 -0.20 -0.20
#> [17,] -0.24 -0.24 -0.2800000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24
#> [18,] -0.28 -0.28  0.6733333 -0.28 -0.28 -0.2333333 -0.28 -0.28 -0.28 -0.28
#> [19,] -0.24 -0.24 -0.2800000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24
#> [20,] -0.24 -0.24 -0.2800000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24
#> [21,] -0.28 -0.28 -0.3266667 -0.28  0.72 -0.2333333 -0.28 -0.28 -0.28 -0.28
#> [22,] -0.24 -0.24 -0.2800000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24
#> [23,]  0.76 -0.24 -0.2800000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24
#> [24,]  0.76  0.76 -0.2800000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24
#> [25,]  0.76  0.76  0.7200000 -0.24 -0.24 -0.2000000 -0.24 -0.24 -0.24 -0.24
#>       [,11] [,12] [,13] [,14]      [,15]      [,16] [,17]      [,18] [,19]
#>  [1,] -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.2800000 -0.24
#>  [2,] -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.2800000 -0.24
#>  [3,] -0.28 -0.28 -0.28 -0.28 -0.2333333 -0.2333333 -0.28  0.6733333 -0.28
#>  [4,] -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.2800000 -0.24
#>  [5,] -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.2800000 -0.24
#>  [6,] -0.20 -0.20 -0.20 -0.20 -0.1666667 -0.1666667 -0.20 -0.2333333 -0.20
#>  [7,] -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.2800000 -0.24
#>  [8,]  0.76 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.2800000 -0.24
#>  [9,]  0.76  0.76 -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.2800000 -0.24
#> [10,]  0.76  0.76  0.76 -0.24 -0.2000000 -0.2000000 -0.24 -0.2800000 -0.24
#> [11,] -0.24  0.76  0.76  0.76 -0.2000000 -0.2000000 -0.24 -0.2800000 -0.24
#> [12,]  0.76 -0.24  0.76  0.76  0.8000000 -0.2000000 -0.24 -0.2800000 -0.24
#> [13,]  0.76  0.76 -0.24 -0.24  0.8000000  0.8000000 -0.24  0.7200000 -0.24
#> [14,]  0.76  0.76 -0.24 -0.24  0.8000000  0.8000000  0.76 -0.2800000  0.76
#> [15,] -0.20  0.80  0.80  0.80 -0.1666667  0.8333333  0.80 -0.2333333 -0.20
#> [16,] -0.20 -0.20  0.80  0.80  0.8333333 -0.1666667  0.80  0.7666667 -0.20
#> [17,] -0.24 -0.24 -0.24  0.76  0.8000000  0.8000000 -0.24  0.7200000  0.76
#> [18,] -0.28 -0.28  0.72 -0.28 -0.2333333  0.7666667  0.72 -0.3266667  0.72
#> [19,] -0.24 -0.24 -0.24  0.76 -0.2000000 -0.2000000  0.76  0.7200000 -0.24
#> [20,] -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.2000000  0.76  0.7200000  0.76
#> [21,] -0.28 -0.28 -0.28 -0.28 -0.2333333 -0.2333333 -0.28  0.6733333  0.72
#> [22,] -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.2800000  0.76
#> [23,] -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.2800000 -0.24
#> [24,] -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.2800000 -0.24
#> [25,] -0.24 -0.24 -0.24 -0.24 -0.2000000 -0.2000000 -0.24 -0.2800000 -0.24
#>       [,20]      [,21] [,22] [,23] [,24] [,25]
#>  [1,] -0.24 -0.2800000 -0.24  0.76  0.76  0.76
#>  [2,] -0.24 -0.2800000 -0.24 -0.24  0.76  0.76
#>  [3,] -0.28 -0.3266667 -0.28 -0.28 -0.28  0.72
#>  [4,] -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.24
#>  [5,] -0.24  0.7200000 -0.24 -0.24 -0.24 -0.24
#>  [6,] -0.20 -0.2333333 -0.20 -0.20 -0.20 -0.20
#>  [7,] -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.24
#>  [8,] -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.24
#>  [9,] -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.24
#> [10,] -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.24
#> [11,] -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.24
#> [12,] -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.24
#> [13,] -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.24
#> [14,] -0.24 -0.2800000 -0.24 -0.24 -0.24 -0.24
#> [15,] -0.20 -0.2333333 -0.20 -0.20 -0.20 -0.20
#> [16,] -0.20 -0.2333333 -0.20 -0.20 -0.20 -0.20
#> [17,]  0.76 -0.2800000 -0.24 -0.24 -0.24 -0.24
#> [18,]  0.72  0.6733333 -0.28 -0.28 -0.28 -0.28
#> [19,]  0.76  0.7200000  0.76 -0.24 -0.24 -0.24
#> [20,] -0.24  0.7200000  0.76  0.76 -0.24 -0.24
#> [21,]  0.72 -0.3266667  0.72  0.72  0.72 -0.28
#> [22,]  0.76  0.7200000 -0.24  0.76  0.76  0.76
#> [23,]  0.76  0.7200000  0.76 -0.24  0.76  0.76
#> [24,] -0.24  0.7200000  0.76  0.76 -0.24  0.76
#> [25,] -0.24 -0.2800000  0.76  0.76  0.76 -0.24
#> attr(,"representation")
#> [1] "modularity"
repr_graphon(g)
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.55555556 0.69444444 0.66666667 0.55555556 0.27777778 0.05555556
#>  [2,] 0.69444444 0.66666667 0.66666667 0.58333333 0.47222222 0.29166667
#>  [3,] 0.66666667 0.66666667 0.50000000 0.66666667 0.66666667 0.50000000
#>  [4,] 0.55555556 0.58333333 0.66666667 0.50000000 0.61111111 0.62500000
#>  [5,] 0.27777778 0.47222222 0.66666667 0.61111111 0.33333333 0.29166667
#>  [6,] 0.05555556 0.29166667 0.50000000 0.62500000 0.29166667 0.25000000
#>  [7,] 0.16666667 0.08333333 0.18333333 0.45000000 0.47777778 0.67500000
#>  [8,] 0.16666667 0.05555556 0.19444444 0.19444444 0.44444444 0.72222222
#>  [9,] 0.05555556 0.00000000 0.08333333 0.16666667 0.33333333 0.62500000
#> [10,] 0.00000000 0.00000000 0.16666667 0.16666667 0.19444444 0.29166667
#> [11,] 0.00000000 0.00000000 0.00000000 0.16666667 0.19444444 0.29166667
#> [12,] 0.00000000 0.00000000 0.00000000 0.00000000 0.19444444 0.20833333
#> [13,] 0.00000000 0.00000000 0.00000000 0.00000000 0.05555556 0.22500000
#> [14,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.05555556
#> [15,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [16,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [17,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [18,] 0.00000000 0.08333333 0.15476190 0.08333333 0.00000000 0.00000000
#> [19,] 0.18055556 0.06250000 0.00000000 0.00000000 0.00000000 0.00000000
#> [20,] 0.21111111 0.18333333 0.00000000 0.00000000 0.10000000 0.00000000
#> [21,] 0.36507937 0.15476190 0.15476190 0.08333333 0.12698413 0.12500000
#> [22,] 0.33333333 0.33333333 0.16666667 0.16666667 0.13888889 0.00000000
#> [23,] 0.44444444 0.33333333 0.33333333 0.19444444 0.22222222 0.00000000
#> [24,] 0.55555556 0.55555556 0.38888889 0.27777778 0.27777778 0.05555556
#> [25,] 0.55555556 0.61111111 0.61111111 0.33333333 0.22222222 0.05555556
#>             [,7]       [,8]       [,9]     [,10]      [,11]     [,12]
#>  [1,] 0.16666667 0.16666667 0.05555556 0.0000000 0.00000000 0.0000000
#>  [2,] 0.08333333 0.05555556 0.00000000 0.0000000 0.00000000 0.0000000
#>  [3,] 0.18333333 0.19444444 0.08333333 0.1666667 0.00000000 0.0000000
#>  [4,] 0.45000000 0.19444444 0.16666667 0.1666667 0.16666667 0.0000000
#>  [5,] 0.47777778 0.44444444 0.33333333 0.1944444 0.19444444 0.1944444
#>  [6,] 0.67500000 0.72222222 0.62500000 0.2916667 0.29166667 0.2083333
#>  [7,] 0.60000000 0.67777778 0.65000000 0.7333333 0.36666667 0.3666667
#>  [8,] 0.67777778 0.55555556 0.66666667 0.6666667 0.61111111 0.3888889
#>  [9,] 0.65000000 0.66666667 0.66666667 0.8333333 0.66666667 0.6666667
#> [10,] 0.73333333 0.66666667 0.83333333 0.6666667 0.66666667 0.5833333
#> [11,] 0.36666667 0.61111111 0.66666667 0.6666667 0.50000000 0.6666667
#> [12,] 0.36666667 0.38888889 0.66666667 0.5833333 0.66666667 0.5000000
#> [13,] 0.20000000 0.26666667 0.26666667 0.5500000 0.73333333 0.6333333
#> [14,] 0.21111111 0.22222222 0.19444444 0.3333333 0.52777778 0.5277778
#> [15,] 0.05555556 0.22222222 0.25000000 0.2500000 0.30555556 0.5277778
#> [16,] 0.00000000 0.11111111 0.19444444 0.1666667 0.25000000 0.3055556
#> [17,] 0.00000000 0.00000000 0.05555556 0.1111111 0.25000000 0.2500000
#> [18,] 0.00000000 0.05555556 0.08333333 0.0000000 0.07142857 0.3095238
#> [19,] 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000 0.0625000
#> [20,] 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000 0.0000000
#> [21,] 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000 0.0000000
#> [22,] 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000 0.0000000
#> [23,] 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000 0.0000000
#> [24,] 0.05555556 0.00000000 0.00000000 0.0000000 0.00000000 0.0000000
#> [25,] 0.11111111 0.05555556 0.00000000 0.0000000 0.00000000 0.0000000
#>            [,13]      [,14]      [,15]      [,16]      [,17]      [,18]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.08333333
#>  [3,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.15476190
#>  [4,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.08333333
#>  [5,] 0.05555556 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [6,] 0.22500000 0.05555556 0.00000000 0.00000000 0.00000000 0.00000000
#>  [7,] 0.20000000 0.21111111 0.05555556 0.00000000 0.00000000 0.00000000
#>  [8,] 0.26666667 0.22222222 0.22222222 0.11111111 0.00000000 0.05555556
#>  [9,] 0.26666667 0.19444444 0.25000000 0.19444444 0.05555556 0.08333333
#> [10,] 0.55000000 0.33333333 0.25000000 0.16666667 0.11111111 0.00000000
#> [11,] 0.73333333 0.52777778 0.30555556 0.25000000 0.25000000 0.07142857
#> [12,] 0.63333333 0.52777778 0.52777778 0.30555556 0.25000000 0.30952381
#> [13,] 0.20000000 0.42222222 0.67777778 0.52222222 0.26666667 0.34285714
#> [14,] 0.42222222 0.44444444 0.61111111 0.55555556 0.50000000 0.43650794
#> [15,] 0.67777778 0.61111111 0.55555556 0.50000000 0.50000000 0.36507937
#> [16,] 0.52222222 0.55555556 0.50000000 0.44444444 0.50000000 0.56349206
#> [17,] 0.26666667 0.50000000 0.50000000 0.50000000 0.55555556 0.63492063
#> [18,] 0.34285714 0.43650794 0.36507937 0.56349206 0.63492063 0.42857143
#> [19,] 0.32500000 0.41666667 0.23611111 0.29166667 0.47222222 0.60714286
#> [20,] 0.10000000 0.15555556 0.11111111 0.11111111 0.36666667 0.58571429
#> [21,] 0.00000000 0.19841270 0.12698413 0.18253968 0.30952381 0.57142857
#> [22,] 0.00000000 0.05555556 0.05555556 0.11111111 0.19444444 0.30952381
#> [23,] 0.00000000 0.11111111 0.00000000 0.05555556 0.22222222 0.30952381
#> [24,] 0.00000000 0.05555556 0.00000000 0.00000000 0.16666667 0.18253968
#> [25,] 0.00000000 0.00000000 0.00000000 0.00000000 0.11111111 0.12698413
#>           [,19]     [,20]      [,21]      [,22]      [,23]      [,24]
#>  [1,] 0.1805556 0.2111111 0.36507937 0.33333333 0.44444444 0.55555556
#>  [2,] 0.0625000 0.1833333 0.15476190 0.33333333 0.33333333 0.55555556
#>  [3,] 0.0000000 0.0000000 0.15476190 0.16666667 0.33333333 0.38888889
#>  [4,] 0.0000000 0.0000000 0.08333333 0.16666667 0.19444444 0.27777778
#>  [5,] 0.0000000 0.1000000 0.12698413 0.13888889 0.22222222 0.27777778
#>  [6,] 0.0000000 0.0000000 0.12500000 0.00000000 0.00000000 0.05555556
#>  [7,] 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000 0.05555556
#>  [8,] 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [9,] 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [10,] 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [11,] 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [12,] 0.0625000 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [13,] 0.3250000 0.1000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [14,] 0.4166667 0.1555556 0.19841270 0.05555556 0.11111111 0.05555556
#> [15,] 0.2361111 0.1111111 0.12698413 0.05555556 0.00000000 0.00000000
#> [16,] 0.2916667 0.1111111 0.18253968 0.11111111 0.05555556 0.00000000
#> [17,] 0.4722222 0.3666667 0.30952381 0.19444444 0.22222222 0.16666667
#> [18,] 0.6071429 0.5857143 0.57142857 0.30952381 0.30952381 0.18253968
#> [19,] 0.3750000 0.5500000 0.59821429 0.50000000 0.35416667 0.23611111
#> [20,] 0.5500000 0.6000000 0.75714286 0.65000000 0.62222222 0.36666667
#> [21,] 0.5982143 0.7571429 0.57142857 0.69047619 0.56349206 0.56349206
#> [22,] 0.5000000 0.6500000 0.69047619 0.66666667 0.75000000 0.61111111
#> [23,] 0.3541667 0.6222222 0.56349206 0.75000000 0.66666667 0.61111111
#> [24,] 0.2361111 0.3666667 0.56349206 0.61111111 0.61111111 0.55555556
#> [25,] 0.2986111 0.3111111 0.42063492 0.55555556 0.55555556 0.55555556
#>            [,25]
#>  [1,] 0.55555556
#>  [2,] 0.61111111
#>  [3,] 0.61111111
#>  [4,] 0.33333333
#>  [5,] 0.22222222
#>  [6,] 0.05555556
#>  [7,] 0.11111111
#>  [8,] 0.05555556
#>  [9,] 0.00000000
#> [10,] 0.00000000
#> [11,] 0.00000000
#> [12,] 0.00000000
#> [13,] 0.00000000
#> [14,] 0.00000000
#> [15,] 0.00000000
#> [16,] 0.00000000
#> [17,] 0.11111111
#> [18,] 0.12698413
#> [19,] 0.29861111
#> [20,] 0.31111111
#> [21,] 0.42063492
#> [22,] 0.55555556
#> [23,] 0.55555556
#> [24,] 0.55555556
#> [25,] 0.55555556
#> attr(,"representation")
#> [1] "graphon"