Test Statistics for Network Populations

This is a collection of functions that provide statistics for testing equality in distribution between samples of networks.

stat_lot(d, indices)

stat_sot(d, indices)

stat_biswas(d, indices)

stat_energy(d, indices, alpha = 1)

stat_mod(d, indices)

stat_dom(d, indices, standardize = TRUE)

stat_student_euclidean(d, indices)

stat_welch_euclidean(d, indices)

stat_edge_count(d, indices, type = "generalized")

Arguments

d
Either a matrix of dimension \((n1+n2)x(n1+n2)\) containing the distances between all the elements of the two samples put together (for distance-based statistics) or the concatenation of the lists of matrix representations of networks in samples 1 and 2 for Euclidean t-Statistics or a list of edge properties of a similarity graph for the graph-based edge count statistics.

indices
A vector of dimension \(n1\) containing the indices of the elements of the first sample.

alpha
An integer specifying to which power elevating the Euclidean distance of the energy-based statistic (default: 1L).

standardize
If TRUE (default), standardize the DoM statistic using the pooled sample variance estimator.

type
A string specifying the version of the edge count test statistic to be used. Choices are "original", "generalized" or "weighted". Default is "generalized".

Value

A scalar giving the value of the desired test statistic.

Details

In details, there are three main categories of statistics:

distance-based statistics: such as stat_lot which is a distance-based version of the t statistic, stat_sot which is a distance-based version of the F statistic, stat_energy which is the class of energy-based statistics and stat_biswas which is the Biswas-Ghosh statistic.

Euclidean t-Statistics: both Student stat_student_euclidean version for equal variances and Welch stat_welch_euclidean version for unequal variances,

Statistics based on similarity graphs: three types of edge count statistics.

Examples

n1 <- 30L
n2 <- 10L
x <- nvd("smallworld", n1)
y <- nvd("pa", n2)
d <- dist_nvd(x, y, representation = "laplacian", distance = "frobenius")
stat_lot(d, 1:n1)

#> [1] 16.10669

stat_sot(d, 1:n1)

#> [1] 1.964544

stat_biswas(d, 1:n1)

#> [1] 212.6454

stat_energy(d, 1:n1)

#> [1] 11.19189

r <- repr_nvd(x, y, representation = "laplacian")
stat_student_euclidean(r, 1:n1)

#> [1] 4.898908

stat_welch_euclidean(r, 1:n1)

#> [1] 4.136024

e <- edge_count_global_variables(d, n1, k = 5L)
stat_edge_count(e, 1:n1, type = "original")

#> [1] 8.625382

stat_edge_count(e, 1:n1, type = "generalized")

#> [1] 242.7804

stat_edge_count(e, 1:n1, type = "weighted")

#> [1] 15.5569