Global Two-Sample Test for Network-Valued Data

This function carries out an hypothesis test where the null hypothesis is that the two populations of networks share the same underlying probabilistic distribution against the alternative hypothesis that the two populations come from different distributions. The test is performed in a non-parametric fashion using a permutational framework in which several statistics can be used, together with several choices of network matrix representations and distances between networks.

test2_global(
  x,
  y,
  representation = "adjacency",
  distance = "frobenius",
  statistic = "lot",
  B = 1000L,
  test = "exact",
  k = 5L,
  seed = NULL
)

Arguments

x
An nvd object listing networks in sample 1.

y
An nvd object listing networks in sample 2.

representation
A string specifying the desired type of representation, among: "adjacency", "laplacian" and "modularity". Default is "adjacency".

distance
A string specifying the chosen distance for calculating the test statistic, among: "hamming", "frobenius", "spectral" and "root-euclidean". Default is "frobenius".

statistic
A string specifying the chosen test statistic(s), among: "lot", "sot", "biswas", "energy", "student", "welch", "original", "generalized", "weighted" or a combination from c("lot", "sot", "biswas", "energy"). Default is "lot".

B
The number of permutation or the tolerance (default: 1000L). If this number is lower than 1, it is intended as a tolerance. Otherwise, it is intended as the number of required permutations.

test
A character string specifying if performing an exact test through the use of Phipson-Smyth estimate of the p-value or an approximate test through a Monte-Carlo estimate of the p-value (default: "exact").

k
An integer specifying the density of the minimum spanning tree used for the edge count statistics (default: 5L).

seed
An integer for specifying the seed of the random generator for result reproducibility (default: NULL).

Value

A list with three components: the value of the statistic for the original two samples, the p-value of the resulting permutation test and a numeric vector storing the values of the permuted statistics.

Examples

n <- 10L

# Two different models for the two populations
x <- nvd("smallworld", n)
y <- nvd("pa", n)
t1 <- test2_global(x, y, "modularity")
t1$pvalue

#> [1] 0.0009936031

# Same model for the two populations
x <- nvd("smallworld", n)
y <- nvd("smallworld", n)
t2 <- test2_global(x, y, "modularity")
t2$pvalue

#> [1] 0.5194751