Local Two-Sample Test for Network-Valued Data

Local Two-Sample Test for Network-Valued Data

test2_local(
  x,
  y,
  partition,
  representation = "adjacency",
  distance = "frobenius",
  statistic = "lot",
  B = 1000L,
  alpha = 0.05,
  test = "exact",
  k = 5L,
  seed = NULL,
  verbose = FALSE
)

Arguments

x
An nvd object listing networks in sample 1.

y
An nvd object listing networks in sample 2.

partition
Either a list or a vector specifying vertex memberships into partition elements.

representation
A string specifying the desired type of representation, among: "adjacency", "laplacian" and "modularity". Default is "adjacency".

distance
A string specifying the chosen distance for calculating the test statistic, among: "hamming", "frobenius", "spectral" and "root-euclidean". Default is "frobenius".

statistic
A string specifying the chosen test statistic(s), among: "lot", "sot", "biswas", "energy", "student", "welch", "original", "generalized", "weighted" or a combination from c("lot", "sot", "biswas", "energy"). Default is "lot".

B
The number of permutation or the tolerance (default: 1000L). If this number is lower than 1, it is intended as a tolerance. Otherwise, it is intended as the number of required permutations.

alpha
Significance level for hypothesis testing (default: 0.05). If set to 1, the function outputs properly adjusted p-values. If lower than 1, then only p-values lower than alpha are properly adjusted.

test
A character string specifying if performing an exact test through the use of Phipson-Smyth estimate of the p-value or an approximate test through a Monte-Carlo estimate of the p-value (default: "exact").

k
An integer specifying the density of the minimum spanning tree used for the edge count statistics (default: 5L).

seed
An integer for specifying the seed of the random generator for result reproducibility (default: NULL).

verbose
Boolean specifying whether information on intermediate tests should be printed in the process (default: FALSE).

Value

A length-2 list reporting the adjusted p-values of each element of the partition for the intra- and inter-tests.

Examples

n <- 10
p1 <- matrix(
  data = c(0.1, 0.4, 0.1, 0.4,
           0.4, 0.4, 0.1, 0.4,
           0.1, 0.1, 0.4, 0.4,
           0.4, 0.4, 0.4, 0.4),
  nrow = 4,
  ncol = 4,
  byrow = TRUE
)
p2 <- matrix(
  data = c(0.1, 0.4, 0.4, 0.4,
           0.4, 0.4, 0.4, 0.4,
           0.4, 0.4, 0.1, 0.1,
           0.4, 0.4, 0.1, 0.4),
  nrow = 4,
  ncol = 4,
  byrow = TRUE
)
sim <- sample2_sbm(n, 68, p1, c(17, 17, 17, 17), p2, seed = 1234)
m <- as.integer(c(rep(1, 17), rep(2, 17), rep(3, 17), rep(4, 17)))
test2_local(sim$x, sim$y, m,
            statistic = c("lot", "sot"),
            seed = 1234,
            alpha = 0.05,
            B = 100)

#> $intra
#> # A tibble: 4 x 3
#>   E      pvalue truncated
#>   <chr>   <dbl> <lgl>    
#> 1 P1    0.0792  TRUE     
#> 2 P2    0.0792  TRUE     
#> 3 P3    0.00990 FALSE    
#> 4 P4    0.0792  TRUE     
#> 
#> $inter
#> # A tibble: 6 x 4
#>   E1    E2    pvalue truncated
#>   <chr> <chr>  <dbl> <lgl>    
#> 1 P1    P2    0.0693 TRUE     
#> 2 P1    P3    0.0198 FALSE    
#> 3 P1    P4    0.0693 TRUE     
#> 4 P2    P3    0.0198 FALSE    
#> 5 P2    P4    0.0693 TRUE     
#> 6 P3    P4    0.0198 FALSE    
#>